% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Add.DE.combined.score}
\alias{Add.DE.combined.score}
\title{Add.DE.combined.score}
\usage{
Add.DE.combined.score(
  df = df.markers,
  p_val_min = 1e-25,
  pval_scaling = 0.001,
  colP = "p_val",
  colLFC = grepv(pattern = c("avg_logFC|avg_log2FC"), x = colnames(df), perl = T)
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION, Default: df.markers}

\item{p_val_min}{PARAM_DESCRIPTION, Default: 1e-25}

\item{pval_scaling}{PARAM_DESCRIPTION, Default: 0.001}

\item{colP}{PARAM_DESCRIPTION, Default: 'p_val'}

\item{colLFC}{PARAM_DESCRIPTION, Default: grepv(pattern = c("avg_logFC|avg_log2FC"), x = colnames(df),
perl = T)}
}
\description{
Add.DE.combined.score
}
\examples{
\dontrun{
if(interactive()){
 df.markers <- Add.DE.combined.score(df.markers)
 }
}
}
