% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{add.meta.fraction}
\alias{add.meta.fraction}
\title{add.meta.fraction}
\usage{
add.meta.fraction(
  col.name = "percent.mito",
  gene.symbol.pattern = c("^MT\\\\.|^MT-", F)[1],
  gene.set = F,
  obj = ls.Seurat[[1]],
  verbose = T
)
}
\arguments{
\item{col.name}{PARAM_DESCRIPTION, Default: 'percent.mito'}

\item{gene.symbol.pattern}{PARAM_DESCRIPTION, Default: c("^MT\\.|^MT-", F)\link{1}}

\item{gene.set}{PARAM_DESCRIPTION, Default: F}

\item{obj}{Seurat object, Default: ls.Seurat[\link{1}]}

\item{verbose}{PARAM_DESCRIPTION, Default: T}
}
\description{
Add a new metadata column, with the fraction of gene set in the transcripome (percentage).
}
\examples{
\dontrun{
if(interactive()){
 ls.Seurat[[1]] <- add.meta.fraction(col.name = "percent.mito", gene.symbol.pattern = "^MT\\\\.|^MT-")
 ls.Seurat[[1]] <- add.meta.fraction(col.name = "percent.ribo", gene.symbol.pattern = "^RPL|^RPS")
 ls.Seurat[[1]] <- add.meta.fraction(col.name = "percent.AC.GenBank", gene.symbol.pattern = "^AC[0-9]{6}\\\\.")
 ls.Seurat[[1]] <- add.meta.fraction(col.name = "percent.AL.EMBL", gene.symbol.pattern = "^AL[0-9]{6}\\\\.")
 ls.Seurat[[1]] <- add.meta.fraction(col.name = "percent.LINC", gene.symbol.pattern = "^LINC0")
 ls.Seurat[[1]] <- add.meta.fraction(col.name = "percent.MALAT1", gene.symbol.pattern = "^MALAT1")
 colnames(ls.Seurat[[1]]@meta.data)
 HGA_MarkerGenes <- c("ENO1", "IGFBP2", "WSB1", "DDIT4", "PGK1", "BNIP3", "FAM162A", "TPI1",
 "VEGFA", "PDK1", "PGAM1", "IER2", "FOS", "BTG1", "EPB41L4A-AS1","NPAS4", "HK2", "BNIP3L",
 "JUN", "ENO2", "GAPDH", "ANKRD37", "ALDOA", "GADD45G", "TXNIP")
 sobj <- add.meta.fraction(col.name = "percent.HGA", gene.set = HGA_MarkerGenes, obj =  sobj)
 }
}
}
\seealso{
\code{\link[Matrix]{colSums}}
}
