% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{add.meta.tags}
\alias{add.meta.tags}
\title{add.meta.tags}
\usage{
add.meta.tags(list.of.tags = tags, obj = ls.Seurat[[1]], n = 1)
}
\arguments{
\item{list.of.tags}{A list of tags to be added as metadata. Default: tags}

\item{obj}{A Seurat object to which the metadata tags are to be added. Default: \code{ls.Seurat[[1]]}}

\item{n}{The index specifying the dataset for which the tags should be applied. Default: 1}
}
\description{
Add metadata tags to a Seurat object dataset.
}
\examples{
\dontrun{
if (interactive()) {
  ls.Seurat[[1]] <- add.meta.tags(list.of.tags = tags, obj = ls.Seurat[[1]], n = 1)
}
}
}
