% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{addGeneClassFractions}
\alias{addGeneClassFractions}
\title{Add Metadata for Gene-Class Fractions}
\usage{
addGeneClassFractions(
  obj,
  gene_fractions = list(percent.mito = "^MT\\\\.|^MT-", percent.ribo = "^RPL|^RPS",
    percent.AC.GenBank = "^AC[0-9]{6}\\\\.", percent.AL.EMBL = "^AL[0-9]{6}\\\\.",
    percent.LINC = "^LINC0", percent.MALAT1 = "^MALAT1"),
  add_hga = TRUE
)
}
\arguments{
\item{obj}{A Seurat object to be updated. Default: None.}

\item{gene_fractions}{A named list containing gene symbol patterns for each meta column name.
Default: List of predefined gene fractions.}

\item{add_hga}{A logical value indicating whether to add percent.HGA metadata. Default: \code{TRUE}.}
}
\value{
An updated Seurat object.
}
\description{
This function adds metadata for various gene-class fractions such as percent.mito, percent.ribo,
percent.AC.GenBank, percent.AL.EMBL, percent.LINC, percent.MALAT1, and percent.HGA to a Seurat object.
If the metadata already exists, a message will be displayed.
}
