% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{addMetaDataSafe}
\alias{addMetaDataSafe}
\title{Add Metadata to a Seurat object, safely with Checks}
\usage{
addMetaDataSafe(obj, metadata, col.name, overwrite = FALSE, verbose = F)
}
\arguments{
\item{obj}{Seurat object to which metadata will be added.}

\item{metadata}{The metadata to be added.}

\item{col.name}{The name of the new metadata column.}

\item{overwrite}{Logical; if TRUE, overwrites the existing column.}

\item{verbose}{Logical; if TRUE, prints additional information.}
}
\value{
Modified Seurat object with additional metadata.
}
\description{
Wrapper function for \code{AddMetaData} that includes additional checks and assertions.
}
