% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{AddNewAnnotation}
\alias{AddNewAnnotation}
\title{AddNewAnnotation}
\usage{
AddNewAnnotation(
  obj = obj,
  source = "RNA_snn_res.0.5",
  named.list.of.identities = ls.Subset.ClusterLists
)
}
\arguments{
\item{obj}{A Seurat object for which the new annotation is to be created. Default is 'obj'.}

\item{source}{A character string specifying the existing metadata column to be used as the
basis for the new annotation. Default is 'RNA_snn_res.0.5'.}

\item{named.list.of.identities}{A named list providing the mappings for the new annotation.
Default is 'ls.Subset.ClusterLists'.}
}
\value{
A character vector representing the new metadata column.
}
\description{
This function creates a new metadata column based on an existing metadata column
and a list of mappings (name <- IDs).
}
\examples{
\dontrun{
if (interactive()) {
  ls.Subset.ClusterLists <- list("hESC.h9" = c("4", "10", "14"), "hESC.176" = c("0", "1", "2"))
  AddNewAnnotation()
}
}
}
