% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{addToMiscOrToolsSlot}
\alias{addToMiscOrToolsSlot}
\title{Add to Misc or Tools Slot}
\usage{
addToMiscOrToolsSlot(
  obj,
  pocket_name = "misc",
  slot_value = NULL,
  slot_name = Stringendo::substitute_deparse(slot_value),
  sub_slot_value = NULL,
  sub_slot_name = Stringendo::substitute_deparse(sub_slot_value),
  overwrite = FALSE
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{pocket_name}{Which main pocket to use: 'misc' or 'tools'. Default: 'misc'.}

\item{slot_value}{The value to be assigned to the sub-slot.}

\item{slot_name}{The name of the sub-slot. Automatically derived from 'sub_slot_value' if not provided.}

\item{sub_slot_value}{The value to be assigned to the sub-slot.}

\item{sub_slot_name}{The name of the sub-slot. Automatically derived from 'sub_slot_value' if not provided.}

\item{overwrite}{A boolean indicating whether to overwrite an existing sub-slot with the same name.}
}
\value{
The modified Seurat object with the new or updated sub-slot.
}
\description{
This function creates and adds a sub-slot to either the 'misc' or 'tools' slot of a
Seurat object. If the sub-slot already exists, it can either be overwritten or a warning will be issued.
}
