% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Annotate4Plotly3D}
\alias{Annotate4Plotly3D}
\title{Annotate4Plotly3D}
\usage{
Annotate4Plotly3D(
  obj = combined.obj,
  plotting.data. = plotting.data,
  AnnotCateg = AutoAnnotBy
)
}
\arguments{
\item{obj}{The Seurat object for which the 3D plot annotations will be generated. Default: combined.obj}

\item{plotting.data.}{The data frame containing plotting data. Default: plotting.data}

\item{AnnotCateg}{The category for which the annotation is generated. Default: AutoAnnotBy}
}
\description{
Create annotation labels for 3D plots. Source https://plot.ly/r/text-and-annotations/#3d-annotations.
}
