% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{AutoLabel.KnownMarkers}
\alias{AutoLabel.KnownMarkers}
\title{AutoLabel.KnownMarkers}
\usage{
AutoLabel.KnownMarkers(
  obj = combined.obj,
  topN = 1,
  res = 0.5,
  KnownMarkers = c(`dl-EN` = "KAZN", `ul-EN` = "SATB2", `Immature neurons` = "SLA",
    Interneurons = "DLX6-AS1", Interneurons = "ERBB4", InterN_CGE = "SCGN",
    `Intermediate progenitor` = "EOMES", `S-phase` = "TOP2A", `G2M-phase` = "H4C3", oRG =
    "HOPX", Astrocyte = "S100B", `Hypoxia/Stress` = "DDIT4", Glycolytic = "PDK1",
    Choroid.Plexus = "TTR", `Low-Quality` = "POLR2A", Mesenchyme = "DCN", Choroid.Plexus
    = "TTR"),
  order.by = c("combined.score", "avg_log2FC", "p_val_adj")[1],
  df_markers = obj@misc$df.markers[[paste0("res.", res)]]
)
}
\arguments{
\item{obj}{A Seurat object to work with. Default: \code{combined.obj}.}

\item{topN}{The top 'N' genes to consider. Default: 1.}

\item{res}{The clustering resolution to use. Default: 0.5.}

\item{KnownMarkers}{A character vector containing known marker genes to be used for annotation.
Default: \verb{c("TOP2A", "EOMES", "SLA", "HOPX", "S100B", "DLX6-AS1", "POU5F1", "SALL4", "DDIT4",}
\verb{"PDK1", "SATB2", "FEZF2")}.}

\item{order.by}{Specifies the column to sort the output tibble by.
Default: 'combined.score' (First among "combined.score", "avg_log2FC", "p_val_adj").}

\item{df_markers}{The data frame of markers. By default, it is stored under
\verb{@misc$df.markers$res...} in the provided Seurat object.
Default: \code{combined.obj@misc$df.markers[[paste0("res.", res)]]}.}
}
\description{
Creates a new "named identity" column in the metadata of a Seurat object,
setting 'Ident' to a clustering output matching the 'res' parameter.
This function requires the output table of \code{FindAllMarkers()}.
If you used \code{StoreAllMarkers()}, the output is stored under \verb{@misc$df.markers$res...},
which is the default location.
}
\examples{
\dontrun{
if (interactive()) {
  combined.obj <- AutoLabel.KnownMarkers()
  DimPlot.ClusterNames(ident = "cl.names.KnownMarkers.0.5")
}
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}
}
