% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{AutoLabelTop.logFC}
\alias{AutoLabelTop.logFC}
\title{AutoLabelTop.logFC}
\usage{
AutoLabelTop.logFC(
  obj = combined.obj,
  group.by,
  res = stringr::str_extract(group.by, "\\\\d+\\\\.\\\\d+"),
  plot.top.genes = TRUE,
  suffix = res,
  order.by = c("combined.score", "avg_log2FC", "p_val_adj")[2],
  exclude = c("^A[CFLP][0-9]{6}", "^Z[0-9]{5}", "^LINC0[0-9]{4}", "^C[1-9]+orf[1-9]+",
    "[-|\\\\.]AS[1-9]*$", "[-|\\\\.]DT[1-9]*$", "^MIR[1-9]", "^SNHG[1-9]"),
  df_markers = obj@misc$df.markers[[paste0("res.", res)]],
  plotEnrichment = TRUE
)
}
\arguments{
\item{obj}{A Seurat object, with default value \code{combined.obj}.}

\item{group.by}{The clustering group to be used, defaults to the first entry by
\code{GetClusteringRuns()}.}

\item{res}{Clustering resolution tag. Default: extracted from \code{group.by}.}

\item{plot.top.genes}{Logical indicating whether to show a plot, default is \code{TRUE}.}

\item{suffix}{Suffix for the naming, defaults to the value of \code{res}.}

\item{order.by}{Sorting criterion for the output tibble, defaults to the second element
of \code{c("combined.score", "avg_log2FC", "p_val_adj")}.}

\item{exclude}{A vector of regular expressions to specify genes to exclude, with
default value \code{c("^AL*|^AC*|^LINC*|^C[0-9]+orf[0-9]*")}.}

\item{df_markers}{Data frame resulting from DGEA analysis (\code{FindAllMarkers}). The default
is \code{combined.obj@misc$df.markers[[paste0("res.", res)]]}.}

\item{plotEnrichment}{Logical indicating whether to plot enrichment, default is \code{TRUE}.}
}
\description{
Create a new "named identity" column in the metadata of a Seurat object,
with \code{Ident} set to a clustering output matching the \code{res} parameter of the function.
It requires the output table of \code{FindAllMarkers()}.  If you used \code{StoreAllMarkers()}
is stored under \verb{@misc$df.markers$res...}, which location is assumed by default.
}
\examples{
\dontrun{
if (interactive()) {
  combined.obj <- AutoLabelTop.logFC()
  combined.obj$"cl.names.top.gene.res.0.5"
}
}
}
