% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{aux_plotAllMseqBCs}
\alias{aux_plotAllMseqBCs}
\title{aux_plotAllMseqBCs}
\usage{
aux_plotAllMseqBCs(
  bar.table = bar.table[, 1:96],
  barcodes.used = BCs.used,
  plotname = "Barcode seq depth"
)
}
\arguments{
\item{bar.table}{PARAM_DESCRIPTION, Default: bar.table\link{, 1:96}}

\item{barcodes.used}{PARAM_DESCRIPTION, Default: BCs.used}

\item{plotname}{Title of the plot, Default: 'Barcode seq depth'}
}
\description{
aux_plotAllMseqBCs
}
\examples{
\dontrun{
if(interactive()){
 aux_plotAllMseqBCs(bar.table = bar.table[,1:96], barcodes.used = BCs.used, plotname = "Barcode seq depth")
 }
}
}
