% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MULTI-seq.Utils.R
\name{BarTableSweepList}
\alias{BarTableSweepList}
\title{BarTableSweepList}
\usage{
BarTableSweepList(min = 0.01, max = 0.99, step = 0.02, bar_table = bar.table)
}
\arguments{
\item{min}{The minimum quantile value. Default: 0.01.}

\item{max}{The maximum quantile value. Default: 0.99.}

\item{step}{The step size to increment the quantile. Default: 0.02.}

\item{bar_table}{The barcode table to be used. Default: bar.table.}
}
\value{
A list of classified cells for each quantile in the specified range.
}
\description{
Generate a list of classified cells for a range of quantiles.
}
\examples{
\dontrun{
if(interactive()){
 bar.table_sweep.list <- BarTableSweepList(bar_table = bar.table.solo) # Quantile Sweep List
 }
}
}
