% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{BulkGEScatterPlot}
\alias{BulkGEScatterPlot}
\title{BulkGEScatterPlot}
\usage{
BulkGEScatterPlot(
  obj = combined.obj,
  clusters = "cl.names.KnownMarkers.0.2",
  TwoCategIdent = "age",
  genes.from.bulk.DE = rownames(df.markers.per.AGE)
)
}
\arguments{
\item{obj}{The Seurat object to use for the analysis. Default: combined.obj.}

\item{clusters}{A string specifying the identity class in the Seurat object to use for cluster assignment. Default: 'cl.names.KnownMarkers.0.2'.}

\item{TwoCategIdent}{A string specifying the binary categorical identity to split the data for comparison. Default: 'age'.}

\item{genes.from.bulk.DE}{A character vector specifying the genes obtained from bulk differential expression analysis to be highlighted in the scatterplots. Default: rownames(df.markers.per.AGE).}
}
\description{
Plots scatterplots of bulk gene expression to identify differentially expressed genes across conditions.
}
\examples{
\dontrun{
if (interactive()) {
  BulkGEScatterPlot(obj = combined.obj, clusters = "cl.names.KnownMarkers.0.2", TwoCategIdent = "age", genes.from.bulk.DE = rownames(df.markers.per.AGE))
}
}
}
