% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{calc.cluster.averages}
\alias{calc.cluster.averages}
\title{calc.cluster.averages}
\usage{
calc.cluster.averages(
  col_name = "Score.GO.0006096",
  plot.UMAP.too = TRUE,
  return.plot = FALSE,
  obj = combined.obj,
  split_by = GetNamedClusteringRuns()[1],
  scale.zscore = FALSE,
  simplify = TRUE,
  plotit = TRUE,
  histogram = FALSE,
  nbins = 50,
  suffix = NULL,
  stat = c("mean", "median")[2],
  quantile.thr = 0.9,
  absolute.thr = FALSE,
  filter = c(FALSE, "above", "below")[1],
  ylab.text = paste("Cluster", stat, "score"),
  title = paste("Cluster", stat, col_name),
  prefix.cl.names = FALSE,
  report = TRUE,
  subtitle = NULL,
  width = 8,
  height = 6,
  ...,
  xlb = if (absolute.thr) {
     paste("Threshold at", absolute.thr)
 } else {
    
    paste("Black lines: ", Stringendo::kppd(Stringendo::percentage_formatter(c(1 -
    quantile.thr, quantile.thr))), "quantiles |", "Cl. >",
    Stringendo::percentage_formatter(quantile.thr), "are highlighted. |", split_by)
 },
  fname = Stringendo::ppp(col_name, split_by, "cluster.average.barplot.pdf", ...)
)
}
\arguments{
\item{col_name}{The name of the column for which the average is calculated. Default: 'Score.GO.0006096'.}

\item{plot.UMAP.too}{Whether to plot a UMAP as well. Default: \code{TRUE}.}

\item{return.plot}{Whether to return the plot. Default: \code{FALSE}.}

\item{obj}{The main Seurat object used for calculations. Default: \code{combined.obj}.}

\item{split_by}{Cluster to split by. Default: First entry of GetNamedClusteringRuns().}

\item{scale.zscore}{Whether to scale z-scores. Default: \code{FALSE}.}

\item{simplify}{Whether to simplify the result. Default: \code{TRUE}.}

\item{plotit}{Whether to plot the results. Default: \code{TRUE}.}

\item{histogram}{Whether to produce a histogram. Default: \code{FALSE}.}

\item{nbins}{The number of bins for the histogram. Default: 50.}

\item{suffix}{Suffix added to the filename. Default: NULL.}

\item{stat}{Statistical method applied, "mean" or "median". Default: "median".}

\item{quantile.thr}{The threshold for quantiles. Default: 0.9.}

\item{absolute.thr}{Absolute threshold used in computations. Default: \code{FALSE}.}

\item{filter}{The filter mode: 'above', 'below', or FALSE. Default: \code{FALSE}.}

\item{ylab.text}{Text for the y-axis label. Default: "Cluster" followed by the statistical method and "score".}

\item{title}{Title for the plot. Default: "Cluster" followed by the statistical method and column name.}

\item{prefix.cl.names}{Whether to prefix cluster names with the split_by column name. Default: \code{FALSE}.}

\item{report}{Whether to print a message reporting the object and split_by column. Default: \code{TRUE}.}

\item{subtitle}{The subtitle for the plot. Default: NULL.}

\item{width}{The width of the plot. Default: 8.}

\item{height}{The height of the plot. Default: 6.}

\item{...}{Additional parameters passed to the internally called functions.}

\item{xlb}{The label for the x-axis. Default depends on the 'absolute.thr' parameter.}

\item{fname}{The filename for the plot. Default: based on column name and split_by value.}
}
\description{
Calculates the average of a metadata column (numeric) per cluster.
}
