% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{calc.q90.Expression.and.set.all.genes}
\alias{calc.q90.Expression.and.set.all.genes}
\title{calc.q90.Expression.and.set.all.genes}
\usage{
calc.q90.Expression.and.set.all.genes(
  obj = combined.obj,
  quantileX = 0.9,
  max.cells = 1e+05,
  slot = "data",
  assay = c("RNA", "integrated")[1],
  set.all.genes = TRUE,
  show = TRUE
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{quantileX}{Quantile level, Default: 0.9}

\item{max.cells}{PARAM_DESCRIPTION, Default: 1e+05}

\item{slot}{PARAM_DESCRIPTION, Default: 'data'}

\item{assay}{PARAM_DESCRIPTION, Default: c("RNA", "integrated")\link{1}}

\item{set.all.genes}{PARAM_DESCRIPTION, Default: TRUE}

\item{show}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Calculate the gene expression of the e.g.: 90th quantile (expression in the top 10\% cells). #
}
\examples{
\dontrun{
if(interactive()){
 combined.obj <- calc.q90.Expression.and.set.all.genes(obj = combined.obj, quantileX=0.9,
 max.cells =  25000)
 head(sort(as.numeric.wNames(obj@misc$expr.q90), decreasing = T))
 combined.obj <- calc.q90.Expression.and.set.all.genes(obj = combined.obj, quantileX=0.95,
 max.cells =  25000, set.all.genes = FALSE)
 }
}
}
\seealso{
\code{\link[sparseMatrixStats]{character(0)}}
}
