% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{calculate.observable.multiplet.rate.10X}
\alias{calculate.observable.multiplet.rate.10X}
\title{calculate.observable.multiplet.rate.10X}
\usage{
calculate.observable.multiplet.rate.10X(
  fractions = c(`176` = 0.07, h9 = 0.93),
  multiplet.rate = FALSE,
  empir.factor = c(ST = 131000, HT = 250000)[1],
  draw_plots = TRUE,
  cell.count,
  suffix = "",
  ...
)
}
\arguments{
\item{fractions}{Genotype fractions observed or loaded on the machine.}

\item{multiplet.rate}{Expected multiplet rate for your number of recovered cells. Default FALSE. If FALSE, it auto-calculates it based on an inferred FMR: Factor for calculating multiplet rate: Cell count / FMR = multiplet rate.}

\item{empir.factor}{Factor for calculating Multiplet Rate: Cell count / FMR = multiplet rate. (.r stands for 'estimating from Recovered cell count.')}

\item{cell.count}{Number of total cells recovered from a 10X lane.}

\item{suffix}{A suffix added to the filename, Default: ''}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Calculate the observable multiplet rate for 10X standard lane.
}
