% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{calculate.observable.multiplet.rate.10X_obj_based}
\alias{calculate.observable.multiplet.rate.10X_obj_based}
\title{Calculate Observable Multiplet Rate for 10X Objects}
\usage{
calculate.observable.multiplet.rate.10X_obj_based(
  obj,
  chipset,
  empir.factors = c(ST = 131000, HT = 250000),
  draw_plots = TRUE,
  gtt = "Genotype"
)
}
\arguments{
\item{obj}{An object for which the observable multiplet rate will be calculated. No default value is provided.}

\item{chipset}{The chipset used in the experiment. No default value is provided.}

\item{empir.factors}{A named vector of empirical factors. Default is \code{c('ST' = 131000, 'HT' = 250000)}.}

\item{draw_plots}{A logical flag indicating whether to draw plots. Default is \code{TRUE}.}

\item{gtt}{A string representing the genotype tag. Default is \code{'Genotype'}.}
}
\description{
This function calculates the observable multiplet rate for 10X Genomics objects based on the given parameters and returns the result. It requires a valid object, chipset, empirical factors, a flag for drawing plots, and a genotype tag. The default empirical factors are \code{c('ST' = 131000, 'HT' = 250000)}, and by default, it draws plots.
}
\examples{
\dontrun{
calculate.observable.multiplet.rate.10X_obj_based(obj = my_obj, chipset = "my_chipset", draw_plots = FALSE, gtt = "MyGenotype")
}
}
