% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{CalculateFractionInTrome}
\alias{CalculateFractionInTrome}
\title{CalculateFractionInTranscriptome}
\usage{
CalculateFractionInTrome(
  genesCalc.Cor.Seuratet = c("MALAT1"),
  obj = combined.obj,
  dataslot = c("counts", "data")[2]
)
}
\arguments{
\item{obj}{A Seurat object from which the gene data is extracted. Default is 'combined.obj'.}

\item{dataslot}{A character vector specifying the data slot to be used in the calculation.
Default is 'data' (second element in the vector c("counts", "data")).}

\item{geneset}{A character vector specifying the set of genes for which the fraction in the
transcriptome is to be calculated. Default is c("MALAT1").}
}
\value{
A numeric vector containing the fraction of the specified genes in the transcriptome of each cell.
}
\description{
This function calculates the fraction of a set of genes within the full transcriptome of each cell.
}
