% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{calculatePercentageMatch}
\alias{calculatePercentageMatch}
\title{Calculate the Percentage of Matches per Category}
\usage{
calculatePercentageMatch(
  obj,
  ident = GetClusteringRuns()[1],
  meta.features = c("AAV.detected.min2", "AAV.detected"),
  match.values = c(AAV.detected.min2 = "AAV", AAV.detected = "AAV"),
  verbose = TRUE,
  max.categ = 100
)
}
\arguments{
\item{obj}{A Seurat object containing the data to be analyzed. Default: combined.obj.}

\item{ident}{A string specifying the column in the metadata that identifies the categories.
Default: first element of \code{GetClusteringRuns()}.}

\item{meta.features}{A vector of strings specifying which metadata features to analyze.
Default: c("AAV.detected.min2", "AAV.detected").}

\item{match.values}{A named vector where names correspond to \code{meta.features} and values are
the strings to match against. Default: c("AAV.detected.min2" = "AAV", "AAV.detected" = "AAV").}

\item{verbose}{A logical value indicating whether to print detailed output. Default: TRUE.}

\item{max.categ}{The maximum number of categories allowed before stopping. Default: 30.}
}
\value{
A data frame with the category as the first column and the subsequent columns showing
the percentage of matches for each metadata feature.
}
\description{
This function calculates the percentage of matches for specified metadata features
against provided match values within each category of an identifier in a Seurat object.
}
\examples{
calculatePercentageMatch(obj = combined.obj, ident = "Simple_Celltypes")
}
