% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slingshot.Utils.R
\name{cell_pal}
\alias{cell_pal}
\title{cell_pal}
\usage{
cell_pal(cell_vars, pal_fun, ...)
}
\arguments{
\item{cell_vars}{A vector of cell variables. Default: PARAM_DEFAULT_VALUE.}

\item{pal_fun}{A function to generate the color palette. Default: PARAM_DEFAULT_VALUE.}
}
\value{
A color palette.
}
\description{
Generate a color palette for cell variables.
}
