% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{CellFractionsBarplot2}
\alias{CellFractionsBarplot2}
\title{CellFractionsBarplot2}
\usage{
CellFractionsBarplot2(
  obj = combined.obj,
  group.by = "integrated_snn_res.0.5.ordered",
  fill.by = "age",
  downsample = T,
  plotname = paste(tools::toTitleCase(fill.by), "proportions"),
  hlines = c(0.25, 0.5, 0.75),
  seedNr = 1989
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{group.by}{PARAM_DESCRIPTION, Default: 'integrated_snn_res.0.5.ordered'}

\item{fill.by}{PARAM_DESCRIPTION, Default: 'age'}

\item{downsample}{PARAM_DESCRIPTION, Default: T}

\item{plotname}{Title of the plot, Default: paste(tools::toTitleCase(fill.by), "proportions")}

\item{hlines}{PARAM_DESCRIPTION, Default: c(0.25, 0.5, 0.75)}

\item{seedNr}{PARAM_DESCRIPTION, Default: 1989}
}
\description{
CellFractionsBarplot2
}
\examples{
\dontrun{
if(interactive()){
 CellFractionsBarplot2(obj = combined.obj, group.by = "integrated_snn_res.0.1", fill.by = "Phase", downsample = T)
 CellFractionsBarplot2(obj = combined.obj, group.by = "integrated_snn_res.0.1", fill.by = "Phase", downsample = F)
 }
}
}
\seealso{
\code{\link[tools]{toTitleCase}}
}
