% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{check.genes}
\alias{check.genes}
\title{Check if Gene Names exist in Seurat Object or HGNC Database}
\usage{
check.genes(
  genes,
  makeuppercase = FALSE,
  HGNC.lookup = FALSE,
  obj,
  assay.slot = c("RNA", "integrated")[1],
  data.slot = c("counts", "data")[2],
  verbose = TRUE,
  ...
)
}
\arguments{
\item{genes}{A vector of gene names to be checked.}

\item{makeuppercase}{If \code{TRUE}, converts all gene names to uppercase before checking. Default: \code{FALSE}.}

\item{HGNC.lookup}{If \code{TRUE}, attempts to look up any missing genes in the HGNC database to
verify their existence. Default: \code{FALSE}.}

\item{obj}{The Seurat object against which the gene names will be checked.}

\item{assay.slot}{Assay slot of the Seurat object to check for gene names. Default: \code{'RNA'}.}

\item{data.slot}{Data slot of the assay to check for gene names. Default: \code{'data'}.}

\item{verbose}{If \code{TRUE}, prints information about any missing genes. Default: \code{TRUE}.}

\item{...}{Additional arguments, not used.}
}
\description{
Verifies the presence of specified gene names within a Seurat object or
queries them against the HGNC database. This function is useful for ensuring gene names are
correctly formatted and exist within the dataset or are recognized gene symbols.
}
\examples{
\dontrun{
if (interactive()) {
  # Check for the presence of a gene name in uppercase
  check.genes(genes = "top2a", makeuppercase = TRUE, obj = combined.obj)

  # Check for a gene name with verbose output and HGNC lookup
  check.genes(genes = "VGLUT2", verbose = TRUE, HGNC.lookup = TRUE, obj = combined.obj)
}
}

}
\seealso{
\code{\link[DatabaseLinke.R]{qHGNC}}
}
