% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{clip10Xcellname}
\alias{clip10Xcellname}
\title{clip10Xcellname}
\usage{
clip10Xcellname(cellnames)
}
\arguments{
\item{cellnames}{Character vector containing the cellIDs (with numeric suffixes).}
}
\description{
Clip all suffices after underscore (10X adds it per chip-lane, Seurat adds in during integration). #
}
