% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Convert10Xfolders.old}
\alias{Convert10Xfolders.old}
\title{Convert10Xfolders.old}
\usage{
Convert10Xfolders.old(
  InputDir,
  folderPattern = c("filtered", "SoupX_decont")[1],
  min.cells = 10,
  min.features = 200,
  updateHGNC = T,
  ShowStats = T
)
}
\arguments{
\item{InputDir}{Input directory}

\item{folderPattern}{PARAM_DESCRIPTION, Default: c("filtered", "SoupX_decont")\link{1}}

\item{min.cells}{PARAM_DESCRIPTION, Default: 10}

\item{min.features}{PARAM_DESCRIPTION, Default: 200}

\item{updateHGNC}{PARAM_DESCRIPTION, Default: T}

\item{ShowStats}{PARAM_DESCRIPTION, Default: T}
}
\description{
Take a parent directory with a number of subfolders, each containing the standard output of 10X Cell Ranger. (1.) It loads the filtered data matrices; (2.) converts them to Seurat objects, and (3.) saves them as *.RDS files. #
}
\examples{
\dontrun{
if(interactive()){
 Convert10Xfolders(InputDir = InputDir)
 }
}
}
