% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{copyCompleteToolsSlots}
\alias{copyCompleteToolsSlots}
\title{Copy Tools Slots from Multiple Seurat Objects}
\usage{
copyCompleteToolsSlots(
  ls.obj,
  obj.to,
  overwrite = TRUE,
  new.slot = "per.experiment"
)
}
\arguments{
\item{ls.obj}{A list of Seurat objects from which the \verb{@tools} slots will be copied.}

\item{obj.to}{The target Seurat object to which the \verb{@tools} slots will be added.}

\item{overwrite}{A logical parameter that is kept for compatibility but not used in this version.
Its presence does not affect the function's behavior.}

\item{new.slot}{The name of the new slot within \code{obj.to@tools} where the copied \verb{@tools} information
will be stored. This allows for the organization of copied tools under a specific label, facilitating
easy access and interpretation.}
}
\value{
Returns the modified target Seurat object (\code{obj.to}) with a new \verb{@tools} slot containing
the copied information from the list of Seurat objects.
}
\description{
This function copies the \verb{@tools} slots from a list of Seurat objects into a new slot
of a target Seurat object. This allows for the aggregation of tools information from multiple
experiments or datasets into a single, consolidated Seurat object.
}
\examples{
# Assuming `ls.obj` is a list of Seurat objects and `obj.to` is a target Seurat object
obj.to <- copyCompleteToolsSlots(ls.obj, obj.to, overwrite = TRUE, new.slot = "per.experiment")
}
