% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{copyMiscElements}
\alias{copyMiscElements}
\title{Copy Specified Elements from One Seurat Object's @misc to Another's}
\usage{
copyMiscElements(obj.from, obj.to, elements.needed, overwrite = TRUE)
}
\arguments{
\item{obj.from}{The source Seurat object from which elements in the \verb{@misc} slot are to be copied.}

\item{obj.to}{The destination Seurat object to which elements in the \verb{@misc} slot are to be copied.}

\item{elements.needed}{A vector of strings specifying the names of the elements in the \verb{@misc} slot of
\code{obj.from} that should be copied to \code{obj.to}.}

\item{overwrite}{Logical indicating whether to overwrite elements in \code{obj.to} that already exist.
If \code{TRUE}, existing elements will be overwritten with a warning; if \code{FALSE}, the function will
stop with an error if it tries to copy an element that already exists in \code{obj.to}.}
}
\value{
Returns the modified destination Seurat object (\code{obj.to}) with the specified elements
added to or updated in its \verb{@misc} slot.
}
\description{
Copies specified elements from the \verb{@misc} slot of one Seurat object to the \verb{@misc} slot
of another. It warns if some specified elements are missing in the source object or if elements are
overwritten in the destination object, depending on the \code{overwrite} argument.
}
\examples{
# Assuming `obj1` and `obj2` are Seurat objects and you wish to copy specific elements
# from obj1 to obj2, possibly overwriting existing elements in obj2
obj2 <- copyMiscElements(obj1, obj2, c("element1", "element2"), overwrite = TRUE)

}
