% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.utils.less.used.R
\name{Create.MiscSlot}
\alias{Create.MiscSlot}
\title{Create.MiscSlot}
\usage{
Create.MiscSlot(obj, NewSlotName = "UVI.tables", SubSlotName = NULL)
}
\arguments{
\item{obj}{Seurat object}

\item{NewSlotName}{Name of the new element inside obj@misc.}

\item{SubSlotName}{Optional name of a sub-element inside \code{obj@misc[[NewSlotName]]}. Default: \code{NULL}.}
}
\description{
Create a new slot in the 'misc' slot of a Seurat object.
}
