% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{create.overlap.vector}
\alias{create.overlap.vector}
\title{create.overlap.vector}
\usage{
create.overlap.vector(vec = All.UVI, obj = combined.obj, min.intersect = 100)
}
\arguments{
\item{vec}{cell - ID vector with partial overlap to cells in a Seurat object.}

\item{obj}{Seurat object}

\item{min.intersect}{Min number of cells to find in both.}
}
\description{
create.overlap.vector
}
