% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{create_scCombinedMeta}
\alias{create_scCombinedMeta}
\title{Create Single-Cell Metadata Object for a collection of Seurat Objects}
\usage{
create_scCombinedMeta(experiment, project_ = CodeAndRoll2::getProject())
}
\arguments{
\item{experiment}{The name of the experiment for which metadata is being created.}

\item{project_}{The project information to be associated with the metadata object.
This defaults to the current project obtained using Seurat.utils::getProject().}
}
\value{
An 'scCollectionMetadata_class' object containing the metadata for a collection of experiment.
}
\description{
This function creates a metadata object to correspond to a list of
single-cell experiments, for storing parent level information.
It initializes the object with the experiment and project name, and the
creation date. The created object is of class 'scMetadata_class'.
}
\examples{
sc_meta <- create_scCombinedMeta(experiment = "Experiment1")
}
