% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{DimPlot.ClusterNames}
\alias{DimPlot.ClusterNames}
\title{DimPlot.ClusterNames}
\usage{
DimPlot.ClusterNames(
  obj = combined.obj,
  ident = "cl.names.top.gene.res.0.5",
  reduction = "umap",
  title = ident,
  ...
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{ident}{identity used, Default: 'cl.names.top.gene.res.0.5'}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'umap'}

\item{title}{Title of the plot, Default: ident}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot UMAP with Cluster names. #
}
\examples{
\dontrun{
if (interactive()) {
  DimPlot.ClusterNames()
}
}
}
