% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{DiscretePaletteSafe}
\alias{DiscretePaletteSafe}
\title{Safely generate a discrete color palette.}
\usage{
DiscretePaletteSafe(
  n,
  palette.used = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[2],
  show.colors = FALSE,
  seed = 1989
)
}
\arguments{
\item{n}{The number of colors to generate.}

\item{palette.used}{The name of the palette to use, Default: "alphabet".}

\item{show.colors}{Whether to display the colors in the palette, Default: FALSE.}

\item{seed}{An integer value to set the seed for reproducibility, Default: 1989.}
}
\value{
A vector of color values.
}
\description{
Generates a discrete color palette safely without NA values.
}
\examples{
\dontrun{
if (interactive()) {
  DiscretePaletteSafe(n = 10)
}
}
}
