% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{.check_and_rename}
\alias{.check_and_rename}
\title{Check and Rename Gene Names in Seurat Assay Object}
\usage{
.check_and_rename(obj, assay, newnames, slotname)
}
\arguments{
\item{obj}{A Seurat object.}

\item{assay}{An Assay name in a Seurat object.}

\item{newnames}{A character vector of new gene names to be assigned.}

\item{slotname}{A string specifying the slot in the Assay object to be updated.
Valid options typically include 'counts', 'data', or 'scale.data'.}
}
\value{
An Assay object with updated gene names in the specified slot.
}
\description{
This function renames rows (genes) in a specified slot of a Seurat assay object.
It supports slots storing data as either a dense or a sparse matrix (dgCMatrix) or data.frame.
}
\examples{
\dontrun{
  # Assuming 'seurat_obj' is a Seurat object and 'new_gene_names' is a vector of gene names
  updated_assay <- check_and_rename(assayobj = seurat_obj[["RNA"]],
                                    newnames = new_gene_names,
                                    slotname = "counts")
}
}
