% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{.getCPUsCBE}
\alias{.getCPUsCBE}
\title{Get the number of CPUs to use for CBE processing}
\usage{
.getCPUsCBE(n.cpus.def = 8)
}
\arguments{
\item{n.cpus.def}{The default number of CPUs to return if \code{CBE.params} does not exist
or does not contain a \code{cpus} entry. Defaults to 8.}
}
\value{
The number of CPUs to use for CBE processing. If \code{CBE.params$cpus} is set,
returns \code{CBE.params$cpus - 1}, ensuring at least 1 CPU is returned. Otherwise, returns \code{n.cpus.def}.
}
\description{
This function checks for the presence of a global \code{CBE.params} list and,
if found and contains a \code{cpus} entry, returns the number of CPUs specified by \code{cpus} minus one.
Otherwise, it returns a default number of CPUs.
}
\examples{
# Assuming CBE.params does not exist or does not have a `cpus` entry
getCPUsCBE() # returns 8 by default

# Assuming CBE.params exists and has a `cpus` entry of 4
getCPUsCBE() # returns 3
}
