% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{.getNrPCs}
\alias{.getNrPCs}
\title{Get number of principal components}
\usage{
.getNrPCs(obj, v = TRUE, reduc = "pca")
}
\arguments{
\item{obj}{A Seurat object containing PCA cell embeddings in \code{reductions$pca@cell.embeddings}}

\item{v}{Verbose? Default: \code{TRUE}.}

\item{reduc}{Name of the dimensionality reduction to use. Default: \code{'pca'}.}
}
\value{
Integer representing the number of principal components
}
\description{
Get number of principal components
}
