% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{.getNrScaledFeatures}
\alias{.getNrScaledFeatures}
\title{Get number of scaled features}
\usage{
.getNrScaledFeatures(
  obj,
  assay = Seurat::DefaultAssay(obj),
  obj.version = obj@version,
  v = TRUE
)
}
\arguments{
\item{obj}{A Seurat object containing scaled data in  \code{obj@assays$RNA@scale.data}.}

\item{assay}{The name of the assay to search for scaled data. Default: \code{RNA}.}

\item{obj.version}{Seurat object version, used to determine which layer/slot API to use. Default: \code{obj@version}.}

\item{v}{Verbose? Default: \code{TRUE}.}
}
\value{
Integer representing the number of scaled features
}
\description{
Get number of scaled features
}
