% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{.getRegressionVariablesForScaleData}
\alias{.getRegressionVariablesForScaleData}
\title{Parse regression variables for name tagging}
\usage{
.getRegressionVariablesForScaleData(
  obj,
  assay = Seurat::DefaultAssay(obj),
  v = TRUE,
  ...
)
}
\arguments{
\item{obj}{A Seurat object}

\item{assay}{The name of the assay to search for scaled data. Default: \code{DefaultAssay()}.}

\item{v}{Verbose? Default: \code{TRUE}.}

\item{...}{Additional arguments, not used.}
}
\value{
Integer representing the number of principal components
}
\description{
This function extracts the regression variables from the \verb{@commands} slot of a Seurat object.
If no regression variables are found, a message is printed.
}
