% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{.parseBasicObjStats}
\alias{.parseBasicObjStats}
\title{Parse basic obj stats}
\usage{
.parseBasicObjStats(
  obj,
  sep = " ",
  assay = DefaultAssay(obj),
  simple = FALSE,
  suffix = NULL
)
}
\arguments{
\item{obj}{An object to extract information from.}

\item{sep}{Character used as the big-mark separator in formatted numbers. Default: \code{' '}.}

\item{assay}{Assay used to determine the feature count. Default: \code{DefaultAssay(obj)}.}

\item{simple}{If \code{TRUE}, only report the cell count. Default: \code{FALSE}.}

\item{suffix}{Optional character string appended to the summary. Default: \code{NULL}.}
}
\value{
A character string summarizing the key parameters.
}
\description{
Parse cell and feature count to a string.
}
