% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{.parseKeyParams}
\alias{.parseKeyParams}
\title{Parse key parameters from an object and format as a string}
\usage{
.parseKeyParams(
  obj,
  regressionVariables = p$variables.2.regress.combined,
  suffix = NULL
)
}
\arguments{
\item{obj}{An object to extract information from.}

\item{regressionVariables}{A list or vector containing variables for regression.}

\item{suffix}{A suffix string to add.}
}
\value{
A character string summarizing the key parameters.
}
\description{
This function extracts the number of scaled features, the number of principal components,
and formats additional information including regression variables.
}
