% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{.parseRegressionVariablesForScaleData}
\alias{.parseRegressionVariablesForScaleData}
\title{Parse regression variables for name tagging}
\usage{
.parseRegressionVariablesForScaleData(
  p,
  element = "variables.2.regress.combined"
)
}
\arguments{
\item{p}{list of parameters}
}
\value{
Integer representing the number of principal components
}
\description{
Parse regression variables for name tagging
}
