% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{.saveRDS.compress.in.BG}
\alias{.saveRDS.compress.in.BG}
\title{.saveRDS.compress.in.BG}
\usage{
.saveRDS.compress.in.BG(
  obj,
  compr = FALSE,
  fname,
  compress_internally = FALSE,
  ...
)
}
\arguments{
\item{obj}{Seurat object.}

\item{compr}{Compress at all? Default: TRUE}

\item{fname}{File name}

\item{compress_internally}{Compress by R? Default: FALSE (still compressed in background via CLI).}

\item{...}{Additional parameters passed to saveRDS() function.}
}
\description{
Save and RDS object and compress resulting file in the background using system(gzip). OS X or unix.
}
\seealso{
\code{\link[tictoc]{tic}}
}
