% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Downsample.Seurat.Objects}
\alias{Downsample.Seurat.Objects}
\title{Downsample.Seurat.Objects}
\usage{
Downsample.Seurat.Objects(ls.obj = ls.Seurat, NrCells = p$dSample.Organoids)
}
\arguments{
\item{ls.obj}{List of Seurat objects, Default: ls.Seurat}

\item{NrCells}{PARAM_DESCRIPTION, Default: p$dSample.Organoids}
}
\description{
Downsample.Seurat.Objects
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
