% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{downsampleListSeuObjs}
\alias{downsampleListSeuObjs}
\title{downsampleListSeuObjs}
\usage{
downsampleListSeuObjs(
  ls.obj = ls.Seurat,
  NrCells = p$dSample.Organoids,
  save_object = FALSE
)
}
\arguments{
\item{ls.obj}{List of Seurat objects. Default: ls.Seurat}

\item{NrCells}{Number of cells to downsample to.}

\item{save_object}{save object by isaveRDS, otherwise return it. Default: FALSE}
}
\description{
downsample a list of Seurat objects
}
\examples{
\dontrun{
if (interactive()) {
  downsampleListSeuObjs(NrCells = 2000)
  downsampleListSeuObjs(NrCells = 200)
}
}
}
