% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{downsampleSeuObj}
\alias{downsampleSeuObj}
\title{downsampleSeuObj}
\usage{
downsampleSeuObj(
  obj = ls.Seurat[[i]],
  fractionCells = 0.25,
  nCells = FALSE,
  seed = 1989
)
}
\arguments{
\item{obj}{A Seurat object to subset. Default: the i-th element of the list 'ls.Seurat'.}

\item{fractionCells}{The fraction of the object's data to keep. Default: 0.25.}

\item{nCells}{If set to a number greater than 1, indicates the absolute number of cells to keep.
If FALSE, the function uses 'fractionCells' to determine the number of cells. Default: FALSE.}

\item{seed}{A seed for random number generation to ensure reproducible results. Default: 1989.}
}
\description{
Subset a compressed Seurat object and save it in the working directory.
}
