% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{drop.levels.Seurat}
\alias{drop.levels.Seurat}
\title{drop.levels.Seurat}
\usage{
drop.levels.Seurat(obj = combined.obj)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}
}
\description{
Drop levels in clustering vectors in metadata (e.g. after subsetting)
}
