% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{filterCodingGenes}
\alias{filterCodingGenes}
\title{Filter Coding Gene Symbols (or any matching input Patterns)}
\usage{
filterCodingGenes(
  genes,
  pattern_NC = c("^A[CFLP][0-9]{6}", "^Z[0-9]{5}", "^LINC0[0-9]{4}", "^C[1-9]+orf[1-9]+",
    "^C[1-9][0-9]+orf[1-9]+", "^CXorf[1-9]+", "[-|\\\\.]AS[1-9]*$", "[-|\\\\.]DT[1-9]*$",
    "^MIR[1-9]", "^SNHG[1-9]", "^CU[0-9]{6}", "^BX[0-9]{6}", "^FP[0-9]{6}",
    "^AC[0-9]{6}"),
  v = TRUE,
  unique = TRUE,
  ...
)
}
\arguments{
\item{genes}{A character vector of gene symbols.}

\item{pattern_NC}{A character vector of patterns to filter out non-coding gene symbols.
Default: \verb{c("^AC.", "^AL.", "^c[1-9]orf", "\\\\.AS[1-9]$", ... etc.)}}

\item{v}{"verbose" Whether to print the number of genes before and after filtering.}

\item{unique}{Whether to return unique gene symbols. Default: \code{TRUE}.}

\item{...}{Additional arguments to pass to \code{\link[stringr]{str_detect}}.}
}
\value{
A character vector of filtered gene symbols.
}
\description{
This function filters out gene names that match specified patterns. It reports
the original and final number of gene symbols and the percentage remaining after filtering.
It filters out non-coding gene symbols by default.
}
\examples{
genes <- c("AC123", "AL456", "c1orf7", "TP53", "BRCA1", "X1.AS1", "MYC")
genes_kept <- filterCodingGenes(genes)
print(genes_kept)

}
