% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{filterNcGenes}
\alias{filterNcGenes}
\title{Filter Coding Gene Symbols (or any matching input Patterns)}
\usage{
filterNcGenes(
  genes,
  pattern_NC = c("^AC[0-9].+", "^AL[0-9].+", "^AP[0-9].+", "^AF[0-9].+", "^Z[0-9]+.+",
    "^LINC0.+", "^C[1-9]orf.+", "^MIR[1-9].+", ".+\\\\.AS[1-9]$", ".+\\\\.DT[1-9]$"),
  unique = TRUE,
  ...
)
}
\arguments{
\item{genes}{A character vector of gene symbols.}

\item{pattern_NC}{A character vector of patterns to filter out non-coding gene symbols.
Default: c("^AC.", "^AL.", "^c\link{1-9}orf", "\\.AS\link{1-9}$").}
}
\value{
A character vector of filtered gene symbols.
}
\description{
This function filters out gene names that match specified patterns. It reports
the original and final number of gene symbols and the percentage remaining after filtering.
It filters out non-coding gene symbols by default.
}
\examples{
genes <- c("AC123", "AL456", "c1orf7", "TP53", "BRCA1", "X1.AS1", "MYC")
genes_kept <- filterNcGenes(genes)
print(genes_kept)

}
