% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{find10XoutputFolders}
\alias{find10XoutputFolders}
\title{Find 'Outs' Subdirectories in Specified Subdirectories}
\usage{
find10XoutputFolders(root_dir, subdir, recursive = TRUE)
}
\arguments{
\item{root_dir}{The root directory.}

\item{subdir}{A character vector of subdirectory names within the root directory to be scanned.}

\item{recursive}{Boolean indicating whether to search recursively within subdirectories.}
}
\value{
A character vector containing the full paths to the 'outs' subdirectories.
}
\description{
This function searches through specified subdirectories within a root directory
to find all subdirectories named 'outs' and returns a character vector with their full paths.
}
