% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{find_prefix_in_cell_IDs}
\alias{find_prefix_in_cell_IDs}
\title{Check Prefix in Seurat Object Cell IDs}
\usage{
find_prefix_in_cell_IDs(obj, cell_ID_pattern = "[ATCG]{16}.*$")
}
\arguments{
\item{obj}{A Seurat object with cell IDs possibly prefixed.}

\item{cell_ID_pattern}{Pattern to match cellIDs (with any suffix).}
}
\value{
A character vector of the identified prefix(es).
}
\description{
This function checks if a prefix has been added to the standard
cell-IDs (16 characters of A,TRUE,C,G) in a Seurat object. If so, it prints the number of unique prefixes found,
issues a warning if more than one unique prefix is found, and returns the identified prefix(es).
}
\examples{
# Assuming 'obj' is your Seurat object
# prefix <- find_prefix_in_cell_IDs(obj)

}
