% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{FindCorrelatedGenes}
\alias{FindCorrelatedGenes}
\title{FindCorrelatedGenes}
\usage{
FindCorrelatedGenes(
  gene = "TOP2A",
  obj = combined.obj,
  assay = "RNA",
  slot = "data",
  HEonly = FALSE,
  minExpr = 1,
  minCells = 1000,
  trailingNgenes = 1000
)
}
\arguments{
\item{gene}{Gene of interest. Default: 'TOP2A'}

\item{obj}{Seurat object to find the correlated genes from. Default: combined.obj}

\item{assay}{Assay to be used from the Seurat object. Default: 'RNA'}

\item{slot}{Slot to be used from the specified assay in the Seurat object. Default: 'data'}

\item{HEonly}{Logical, if TRUE, filters matrix to high-expressing genes only. Default: FALSE}

\item{minExpr}{Minimum expression level for a gene to be considered. Default: 1}

\item{minCells}{Minimum number of cells expressing a gene for the gene to be considered. Default: 1000}

\item{trailingNgenes}{Number of top genes to consider based on their correlation. Default: 1000}
}
\description{
Find correlated genes in a Seurat object
}
\examples{
\dontrun{
if (interactive()) {
  FindCorrelatedGenes(gene = "TOP2A", obj = combined.obj)
  write_clip(names(head(topGenes[-(1:6)], n = 50)))
}
}
}
\seealso{
\code{\link[matrixStats]{rowSums2}}
}
