% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{fix.orig.ident}
\alias{fix.orig.ident}
\title{fix.orig.ident}
\usage{
fix.orig.ident(obj = merged.obj)
}
\arguments{
\item{obj}{Seurat object, Default: merged.obj}
}
\description{
fix.orig.ident
}
\examples{
\dontrun{
if(interactive()){
 merged.obj$orig.ident <- fix.orig.ident(obj = merged.obj); table(merged.obj$orig.ident)
 }
}
}
