% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{FlipReductionCoordinates}
\alias{FlipReductionCoordinates}
\title{Flip Reduction Coordinates}
\usage{
FlipReductionCoordinates(
  obj = combined.obj,
  dim = 2,
  reduction = "umap",
  flip = c("x", "y", "xy", NULL)[1],
  FlipReductionBackupToo = TRUE
)
}
\arguments{
\item{obj}{Seurat object to modify; Default: \code{combined.obj}.}

\item{dim}{Number of dimensions in the reduction to consider; Default: 2.}

\item{reduction}{Dimension reduction technique to modify ('umap', 'tsne', or 'pca'); Default: 'umap'.}

\item{flip}{Axis (or axes) to flip; can be 'x', 'y', or 'xy' to flip both; Default: "x".}

\item{FlipReductionBackupToo}{Boolean indicating whether to also flip coordinates in the backup slot; Default: TRUE.}
}
\description{
Flips dimensionality reduction coordinates (such as UMAP or tSNE) vertically or
horizontally to change the visualization perspective.
}
\examples{
\dontrun{
  # Before flipping UMAP coordinates
  clUMAP()
  # Flip UMAP coordinates and visualize again
  combined.obj <- FlipReductionCoordinates(combined.obj)
  clUMAP()
}

}
