% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{get.clustercomposition}
\alias{get.clustercomposition}
\title{Cluster Composition Analysis}
\usage{
get.clustercomposition(
  obj = combined.obj,
  ident = "integrated_snn_res.0.3",
  splitby = "ShortNames",
  color = y,
  plot = TRUE,
  ScaleTo100pc = TRUE,
  ...
)
}
\arguments{
\item{obj}{Seurat object to analyze. Default: \code{combined.obj}.}

\item{ident}{Cluster identity resolution to use. Default: 'integrated_snn_res.0.3'.}

\item{splitby}{Variable to split the data by, typically a project or dataset identifier.
Default: 'ShortNames'.}

\item{color}{Bar color. Default: as defined by \code{splitby}.}

\item{plot}{Whether to display the plot. Default: TRUE.}

\item{ScaleTo100pc}{Whether to scale Y axis to 100\%. Default: TRUE.}

\item{...}{Additional parameters for plotting functions.}
}
\value{
If \code{plot} is TRUE, displays a bar plot showing the composition of each cluster. Otherwise,
performs the analysis without plotting.
}
\description{
Analyzes and visualizes the composition of clusters in a Seurat object, indicating
the contribution of different datasets to each cluster.
}
\examples{
get.clustercomposition()

}
