% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{get_levels_seu}
\alias{get_levels_seu}
\title{Get Unique Levels of a Seurat Object Ident Slot}
\usage{
get_levels_seu(ident, obj = combined.obj, max_levels = 100, dput = T)
}
\arguments{
\item{ident}{A character string representing the name of the slot in the Seurat object.}

\item{obj}{A Seurat object.}

\item{max_levels}{An integer that sets the maximum number of levels allowed. Default is 100.}

\item{dput}{A logical that decides whether to print the R code to recreate the 'Levels' vector. Default is TRUE.}
}
\value{
A vector of unique levels present in the 'ident' slot of the Seurat object.
}
\description{
This function extracts the unique levels present in the 'ident' slot of a Seurat object.
The function throws an error if the number of levels exceeds 'max_levels'.
The function optionally prints the R code to recreate the 'Levels' vector using 'dput'.
}
