% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{GetClusteringRuns}
\alias{GetClusteringRuns}
\title{GetClusteringRuns}
\usage{
GetClusteringRuns(
  obj = combined.obj,
  res = FALSE,
  pat = "*snn_res.[0-9].[0-9]$"
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{res}{Clustering resoluton to use, Default: FALSE}

\item{pat}{Pettern to match, Default: \verb{*snn_res.*[0-9]$}}
}
\description{
The \code{GetClusteringRuns} function retrieves metadata column names associated with
clustering runs, based on a pattern to match, \code{"*snn_res.[0-9].[0-9]$"}, by default.
}
\examples{
\dontrun{
if (interactive()) {
  GetClusteringRuns(obj = combined.obj, pat = "*snn_res.*[0-9]$")
}
}
}
