% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{getClusterNames}
\alias{getClusterNames}
\title{getClusterNames}
\usage{
getClusterNames(obj = combined.obj, ident = GetClusteringRuns(obj)[2])
}
\arguments{
\item{obj}{Seurat object}

\item{ident}{ident}
}
\description{
getClusterNames
}
