% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{getDiscretePalette}
\alias{getDiscretePalette}
\title{Safely generate a discrete color palette (NA).}
\usage{
getDiscretePalette()
}
\description{
Safe wrapper around Seurat's DiscretePalette(), which returns NA's if too many
categories are requested. This function is deprecated; see \code{DiscretePaletteSafe} and
\code{DiscretePaletteObj}.
}
\examples{
\dontrun{
if (interactive()) {
  getDiscretePalette()
}
}

}
