% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{getDiscretePaletteObj}
\alias{getDiscretePaletteObj}
\title{Generate a discrete color palette for a Seurat object.}
\usage{
getDiscretePaletteObj(
  ident.used,
  obj,
  palette.used = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[2],
  show.colors = FALSE,
  seed = 1989
)
}
\arguments{
\item{ident.used}{The identity column used for determining the number of clusters.}

\item{obj}{Seurat object.}

\item{palette.used}{The name of the palette to use, Default: "alphabet".}

\item{show.colors}{Whether to display the colors in the palette, Default: FALSE.}

\item{seed}{An integer value to set the seed for reproducibility, Default: 1989.}
}
\value{
A vector of color values.
}
\description{
Wrapper function that utilizes \code{DiscretePaletteSafe} to generate a color
palette based on the number of unique clusters in a Seurat object.
}
\examples{
\dontrun{
if (interactive()) {
  getDiscretePaletteObj(ident.used = "ident", obj = yourSeuratObj)
}
}
}
