% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{getMedianMetric}
\alias{getMedianMetric}
\title{getMedianMetric}
\usage{
getMedianMetric(
  ls.obj = ls.Seurat,
  n.datasets = length(ls.Seurat),
  mColname = "percent.mito"
)
}
\arguments{
\item{ls.obj}{List of Seurat objects, Default: ls.Seurat}

\item{n.datasets}{PARAM_DESCRIPTION, Default: length(ls.Seurat)}

\item{mColname}{PARAM_DESCRIPTION, Default: 'percent.mito'}
}
\description{
getMedianMetric
}
\examples{
\dontrun{
if(interactive()){
 ls.Seurat <- getMedianMetric(ls.obj = ls.Seurat, n.datasets = length(ls.Seurat), mColname = "percent.mito")
 }
}
}
