% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{getMetaColnames}
\alias{getMetaColnames}
\title{Get Metadata Column Names Matching Pattern}
\usage{
getMetaColnames(obj = combined.obj, pattern = "RNA")
}
\arguments{
\item{obj}{An object containing a \code{meta.data} slot, typically from combined datasets.
Default: \code{combined.obj}.}

\item{pattern}{A character string containing a regular expression to match against the
column names. Default: "RNA".}
}
\value{
A character vector of column names matching the pattern.
}
\description{
Retrieves column names from an object's metadata that match a specified pattern.
}
\examples{
# Assuming `combined.obj` is an object with a meta.data slot
getMetaColnames()

}
