% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{GetMostVarGenes}
\alias{GetMostVarGenes}
\title{GetMostVarGenes}
\usage{
GetMostVarGenes(obj, nGenes = p$nVarGenes)
}
\arguments{
\item{obj}{A Seurat object.}

\item{nGenes}{Number of genes, Default: p$nVarGenes}
}
\description{
Get the N most variable Genes
}
