% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{GetNumberOfClusters}
\alias{GetNumberOfClusters}
\title{GetNumberOfClusters}
\usage{
GetNumberOfClusters(obj = combined.obj)
}
\arguments{
\item{obj}{Seurat object, Default: \code{combined.obj}}
}
\description{
Get Number Of Clusters #
}
\examples{
\dontrun{
if (interactive()) {
  GetNumberOfClusters()
}
}
}
