% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{GetOrderedClusteringRuns}
\alias{GetOrderedClusteringRuns}
\title{GetOrderedClusteringRuns}
\usage{
GetOrderedClusteringRuns(
  obj = combined.obj,
  res = FALSE,
  pat = "*snn_res.*[0,1]\\\\.[0-9]\\\\.ordered$"
)
}
\arguments{
\item{obj}{Seurat object, Default: \code{combined.obj}.}

\item{res}{Clustering resolution to use, Default: \code{FALSE}.}

\item{pat}{Pattern to match, Default: \code{'*snn_res.*[0,1]\\.[0-9]\\.ordered$'}}
}
\description{
Get Clustering Runs: metadata column names.
}
\examples{
\dontrun{
if (interactive()) {
  GetOrderedClusteringRuns()
  GetOrderedClusteringRuns(res = 0.5)
}
}
}
