% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{getProject}
\alias{getProject}
\title{getProject}
\usage{
getProject()
}
\value{
The final subfolder of your project, or NULL, if you are not running one
}
\description{
Try to get the project name you are wokring on in Rstudio.
}
\examples{
getProject()
}
