% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{GetTopMarkers}
\alias{GetTopMarkers}
\title{Get Top Differential Expression Markers from DGEA Results}
\usage{
GetTopMarkers(
  dfDE = df.markers,
  n = p$n.markers,
  order.by = c("combined.score", "avg_log2FC", "p_val_adj")[2]
)
}
\arguments{
\item{dfDE}{Data frame containing differential expression analysis results; Default: \code{df.markers}.}

\item{n}{Number of top markers to retrieve for each cluster; Default: \code{p$n.markers}.}

\item{order.by}{Column by which to sort the markers before selection, typically prioritizing
markers by significance or effect size; Default: \code{"avg_log2FC"}.}
}
\description{
Retrieves the top N differentially expressed genes from the results of a differential
gene expression analysis, such as that provided by \code{FindAllMarkers()}.
}
\examples{
\dontrun{
if (interactive()) {
  topMarkers <- GetTopMarkers(df = df.markers, n = 3)
}
}

}
\seealso{
\code{\link[Seurat]{FindAllMarkers}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{group_by}}
}
