% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Reproduce the ggplot2 default color palette}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Integer specifying the number of distinct colors to generate.}
}
\value{
A character vector of color values in hexadecimal format, replicating the default hue
color scale of ggplot2.
}
\description{
Generates a vector of colors that emulates the default color palette used by ggplot2.
This function is useful for creating color sets for custom plotting functions or for applications
outside of ggplot2 where a similar aesthetic is desired.
}
\examples{
\dontrun{
# Generate a palette of 5 colors
print(gg_color_hue(5))
}

}
