% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slingshot.Utils.R
\name{gg_plot}
\alias{gg_plot}
\title{gg_plot}
\usage{
gg_plot(
  sds,
  col = NULL,
  title = NULL,
  lineSize = 1,
  reduction = "UMAP",
  titleFsize = 20,
  line.colors = gray.colors(n = length(sds@curves), start = 0, end = 0.6, alpha = 1),
  ...
)
}
\arguments{
\item{sds}{A Slingshot dataset. Default: PARAM_DEFAULT_VALUE.}

\item{col}{The color to be used for points. Default: NULL.}

\item{title}{The title of the plot. Default: NULL.}

\item{lineSize}{The size of the lines drawn. Default: 1.}

\item{reduction}{The dimension reduction method used ('UMAP', 'tSNE', or 'PCA'). Default: 'UMAP'.}

\item{titleFsize}{The font size for the title. Default: 20.}

\item{line.colors}{The colors to be used for lines, specified as a vector of colors. Default: gray scale colors.}
}
\value{
A ggplot object.
}
\description{
An adjusted ggplot function for visualizing Slingshot datasets.
}
\details{
This function creates a ggplot object that visualizes a Slingshot dataset, with curves representing the pseudotime trajectories.
}
\examples{
\dontrun{
if(interactive()){
 plotFittedGenePseudotime(data = sds, gene ="SST", expr = EXPR, loessCI = T, col = colz, pch = 20,
 panel_first = grid(NULL) )
 }
}
}
\seealso{
\code{\link[slingshot]{SlingshotDataSet}}, \code{\link[ggplot2]{ggplot}}
}
