% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slingshot.Utils.R
\name{ggplotColours}
\alias{ggplotColours}
\title{ggplotColours}
\usage{
ggplotColours(n = 6, h = c(0, 360) + 15)
}
\arguments{
\item{n}{The number of colors to be generated. Default: 6.}

\item{h}{The range of hues to be used in the generation of the color palette. Default: c(0, 360) + 15.}
}
\value{
A color palette.
}
\description{
Generate a color palette suitable for ggplot2.
}
\seealso{
\code{\link[ggplot2]{scale_color_manual}}
}
