% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{HGNC.EnforceUnique}
\alias{HGNC.EnforceUnique}
\title{HGNC.EnforceUnique}
\usage{
HGNC.EnforceUnique(updatedSymbols)
}
\arguments{
\item{updatedSymbols}{PARAM_DESCRIPTION}
}
\description{
Enforce Unique names after HGNC symbol update. updatedSymbols is the output of HGNChelper::checkGeneSymbols. #
}
\examples{
\dontrun{
if(interactive()){
 x <- HGNC.EnforceUnique(updatedSymbols = SymUpd)
 # While "make.unique" is not the ideal solution, because it generates mismatched, in my integration example it does reduce the mismatching genes from ~800 to 4
 }
}
}
