% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{IntersectWithExpressed}
\alias{IntersectWithExpressed}
\title{IntersectWithExpressed}
\usage{
IntersectWithExpressed(genes, obj = combined.obj, genes.shown = 10)
}
\arguments{
\item{obj}{Seurat object}
}
\description{
IntersectWithExpressed
}
