% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{isave.image}
\alias{isave.image}
\title{isave.image}
\usage{
isave.image(
  ...,
  path_rdata = paste0("~/Dropbox/Abel.IMBA/AnalysisD/_Rdata.files/", basename(OutDir)),
  showMemObject = T,
  options = c("--force", NULL)[1]
)
}
\arguments{
\item{...}{Pass any other parameter to the internally called functions (most of them should work).}

\item{path_rdata}{PARAM_DESCRIPTION, Default: paste0("~/Dropbox/Abel.IMBA/AnalysisD/_Rdata.files/", basename(OutDir))}

\item{showMemObject}{PARAM_DESCRIPTION, Default: T}

\item{options}{PARAM_DESCRIPTION, Default: c("--force", NULL)\link{1}}
}
\description{
Save and RData image.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Stringendo]{kollapse}}, \code{\link[MarkdownReports]{iprint}}
}
