% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{isave.RDS}
\alias{isave.RDS}
\title{isave.RDS}
\usage{
isave.RDS(
  obj,
  prefix = NULL,
  suffix = NULL,
  inOutDir = TRUE,
  project = CodeAndRoll2::getProject(),
  alternative_path_rdata = paste0("~/Dropbox (VBC)/Abel.IMBA/AnalysisD/_RDS.files/",
    basename(OutDir)),
  homepath = if (Sys.info()[1] == "Darwin") "~/" else "/users/abel.vertesy/",
  showMemObject = TRUE,
  saveParams = TRUE,
  compress = TRUE,
  test_read = FALSE
)
}
\arguments{
\item{obj}{The object to be saved, typically a Seurat object.}

\item{prefix}{A string prefix added to the filename. Default: NULL.}

\item{suffix}{A string suffix added to the filename. Default: NULL.}

\item{inOutDir}{A boolean flag, if TRUE the OutDir is used as save directory, if FALSE the
alternative_path_rdata is used. Default: \code{TRUE}.}

\item{project}{A string representing the project code. This is appended to the saved file name.
Default: the active project determined by getProject().}

\item{alternative_path_rdata}{A string that specifies the alternative path for storing the
RDS file if inOutDir is FALSE. Default: "~/Dropbox (VBC)/Abel.IMBA/AnalysisD/_RDS.files/"
appended with the basename of OutDir.}

\item{homepath}{A string representing the homepath. Will be replaced by '~' in the file path. Default: '~/'.}

\item{showMemObject}{A boolean flag, if TRUE the function will print out the memory size of the
largest objects in the workspace. Default: \code{TRUE}.}

\item{saveParams}{A boolean flag, if TRUE the parameters 'p' and 'all.genes' are added to the
'misc' slot of the Seurat object if the object is of class Seurat. Default: \code{TRUE}.}

\item{compress}{Compress .Rds file after writing? Default: \code{TRUE}.}

\item{test_read}{Provide command to test validity by reading in the object just written.}
}
\description{
Save an RDS object, using a faster and efficient compression method that runs in the background.
}
\examples{
\dontrun{
if (interactive()) {
  isave.RDS(my.R.object)
}
}
}
