% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{jJaccardIndexVec}
\alias{jJaccardIndexVec}
\title{jJaccardIndexVec}
\usage{
jJaccardIndexVec(A = 1:3, B = 2:4)
}
\arguments{
\item{A}{PARAM_DESCRIPTION, Default: 1:3}

\item{B}{PARAM_DESCRIPTION, Default: 2:4}
}
\description{
Calculate jaccard similarity for 2 vecotrs. Helper to jPairwiseJaccardIndexList.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
