% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{jPairwiseJaccardIndex}
\alias{jPairwiseJaccardIndex}
\title{jPairwiseJaccardIndex}
\usage{
jPairwiseJaccardIndex(binary.presence.matrix = df.presence)
}
\arguments{
\item{binary.presence.matrix}{A boolean matrix. Default: df.presence}
}
\description{
Create a pairwise jaccard similarity matrix across all combinations of columns in binary.presence.matrix. Modified from: https://www.displayr.com/how-to-calculate-jaccard-coefficients-in-displayr-using-r/ #
}
\examples{
\dontrun{
if (interactive()) {
  PairwiseJaccardIndices <- jPairwiseJaccardIndex(binary.presence.matrix = df.presence)
}
}
}
