% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{load10Xv3}
\alias{load10Xv3}
\title{load10Xv3}
\usage{
load10Xv3(
  dataDir,
  cellIDs = NULL,
  channelName = NULL,
  readArgs = list(),
  includeFeatures = c("Gene Expression"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dataDir}{PARAM_DESCRIPTION}

\item{cellIDs}{PARAM_DESCRIPTION, Default: NULL}

\item{channelName}{PARAM_DESCRIPTION, Default: NULL}

\item{readArgs}{PARAM_DESCRIPTION, Default: list()}

\item{includeFeatures}{PARAM_DESCRIPTION, Default: c("Gene Expression")}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Load 10X output folders.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[SoupX]{SoupChannel}}
}
