% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{m3.backup.umap}
\alias{m3.backup.umap}
\title{m3.backup.umap}
\usage{
m3.backup.umap(obj = cds_from_seurat, slot = "UMAP", dim = (2:3)[2], ...)
}
\arguments{
\item{obj}{Seurat object, Default: cds_from_seurat}

\item{slot}{PARAM_DESCRIPTION, Default: 'UMAP'}

\item{dim}{Numer of dimensions used, Default: (2:3)\link{2}}

\item{...}{Pass any other parameter to the internally called functions (most of them should  work).}
}
\description{
m3.backup.umap
}
\examples{
\dontrun{
if(interactive()){
 cds_from_seurat <- m3.backup.umap(obj = cds_from_seurat, slot = 'UMAP', dim = (2:3)[2])
 }
}
}
