% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monocle.Utils.R
\name{m3.export.umap.2.Seurat}
\alias{m3.export.umap.2.Seurat}
\title{m3.export.umap.2.Seurat}
\usage{
m3.export.umap.2.Seurat(
  mobj = cds_from_seurat,
  sobj = combined.obj,
  def.dim = 2
)
}
\arguments{
\item{mobj}{The Monocle object from which to extract UMAP coordinates. Default: cds_from_seurat.}

\item{sobj}{The Seurat object to which UMAP coordinates will be added. Default: combined.obj.}

\item{def.dim}{The dimension of the UMAP coordinates to extract (2D or 3D). Default: 2.}
}
\value{
A Seurat object with the UMAP coordinates added.
}
\description{
Exports UMAP coordinates from a monocle object to a Seurat object.
}
\details{
This function extracts UMAP coordinates from a Monocle object and adds them to a Seurat object. The dimensions of the UMAP coordinates can be specified.
}
\examples{
\dontrun{
if(interactive()){
 combined.obj.bac <- combined.obj; # combined.obj <- combined.obj.bac
 combined.obj <- m3.export.umap.2.Seurat(mobj = cds_from_seurat, sobj = combined.obj); qUMAP("DDIT4")
 }
}
}
\seealso{
\code{\link[m3.get.umap]{m3.get.umap}}
}
