% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{m3.get.umap}
\alias{m3.get.umap}
\title{m3.get.umap}
\usage{
m3.get.umap(obj = cds_from_seurat, slot = "UMAP", dim = (2:3)[2])
}
\arguments{
\item{obj}{Seurat object, Default: cds_from_seurat}

\item{slot}{PARAM_DESCRIPTION, Default: 'UMAP'}

\item{dim}{Numer of dimensions used, Default: (2:3)\link{2}}
}
\description{
Fetch the umap coordinates from obj@int_colData@listData$reducedDims[\link{slot}]
}
\examples{
\dontrun{
if(interactive()){
 m3.get.umap(obj = cds_from_seurat, slot = 'UMAP', dim = (2:3)[2])
 }
}
}
