% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{m3DplotGene}
\alias{m3DplotGene}
\title{m3DplotGene}
\usage{
m3DplotGene(
  gene = "PGK1",
  reduction = "UMAP",
  obj = cds.10pc,
  ttl.suffix = "expression",
  suffix = "",
  cex = 5
)
}
\arguments{
\item{gene}{gene of interest, Default: 'PGK1'}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'UMAP'}

\item{obj}{Seurat object, Default: cds.10pc}

\item{ttl.suffix}{PARAM_DESCRIPTION, Default: 'expression'}

\item{suffix}{A suffix added to the filename, Default: ''}

\item{cex}{Point size, Default: 5}
}
\description{
m3DplotGene
}
\examples{
\dontrun{
if(interactive()){
 m3DplotGene("DDIT4")
 }
}
}
