% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monocle.Utils.R
\name{m3DplotKeyGenes}
\alias{m3DplotKeyGenes}
\title{m3DplotKeyGenes}
\usage{
m3DplotKeyGenes(
  obj = cds.10pc,
  cex = iround(log10(idim(obj)[2])),
  ls.genes = c(`S-phase` = "TOP2A", `G2M-phase` = "HIST1H4C", oRG = "ID4", oRG = "HOPX",
    `Intermediate progenitor` = "EOMES", `Intermediate progenitor1` = "TAC3", Astroglia =
    "GFAP", Astrocyte = "S100B", `Immature neurons` = "SLA", Interneurons = "DLX6-AS1",
    `Hypoxia/Stress` = "DDIT4", Glycolytic = "PDK1", `Low-Quality` = "POLR2A", PGC =
    "DCN", `dl-EN` = "KAZN", `ul-EN` = "SATB2"),
  reduction = "UMAP",
  suffix = ""
)
}
\arguments{
\item{obj}{Seurat object, Default: cds.10pc}

\item{cex}{Point size, Default: iround(log10(idim(obj)\link{2}))}

\item{ls.genes}{PARAM_DESCRIPTION, Default: c(\code{S-phase} = "TOP2A", \code{G2M-phase} = "HIST1H4C", oRG = "ID4",
oRG = "HOPX", \verb{Intermediate progenitor} = "EOMES", \verb{Intermediate progenitor1} = "TAC3",
Astroglia = "GFAP", Astrocyte = "S100B", \verb{Immature neurons} = "SLA",
Interneurons = "DLX6-AS1", \code{Hypoxia/Stress} = "DDIT4", Glycolytic = "PDK1",
\code{Low-Quality} = "POLR2A", PGC = "DCN", \code{dl-EN} = "KAZN",
\code{ul-EN} = "SATB2")}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'UMAP'}

\item{suffix}{A suffix added to the filename, Default: ''}
}
\description{
Plot many genes in 3D in Monocle.
}
\examples{
\dontrun{
if(interactive()){
 m3DplotKeyGenes()
 }
}
}
