% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{make10Xcellname}
\alias{make10Xcellname}
\title{make10Xcellname}
\usage{
make10Xcellname(cellnames, suffix = "_1")
}
\arguments{
\item{cellnames}{PARAM_DESCRIPTION}

\item{suffix}{A suffix added to the filename, Default: '_1'}
}
\description{
Add a suffix to cell names, so that it mimics the lane-suffix, e.g.: "_1". #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
