% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{match_best_identity}
\alias{match_best_identity}
\title{Match and Translate Best Identity}
\usage{
match_best_identity(
  obj,
  ident_from,
  ident_to = gsub(pattern = "ordered", replacement = "transferred", x = ident_from),
  to_suffix = FixPlotName(gsub(pattern = "[a-zA-Z_]", replacement = "", x = ident_from)),
  new_ident_name = kpp(ident_from, "best.match", to_suffix),
  ...
)
}
\arguments{
\item{obj}{The object to update. This object must have a \code{meta.data} attribute which is a data frame
containing columns named as \code{ident_from} and \code{ident_to}.}

\item{ident_from}{A string. The name of the column in \code{obj@meta.data} that is used as the source of identities.
There is no default value for this parameter.}

\item{ident_to}{A string. The name of the column in \code{obj@meta.data} that is used as the target of identities.
There is no default value for this parameter.}

\item{to_suffix}{A string. The suffix to add to the new identity name. Default is the output of the \code{FixPlotName}
function applied to the \code{ident_from} string, with all alphabetical and underscore characters removed.}

\item{new_ident_name}{A string. The name for the newly created identity column in \code{obj@meta.data}.
Default is a concatenation of \code{ident_from}, "best.match", and \code{to_suffix} using \code{kpp} function.}

\item{...}{Additional parameters to be passed to \code{.replace_by_most_frequent_categories} function.}
}
\value{
An updated version of \code{obj} with an additional column in \code{obj@meta.data} named as \code{new_ident_name}
representing the new identity. The function also generates a UMAP plot based on this new identity.
}
\description{
This function matches the best identity from \code{ident_from} to \code{ident_to} in an object,
updates the metadata of the object with this new identity and returns the updated object. Additionally,
it generates a UMAP plot based on the new identity. The function replaces original categories with
the most frequent ones, hence helps to filter out less important categories.
}
\details{
Match and Translate Best Identity
}
\examples{
\dontrun{
updated_obj <- match_best_identity(my_obj, "origin_identity", "target_identity")
}
}
\seealso{
\code{\link[clUMAP]{clUMAP}}, \code{\link[kpp]{kpp}}, \code{\link[FixPlotName]{FixPlotName}},
\code{\link[.replace_by_most_frequent_categories]{.replace_by_most_frequent_categories}}
}
