% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{merge_seurat_metadata}
\alias{merge_seurat_metadata}
\title{Merge Seurat Metadata}
\usage{
merge_seurat_metadata(ls_obj, include_cols = NULL, exclude_cols = NULL)
}
\arguments{
\item{ls_obj}{A list of Seurat objects.}

\item{include_cols}{A character vector of column names to include (default NULL for all columns).}

\item{exclude_cols}{A character vector of column names to exclude (default NULL for no exclusions).}
}
\value{
A merged dataframe of metadata from all Seurat objects.
}
\description{
Merges the \verb{@metadata} from a list of Seurat objects, binds them by row, and applies optional inclusion/exclusion of columns.
}
