% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monocle.Utils.R
\name{mplotGene}
\alias{mplotGene}
\title{mplotGene}
\usage{
mplotGene(gene = "PGK1", reduction = "UMAP", obj = cds_from_seurat)
}
\arguments{
\item{gene}{A string representing the gene of interest. Default: 'PGK1'.}

\item{reduction}{A string representing the dimensionality reduction method to use. Can be 'UMAP', 'tSNE', or 'PCA'. Default: 'UMAP'.}

\item{obj}{A Seurat object that will be plotted. Default: cds_from_seurat.}
}
\value{
A plot showing the gene of interest.
}
\description{
Plots a specific gene in Monocle using a chosen dimensionality reduction method.
}
\details{
This function takes a specific gene and plots it in Monocle using a specified dimensionality reduction method. The plot is then saved using the ggExpress::qqSave function.
}
\examples{
\dontrun{
if(interactive()){
 mplotGene(); mplotGene("GAPDH")
 }
}
}
\seealso{
\code{\link[plot_cells]{plot_cells}}, \code{\link[qqSave]{qqSave}}
}
