% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{mplotManyGenes}
\alias{mplotManyGenes}
\title{mplotManyGenes}
\usage{
mplotManyGenes(
  ls.genes = c(`S-phase` = "TOP2A", `G2M-phase` = "HIST1H4C", oRG = "ID4", oRG =
    "HOPX", `Intermediate progenitor` = "EOMES", `Intermediate progenitor1` = "TAC3",
    Astroglia = "GFAP", Astrocyte = "S100B", `Immature neurons` = "SLA", Interneurons =
    "DLX6-AS1", `Hypoxia/Stress` = "DDIT4", Glycolytic = "PDK1", `Low-Quality` =
    "POLR2A", PGC = "DCN", `dl-EN` = "KAZN", `ul-EN` = "SATB2")
)
}
\arguments{
\item{ls.genes}{PARAM_DESCRIPTION, Default: c(\code{S-phase} = "TOP2A",
\code{G2M-phase} = "HIST1H4C", oRG = "ID4",
oRG = "HOPX", \verb{Intermediate progenitor} = "EOMES", \verb{Intermediate progenitor1} = "TAC3",
Astroglia = "GFAP", Astrocyte = "S100B", \verb{Immature neurons} = "SLA",
Interneurons = "DLX6-AS1", \code{Hypoxia/Stress} = "DDIT4", Glycolytic = "PDK1",
\code{Low-Quality} = "POLR2A", PGC = "DCN", \code{dl-EN} = "KAZN",
\code{ul-EN} = "SATB2")}
}
\description{
Plot many genes in Monocle.
}
\examples{
\dontrun{
if(interactive()){
 mplotManyGenes()
 }
}
}
