% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MULTI-seq.Utils.R
\name{mSeq.map.all96.BCs}
\alias{mSeq.map.all96.BCs}
\title{mSeq.map.all96.BCs}
\usage{
mSeq.map.all96.BCs(
  readTable = readTable,
  CellIDs = CellIDs,
  path2allBCs =
    "~/Google_Drive/Science/IMBA/MULTI.seq/from.US/All.MULTI-seq_barcodes.Mar2019.tsv"
)
}
\arguments{
\item{readTable}{The table of read counts. Default: readTable.}

\item{CellIDs}{The identifiers for the cells. Default: CellIDs.}

\item{path2allBCs}{The path to the reference barcode sequences file. Default: '~/Google_Drive/Science/IMBA/MULTI.seq/from.US/All.MULTI-seq_barcodes.Mar2019.tsv'.}
}
\value{
The mapped barcode sequences.
}
\description{
Map all 96 barcode sequences.
}
\examples{
\dontrun{
if(interactive()){
 bar.table <-mSeq.map.all96.BCs(readTable = readTable, CellIDs = CellIDs)
 }
}
}
