% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.utils.less.used.R
\name{multi_clUMAP.A4}
\alias{multi_clUMAP.A4}
\title{Plot multiple categorical variables in combined UMAPs}
\usage{
multi_clUMAP.A4(
  obj = combined.obj,
  idents = GetClusteringRuns(obj)[1:4],
  foldername = "clUMAPs_multi",
  plot.reduction = "umap",
  intersectionAssay = c("RNA", "integrated")[1],
  layout = c("tall", "wide", FALSE)[2],
  nr.Col = 2,
  nr.Row = 4,
  cex = round(0.1/(nr.Col * nr.Row), digits = 2),
  label = FALSE,
  legend = !label,
  subdir = TRUE,
  prefix = NULL,
  suffix = NULL,
  background_col = "white",
  aspect.ratio = c(FALSE, 0.6)[2],
  saveGeneList = FALSE,
  w = 8.27,
  h = 11.69,
  scaling = 1,
  format = c("jpg", "pdf", "png")[1],
  ...
)
}
\arguments{
\item{obj}{The Seurat object containing clustering information. Default: \code{combined.obj}.}

\item{idents}{A vector of cluster identities to plot. Default: \code{GetClusteringRuns()[1:4]}.}

\item{foldername}{The name of the folder to save plots. Default: \code{substitute(ident)}.}

\item{plot.reduction}{The dimensionality reduction technique to use for plotting. Default: "umap".}

\item{intersectionAssay}{The assay to use for intersection. Default: "RNA".}

\item{layout}{The layout orientation, either "tall", "wide", or \code{FALSE} to disable. Default: "wide".}

\item{nr.Col}{Number of columns in the plot grid. Default: 2.}

\item{nr.Row}{Number of rows in the plot grid. Default: 4.}

\item{cex}{The character expansion size for plot text, automatically adjusted. Default: \code{round(0.1 / (nr.Col * nr.Row), digits = 2)}.}

\item{label}{Logical indicating if labels should be displayed on the plots. Default: \code{FALSE}.}

\item{legend}{Logical indicating if a legend should be included in the plots. Default: \code{!label}.}

\item{subdir}{Logical indicating if a subdirectory should be created for saving plots. Default: \code{TRUE}.}

\item{prefix}{Optional prefix for plot filenames. Default: \code{NULL}.}

\item{suffix}{Optional suffix for plot filenames. Default: \code{NULL}.}

\item{background_col}{The background color of the plot. Default: "white".}

\item{aspect.ratio}{The aspect ratio of the plot, \code{FALSE} to disable fixed ratio. Default: 0.6.}

\item{saveGeneList}{Logical indicating if a list of genes should be saved. Default: \code{FALSE}.}

\item{w}{The width of the plot in inches. Default: \code{8.27}.}

\item{h}{The height of the plot in inches. Default: \code{11.69}.}

\item{scaling}{The scaling factor to apply to plot dimensions. Default: 1.}

\item{format}{The file format for saving plots. Default: "jpg".}

\item{...}{Additional arguments passed to plotting functions.}
}
\value{
Invisible \code{NULL}. Plots are saved to files.
}
\description{
Generates and saves multiple UMAP plots for clustering results, adjusting the
layout and plot dimensions. Supports the generation of plots in different
formats and customization of the visual appearance.
}
\examples{
\dontrun{
multi_clUMAP.A4(idents = c("S1", "S2"), obj = YourSeuratObject)
}
}
