% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{multiFeatureHeatmap.A4}
\alias{multiFeatureHeatmap.A4}
\title{multiFeatureHeatmap.A4}
\usage{
multiFeatureHeatmap.A4(
  obj = combined.obj,
  list.of.genes,
  gene.per.page = 5,
  group.cells.by = "batch",
  plot.reduction = "umap",
  cex = iround(3/gene.per.page),
  sep_scale = F,
  gene.min.exp = "q5",
  gene.max.exp = "q95",
  jpeg.res = 225,
  jpeg.q = 90,
  ...
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{list.of.genes}{PARAM_DESCRIPTION}

\item{gene.per.page}{PARAM_DESCRIPTION, Default: 5}

\item{group.cells.by}{PARAM_DESCRIPTION, Default: 'batch'}

\item{plot.reduction}{PARAM_DESCRIPTION, Default: 'umap'}

\item{cex}{Point size, Default: iround(3/gene.per.page)}

\item{sep_scale}{PARAM_DESCRIPTION, Default: F}

\item{gene.min.exp}{PARAM_DESCRIPTION, Default: 'q5'}

\item{gene.max.exp}{PARAM_DESCRIPTION, Default: 'q95'}

\item{jpeg.res}{PARAM_DESCRIPTION, Default: 225}

\item{jpeg.q}{PARAM_DESCRIPTION, Default: 90}

\item{...}{Pass any other parameter to the internally called functions (most of them should  work).}
}
\description{
Save multiple FeatureHeatmaps from a list of genes on A4 jpeg #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tictoc]{tic}}
}
