% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{panelCorPearson}
\alias{panelCorPearson}
\title{Display Correlation Values in Pairs Plot}
\usage{
panelCorPearson(x, y, digits = 2, prefix = "", cex.cor = 2, method = "pearson")
}
\arguments{
\item{x}{Numeric vector or the first half of the data pair.}

\item{y}{Numeric vector or the second half of the data pair.}

\item{digits}{Number of significant digits to display in the correlation coefficient.
Default: 2.}

\item{prefix}{A string prefix added before the correlation coefficient. Default: "".}

\item{cex.cor}{The character expansion factor for the correlation coefficient text.
This argument directly influences the text size. Default: 2.}

\item{method}{The method of correlation coefficient calculation. It can be "pearson" (default),
"kendall", or "spearman".}
}
\value{
This function does not return a value but modifies the current plot by adding the
correlation coefficient and its significance level.
}
\description{
This function displays the correlation coefficient and significance level within
a scatterplot generated by the \code{pairs()} function. The default correlation method is Pearson,
but Kendall or Spearman methods can also be selected.
}
\examples{
\dontrun{
pairs(mtcars[, 1:4], panel = panelCorPearson)
}
}
