% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{Percent.in.Trome}
\alias{Percent.in.Trome}
\title{Percent.in.Trome}
\usage{
Percent.in.Trome(
  obj = combined.obj,
  n.genes.barplot = 25,
  width.barplot = round(n.genes.barplot/4)
)
}
\arguments{
\item{obj}{Seurat object}

\item{n.genes.barplot}{number of top genes shows}

\item{width.barplot}{barplot width}
}
\value{
Seurat object
}
\description{
Gene expression as fraction of all UMI's
}
\examples{
# combined.obj <- Percent.in.Trome()
}
