% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot.clust.size.distr}
\alias{plot.clust.size.distr}
\title{plot.clust.size.distr}
\usage{
\method{plot}{clust.size.distr}(
  obj = combined.obj,
  ident = GetClusteringRuns()[2],
  plot = T,
  thr.hist = 30,
  ...
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{ident}{identity used, Default: GetClusteringRuns()\link{2}}

\item{plot}{PARAM_DESCRIPTION, Default: T}

\item{thr.hist}{PARAM_DESCRIPTION, Default: 30}

\item{...}{Pass any other parameter to the internally called functions (most of them should  work).}
}
\description{
plot.clust.size.distr
}
\examples{
\dontrun{
if(interactive()){
 plot.clust.size.distr()
 }
}
}
