% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot.expression.rank.q90}
\alias{plot.expression.rank.q90}
\title{plot.expression.rank.q90}
\usage{
\method{plot}{expression.rank.q90}(obj = combined.obj, gene = "ACTB", filterZero = TRUE)
}
\arguments{
\item{obj}{Seurat object, Default: \code{combined.obj}}

\item{gene}{gene of interest, Default: 'ACTB'}

\item{filterZero}{Remove genes whose quantile-90 expression in 0? Default: \code{TRUE}.}
}
\description{
Plot gene expression based on the expression at the 90th quantile
(so you will not lose genes expressed in few cells).
}
\examples{
\dontrun{
if (interactive()) {
  plot.expression.rank.q90(gene = "SATB2")
}
}
}
