% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot.Gene.Cor.Heatmap}
\alias{plot.Gene.Cor.Heatmap}
\title{plot.Gene.Cor.Heatmap}
\usage{
\method{plot}{Gene.Cor.Heatmap}(
  genes = WU.2017.139.IEGsf,
  assay.use = "RNA",
  slot.use = c("data", "scale.data", "data.imputed")[1],
  quantileX = 0.95,
  min.g.cor = 0.3,
  calc.COR = FALSE,
  cutRows = NULL,
  cutCols = cutRows,
  obj = combined.obj,
  ...
)
}
\arguments{
\item{genes}{The list of genes of interest. Default: WU.2017.139.IEGsf}

\item{assay.use}{The assay to use from the Seurat object. Default: 'RNA'}

\item{slot.use}{The slot to use from the assay in the Seurat object. Default: First item in the vector c("data", "scale.data", "data.imputed")}

\item{quantileX}{The quantile level for the calculation. Default: 0.95}

\item{min.g.cor}{The minimum gene correlation. Genes with correlation above this value or below its negative will be included. Default: 0.3}

\item{calc.COR}{A logical flag, if TRUE, the correlation will be calculated if not found. Default: FALSE}

\item{cutRows}{A number, the dendrogram will be cut at this height into clusters. Default: NULL}

\item{cutCols}{A number, the dendrogram will be cut at this height into clusters. If NULL, cutCols will be the same as cutRows. Default: cutRows}

\item{obj}{The Seurat object to perform calculations on. Default: combined.obj}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot a gene correlation heatmap.
}
