% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot.Metadata.Cor.Heatmap}
\alias{plot.Metadata.Cor.Heatmap}
\title{plot.Metadata.Cor.Heatmap}
\usage{
\method{plot}{Metadata.Cor.Heatmap}(
  columns = c("nCount_RNA", "nFeature_RNA", "percent.mito", "percent.ribo"),
  cormethod = c("pearson", "spearman")[1],
  main = paste("Metadata", cormethod, "correlations"),
  obj = combined.obj,
  w = 10,
  h = w,
  ...
)
}
\arguments{
\item{columns}{PARAM_DESCRIPTION, Default: c("nCount_RNA", "nFeature_RNA", "percent.mito", "percent.ribo")}

\item{cormethod}{PARAM_DESCRIPTION, Default: c("pearson", "spearman")\link{1}}

\item{main}{PARAM_DESCRIPTION, Default: paste("Metadata", cormethod, "correlations")}

\item{obj}{Seurat object, Default: combined.obj}

\item{w}{width of the plot, Default: 10}

\item{h}{height of the plot, Default: w}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot a heatmap with Metadata correlation values.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggcorrplot]{ggcorrplot}}
}
