% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot.Metadata.median.fraction.barplot}
\alias{plot.Metadata.median.fraction.barplot}
\title{plot.Metadata.median.fraction.barplot}
\usage{
\method{plot}{Metadata.median.fraction.barplot}(
  columns = c("percent.mito", "percent.ribo"),
  suffix = NULL,
  group.by = GetClusteringRuns(obj = obj)[2],
  method = c("median", "mean")[1],
  min.thr = 2.5,
  return.matrix = F,
  main = paste(method, "read fractions per transcript class and cluster", suffix),
  ylab = "Fraction of transcriptome (\%)",
  percentify = T,
  subt = NULL,
  position = position_stack(),
  w = 10,
  h = 6,
  obj = combined.obj,
  ...
)
}
\arguments{
\item{columns}{PARAM_DESCRIPTION, Default: c("percent.mito", "percent.ribo")}

\item{suffix}{A suffix added to the filename, Default: NULL}

\item{group.by}{PARAM_DESCRIPTION, Default: GetClusteringRuns(obj = obj)\link{2}}

\item{method}{PARAM_DESCRIPTION, Default: c("median", "mean")\link{1}}

\item{min.thr}{PARAM_DESCRIPTION, Default: 2.5}

\item{return.matrix}{PARAM_DESCRIPTION, Default: F}

\item{main}{PARAM_DESCRIPTION, Default: paste(method, "read fractions per transcript class and cluster",
suffix)}

\item{ylab}{PARAM_DESCRIPTION, Default: 'Fraction of transcriptome (\%)'}

\item{percentify}{PARAM_DESCRIPTION, Default: T}

\item{subt}{PARAM_DESCRIPTION, Default: NULL}

\item{position}{PARAM_DESCRIPTION, Default: position_stack()}

\item{w}{width of the plot, Default: 10}

\item{h}{height of the plot, Default: 6}

\item{obj}{Seurat object, Default: combined.obj}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Barplot Metadata median values
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{summarise_all}}
\code{\link[reshape2]{melt}}
}
