% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot.UMAP.tSNE.sidebyside}
\alias{plot.UMAP.tSNE.sidebyside}
\title{plot.UMAP.tSNE.sidebyside}
\usage{
\method{plot}{UMAP.tSNE.sidebyside}(
  obj = combined.obj,
  grouping = "res.0.6",
  no_legend = F,
  do_return = TRUE,
  do_label = T,
  label_size = 10,
  vector_friendly = TRUE,
  cells_use = NULL,
  no_axes = T,
  pt_size = 0.5,
  name.suffix = NULL,
  width = hA4,
  heigth = 1.75 * wA4,
  filetype = "pdf",
  ...
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{grouping}{PARAM_DESCRIPTION, Default: 'res.0.6'}

\item{no_legend}{PARAM_DESCRIPTION, Default: F}

\item{do_return}{PARAM_DESCRIPTION, Default: TRUE}

\item{do_label}{PARAM_DESCRIPTION, Default: T}

\item{label_size}{PARAM_DESCRIPTION, Default: 10}

\item{vector_friendly}{PARAM_DESCRIPTION, Default: TRUE}

\item{cells_use}{PARAM_DESCRIPTION, Default: NULL}

\item{no_axes}{PARAM_DESCRIPTION, Default: T}

\item{pt_size}{PARAM_DESCRIPTION, Default: 0.5}

\item{name.suffix}{PARAM_DESCRIPTION, Default: NULL}

\item{width}{PARAM_DESCRIPTION, Default: hA4}

\item{heigth}{PARAM_DESCRIPTION, Default: 1.75 * wA4}

\item{filetype}{PARAM_DESCRIPTION, Default: 'pdf'}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot a UMAP and tSNE sidebyside #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cowplot]{save_plot}}
}
