% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{plot3D.umap.gene}
\alias{plot3D.umap.gene}
\title{plot3D.umap.gene}
\usage{
plot3D.umap.gene(
  gene = "TOP2A",
  obj = combined.obj,
  annotate.by = GetNamedClusteringRuns(obj = obj, v = FALSE)[1],
  quantileCutoff = 0.99,
  def.assay = c("integrated", "RNA")[2],
  suffix = NULL,
  alpha = 0.5,
  dotsize = 1.25,
  col.names = c("umap_1", "umap_2", "umap_3"),
  assay = "RNA",
  ...
)
}
\arguments{
\item{gene}{The gene of interest. Default: 'TOP2A'}

\item{obj}{The Seurat object for which the 3D umap plot will be generated. Default: combined.obj}

\item{annotate.by}{The cluster or grouping to be used for automatic annotation. Default: First returned result from GetNamedClusteringRuns(obj) function.}

\item{quantileCutoff}{Cutoff value for the quantile for clipping outliers in the gene expression data. Default: 0.99}

\item{def.assay}{The default assay to be used. Choose between "integrated" and "RNA". Default: "RNA"}

\item{suffix}{A suffix added to the filename. Default: NULL}

\item{alpha}{Opacity of the points in the plot. Default: 0.5}

\item{dotsize}{The size of the dots in the plot. Default: 1.25}

\item{col.names}{Names of the 3D UMAP coordinate columns to use. Default: \code{c("umap_1", "umap_2", "umap_3")}.}

\item{assay}{Assay used to check gene/feature presence in the object. Default: 'RNA'.}

\item{...}{Pass any other parameter to the internally called \code{plotly::plot_ly}.}
}
\description{
Plot a 3D umap with gene expression. Uses plotly. Based on github.com/Dragonmasterx87.
}
\examples{
\dontrun{
if (interactive()) {
  plot3D.umap.gene(obj = combined.obj, gene = "DDIT4", quantileCutoff = .95)
  plot3D.umap.gene(obj = combined.obj, gene = "percent.mito", quantileCutoff = .95) # for continuous meta variables
  plot3D.umap.gene(obj = combined.obj, gene = "nFeature_RNA", quantileCutoff = .95) # for continuous meta variables
}
}
}
