% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{plotClustSizeDistr}
\alias{plotClustSizeDistr}
\title{Cluster Size Distribution Plot (Barplot or Histogram)}
\usage{
plotClustSizeDistr(obj = combined.obj, ident, plot = TRUE, thr.hist = 30, ...)
}
\arguments{
\item{obj}{Seurat object for analysis. Default: \code{combined.obj}.}

\item{ident}{Clustering identity to base the plot on.
Default: The second entry from \code{GetClusteringRuns()}.}

\item{plot}{Whether to display the plot (TRUE) or return cluster sizes (FALSE). Default: TRUE.}

\item{thr.hist}{Threshold for switching from a bar plot to a histogram based on the number of
clusters. Default: 30.}

\item{...}{Extra parameters for the plot.}
}
\description{
Generates a bar plot or histogram to visualize the size distribution of clusters
within a Seurat object, based on the specified clustering identity.
}
\examples{
\dontrun{
plotClustSizeDistr()
}

}
