% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{plotGeneExpHist}
\alias{plotGeneExpHist}
\title{plotGeneExpHist}
\usage{
plotGeneExpHist(
  obj = cobj.H9.L92,
  genes = c("MALAT1", "MT-CO1", "MT-CO2", "MT-CYB", "TMSB4X", "KAZN"),
  assay = "RNA",
  slot_ = "data",
  thr_expr = 10,
  suffix = NULL,
  xlab = paste0("log10(Summed UMI count @", slot_, ")"),
  return_cells_passing = TRUE,
  quantile_thr = 0.95,
  return_quantile,
  ...
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{genes}{A vector of genes to plot.}

\item{assay}{The name of the assay to use.}

\item{slot_}{The slot to use.}

\item{thr_expr}{The expression threshold to use for filtering.}

\item{suffix}{A string to append to the title of the plot.}

\item{xlab}{The x-axis label.}

\item{return_cells_passing}{Whether to return the number of cells passing the filter.}

\item{quantile_thr}{The quantile to use for clipping the counts slot.}

\item{return_quantile}{Whether to return the number of cells passing the quantile filter.}

\item{...}{Additional arguments passed to \code{qhistogram()}.}
}
\value{
A ggplot object.
}
\description{
This function creates a histogram of gene expression for a given set of genes in a Seurat object.
}
