% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{plotGeneExpressionInBackgroundHist}
\alias{plotGeneExpressionInBackgroundHist}
\title{Histogram All Genes' Expression Level and a Highlighted Gene}
\usage{
plotGeneExpressionInBackgroundHist(
  gene = "TOP2A",
  obj = ls.Seurat[[1]],
  assay = "RNA",
  slot = c("counts", "data")[2],
  w = 7,
  h = 4,
  ...
)
}
\arguments{
\item{gene}{The gene of interest for which the expression level distribution is to be plotted.
Default: 'TOP2A'.}

\item{obj}{A Seurat object containing the expression data. Default: The first Seurat object in
\code{ls.Seurat}.}

\item{assay}{The assay from which to retrieve the expression data. Default: "RNA".}

\item{slot}{The slot in the Seurat object from which to retrieve the expression data. Options
include "counts" for raw counts and "data" for normalized (and possibly log-transformed) data.
Default: "data".}

\item{w}{The width of the plot. Default: 7.}

\item{h}{The height of the plot. Default: 4.}

\item{...}{Any other parameter that can be passed to the internally called functions.}
}
\description{
This function generates a histogram to visualize the expression level distribution
of a specified gene across all cells in a Seurat object. It highlights the position of the gene
of interest within the overall distribution.
}
