% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{plotMetadataCategPie}
\alias{plotMetadataCategPie}
\title{Plot Metadata Category Pie Chart}
\usage{
plotMetadataCategPie(
  metacol = "Singlet.status",
  plot_name = paste(metacol, "distribution"),
  obj = combined.obj,
  max.categs = 20,
  both_pc_and_value = TRUE,
  subtitle = NULL,
  labels = NULL,
  LegendSide = FALSE,
  ...
)
}
\arguments{
\item{metacol}{The metadata column to visualize.}

\item{plot_name}{Name of the plot to generate.}

\item{obj}{Seurat object containing the metadata. Default: \code{combined.obj}.}

\item{max.categs}{The maximum number of categories to display in the pie chart.
If the number of categories exceeds this value, an error is thrown.}

\item{both_pc_and_value}{If \code{TRUE}, labels on the pie chart will show both the percentage
and the count of each category. If \code{FALSE}, only the percentage is shown.}

\item{subtitle}{Optional subtitle for the pie chart.}

\item{labels}{Optional labels for the pie chart.}

\item{LegendSide}{Should the legend be shown on the side of the pie chart? Default: \code{FALSE}.}

\item{...}{Additional arguments to pass to the pie chart plotting function.}
}
\value{
A pie chart visualizing the distribution of categories within the specified metadata column.
}
\description{
Generates a pie chart visualizing the distribution of categories within a specified
metadata column of a Seurat object.
}
\examples{
\dontrun{
plotMetadataCategPie(
  metacol = "Singlet.status",
  plot_name = "Singlet Status Distribution",
  obj = combined.obj,
  max.categs = 20,
  both_pc_and_value = TRUE
)
}

}
