% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{plotMetadataMedianFractionBarplot}
\alias{plotMetadataMedianFractionBarplot}
\title{plotMetadataMedianFractionBarplot}
\usage{
plotMetadataMedianFractionBarplot(
  columns = c("percent.mito", "percent.ribo"),
  suffix = NULL,
  group.by = GetClusteringRuns(obj = obj)[2],
  method = c("median", "mean")[1],
  min.thr = 2.5,
  return.matrix = F,
  main = paste(method, "read fractions per transcript class and cluster", suffix),
  ylab = "Fraction of transcriptome (\%)",
  percentify = T,
  subt = NULL,
  position = position_stack(),
  w = 10,
  h = 6,
  obj = combined.obj,
  ...
)
}
\arguments{
\item{columns}{A vector of column names to consider for the barplot. Default: c("percent.mito", "percent.ribo").}

\item{suffix}{A suffix added to the output filename. Default: NULL.}

\item{group.by}{The variable to group by for calculations. Default: Second result of GetClusteringRuns(obj).}

\item{method}{Method used for calculations, either "median" or "mean". Default: "median".}

\item{min.thr}{Minimum threshold percentage for a cluster. Default: 2.5.}

\item{return.matrix}{Logical; if TRUE, returns a matrix. Default: FALSE.}

\item{main}{Main title for the plot. Default: "read fractions per transcript class and cluster" followed by the method and suffix.}

\item{ylab}{Label for the y-axis. Default: "Fraction of transcriptome (\%)".}

\item{percentify}{Logical; if TRUE, multiplies the fraction by 100. Default: TRUE.}

\item{subt}{Subtitle for the plot. Default: NULL.}

\item{position}{Position adjustment for geoms. Default: position_stack().}

\item{w}{The width of the plot. Default: 10.}

\item{h}{The height of the plot. Default: 6.}

\item{obj}{The main Seurat object used for calculations. Default: combined.obj.}

\item{...}{Additional parameters passed to the internally called functions.}
}
\description{
Generates a barplot of metadata median values.
}
\seealso{
\code{\link[dplyr]{summarise_all}}
\code{\link[reshape2]{melt}}
}
