% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{plotQUMAPsInAFolder}
\alias{plotQUMAPsInAFolder}
\title{Plot qUMAPs for Genes in a Folder}
\usage{
plotQUMAPsInAFolder(
  genes,
  obj = combined.obj,
  foldername = NULL,
  intersectionAssay = "RNA",
  plot.reduction = "umap",
  ...
)
}
\arguments{
\item{genes}{A vector of gene names to be plotted.}

\item{obj}{An object containing the UMAP and gene data. Default: combined.obj.}

\item{foldername}{The name of the folder where the plots will be saved. If NULL, the gene set
name is used. Default: NULL.}

\item{intersectionAssay}{The assay slot to use for intersection. Default: 'RNA'.}

\item{plot.reduction}{The type of reduction to plot. Default: 'umap'.}

\item{...}{Additional arguments passed to plotting and directory creation functions.}
}
\value{
Invisible. The function generates plots and saves them in the specified folder.
}
\description{
This function plots qUMAPs for a specified set of genes, storing the results in a
specified folder. If no folder name is provided, it defaults to using the gene set name.
}
\examples{
plotQUMAPsInAFolder(genes = c("Gene1", "Gene2"), obj = combined.obj,
                    foldername = "MyGenePlots", intersectionAssay = 'RNA',
                    plot.reduction = 'umap')

}
