% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{PlotUpdateStats}
\alias{PlotUpdateStats}
\title{PlotUpdateStats}
\usage{
PlotUpdateStats(
  mat = UpdateStatMat,
  column.names = c("Updated (\%)", "Updated (Nr.)")
)
}
\arguments{
\item{mat}{A matrix containing update statistics. Default: UpdateStatMat.}

\item{column.names}{A character vector of column names in the mat parameter. Default: c("Updated (\%)", "Updated (Nr.)").}
}
\value{
A scatter plot displaying update statistics.
}
\description{
Creates a scatter plot of update statistics.
}
\details{
This function takes a matrix containing update statistics and column names to plot
the corresponding statistics. It colorizes the genes and plots the percentage of total genes
updated against the number of genes updated.
}
\examples{
\dontrun{
if (interactive()) {
  PlotUpdateStats(mat = result.of.GetUpdateStats)
}
}
}
\seealso{
\code{MarkdownReports::wplot()}, \code{wcolorize()}
}
