% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slingshot.Utils.R
\name{points_on_curve.principal_curve}
\alias{points_on_curve.principal_curve}
\title{points_on_curve.principal_curve}
\usage{
\method{points_on_curve}{principal_curve}(curve, lambda, ...)
}
\arguments{
\item{curve}{The principal curve on which to visualize points. Default: PARAM_DEFAULT_VALUE.}

\item{lambda}{The parametric location of points along the principal curve. Default: PARAM_DEFAULT_VALUE.}
}
\value{
Coordinates of points on the principal curve.
}
\description{
Helper function to visualize points on a principal curve in Slingshot.
}
\details{
This function interpolates the coordinates of points along the principal curve in Slingshot.
}
\seealso{
\code{\link[slingshot]{SlingshotDataSet}}
}
