% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slingshot.Utils.R
\name{points_on_curve.SlingshotDataSet}
\alias{points_on_curve.SlingshotDataSet}
\title{points_on_curve.SlingshotDataSet}
\usage{
\method{points_on_curve}{SlingshotDataSet}(curve, lambda, ...)
}
\arguments{
\item{curve}{The SlingshotDataSet on which to visualize points. Default: PARAM_DEFAULT_VALUE.}

\item{lambda}{The parametric location of points along the curves in the SlingshotDataSet. Default: PARAM_DEFAULT_VALUE.}
}
\value{
A dataframe of coordinates of points on the curves in the SlingshotDataSet.
}
\description{
Helper function to visualize points on a SlingshotDataSet in Slingshot.
}
\details{
This function interpolates the coordinates of points along the curves in a SlingshotDataSet.
}
\seealso{
\code{\link[slingshot]{SlingshotDataSet}}
}
