% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{PrctCellExpringGene}
\alias{PrctCellExpringGene}
\title{PrctCellExpringGene}
\source{
Adapted from code by Ryan-Zhu on Github (https://github.com/satijalab/seurat/issues/371)
}
\usage{
PrctCellExpringGene(genes, group.by = "all", obj = combined.obj)
}
\arguments{
\item{genes}{A character vector of genes of interest.}

\item{group.by}{Grouping variable, Default: 'all'.}

\item{obj}{A Seurat object containing cell data. Default: combined.obj.}
}
\value{
A data frame with the proportion of cells expressing each gene, grouped by the group.by variable.
}
\description{
Function to calculate the proportion of cells expressing a given set of genes.
}
\examples{
\dontrun{
PrctCellExpringGene(genes = c("Gene1", "Gene2"), obj = seurat_object)
}
}
