% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{processSeuratObject}
\alias{processSeuratObject}
\title{Process Seurat Objects in Parallel}
\usage{
processSeuratObject(
  obj,
  param.list = p,
  add.meta.fractions = FALSE,
  compute = TRUE,
  save = TRUE,
  plot = TRUE,
  nfeatures = param.list$n.var.genes,
  variables.2.regress = param.list$variables.2.regress.combined,
  n.PC = param.list$n.PC,
  resolutions = param.list$snn_res,
  ...
)
}
\arguments{
\item{obj}{A Seurat object to be processed.}

\item{param.list}{A list of parameters used in the processing steps.}

\item{add.meta.fractions}{A boolean indicating whether to add meta data for fractions of cells in each cluster. Default: FALSE.}

\item{compute}{A boolean indicating whether to compute the results. Default: TRUE.}

\item{save}{A boolean indicating whether to save the results. Default: TRUE.}

\item{plot}{A boolean indicating whether to plot the results. Default: TRUE.}

\item{nfeatures}{The number of variable genes to use. Default: 2000.}

\item{variables.2.regress}{A list of variables to regress out. Default: NULL.}

\item{n.PC}{The number of principal components to use. Default: 30.}

\item{resolutions}{A list of resolutions to use for clustering. Default: c(0.1, 0.2, 0.3, 0.4, 0.5).}
}
\value{
A Seurat object after applying scaling, PCA, UMAP, neighbor finding, and clustering.
}
\description{
Applies a series of Seurat processing steps to each Seurat object in a list.
The operations include scaling data, running PCA, UMAP, finding neighbors, and finding clusters.
This is done in parallel using multiple cores.
}
\examples{
# Assuming ls.Seurat is a list of Seurat objects and params is a list of parameters
# results <- mclapply(ls.Seurat, processSeuratObject, params, mc.cores = 4)
}
