% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{qClusteringUMAPS}
\alias{qClusteringUMAPS}
\title{Quick Clustering UMAPs on A4 Page}
\usage{
qClusteringUMAPS(
  obj = combined.obj,
  idents = CodeAndRoll2::na.omit.strip(GetClusteringRuns(obj)[1:4]),
  prefix = "Clustering.UMAP.Res",
  suffix = "",
  nrow = 2,
  ncol = 2,
  w = 2 * 11.69,
  h = 2 * 8.27,
  title = Stringendo::sppu(prefix, as.numeric(stringr::str_extract(idents,
    "\\\\d+\\\\.\\\\d+$")), suffix),
  ...
)
}
\arguments{
\item{obj}{Seurat object to visualize; Default: \code{combined.obj}.}

\item{idents}{Vector of clustering resolution identifiers to plot;
dynamically defaults to the first 4 found by \code{GetClusteringRuns}.}

\item{prefix}{Prefix for plot titles; Default: "Clustering.UMAP.Res".}

\item{suffix}{Suffix for plot titles; Default: "".}

\item{nrow}{Number of rows in the plot grid; Default: 2.}

\item{ncol}{Number of columns in the plot grid; Default: 2.}

\item{w}{Width of the plot; Default: 11.69.}

\item{h}{Height of the plot; Default: 8.27.}

\item{title}{Custom title for the composite plot; dynamically generated from \code{prefix}, \code{idents}, and \code{suffix}.}

\item{...}{Additional parameters for individual UMAP plots.}
}
\description{
Generates and arranges UMAP plots for up to four specified clustering resolutions
from a Seurat object onto an A4 page, facilitating comparative visualization.
}
\examples{
\dontrun{
qClusteringUMAPS()
}

}
