% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qFeatureScatter}
\alias{qFeatureScatter}
\title{qFeatureScatter}
\usage{
qFeatureScatter(
  feature1 = "TOP2A",
  feature2 = "ID2",
  obj = combined.obj,
  ext = "png",
  plot = TRUE,
  logX = F,
  logY = F,
  w = 8,
  h = 5,
  ...
)
}
\arguments{
\item{feature1}{Name of the first feature to plot. Default: 'TOP2A'.}

\item{feature2}{Name of the second feature to plot. Default: 'ID2'.}

\item{obj}{Seurat object from which to extract feature data. Default: combined.obj.}

\item{ext}{File extension for the saved plot. Default: 'png'.}

\item{plot}{Logical indicating whether to display the plot. Default: TRUE.}

\item{logX}{Logical indicating whether to apply a logarithmic transformation to the x-axis. Default: FALSE.}

\item{logY}{Logical indicating whether to apply a logarithmic transformation to the y-axis. Default: FALSE.}

\item{w}{Width of the plot in inches, Default: 6.}

\item{h}{Height of the plot in inches, Default: 9.}

\item{...}{Additional arguments to pass to the FeatureScatter function.}
}
\value{
If \code{plot} is TRUE, a ggplot object containing the feature scatter plot is returned.
}
\description{
Generates and optionally saves a scatter plot of two features from a Seurat object.
}
\examples{
\dontrun{
if(interactive()){
 qFeatureScatter(feature1 = "Gene1", feature2 = "Gene2", obj = seuratObject)
 }
}
}
\seealso{
\code{\link[Seurat]{FeatureScatter}}, \code{\link[ggplot2]{ggplot}}
}
