% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{qQC.plots.BrainOrg}
\alias{qQC.plots.BrainOrg}
\title{Quickly Plot Key QC Markers in Brain Organoids}
\usage{
qQC.plots.BrainOrg(
  obj = combined.obj,
  QC.Features = c("nFeature_RNA", "percent.ribo", "percent.mito", "nuclear.fraction",
    "percent.HGA"),
  prefix = "QC.markers.4.UMAP",
  suffix = "",
  title = Stringendo::sppu(prefix, QC.Features, suffix),
  nrow = 2,
  ncol = 2,
  ...
)
}
\arguments{
\item{obj}{Seurat object to visualize; Default: \code{combined.obj}.}

\item{QC.Features}{Vector of QC feature names to plot; Default:
\code{c("nFeature_RNA", "percent.ribo", "percent.mito", "nuclear.fraction")}.}

\item{prefix}{Prefix for plot titles; Default: "QC.markers.4.UMAP".}

\item{suffix}{Suffix for plot titles; Default: "".}

\item{title}{Custom title for the composite plot; dynamically generated from \code{prefix},
\code{QC.Features}, and \code{suffix}.}

\item{nrow}{Number of rows in the plot grid; Default: 2.}

\item{ncol}{Number of columns in the plot grid; Default: 2.}

\item{...}{Additional parameters for individual UMAP plots.}
}
\description{
Generates and arranges UMAP plots for specified QC features
from a Seurat object on an A4 page, facilitating a quick quality control (QC) overview.
}
\examples{
\dontrun{
qQC.plots.BrainOrg()
}

}
