% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{qqSaveGridA4}
\alias{qqSaveGridA4}
\title{qqSaveGridA4}
\usage{
qqSaveGridA4(
  plotlist = pl,
  plots = 1:2,
  NrPlots = length(plots),
  height = hA4,
  width = wA4,
  fname = "Fractions.Organoid-to-organoid variation.png",
  ...
)
}
\arguments{
\item{plotlist}{A list of ggplot objects. Default: pl.}

\item{plots}{A numeric vector indicating the indices of the plots to save from the 'plotlist'. Default: 1:2.}

\item{NrPlots}{Number of plots to save. Default: length(plots).}

\item{height}{Height for the saved image. Default: hA4.}

\item{width}{Width for the saved image. Default: wA4.}

\item{fname}{File name for the saved image. Default: "Fractions.Organoid-to-organoid variation.png".}

\item{...}{Additional arguments passed to the plot_grid function.}
}
\value{
This function does not return a value. It saves a grid plot of ggplot objects to the specified file.
}
\description{
Saves a grid of 2 or 4 ggplot objects onto an A4 page.
}
\examples{
\dontrun{
if (interactive()) {
  qqSaveGridA4(plotlist = pl, plots = 1:2, fname = "Fractions.per.Cl.png")
  qqSaveGridA4(plotlist = pl, plots = 1:4, fname = "Fractions.per.Cl.4.png")
}
}
}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
