% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qsave.image}
\alias{qsave.image}
\title{Save workspace - qsave.image}
\usage{
qsave.image(..., showMemObject = TRUE, options = c("--force", NULL)[1])
}
\arguments{
\item{...}{Pass any other parameter to the internally called functions (most of them should work).}

\item{showMemObject}{Logical; if TRUE, the function will print out the memory size of the largest
objects in the workspace. Default: \code{TRUE}.}

\item{options}{Options passed on to gzip, via CLI. Default: \code{c("--force", NULL)[1]}}
}
\description{
Save the workspace with external gzip compression (CPU intensive).
}
\seealso{
\code{\link[Stringendo]{kollapse}}, \code{Stringendo::iprint()}
}
