% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{qSeuViolin}
\alias{qSeuViolin}
\title{Create a Violin Plot for a Seurat Object Feature and save the file.}
\usage{
qSeuViolin(
  obj,
  feature = "nFeature_RNA",
  ident = GetNamedClusteringRuns(obj = obj, v = FALSE)[1],
  assay = "RNA",
  slot = "data",
  split.by = NULL,
  colors = NULL,
  clip.outliers = TRUE,
  replace.na = FALSE,
  pt.size = 0,
  sub = NULL,
  suffix = NULL,
  suffix.2.title = FALSE,
  caption = try(.parseKeyParams(obj), silent = TRUE),
  logY = TRUE,
  hline = FALSE,
  ylab = "Expression",
  ylimit = NULL,
  legend = TRUE,
  legend.pos = NULL,
  legend.title = NULL,
  show_plot = TRUE,
  grid = TRUE,
  w = NULL,
  h = 7,
  ...
)
}
\arguments{
\item{obj}{A Seurat object to be plotted.}

\item{feature}{A character string specifying the name of the feature to plot.}

\item{ident}{A character vector specifying the identities to be used in the plot.}

\item{assay}{Assay to fetch the feature from. Default: \code{'RNA'}.}

\item{slot}{Slot/layer of the assay to fetch the feature from. Default: \code{'data'}.}

\item{split.by}{A character string specifying the grouping variable for splitting the plot.}

\item{colors}{A character vector specifying the colors to use for the plot.}

\item{clip.outliers}{A logical indicating whether to clip outliers.}

\item{replace.na}{A logical indicating whether NA values should be replaced.}

\item{pt.size}{The size of the individual datapoints in the plot. Default: 0 for simple violin plot.}

\item{sub}{Subtitle of the plot. Default: feature by ident.}

\item{suffix}{An optional string to append to the title of the plot.}

\item{suffix.2.title}{A logical indicating whether to append the suffix to the plot title.}

\item{caption}{A character string or logical for the plot caption. If FALSE, no caption is displayed.}

\item{logY}{A logical indicating whether to use a logarithmic scale for the y-axis.}

\item{hline}{A numeric or logical value; if numeric, the value where a horizontal line should be drawn.}

\item{ylab}{Y-axis label. Default: "Expression".}

\item{ylimit}{A numeric vector specifying the limits of the y-axis.}

\item{legend}{Show legend; Default: opposite of \code{label}.}

\item{legend.pos}{Position of legend; Default: 'NULL'.}

\item{legend.title}{Title of legend; Default: 'split.by'.}

\item{show_plot}{A logical indicating whether to display the plot.}

\item{grid}{A logical indicating whether to display grid lines.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Additional arguments passed to \code{VlnPlot}.}
}
\value{
A ggplot object representing the violin plot.
}
\description{
Generates a violin plot for a specified feature in a Seurat object,
allowing for the data to be split by a specified grouping variable.
The function supports customization options such as logarithmic scaling, custom titles, and more.
}
\examples{
# Assuming `seurat_obj` is a valid Seurat object
qSeuViolin(obj = seurat_obj, feature = "nFeature_RNA")

}
