% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{read10x}
\alias{read10x}
\title{read10x}
\usage{
read10x(dir)
}
\arguments{
\item{dir}{PARAM_DESCRIPTION}
}
\description{
read10x from gzipped matrix.mtx, features.tsv and barcodes.tsv #
}
\seealso{
\code{\link[tictoc]{tic}}
\code{\link[R.utils]{compressFile}}
\code{\link[Seurat]{Read10X}}
}
