% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{recall.all.genes}
\alias{recall.all.genes}
\title{Recall all.genes global variable from a Seurat object}
\usage{
recall.all.genes(obj = combined.obj, overwrite = FALSE)
}
\arguments{
\item{obj}{Seurat object, Default: \code{combined.obj}}

\item{overwrite}{Overwrite the \code{all.genes} global variable if it already exists? Default: \code{FALSE}.}
}
\description{
Recall \code{all.genes} from a Seurat object's \code{misc} slot,
which is stored by \code{calc.q99.Expression.and.set.all.genes()}, and optionally reset the global variable.
}
\examples{
\dontrun{
if (interactive()) {
  recall.all.genes()
  all.genes
}
}
}
