% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{recall.genes.ls}
\alias{recall.genes.ls}
\title{recall.genes.ls}
\usage{
recall.genes.ls(obj = combined.obj, overwrite = FALSE)
}
\arguments{
\item{obj}{Seurat object, Default: \code{combined.obj}}

\item{overwrite}{Overwrite already existing in environment? Default: \code{FALSE}.}
}
\description{
Recall genes.ls from obj@misc to "genes.ls" in the global environment.
}
\examples{
\dontrun{
if (interactive()) {
  recall.genes.ls()
  genes.ls
}
}
}
