% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{recall.parameters}
\alias{recall.parameters}
\title{recall.parameters}
\usage{
recall.parameters(obj = combined.obj, overwrite = FALSE)
}
\arguments{
\item{obj}{Seurat object, Default: \code{combined.obj}}

\item{overwrite}{Overwrite already existing in environment? Default: \code{FALSE}.}
}
\description{
Recall parameters from obj@misc to "p" in the global environment.
}
\examples{
\dontrun{
if (interactive()) {
  recall.parameters()
  p
}
}
}
