% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{recallAllGenes}
\alias{recallAllGenes}
\title{recallAllGenes}
\usage{
recallAllGenes(obj = combined.obj, overwrite = F)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}
}
\description{
all.genes set by calc.q99.Expression.and.set.all.genes() #
}
\examples{
\dontrun{
if (interactive()) {
  recallAllGenes()
  all.genes
}
}
}
