% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{RecallReduction}
\alias{RecallReduction}
\title{RecallReduction}
\usage{
RecallReduction(obj = combined.obj, dim = 2, reduction = "umap")
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{dim}{Numer of dimensions used, Default: 2}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'umap'}
}
\description{
Set active UMAP to \code{obj@reductions$umap} from \code{obj@misc$reductions.backup}. #
}
\examples{
\dontrun{
if (interactive()) {
  combined.obj <- RecallReduction(obj = combined.obj, dim = 2, reduction = "umap")
  qUMAP()
  combined.obj <- RecallReduction(obj = combined.obj, dim = 3, reduction = "umap")
  qUMAP()
}
}
}
