% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{remove.cells.by.UMAP}
\alias{remove.cells.by.UMAP}
\title{remove.cells.by.UMAP}
\usage{
remove.cells.by.UMAP(
  reduction = "umap",
  umap_dim = 1,
  obj = combined.obj,
  cutoff = 0,
  cut_below = TRUE,
  ...
)
}
\description{
remove.cells.by.UMAP
}
