% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{remove.residual.small.clusters}
\alias{remove.residual.small.clusters}
\title{remove.residual.small.clusters}
\usage{
remove.residual.small.clusters(
  obj = combined.obj,
  identitites = GetClusteringRuns(obj = obj),
  max.cells = max(round((ncol(obj))/2000), 5)
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{identitites}{Identities to scan for residual clusters}

\item{max.cells}{Max number of cells in cluster to be removed. Default: 0.5\% of the dataset, or 5 cells.}
}
\description{
E.g.: after subsetting often some residual cells remain in clusters originally defined in the full dataset.
}
