% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{remove_clusters_and_drop_levels}
\alias{remove_clusters_and_drop_levels}
\title{Remove Clusters and Drop Levels}
\usage{
remove_clusters_and_drop_levels(
  ls_obj,
  object_names = names(ls_obj),
  indices = 2:3,
  ...
)
}
\arguments{
\item{ls_obj}{A list of Seurat objects.}

\item{object_names}{A character vector containing the names of the Seurat objects to process. Default is names of all objects in the \code{ls_obj}.}

\item{indices}{A numeric vector indicating which datasets to process by their position in the \code{object_names} vector. By default, it processes the second and third datasets.}

\item{...}{Additional parameters passed to the \code{remove.residual.small.clusters} function.}
}
\value{
The function returns the modified list of Seurat objects.
}
\description{
This function removes residual small clusters from specified Seurat objects and drops levels in factor-like metadata.
}
\details{
This function applies \code{remove.residual.small.clusters} and \code{dropLevelsSeurat} to the Seurat objects specified by the \code{indices} in the \code{object_names}.
It operates in place, modifying the input \code{ls_obj} list.
}
\examples{
\dontrun{
# Process the 2nd and 3rd datasets
remove_clusters_and_drop_levels(ls_obj, indices = c(2, 3))
}

}
