% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{removeCellsByUmap}
\alias{removeCellsByUmap}
\title{Remove Cells by Dimension Reduction}
\usage{
removeCellsByUmap(
  reduction = "umap",
  umap_dim = 1,
  obj = combined.obj,
  cutoff = 0,
  cut_below = TRUE,
  only_plot_cutoff = FALSE,
  ...
)
}
\arguments{
\item{reduction}{A string specifying the dimension reduction technique to be used ('umap', 'pca', or 'tsne'). Default is 'umap'.}

\item{umap_dim}{An integer specifying which dimension (axis) to apply the cutoff. Default is 1.}

\item{obj}{A Seurat object. Default is 'combined.obj'.}

\item{cutoff}{A numerical value indicating the cutoff value for the specified dimension. Default is 0.}

\item{cut_below}{A logical value indicating whether to remove cells below (TRUE) or above (FALSE) the cutoff line. Default is TRUE.}

\item{only_plot_cutoff}{Simulate and plot cutoff only.}

\item{...}{Any other parameters to be passed to internally called functions.}
}
\value{
A Seurat object with cells removed according to the specified cutoff.
}
\description{
This function applies a cutoff in the specified dimension of a given dimension reduction (UMAP, PCA, or t-SNE) to remove cells.
}
