% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.utils.less.used.R
\name{removeLayersByPattern}
\alias{removeLayersByPattern}
\title{Remove Layers from Seurat Object by Pattern}
\usage{
removeLayersByPattern(obj, pattern = "sc[0-9][0-9]_", perl = TRUE)
}
\arguments{
\item{pattern}{A regular expression pattern to match layer names.}

\item{seuratObj}{A Seurat object.}
}
\value{
A Seurat object with specified layers removed.
}
\description{
This function removes layers from a Seurat object's RNA assay based on a specified regular expression pattern.
It first backs up the object before removing layers that match the pattern.
}
