% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{removeResidualSmallClusters}
\alias{removeResidualSmallClusters}
\title{Remove Residual Small Clusters from Seurat Object}
\usage{
removeResidualSmallClusters(
  obj = combined.obj,
  identitites = GetClusteringRuns(obj.RG, pat = "*snn_res.[0-9].[0-9]$")[1:5],
  max.cells = max(round((ncol(obj))/2000), 5)
)
}
\arguments{
\item{obj}{Seurat object from which small clusters will be removed; Default: \code{combined.obj}.}

\item{identitites}{Vector of clustering identities to examine for small clusters;
Default: \code{GetClusteringRuns(obj)}.}

\item{max.cells}{Maximum number of cells a cluster can contain to still be considered for removal.
Default: The lesser of 0.5\% of the dataset or 5 cells.}
}
\description{
Removes clusters containing fewer cells than specified by \code{max.cells}
from a Seurat object. This function is particularly useful after subsetting a dataset,
where small, possibly unrepresentative clusters may remain.
}
\examples{
\dontrun{
if (interactive()) {
  combined.obj <- removeResidualSmallClusters(obj = combined.obj)
}
}

}
