% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.utils.less.used.R
\name{removeScaleData}
\alias{removeScaleData}
\title{Remove Scale Data from Seurat Objects}
\usage{
removeScaleData(ls.obj)
}
\arguments{
\item{ls.obj}{A list of Seurat objects.}
}
\value{
A list of Seurat objects with \code{scale.data} slot removed from RNA assays.
}
\description{
Remove Scale Data from Seurat Objects
}
\examples{
# Assuming `seuratList` is a list of Seurat objects
seuratList <- removeScaleData(seuratList)
}
