% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{renameAzimuthColumns}
\alias{renameAzimuthColumns}
\title{Rename Azimuth Columns in Seurat Object}
\usage{
renameAzimuthColumns(
  obj,
  ref = c("humancortexref", "fetusref")[1],
  prefix = "azi",
  azim_cols = CodeAndRoll2::grepv(x = tail(colnames(obj@meta.data), 10), pattern =
    "predicted.")
)
}
\arguments{
\item{obj}{A Seurat object containing metadata in \code{meta.data} that needs column names to be renamed.}

\item{ref}{A character string specifying the reference; defaults to "humancortexref".
The "ref" part of the string will be removed in the new column names.}

\item{prefix}{A character string to prefix to the column names, defaulting to "azi".
This prefix is combined with the modified \code{ref} to form the new column names.}

\item{azim_cols}{Azimuth columns}
}
\value{
Returns the Seurat object with renamed metadata columns.
}
\description{
Dynamically renames specified metadata columns in a Seurat object, particularly those
prefixed with "predicted." and the "mapping.score" column, by applying a new prefix
that combines a user-defined prefix and a reference name.
}
\examples{
# Assuming `obj` is a Seurat object with metadata columns following the "predicted." pattern:
obj <- renameAzimuthColumns(obj, ref = "humancortexref", prefix = "azi")
# This will rename columns like "predicted.class" to "azi.humancortex.class"
# and include "mapping.score" as "azi.humancortex.mapping.score"

}
