% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{RenameGenesSeurat}
\alias{RenameGenesSeurat}
\title{RenameGenesSeurat}
\usage{
RenameGenesSeurat(
  obj = ls.Seurat[[i]],
  newnames = HGNC.updated[[i]]$Suggested.Symbol
)
}
\arguments{
\item{obj}{Seurat object, Default: ls.Seurat[\link{i}]}

\item{newnames}{PARAM_DESCRIPTION, Default: HGNC.updated[\link{i}]$Suggested.Symbol}
}
\description{
Replace gene names in different slots of a Seurat object. Run this before integration. Run this before integration. It only changes obj@assays$RNA@counts, @data and @scale.data. #
}
\examples{
\dontrun{
if(interactive()){
 RenameGenesSeurat(obj = SeuratObj, newnames = HGNC.updated.genes$Suggested.Symbol)
 }
}
}
