% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{replace_by_most_frequent_categories}
\alias{replace_by_most_frequent_categories}
\title{replace_by_most_frequent_categories}
\usage{
replace_by_most_frequent_categories(
  df,
  query_col = colnames(df)[1],
  ref_col = colnames(df)[2],
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{query_col}{The name of the column in 'df' whose categories are to be replaced.
By default, the first column of 'df' is used.}

\item{ref_col}{The name of the column in 'df' used as reference for replacement.
By default, the second column of 'df' is used.}

\item{show_plot}{Logical, whether to plot assignment quality. Defaults to TRUE.}

\item{...}{Additional parameters passed to the qbarplot function.}
}
\value{
A data frame with categories in 'query_col' replaced by the most frequent match
from 'ref_col'.
}
\description{
Replace each category in the 'query_col' column of a data frame
with the most frequently corresponding category in the 'ref_col' column.
}
\details{
Replace Categories by the Most Frequent Match

This function replaces each category in a query column of a data frame with the most
frequently corresponding category in a reference column. It calculates the assignment quality,
reports it, and optionally plots it.
}
\examples{
\dontrun{
replace_by_most_frequent_categories(df = my_data)
}

}
