% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{sampleNpc}
\alias{sampleNpc}
\title{sampleNpc}
\usage{
sampleNpc(metaDF = MetaData[which(Pass), ], pc = 0.1)
}
\arguments{
\item{metaDF}{PARAM_DESCRIPTION, Default: MetaData\link{which(Pass), }}

\item{pc}{PARAM_DESCRIPTION, Default: 0.1}
}
\description{
Sample N \% of a dataframe (obj@metadata), and return the cell IDs. #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
