% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{save.parameters}
\alias{save.parameters}
\title{Save Parameters to Seurat Object}
\usage{
save.parameters(obj = combined.obj, params = p, overwrite = TRUE)
}
\arguments{
\item{obj}{Seurat object to update; Default: \code{combined.obj}.}

\item{params}{List of parameters to save; Default: \code{p}.}

\item{overwrite}{Logical indicating if existing parameters should be overwritten; Default: \code{TRUE}.}
}
\description{
Stores a list of parameters within the \verb{@misc$p} slot of a Seurat object,
allowing for easy reference and tracking of analysis parameters used.
}
\examples{
\dontrun{
if (interactive()) {
  save.parameters(obj = combined.obj, params = p)
}
}

}
