% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{save2plots.A4}
\alias{save2plots.A4}
\title{Save Two Plots on One A4 Page}
\usage{
save2plots.A4(
  plot_list,
  pname = FALSE,
  suffix = NULL,
  scale = 1,
  nrow = 2,
  ncol = 1,
  h = hA4 * scale,
  w = wA4 * scale,
  ...
)
}
\arguments{
\item{plot_list}{A list containing ggplot objects to be arranged and saved.}

\item{pname}{Boolean indicating if the plot name should be automatically generated;
if FALSE, the name is based on \code{plot_list} and \code{suffix}; Default: FALSE.}

\item{suffix}{Suffix to be added to the generated filename if \code{pname} is FALSE; Default: NULL.}

\item{scale}{Scaling factor for adjusting the plot size; Default: 1.}

\item{nrow}{Number of rows in the plot arrangement; Default: 2.}

\item{ncol}{Number of columns in the plot arrangement; Default: 1.}

\item{h}{Height of the plot, calculated as A4 height times \code{scale};
calculated dynamically based on \code{scale}.}

\item{w}{Width of the plot, calculated as A4 width times \code{scale};
calculated dynamically based on \code{scale}.}

\item{...}{Additional parameters passed to \code{plot_grid}.}
}
\description{
Arranges and saves two UMAP plots (or any plots) side-by-side or one above
the other on a single A4 page.
}
\examples{
\dontrun{
p1 <- ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point()
p2 <- ggplot(iris, aes(Petal.Length, Petal.Width, color = Species)) +
  geom_point()
save2plots.A4(plot_list = list(p1, p2))
}

}
