% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{save2umaps.A4}
\alias{save2umaps.A4}
\title{save2umaps.A4}
\usage{
save2umaps.A4(
  plot_list,
  pname = FALSE,
  suffix = NULL,
  scale = 1,
  nrow = 2,
  ncol = 1,
  h = hA4 * scale,
  w = wA4 * scale,
  ...
)
}
\arguments{
\item{plot_list}{A list of plots to be saved on an A4 page.}

\item{pname}{Boolean to determine if name is generated automatically. If FALSE, name is generated based on plot_list and suffix. Default: FALSE}

\item{suffix}{A suffix added to the filename. Default: NULL}

\item{scale}{Scaling factor for the size of the plot. Default: 1}

\item{nrow}{Number of rows to arrange the plots in. Default: 2}

\item{ncol}{Number of columns to arrange the plots in. Default: 1}

\item{h}{Height of the plot, calculated as height of A4 page times the scale. Default: hA4 * scale}

\item{w}{Width of the plot, calculated as width of A4 page times the scale. Default: wA4 * scale}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Save 2 umaps on 1 A4
}
