% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{save4plots.A4}
\alias{save4plots.A4}
\title{Save Four Plots on One A4 Page}
\usage{
save4plots.A4(
  plot_list,
  pname = FALSE,
  suffix = NULL,
  scale = 1,
  nrow = 2,
  ncol = 2,
  h = 8.27 * scale,
  w = 11.69 * scale,
  ...
)
}
\arguments{
\item{plot_list}{A list containing ggplot objects to be arranged and saved; each object represents one panel.}

\item{pname}{Plot name; if FALSE, a name is generated automatically based on \code{plot_list} and \code{suffix}; Default: FALSE.}

\item{suffix}{Suffix to be added to the filename; Default: NULL.}

\item{scale}{Scaling factor for adjusting the size of the overall plot canvas; Default: 1.}

\item{nrow}{Number of rows to arrange the plots in; Default: 2.}

\item{ncol}{Number of columns to arrange the plots in; Default: 2.}

\item{h}{Height of the plot canvas, calculated as the height of an A4 page times \code{scale}; Default: \code{8.27 * scale}.}

\item{w}{Width of the plot canvas, calculated as the width of an A4 page times \code{scale}; Default: \code{11.69 * scale}.}

\item{...}{Additional parameters passed to \code{plot_grid}.}
}
\description{
Arranges and saves four plots (e.g. UMAPs) onto a single A4 page, allowing for a
compact comparison of different visualizations or clustering results.
}
\examples{
\dontrun{
p1 <- ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point()
p2 <- ggplot(mtcars, aes(mpg, disp, color = as.factor(cyl))) +
  geom_point()
p3 <- ggplot(mpg, aes(displ, hwy, color = class)) +
  geom_point()
p4 <- ggplot(diamonds, aes(carat, price, color = cut)) +
  geom_point()
save4plots.A4(plot_list = list(p1, p2, p3, p4))
}

}
