% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{save4umaps.A4}
\alias{save4umaps.A4}
\title{save4umaps.A4}
\usage{
save4umaps.A4(
  plot_list,
  pname = F,
  suffix = NULL,
  scale = 1,
  nrow = 2,
  ncol = 2,
  h = wA4 * scale,
  w = hA4 * scale,
  ...
)
}
\arguments{
\item{plot_list}{A list of ggplot objects, each of which is one panel.}

\item{pname}{Plotname, Default: F}

\item{suffix}{A suffix added to the filename, Default: NULL}

\item{scale}{Scaling factor of the canvas, Default: 1}

\item{nrow}{number of rows for panelson the page, Default: 2}

\item{ncol}{number of columns for panelson the page, Default: 2}

\item{h}{height of the plot, Default: wA4 * scale}

\item{w}{width of the plot, Default: hA4 * scale}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Save 4 umaps on 1 A4
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
