% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{saveLsSeuratMetadata}
\alias{saveLsSeuratMetadata}
\title{Save Metadata from a List of Seurat Objects}
\usage{
saveLsSeuratMetadata(ls.obj, suffix)
}
\arguments{
\item{ls.obj}{A list of Seurat objects.}

\item{suffix}{A character string to append to the filename when saving metadata.}
}
\value{
Invisible list of metadata frames
}
\description{
This function takes a list of Seurat objects, extracts their metadata, and saves it to a file with a specified suffix.
}
