% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{saveRDS.compress.in.BG}
\alias{saveRDS.compress.in.BG}
\title{saveRDS.compress.in.BG}
\usage{
saveRDS.compress.in.BG(obj, compr = FALSE, fname)
}
\arguments{
\item{obj}{Seurat object.}

\item{compr}{PARAM_DESCRIPTION, Default: FALSE}

\item{fname}{File name}
}
\description{
Save and RDS object and compress resulting file in the background using system(gzip). OS X or unix.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tictoc]{tic}}
}
