% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{scBarplot.cells.per.cluster}
\alias{scBarplot.cells.per.cluster}
\title{scBarplot.cells.per.cluster}
\usage{
scBarplot.cells.per.cluster(
  ident = GetOrderedClusteringRuns()[1],
  sort = F,
  label = T,
  palette = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[3],
  obj = combined.obj,
  return_table = F,
  ...
)
}
\arguments{
\item{ident}{identity used, Default: 'cl.names.KnownMarkers.0.5'}

\item{sort}{PARAM_DESCRIPTION, Default: F}

\item{obj}{Seurat object, Default: combined.obj}
}
\description{
Barplot the Fraction of cells per cluster. (dupl?)
}
\examples{
\dontrun{
if(interactive()){
 scBarplot.cells.per.cluster(); barplot.cells.per.cluster(sort = T)
 }
}
}
