% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{scBarplotFractionBelowThr}
\alias{scBarplotFractionBelowThr}
\title{scBarplotFractionBelowThr}
\usage{
scBarplotFractionBelowThr(
  thrX = 0.01,
  value.col = "percent.ribo",
  id.col = "cl.names.top.gene.res.0.3",
  obj = combined.obj,
  return.df = F
)
}
\arguments{
\item{thrX}{PARAM_DESCRIPTION, Default: 0.01}

\item{value.col}{PARAM_DESCRIPTION, Default: 'percent.ribo'}

\item{id.col}{PARAM_DESCRIPTION, Default: 'cl.names.top.gene.res.0.3'}

\item{obj}{Seurat object, Default: combined.obj}

\item{return.df}{PARAM_DESCRIPTION, Default: F}
}
\description{
scBarplotFractionBelowThr
}
\examples{
\dontrun{
if(interactive()){
 scBarplotFractionBelowThr(id.col =  'cl.names.top.gene.res.0.3', value.col = 'percent.ribo', thrX = 0.01, return.df = T)
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{se-deprecated}}
}
