% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scBarplotStackedMetaCateg_List}
\alias{scBarplotStackedMetaCateg_List}
\title{Stacked Barplot of Metadata Categories for List of Seurat Objects}
\usage{
scBarplotStackedMetaCateg_List(ls.obj, meta.col, ...)
}
\arguments{
\item{ls.obj}{List of Seurat objects.}

\item{meta.col}{The metadata column name to be used for the barplot.}

\item{...}{Additional arguments passed to \code{ggExpress::qbarplot.df}.}
}
\value{
A ggplot object representing the stacked barplot.
}
\description{
Creates and saves a stacked barplot for a specified metadata category
from a list of Seurat objects.
}
\examples{
\dontrun{
ls.obj <- list(obj, obj)
scBarplotStackedMetaCateg_List(ls.obj, meta.col = orig.ident)
}

}
