% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scGOEnrichment}
\alias{scGOEnrichment}
\title{Perform GO Enrichment Analysis}
\usage{
scGOEnrichment(
  genes,
  universe = NULL,
  org_db = "org.Hs.eg.db",
  key_type = "SYMBOL",
  ont = "BP",
  pAdjustMethod = "BH",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2,
  save = TRUE,
  suffix = NULL,
  check.gene.symbols = TRUE,
  ...
)
}
\arguments{
\item{genes}{Character vector. List of genes for enrichment analysis. Default: NULL.}

\item{universe}{Character vector. Background gene list (universe). Default: NULL.}

\item{org_db}{Character. Organism-specific database to use (e.g., 'org.Hs.eg.db'). Default: 'org.Hs.eg.db'.}

\item{key_type}{Character. Gene identifier type (e.g., 'SYMBOL', 'ENTREZID'). Default: 'SYMBOL'.}

\item{ont}{Character. Ontology type to use (e.g., 'BP', 'MF', 'CC'). Default: 'BP'.}

\item{pAdjustMethod}{Character. Method for p-value adjustment. Default: 'BH' for Benjamini-Hochberg.}

\item{pvalueCutoff}{Numeric. P-value cutoff for significance. Default: 0.05.}

\item{qvalueCutoff}{Numeric. Q-value cutoff for significance. Default: 0.2.}

\item{save}{Logical. Save the results as a data frame. Default: TRUE.}

\item{suffix}{Character. Suffix to append to the output file name. Default: 'GO.Enrichments'.}

\item{check.gene.symbols}{Logical. Check gene symbols for validity. Default: TRUE.}
}
\value{
A data frame with GO enrichment results.
}
\description{
This function performs Gene Ontology (GO) enrichment analysis using the
\code{clusterProfiler::enrichGO} function. It takes the gene list, universe, organism database,
gene identifier type, and ontology type as inputs and returns the enrichment results.
}
\examples{
\dontrun{
gene_list <- rownames(df.m.DL.up.in.TSC)
background_genes <- names(all.genes)
go_results <- performGOEnrichment(gene_list, background_genes, "org.Hs.eg.db", "SYMBOL", "BP")
print(go_results)
}
}
