% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scPieClusterDistribution}
\alias{scPieClusterDistribution}
\title{scPieClusterDistribution}
\usage{
scPieClusterDistribution(
  obj = combined.obj,
  ident = GetClusteringRuns(obj)[1],
  ...
)
}
\arguments{
\item{obj}{A single-cell RNA-seq object. Default: \code{combined.obj}.}

\item{ident}{A character string specifying the clustering identity to be used. Default: the first
clustering run in the object.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A pie chart displaying the cluster distribution.
}
\description{
This function generates a pie chart of cluster distributions for a given clustering
identity in a single-cell RNA-seq object.
}
\examples{
\dontrun{
scPieClusterDistribution(obj = combined.obj, ident = 'cluster_identity')
}
}
