% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scPlotPCAvarExplained}
\alias{scPlotPCAvarExplained}
\title{scPlotPCAvarExplained}
\usage{
scPlotPCAvarExplained(obj = combined.obj, use.MarkdownReports = FALSE)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{use.MarkdownReports}{Use MarkdownReports for plotting, Default: FALSE}
}
\description{
Plot the percent of variation associated with each PC.
}
