% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{set.all.genes}
\alias{set.all.genes}
\title{set.all.genes}
\usage{
set.all.genes(obj = combined.obj)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}
}
\description{
set.all.genes
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
