% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{set.mm}
\alias{set.mm}
\title{set.mm}
\usage{
set.mm(obj = combined.obj)
}
\arguments{
\item{obj}{Seurat object, Default: \code{combined.obj}}
}
\description{
Helps to find metadata columns. It creates a list with the names of of 'obj@meta.data'.
}
\examples{
\dontrun{
if (interactive()) {
  set.mm()
  mm
}
}
}
