% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{SetupReductionsNtoKdimensions}
\alias{SetupReductionsNtoKdimensions}
\title{SetupReductionsNtoKdimensions}
\usage{
SetupReductionsNtoKdimensions(
  obj = combined.obj,
  nPCs = p$n.PC,
  dimensions = 3:2,
  reduction_input = "pca",
  reduction_output = "umap",
  ...
)
}
\arguments{
\item{obj}{A Seurat object. Default: combined.obj}

\item{nPCs}{A numeric value representing the number of principal components to use. Default: p$n.PC}

\item{dimensions}{A numeric vector specifying the dimensions to use for the dimensionality reductions. Default: 3:2}

\item{reduction_input}{The type of dimensionality reduction to use as input. Can be "pca", or
some correctionn results, like harmony pca.}

\item{reduction_output}{Te type of dimensionality reduction to perform.  Can be "umap", "tsne",
or "pca". Default: 'umap'}
}
\value{
The input Seurat object with computed dimensionality reductions and backups of these reductions.
}
\description{
Function to calculate N-to-K dimensional umaps (default = 2:3); and back them up to
slots \code{obj@misc$reductions.backup} from @reductions$umap
}
\examples{
\dontrun{
if (interactive()) {
  combined.obj <- SetupReductionsNtoKdimensions(obj = combined.obj, nPCs = 10, dimensions = 2:3, reduction = "umap")
}
}
}
