% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{seu.add.meta.from.table}
\alias{seu.add.meta.from.table}
\title{seu.add.meta.from.table}
\usage{
seu.add.meta.from.table(
  obj = combined.obj,
  meta = MetaData.ORC,
  suffix = ".fromMeta"
)
}
\arguments{
\item{obj}{Seurat object, Default: seu.ORC}

\item{meta}{PARAM_DESCRIPTION, Default: MetaData.ORC}

\item{suffix}{A suffix added to the filename, Default: '.fromMeta'}
}
\description{
Add multiple new metadata columns to a Seurat object from a table. #
}
\examples{
\dontrun{
if(interactive()){
 combined.obj <- seu.add.meta.from.table()
 }
}
}
