% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{seu.add.meta.from.vector}
\alias{seu.add.meta.from.vector}
\title{seu.add.meta.from.vector}
\usage{
seu.add.meta.from.vector(
  obj = combined.obj,
  metaD.colname = metaD.colname.labeled,
  Label.per.cell = Cl.Label.per.cell
)
}
\arguments{
\item{obj}{A Seurat object to which the new metadata column will be added. Default: combined.obj.}

\item{metaD.colname}{A string specifying the name of the new metadata column. Default: metaD.colname.labeled.}

\item{Label.per.cell}{A vector of labels for each cell, to be added as new metadata. Default: Cl.Label.per.cell.}
}
\value{
A Seurat object with the new metadata column added.
}
\description{
Adds a new metadata column to a Seurat object.
}
\examples{
\dontrun{
if (interactive()) {
  # Example usage:
  combined.obj <- seu.add.meta.from.vector(
    obj = combined.obj,
    metaD.colname = metaD.colname.labeled,
    Label.per.cell = Cl.Label.per.cell
  )
}
}
}
