% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{seu.map.and.add.new.ident.to.meta}
\alias{seu.map.and.add.new.ident.to.meta}
\title{seu.map.and.add.new.ident.to.meta}
\usage{
seu.map.and.add.new.ident.to.meta(
  obj = combined.obj,
  ident.table = clusterIDs.GO.process,
  orig.ident = Idents(obj),
  metaD.colname = Stringendo::substitute_deparse(ident.table)
)
}
\arguments{
\item{obj}{The Seurat object to which the new metadata column will be added. Default: combined.obj.}

\item{ident.table}{A data frame or matrix with identity mapping data. This parameter is used to map the old identities to the new ones. Default: clusterIDs.GO.process.}

\item{orig.ident}{The original identities of the Seurat object. Default: Idents(obj).}

\item{metaD.colname}{A string specifying the name of the new metadata column. The default value is the name of the provided ident.table.}
}
\value{
A Seurat object with the new metadata column added.
}
\description{
Adds a new metadata column to a Seurat object based on an identity mapping table.
}
\examples{
\dontrun{
if (interactive()) {
  # Example usage:
  combined.obj <- seu.map.and.add.new.ident.to.meta(
    obj = combined.obj,
    ident.table = clusterIDs.GO.process
  )
}
}
}
