% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{seu.PC.var.explained}
\alias{seu.PC.var.explained}
\title{seu.PC.var.explained}
\usage{
seu.PC.var.explained(obj = combined.obj)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}
}
\description{
Determine percent of variation associated with each PC. #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
