% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{seu_reclassify_doublets_by_q95_nFeature_of_unassigned}
\alias{seu_reclassify_doublets_by_q95_nFeature_of_unassigned}
\title{Reclassify doublets based on 95th percentile of a feature for unassigned cells in a Seurat object}
\usage{
seu_reclassify_doublets_by_q95_nFeature_of_unassigned(
  obj,
  genotype_col = "Genotype",
  features_col = "nFeature_RNA",
  genotype_orig_col = "Genotype.orig",
  unassigned_str = "^unassigned",
  doublet_str = "^doublet",
  true_doublet_str = "TrueDoublet",
  unassigned_doublet_str = "UnassignedDoublet",
  quantile_val = 0.95,
  suffix = "Reassigned Doublets",
  plot_vio = TRUE,
  show_plot = FALSE,
  min_unassigned_cells = 20,
  fixed_cutoff = 2500
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{genotype_col}{A string indicating the metadata column storing genotype information. Default is "Genotype".}

\item{features_col}{A string indicating the metadata column with the feature (e.g., nFeature_RNA) to use for reclassification. Default is "nFeature_RNA".}

\item{genotype_orig_col}{A string indicating the column where the original genotype information will be stored. Default is "Genotype.orig".}

\item{unassigned_str}{A regular expression string to match cells that are unassigned. Default is "^unassigned".}

\item{doublet_str}{A regular expression string to match cells that are doublets. Default is "^doublet".}

\item{true_doublet_str}{A string indicating the name for true doublets. Default is "TrueDoublet".}

\item{unassigned_doublet_str}{A string indicating the name for unassigned doublets. Default is "UnassignedDoublet".}

\item{quantile_val}{The quantile value to use for reclassification. Default is 0.95.}

\item{plot_vio}{A boolean indicating if a violin plot should be generated. Default is TRUE.}
}
\value{
A modified Seurat object.
}
\description{
This function modifies the \code{genotype_col} metadata column of a Seurat object based on
a given quantile of the \code{features_col} for cells that match the \code{unassigned_str} pattern.
If the \code{plot_vio} flag is set to TRUE, a violin plot is generated for visualization.
}
\examples{
# Assuming `seurat_obj` is a valid Seurat object
modified_seurat_obj <- seu_reclassify_doublets_by_q95_nFeature_of_unassigned(seurat_obj)

}
