% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{SeuratColorVector}
\alias{SeuratColorVector}
\title{SeuratColorVector}
\usage{
SeuratColorVector(
  ident = NULL,
  obj = combined.obj,
  plot.colors = FALSE,
  simple = FALSE
)
}
\arguments{
\item{ident}{identity used, Default: NULL}

\item{obj}{Seurat object, Default: combined.obj}

\item{plot.colors}{Show colors? Default: FALSE}

\item{simple}{Return simply the unique colors, in order? Default: FALSE}
}
\description{
Recall a Seurat color vector.
}
\examples{
\dontrun{
if (interactive()) {
  SeuratColorVector()
  SeuratColorVector(ident = GetNamedClusteringRuns()[2], plot.colors = TRUE)
}
}
}
\seealso{
\code{\link[scales]{hue_pal}}
}
