% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{shorten_clustering_names}
\alias{shorten_clustering_names}
\title{Shorten Clustering Names}
\usage{
shorten_clustering_names(str)
}
\arguments{
\item{str}{A character string representing the clustering name to be shortened.}
}
\value{
A character string representing the shortened clustering name.
}
\description{
This function takes in a string representing a clustering name,
and shortens it according to specific rules. It replaces "snn_res." with "",
"best.matching.names" with "bmatch", "ordered" with "ord",
"ManualNames" with "mNames", and ".long" at the end of the string with ".L".
}
\examples{
\dontrun{
shorten_clustering_names("RNA_snn_res.0.5.ordered.ManualNames") # Returns 'RNA.0.5.ord.mNames'
shorten_clustering_names("RNA_snn_res.0.3.best.matching.names.ManualNames.long") # Returns 'RNA.0.3.bmatch.mNames.L'
shorten_clustering_names("RNA_snn_res.1.7.ordered.ManualNames.Simplest") # Returns 'RNA.1.7.ord.mNames.Simplest'
shorten_clustering_names("RNA_snn_res.0.5.ordered.ManualNames.Simpler") # Returns 'RNA.0.5.ord.mNames.Simpler'
}

}
