% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{showToolsSlots}
\alias{showToolsSlots}
\title{Display Slots in the @tools of an Seurat Object}
\usage{
showToolsSlots(obj, indent = 0)
}
\arguments{
\item{obj}{An object whose \verb{@tools} slot needs to be examined.}

\item{indent}{An integer specifying the level of indentation for the output.
Default is 0. Currently not used in the function implementation.}
}
\description{
\code{showToolsSlots} prints the names of slots in the \verb{@tools} of a given object.
It specifically targets list elements, skipping over data frames and other non-list objects.
}
\details{
The function iterates over the slots in the \verb{@tools} of \code{obj}. If a slot is a list
(and not a data frame), it prints the names of elements within this list. If the slot
is not a list or is a data frame, it skips printing the names. The function currently
does not use the \code{indent} parameter but it could be incorporated in future enhancements
to control the formatting of the output.
}
\note{
This function is specifically designed for S4 objects with a 'tools' slot.
The behavior with other object types or slots may not be as expected.
}
\examples{
# Assuming obj is an S4 object with a 'tools' slot
showToolsSlots(obj)

}
