% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{SNP.demux.fix.GT.table}
\alias{SNP.demux.fix.GT.table}
\title{SNP.demux.fix.GT.table}
\usage{
SNP.demux.fix.GT.table(
  GT.table = Genotypes.37.named,
  obj = obj.Neurons.37,
  col1 = "assignment.named",
  col1.new.name = "Genotype",
  col2 = "status",
  col2.new.name = "Singlet.status",
  suffix = "lib.37",
  cellname_prefix = FALSE,
  return_tbl_for_cells_found_in_object = TRUE
)
}
\arguments{
\item{GT.table}{A 2 column Genotype assignment table with singlet status}

\item{obj}{Seurat object.}

\item{col1}{'assignment.named'}

\item{col1.new.name}{New name 'Genotype'}

\item{col2}{'status'}

\item{col2.new.name}{New name 'Singlet.status'}

\item{suffix}{Add this to 'Singlet.status', then overwrite 'Genotype'}

\item{cellname_prefix}{prefix for cell names added by Seurat to distingish 10X lanes}

\item{return_tbl_for_cells_found_in_object}{cells.found.in.both}
}
\description{
Fix SNP demux table results coming from SoupOrCell
}
