% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{sparse.cor}
\alias{sparse.cor}
\title{sparse.cor}
\usage{
sparse.cor(smat)
}
\arguments{
\item{smat}{PARAM_DESCRIPTION}
}
\description{
sparse.cor
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
