% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{StoreAllMarkers}
\alias{StoreAllMarkers}
\title{Store All Differential Expression Markers}
\usage{
StoreAllMarkers(
  obj = combined.obj,
  df_markers = df.markers,
  res = 0.5,
  digit = c(0, 3)[2]
)
}
\arguments{
\item{obj}{Seurat object to update with differential expression markers. Default: \code{combined.obj}.}

\item{df_markers}{Data frame containing the results from differential gene expression analysis
(\code{FindAllMarkers()} output). Default: \code{df.markers}.}

\item{res}{Clustering resolution identifier for storing and referencing the markers. Default: 0.5.}

\item{digit}{Number of significant digits to retain in numerical values. Default: 3.}
}
\description{
Saves the complete output table from \code{FindAllMarkers()} to a Seurat object, facilitating
easy access to differential expression analysis results. This function rounds numerical values to a
specified number of digits to maintain readability and manage file sizes.
}
\examples{
\dontrun{
if (interactive()) {
  combined.obj <- StoreAllMarkers(obj = combined.obj, df_markers = df.markers, res = 0.5)
}
}

}
