% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{StoreTop25Markers}
\alias{StoreTop25Markers}
\title{StoreTop25Markers}
\usage{
StoreTop25Markers(obj = combined.obj, df_markers = df.markers, res = 0.5)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{df_markers}{PARAM_DESCRIPTION, Default: df.markers}

\item{res}{Clustering resoluton to use, Default: 0.5}
}
\description{
Save the top 25 makers based on \code{avg_log2FC} output table of \code{FindAllMarkers()} (df_markers) under \verb{@misc$df.markers$res...}. By default, it rounds up insignificant digits up to 3. #
}
\examples{
\dontrun{
if(interactive()){
 combined.obj <- StoreTop25Markers(df_markers = df.markers, res = 0.)
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
}
