% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monocle.Utils.R
\name{subsetMonocleObject}
\alias{subsetMonocleObject}
\title{subsetMonocleObject}
\usage{
subsetMonocleObject(
  obj = cds_from_seurat,
  fraction_ = 0.1,
  nCells = F,
  seed_ = 1989
)
}
\arguments{
\item{obj}{A Seurat object to be subsetted. Default: cds_from_seurat.}

\item{fraction_}{A numeric value indicating the fraction of the object to be kept. Default: 0.1.}

\item{nCells}{A numeric or boolean value indicating the number of cells to keep. If FALSE, the function calculates the number of cells to keep based on the fraction parameter. Default: FALSE.}

\item{seed_}{A numeric value specifying the seed for random number generation. Default: 1989.}
}
\value{
A subset of the original Seurat object.
}
\description{
Subsets a compressed Seurat object and saves it in the working directory.
}
\details{
This function subsets a Seurat object by either a specified fraction or a specific number of cells. The subset object is then returned by the function.
}
\examples{
\dontrun{
if(interactive()){
 cds.10pc <- subsetMonocleObject(cds_from_seurat, fraction_ = .1);
 cds.25pc <- subsetMonocleObject(cds_from_seurat, fraction_ = .25)
 }
}
}
\seealso{
\code{\link[sample]{sample}}, \code{\link[percentage_formatter]{percentage_formatter}}
}
