% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{subsetSeuObj.ident.class}
\alias{subsetSeuObj.ident.class}
\title{subsetSeuObj.ident.class}
\usage{
subsetSeuObj.ident.class(
  obj = combined.obj,
  ident = "RNA_snn_res.0.5.ordered.ManualNames",
  clusters = "Neuron, unclear"
)
}
\arguments{
\item{obj}{Seurat object, Default: ORC}

\item{ident}{identity}

\item{clusters}{which value to match}
}
\description{
Subset a Seurat Obj to a given column
}
