% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{subsetSeuObj}
\alias{subsetSeuObj}
\title{subsetSeuObj}
\usage{
subsetSeuObj(obj = ls.Seurat[[i]], fraction_ = 0.25, nCells = F, seed_ = 1989)
}
\arguments{
\item{obj}{Seurat object, Default: ls.Seurat[\link{i}]}

\item{fraction_}{PARAM_DESCRIPTION, Default: 0.25}

\item{nCells}{PARAM_DESCRIPTION, Default: F}

\item{seed_}{PARAM_DESCRIPTION, Default: 1989}
}
\description{
Subset a compressed Seurat Obj and save it in wd. #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
