% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{subsetSeuObjByIdent}
\alias{subsetSeuObjByIdent}
\title{Subset a Seurat Object by Identity}
\usage{
subsetSeuObjByIdent(
  obj = combined.obj,
  ident = GetClusteringRuns()[1],
  clusters,
  invert = FALSE
)
}
\arguments{
\item{obj}{A Seurat object. Default: \code{NULL}.}

\item{ident}{The name of the identity column to use for subsetting. It is recommended to
specify this explicitly. Default: First entry from the result of \code{GetClusteringRuns()}.}

\item{clusters}{A vector of cluster values for which cells should be matched and retained.
This parameter does not have a default value and must be specified.}

\item{invert}{A logical indicating whether to invert the selection, keeping cells that do
not match the specified clusters. Default: \code{FALSE}.}
}
\value{
A Seurat object subsetted based on the specified identity and clusters.
}
\description{
Subsets a Seurat object based on a specified identity column and values. It allows
for an optional inversion of the selection.
}
\examples{
# Assuming `seurat_obj` is your Seurat object and you want to subset based on cluster 1
subsetted_obj <- subsetSeuObjByIdent(
  obj = seurat_obj, ident = "your_ident_column",
  clusters = c(1), invert = FALSE
)

}
