% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{transfer_labels_seurat}
\alias{transfer_labels_seurat}
\title{Transfer Labels in Seurat}
\usage{
transfer_labels_seurat(
  query_obj,
  reference_path,
  named_ident = "RNA_snn_res.0.3.ordered.ManualNames",
  new_ident = gsub(pattern = "ordered", replacement = "transferred", x = named_ident),
  predictions_col = "predicted.id",
  suffix = "NEW",
  ...
)
}
\arguments{
\item{query_obj}{A Seurat object for which the labels are to be transferred.}

\item{reference_path}{A character string indicating the file path to the reference Seurat object. The path must exist.}

\item{named_ident}{A character string specifying the name of the identity class to be used from the reference Seurat object. Default is 'RNA_snn_res.0.3.ordered.ManualNames'.}

\item{new_ident}{A character string specifying the name of the new identity class to be created in the query Seurat object. Default is obtained by replacing 'ordered' with 'transferred' in named_ident.}

\item{predictions_col}{A character string specifying the column in the metadata of the transferred Seurat object containing the transferred labels. Default is 'predicted.id'.}

\item{suffix}{A character string to be used as a suffix in the visualization. Default is 'NEW'.}

\item{...}{Additional arguments passed to the Seurat.utils::clUMAP function.}
}
\value{
The modified query Seurat object with the transferred labels as a new identity class.
}
\description{
Function to transfer labels from a reference Seurat object to a query Seurat object using anchoring and transfer data methods from the Seurat package.
It then visualizes the reference and the combined objects using Uniform Manifold Approximation and Projection (UMAP).
}
\details{
Transfer labels from a reference Seurat object to a query Seurat object

This function takes a query Seurat object and a path to a reference Seurat object,
finds transfer anchors, transfers labels, and visualizes the combined object.
}
\examples{
# combined.objX <- transfer_labels_seurat(named_ident = 'RNA_snn_res.0.3.ordered.ManualNames',
#                                     reference_path = '~/Dropbox (VBC)/Abel.IMBA/Metadata.D/CON.meta/label.transfer/sc6/reference.obj.sc6.DIET.2023.07.19_13.24.Rds.gz',
#                                     query_obj = combined.obj)
}
