% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{umapHiLightSel}
\alias{umapHiLightSel}
\title{Highlight Selected Clusters on UMAP}
\usage{
umapHiLightSel(
  obj = combined.obj,
  COI = c("0", "2", "4"),
  ident = GetClusteringRuns()[1],
  h = 7,
  w = 5,
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{obj}{Seurat object to be visualized; Default: \code{combined.obj}.}

\item{COI}{Vector of cluster IDs to highlight on the UMAP plot;
Default: \code{c("0", "2", "4")}.}

\item{ident}{Name of the metadata column containing cluster IDs;
Default: \code{GetClusteringRuns()[1]}.}

\item{h}{Height of the plot; Default: \code{7}.}

\item{w}{Width of the plot; Default: \code{5}.}

\item{show_plot}{Logical; if \code{TRUE}, the plot will be displayed in the RStudio viewer;
Default: \code{TRUE}.}

\item{...}{Additional arguments to be passed to the \code{DimPlot} function.#'}
}
\value{
Saves a UMAP plot highlighting specified clusters to the current working directory.
The function itself does not return an object within R.
}
\description{
Generates a UMAP plot from a Seurat object with specified clusters highlighted.
It saves the resulting UMAP plot directly to the current working directory.
}
\examples{
\dontrun{
if (interactive()) {
  # GetClusteringRuns()[1] "integrated_snn_res.0.1"
  umapHiLightSel(obj = combined.obj, COI = c("0", "1"), ident = GetClusteringRuns()[1])
}
}

}
\seealso{
\code{\link[Seurat]{DimPlot}}
}
