% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.utils.less.used.R
\name{umapNamedClusters}
\alias{umapNamedClusters}
\title{Plot and Save UMAP without legend}
\usage{
umapNamedClusters(
  obj = combined.obj,
  metaD.colname = metaD.colname.labeled,
  ext = "png",
  ...
)
}
\arguments{
\item{obj}{Seurat object to be visualized; Default: \code{combined.obj}.}

\item{metaD.colname}{Metadata column name to color the UMAP by; Default: 'metaD.colname.labeled'.}

\item{ext}{File extension for the saved plot, supports 'png', 'pdf', etc.; Default: 'png'.}

\item{...}{Additional arguments passed to Seurat's \code{DimPlot}.}
}
\value{
Displays a UMAP plot and saves it to the current working directory.
}
\description{
Generates a UMAP plot colored by a specified metadata column and saves the plot to a file.
}
\examples{
\dontrun{
  if (interactive()) {
    umapNamedClusters(obj = combined.obj, metaD.colname = "metaD.colname.labeled")
  }
}

}
