% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{UpdateSeuratObjectProperly}
\alias{UpdateSeuratObjectProperly}
\title{Update Seurat Object Properly, including Assays and DimReducs}
\usage{
UpdateSeuratObjectProperly(obj, update.gene.symbols = TRUE)
}
\arguments{
\item{obj}{A Seurat object to be updated. Default: None.}

\item{update.gene.symbols}{Logical. If TRUE, gene symbols are updated to the latest version.}
}
\value{
An updated Seurat object.
}
\description{
This function is an extension on \code{SeuratObject::UpdateSeuratObject()}. It
first calls \code{UpdateSeuratObject()}, to updates the class definitions of of a (v3) Seurat object,
then it updates its assays to the 'Assay5' class, and updates the UMAP DimReduc to keys.
}
\examples{
\dontrun{
combined.obj <- UpdateSeuratObjectProperly(combined.obj)
}

}
