% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{write_metadata_to_tsv}
\alias{write_metadata_to_tsv}
\title{Write Metadata to a (list of) TSV file(s)}
\usage{
write_metadata_to_tsv(
  obj_list = ls.Seurat.Per.Tracing.GT,
  out_dir = pps(OutDirOrig, "Cell_Meta"),
  combined = FALSE
)
}
\arguments{
\item{obj_list}{A list of Seurat objects from which metadata will be extracted.
Default is \code{ls.Seurat.Per.Tracing.GT}.}

\item{out_dir}{Directory where the TSV and CSV files will be saved.
Default is a "Cell_Meta" directory within the current working directory.}

\item{...}{Additional arguments passed to \code{ReadWriter::write.simple.tsv}.}
}
\value{
Invisibly returns \code{NULL}. The function writes metadata to specified directory.
}
\description{
This function extracts the metadata from a list of Seurat objects and writes
each metadata as an individual TSV file. Additionally, it combines all metadata
and writes them as a single CSV file.
}
\examples{
# write_metadata_to_csv(obj_list = list_of_seurat_objects)

}
