% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{writeCombinedMetadataToTsvFromLsObj}
\alias{writeCombinedMetadataToTsvFromLsObj}
\title{Combine Metadata from a list of Seurat objects and Write to TSV}
\usage{
writeCombinedMetadataToTsvFromLsObj(
  ls.Obj,
  cols.remove = character(),
  save_as_qs = TRUE,
  save_as_tsv = TRUE,
  ...
)
}
\arguments{
\item{ls.Obj}{A list of objects, each containing a \verb{@meta.data} slot.}

\item{cols.remove}{A character vector of column names to be removed from each metadata data frame.
Default is an empty character vector, meaning no columns will be removed.}

\item{save_as_qs}{A logical indicating whether to save the merged metadata as a .qs object.}

\item{save_as_tsv}{A logical indicating whether to save the merged metadata as a .tsv file.}

\item{...}{Additional arguments to be passed to \code{write.table}.}
}
\value{
A large data frame that is the row-wise merge of all \verb{@meta.data} data frames.
}
\description{
Formerly \code{writeMetadataToTsv}. \code{writeCombinedMetadataToTsvFromLsObj} takes a list of ls.Obj, extracts their \verb{@meta.data} slots,
removes specified columns, checks for column consistency, creates a barplot showing the number
of rows per object, and finally merges these into one large data frame.
}
\details{
The function starts by validating the input to ensure it's a list. It then extracts the \verb{@meta.data}
from each object, removing the specified columns. It checks if all data frames have the same columns
and issues a warning if not. A barplot is created to visualize the number of rows (cells) per object.
Finally, it merges all the metadata into one large data frame and prints its dimensions.
}
\note{
This function is intended for use with S4 objects that have a \verb{@meta.data} slot.
}
\examples{
# Assuming a list of Seurat objects with meta.data
mergedMetaData <- writeCombinedMetadataToTsvFromLsObj(
  seuratObjectsList,
  cols.remove = c("column1", "column2")
)

}
