% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{writeMetadataToTsv}
\alias{writeMetadataToTsv}
\title{Combine Metadata and Write to TSV}
\usage{
writeMetadataToTsv(objects, cols.remove = character(), write_out = T, ...)
}
\arguments{
\item{objects}{A list of objects, each containing a \verb{@meta.data} slot.}

\item{cols.remove}{A character vector of column names to be removed from each metadata data frame.
Default is an empty character vector, meaning no columns will be removed.}
}
\value{
A large data frame that is the row-wise merge of all \verb{@meta.data} data frames.
}
\description{
\code{writeMetadataToTsv} takes a list of objects, extracts their \verb{@meta.data} slots,
removes specified columns, checks for column consistency, creates a barplot showing the number
of rows per object, and finally merges these into one large data frame.
}
\details{
The function starts by validating the input to ensure it's a list. It then extracts the \verb{@meta.data}
from each object, removing the specified columns. It checks if all data frames have the same columns
and issues a warning if not. A barplot is created to visualize the number of rows (cells) per object.
Finally, it merges all the metadata into one large data frame and prints its dimensions.
}
\note{
This function is intended for use with S4 objects that have a \verb{@meta.data} slot.
The function currently contains a \code{browser()} call for debugging purposes, which should be removed in production.
}
\examples{
# Assuming a list of Seurat objects with meta.data
mergedMetaData <- writeMetadataToTsv(seuratObjectsList, cols.remove = c("column1", "column2"))

}
