% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{ww.calc_helper}
\alias{ww.calc_helper}
\title{ww.calc_helper}
\usage{
ww.calc_helper(obj, genes)
}
\arguments{
\item{obj}{Seurat object}

\item{genes}{Genes of iinterest}
}
\description{
From Github/Ryan-Zhu https://github.com/satijalab/seurat/issues/371 #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
