% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{ww.check.if.3D.reduction.exist}
\alias{ww.check.if.3D.reduction.exist}
\title{ww.check.if.3D.reduction.exist}
\usage{
ww.check.if.3D.reduction.exist(obj = obj)
}
\arguments{
\item{obj}{Seurat object, Default: obj}
}
\description{
ww.check.if.3D.reduction.exist in backup slot #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
