% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{ww.get.1st.Seur.element}
\alias{ww.get.1st.Seur.element}
\title{ww.get.1st.Seur.element}
\usage{
ww.get.1st.Seur.element(obj)
}
\arguments{
\item{obj}{Seurat object OR list of Seurat objects OR any other list}
}
\description{
Get the first Seurat object from a list of Seurat objects. Returns itself if not a list.
}
