% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{xread}
\alias{xread}
\title{Read an R Object Using 'qs2' Package for Fast Decompression}
\usage{
xread(
  file,
  nthreads = if (file.size(file) < 1e+07) 1 else 4,
  loadParamsAndAllGenes = TRUE,
  overwriteParams = FALSE,
  overwriteAllGenes = FALSE,
  set_m = TRUE,
  ...
)
}
\arguments{
\item{file}{A character string specifying the path to the file where the R object is saved.}

\item{nthreads}{The number of threads to use when reading the object, defaults to 4.}

\item{loadParamsAndAllGenes}{Logical; if TRUE and if the object is a Seurat object, additional parameters
are loaded from within it.}

\item{overwriteParams}{Logical; if TRUE and if the object is a Seurat object, the parameters are overwritten.}

\item{overwriteAllGenes}{Logical; if TRUE and if the object is a Seurat object, the all genes are overwritten.}

\item{set_m}{Logical; if TRUE, the variable 'm', a list of @meta.data colnames, is assigned to
the global environment.}

\item{...}{Further arguments passed on to the 'qs2::qs_read' function.}
}
\value{
The R object that was saved in the specified file.
}
\description{
This function reads an R object from a file saved in a format specific to the 'qs2' package,
which is designed for quick and efficient compression and decompression of R objects.
It also times the read operation, providing feedback on the duration of the operation.
}
\note{
The function uses the 'qs2' package for fast and efficient deserialization of objects
and includes a timing feature from the 'tictoc' package.
}
\seealso{
\code{\link[qs2]{qs_read}} for the underlying read function used.
}
