% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{add.meta.tags}
\alias{add.meta.tags}
\title{add.meta.tags}
\usage{
add.meta.tags(list.of.tags = tags, obj = ls.Seurat[[1]], n = 1)
}
\arguments{
\item{list.of.tags}{PARAM_DESCRIPTION, Default: tags}

\item{obj}{Seurat object, Default: ls.Seurat[\link{1}]}

\item{n}{PARAM_DESCRIPTION, Default: 1}
}
\description{
N is the for which dataset #
}
\examples{
\dontrun{
if(interactive()){
 ls.Seurat[[1]] <- add.meta.tags(list.of.tags = tags, obj = ls.Seurat[[1]], n = 1)
 }
}
}
