% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{addMetaFraction}
\alias{addMetaFraction}
\title{addMetaFraction}
\usage{
addMetaFraction(
  obj,
  col.name = "percent.mito",
  gene.symbol.pattern = c("^MT\\\\.|^MT-", FALSE)[1],
  assay = "RNA",
  layer = "counts",
  gene.set = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{Seurat object to which the new metadata column will be added. Default: ls.Seurat[\link{1}]}

\item{col.name}{Name of the new metadata column to be added. Default: 'percent.mito'}

\item{gene.symbol.pattern}{Regular expression pattern to match gene symbols. Default: c("^MT\\.|^MT-", FALSE)\link{1}}

\item{assay}{Name of the assay to be used. Default: 'RNA'}

\item{layer}{Name of the layer to be used. Default: 'data'}

\item{gene.set}{A set of gene symbols. If specified, it will be used instead of gene.symbol.pattern. Default: \code{FALSE}.}

\item{verbose}{Logical indicating whether to display detailed messages (TRUE) or not (FALSE). Default: \code{TRUE}.}
}
\description{
Add a new metadata column to a Seurat object, representing the fraction of a gene set in the transcriptome (expressed as a percentage).
}
\examples{
\dontrun{
if (interactive()) {
  ls.Seurat[[1]] <- addMetaFraction(col.name = "percent.mito", gene.symbol.pattern = "^MT\\\\.|^MT-", obj = ls.Seurat[[1]])
  ls.Seurat[[1]] <- addMetaFraction(col.name = "percent.ribo", gene.symbol.pattern = "^RPL|^RPS", obj = ls.Seurat[[1]])
  ls.Seurat[[1]] <- addMetaFraction(col.name = "percent.AC.GenBank", gene.symbol.pattern = "^AC[0-9]{6}\\\\.", obj = ls.Seurat[[1]])
  ls.Seurat[[1]] <- addMetaFraction(col.name = "percent.AL.EMBL", gene.symbol.pattern = "^AL[0-9]{6}\\\\.", obj = ls.Seurat[[1]])
  ls.Seurat[[1]] <- addMetaFraction(col.name = "percent.LINC", gene.symbol.pattern = "^LINC0", obj = ls.Seurat[[1]])
  ls.Seurat[[1]] <- addMetaFraction(col.name = "percent.MALAT1", gene.symbol.pattern = "^MALAT1", obj = ls.Seurat[[1]])
  colnames(ls.Seurat[[1]]@meta.data)
  HGA_MarkerGenes <- c(
    "ENO1", "IGFBP2", "WSB1", "DDIT4", "PGK1", "BNIP3", "FAM162A", "TPI1",
    "VEGFA", "PDK1", "PGAM1", "IER2", "FOS", "BTG1", "EPB41L4A-AS1", "NPAS4", "HK2", "BNIP3L",
    "JUN", "ENO2", "GAPDH", "ANKRD37", "ALDOA", "GADD45G", "TXNIP"
  )
  sobj <- addMetaFraction(col.name = "percent.HGA", gene.set = HGA_MarkerGenes, obj = sobj)
}
}
}
\seealso{
\code{\link[Matrix]{colSums}}
}
