% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{AddNewAnnotation}
\alias{AddNewAnnotation}
\title{AddNewAnnotation}
\usage{
AddNewAnnotation(
  obj = obj,
  source = "RNA_snn_res.0.5",
  named.list.of.identities = ls.Subset.ClusterLists
)
}
\arguments{
\item{obj}{Seurat object, Default: obj}

\item{source}{PARAM_DESCRIPTION, Default: 'RNA_snn_res.0.5'}

\item{named.list.of.identities}{PARAM_DESCRIPTION, Default: ls.Subset.ClusterLists}
}
\description{
Create a new metadata column based on an exisiting metadata column and a list of mappings (name <- IDs). #
}
\examples{
\dontrun{
if(interactive()){
 ls.Subset.ClusterLists = list( "hESC.h9" = c("4", "10", "14"), "hESC.176" = c("0", "1", "2")); AddNewAnnotation()
 }
}
}
