% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{addToMiscOrToolsSlot}
\alias{addToMiscOrToolsSlot}
\title{Create Misc or Tools Slot in Seurat Object}
\usage{
addToMiscOrToolsSlot(
  obj,
  slot_name = "misc",
  sub_slot_value = NULL,
  sub_slot_name = substitute(sub_slot_value),
  overwrite = FALSE
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{slot_name}{Name of the primary slot (\verb{@misc} or \verb{@tools}).}

\item{sub_slot_value}{what value to add?}

\item{sub_slot_name}{Name of the new element inside the primary slot.}

\item{overwrite}{overwrite?}
}
\description{
Adds a new sub-slot to either the \verb{@misc} or \verb{@tools}  slots of a Seurat object.
}
