% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{addTranslatedMetadata}
\alias{addTranslatedMetadata}
\title{Add Translated Metadata to a Seurat Object}
\usage{
addTranslatedMetadata(
  obj = combined.obj,
  orig.ident = "RNA_snn_res.0.4",
  translation_as_named_vec,
  new_col_name = substitute_deparse(translation_as_named_vec),
  suffix = NULL,
  plot = FALSE,
  ...
)
}
\arguments{
\item{obj}{A Seurat object to be updated. Default: combined.obj.}

\item{orig.ident}{A character string specifying the original metadata to be translated. Default: "RNA_snn_res.0.4".}

\item{translation_as_named_vec}{A named vector where names are old values and values are new translations. Default: None.}

\item{new_col_name}{A character string specifying the base name of the new metadata column.}

\item{suffix}{A character string appended to \code{new_col_name} to form the final metadata column name. Default: \code{NULL}.}

\item{plot}{A logical indicating whether to plot the UMAP for the new metadata. Default: FALSE.}

\item{...}{Additional arguments, not used.}
}
\value{
An updated Seurat object.
}
\description{
This function translates a specified metadata vector in a Seurat object using a named vector of old
and new values and adds it to the Seurat object. A suffix can optionally be appended to the name of
the newly created metadata column. The function also generates UMAP plots for the new metadata.
}
