% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{AreTheseCellNamesTheSame}
\alias{AreTheseCellNamesTheSame}
\title{AreTheseCellNamesTheSame}
\usage{
AreTheseCellNamesTheSame(
  vec1 = names(UVI.annot),
  vec2 = names(nr_UVI),
  names = c("Cells in Targ.Ampl", "Cells in GEX"),
  min.overlap = 0.33
)
}
\arguments{
\item{vec1}{Character vector, eg. with cell names}

\item{vec2}{Character vector, eg. with cell names}

\item{names}{Names for plotting}

\item{min.overlap}{Threshold below there is no there is no meaningful overlap between the two vectors.}
}
\description{
Assert and compare two character vectors (e.g.: cell IDs) how much they overlap and
plot a Venn Diagram. The function aborts with an error if overlap is too small.
}
\examples{
# reTheseCellNamesTheSame()
}
