% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{AutoLabel.KnownMarkers}
\alias{AutoLabel.KnownMarkers}
\title{AutoLabel.KnownMarkers}
\usage{
AutoLabel.KnownMarkers(
  obj = combined.obj,
  topN = 1,
  res = 0.5,
  KnownMarkers = c("TOP2A", "EOMES", "SLA", "HOPX", "S100B", "DLX6-AS1", "POU5F1",
    "SALL4", "DDIT4", "PDK1", "SATB2", "FEZF2"),
  order.by = c("combined.score", "avg_log2FC", "p_val_adj")[1],
  df_markers = obj@misc$df.markers[[paste0("res.", res)]]
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{topN}{PARAM_DESCRIPTION, Default: 1}

\item{res}{Clustering resoluton to use, Default: 0.5}

\item{KnownMarkers}{PARAM_DESCRIPTION, Default: c("TOP2A", "EOMES", "SLA", "HOPX", "S100B", "DLX6-AS1", "POU5F1",
"SALL4", "DDIT4", "PDK1", "SATB2", "FEZF2")}

\item{order.by}{Sort output tibble by which column, Default: c("combined.score", "avg_log2FC", "p_val_adj")\link{1}}

\item{df_markers}{PARAM_DESCRIPTION, Default: combined.obj@misc$df.markers[\link{paste0("res.", res)}]}
}
\description{
Create a new "named identity" column in the metadata of a Seurat object, with \code{Ident} set to a clustering output matching the \code{res} parameter of the function. It requires the output table of \code{FindAllMarkers()}. If you used \code{StoreAllMarkers()} is stored under \verb{@misc$df.markers$res...}, which location is assumed by default. #
}
\examples{
\dontrun{
if(interactive()){
 combined.obj <- AutoLabel.KnownMarkers(); # combined.obj$cl.names.KnownMarkers.0.5
 DimPlot.ClusterNames(ident = "cl.names.KnownMarkers.0.5")
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}
}
