% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{AutoLabelTop.logFC}
\alias{AutoLabelTop.logFC}
\title{AutoLabelTop.logFC}
\usage{
AutoLabelTop.logFC(
  obj = combined.obj,
  ident = GetClusteringRuns()[1],
  res = 0.2,
  plot.top.genes = T,
  order.by = c("combined.score", "avg_log2FC", "p_val_adj")[2],
  df_markers = obj@misc$df.markers[[paste0("res.", res)]]
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{res}{Clustering resoluton to use, Default: 0.2}

\item{plot.top.genes}{PARAM_DESCRIPTION, Default: T}

\item{order.by}{Sort output tibble by which column, Default: c("combined.score", "avg_logFC", "p_val_adj")\link{1}}

\item{df_markers}{PARAM_DESCRIPTION, Default: combined.obj@misc$df.markers[\link{paste0("res.", res)}]}
}
\description{
Create a new "named identity" column in the metadata of a Seurat object, with \code{Ident} set to a clustering output matching the \code{res} parameter of the function. It requires the output table of \code{FindAllMarkers()}. If you used \code{StoreAllMarkers()} is stored under \verb{@misc$df.markers$res...}, which location is assumed by default. #
}
\examples{
\dontrun{
if(interactive()){
 combined.obj <- AutoLabelTop.logFC(); combined.obj$"cl.names.top.gene.res.0.5"
 }
}
}
