% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.utils.less.used.R
\name{AutoNumber.by.PrinCurve}
\alias{AutoNumber.by.PrinCurve}
\title{AutoNumber.by.PrinCurve}
\usage{
AutoNumber.by.PrinCurve(
  obj = combined.obj,
  dim = 1:2,
  plotit = TRUE,
  swap = -1,
  reduction = "umap",
  res = "integrated_snn_res.0.5"
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{dim}{Dimensions to use, Default: 1:2}

\item{plotit}{Plot results (& show it), Default: \code{TRUE}.}

\item{swap}{Swap Lambda parameter (multiplied with this) , Default: -1}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'umap'}

\item{res}{Clustering resolution to use, Default: 'integrated_snn_res.0.5'}
}
\description{
Relabel cluster numbers along the principal curve of 2 UMAP (or tSNE) dimensions. #
}
\examples{
\dontrun{
if (interactive()) {
  DimPlot.ClusterNames(ident = "integrated_snn_res.0.5")
  combined.obj <- AutoNumber.by.PrinCurve(
    obj = combined.obj, dim = 1:2, reduction = "umap", plotit = TRUE,
    swap = -1, res = "integrated_snn_res.0.5"
  )
  DimPlot.ClusterNames(ident = "integrated_snn_res.0.5.prin.curve")
}
}
}
\seealso{
\code{\link[princurve]{principal_curve}}
}
