% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{AutoNumber.by.UMAP}
\alias{AutoNumber.by.UMAP}
\title{AutoNumber.by.UMAP}
\usage{
AutoNumber.by.UMAP(
  obj = combined.obj,
  dim = 1,
  swap = F,
  reduction = "umap",
  res = "RNA_snn_res.0.5"
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{dim}{Which dimension? Default: 1}

\item{swap}{Swap direction? Default: F}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'umap'}

\item{res}{Clustering resoluton to use, Default: 'integrated_snn_res.0.5'}
}
\description{
Relabel cluster numbers along a UMAP (or tSNE) axis #
}
\examples{
\dontrun{
if (interactive()) {
  combined.obj <- AutoNumber.by.UMAP(obj = combined.obj, dim = 1, reduction = "umap", res = "integrated_snn_res.0.5")
  DimPlot.ClusterNames(ident = "integrated_snn_res.0.5.ordered")
}
}
}
