% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{AutoNumber.by.UMAP}
\alias{AutoNumber.by.UMAP}
\title{Relabel Cluster Numbers Along a UMAP (or tSNE) Axis}
\usage{
AutoNumber.by.UMAP(
  obj = combined.obj,
  reduction = "umap",
  dim = 1,
  swap = FALSE,
  ident = GetClusteringRuns(obj = obj)[1],
  plot = TRUE,
  obj.version = obj@version
)
}
\arguments{
\item{obj}{Seurat object containing clustering and UMAP (or other dimensional reduction) data;
Default: \code{combined.obj}.}

\item{reduction}{Dimension reduction technique used for cluster positioning ('umap', 'tsne', or 'pca');
Default: 'umap'.}

\item{dim}{Dimension along which to order clusters (1 for the first dimension, typically horizontal);
Default: 1.}

\item{swap}{If TRUE, reverses the ordering direction; Default: \code{FALSE}.}

\item{ident}{Clustering resolution identifier used to fetch cluster labels from \code{obj} metadata;
Default: 'integrated_snn_res.0.5'.}

\item{plot}{If TRUE, plots the UMAP with new cluster names; Default: \code{TRUE}.}
}
\description{
Automatically renumbers clusters based on their position along a specified dimension
in a UMAP (or tSNE or PCA) plot, potentially enhancing interpretability by ordering clusters.
}
\examples{
\dontrun{
combined.obj <- AutoNumber.by.UMAP(
  obj = combined.obj, dim = 1, reduction = "umap",
  ident = "integrated_snn_res.0.5"
)
DimPlot.ClusterNames(combined.obj, ident = GetClusteringRuns(combined.obj)[1])
}

}
