% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MULTI-seq.Utils.R
\name{aux_plotAllMseqBCs}
\alias{aux_plotAllMseqBCs}
\title{aux_plotAllMseqBCs}
\usage{
aux_plotAllMseqBCs(
  bar.table = bar.table[, 1:96],
  barcodes.used = BCs.used,
  plotname = "Barcode seq depth"
)
}
\arguments{
\item{bar.table}{The barcode table to be used. Default: bar.table\link{,1:96}.}

\item{barcodes.used}{The barcodes that have been used. Default: BCs.used.}

\item{plotname}{The title of the plot. Default: 'Barcode seq depth'.}
}
\value{
A plot of all MULTI-seq barcodes.
}
\description{
Plot all MULTI-seq barcodes.
}
\examples{
\dontrun{
if(interactive()){
 aux_plotAllMseqBCs(bar.table = bar.table[,1:96], barcodes.used = BCs.used, plotname = "Barcode seq depth")
 }
}
}
