% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{BackupReduction}
\alias{BackupReduction}
\title{Backup Dimensionality Reduction Data}
\usage{
BackupReduction(obj = combined.obj, dim = 2, reduction = "umap")
}
\arguments{
\item{obj}{Seurat object containing dimensionality reduction data; Default: \code{combined.obj}.}

\item{dim}{Number of dimensions to include in the backup; Default: 2.}

\item{reduction}{Type of dimensionality reduction to backup ('umap', 'tsne', 'pca'); Default: 'umap'.}
}
\description{
This function is mostly used internally.It stores a backup of specified
dimensionality reduction data (e.g., UMAP, tSNE, PCA)
within the Seurat object, from \code{obj@reductions$umap} to the \verb{@misc$reductions.backup} slot. This
allows to store 2D and 3D UMAP visualizations in parallel and easily switch between them via
the \code{RecallReduction} function.
}
\examples{
\dontrun{
if (interactive()) {
  obj <- BackupReduction(obj = obj, dim = 2, reduction = "umap")
}
}

}
