% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{barplot.cells.per.cluster}
\alias{barplot.cells.per.cluster}
\title{barplot.cells.per.cluster}
\usage{
\method{barplot}{cells.per.cluster}(obj = combined.obj, ident = "cl.names.KnownMarkers.0.5", sort = F)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{ident}{identity used, Default: 'cl.names.KnownMarkers.0.5'}

\item{sort}{PARAM_DESCRIPTION, Default: F}
}
\description{
Barplot the Fraction of cells per cluster. (dupl?)
}
\examples{
\dontrun{
if(interactive()){
 barplot.cells.per.cluster(); barplot.cells.per.cluster(sort = T)
 }
}
}
