% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{BarTableSweepList}
\alias{BarTableSweepList}
\title{BarTableSweepList}
\usage{
BarTableSweepList(min = 0.01, max = 0.99, step = 0.02, bar_table = bar.table)
}
\arguments{
\item{min}{PARAM_DESCRIPTION, Default: 0.01}

\item{max}{PARAM_DESCRIPTION, Default: 0.99}

\item{step}{PARAM_DESCRIPTION, Default: 0.02}

\item{bar_table}{PARAM_DESCRIPTION, Default: bar.table}
}
\description{
BarTableSweepList
}
\examples{
\dontrun{
if(interactive()){
 bar.table_sweep.list <- BarTableSweepList(bar_table = bar.table.solo) # Quantile Sweep List
 }
}
}
