% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{BulkGEScatterPlot}
\alias{BulkGEScatterPlot}
\title{BulkGEScatterPlot}
\usage{
BulkGEScatterPlot(
  obj = combined.obj,
  clusters = "cl.names.KnownMarkers.0.2",
  TwoCategIdent = "age",
  genes.from.bulk.DE = rownames(df.markers.per.AGE)
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{clusters}{PARAM_DESCRIPTION, Default: 'cl.names.KnownMarkers.0.2'}

\item{TwoCategIdent}{PARAM_DESCRIPTION, Default: 'age'}

\item{genes.from.bulk.DE}{PARAM_DESCRIPTION, Default: rownames(df.markers.per.AGE)}
}
\description{
Plot bulk scatterplots to identify differential expressed genes across conditions #
}
\examples{
\dontrun{
if(interactive()){
 BulkGEScatterPlot(obj = combined.obj, clusters = "cl.names.KnownMarkers.0.2", TwoCategIdent = 'age', genes.from.bulk.DE = rownames(df.markers.per.AGE))
 }
}
}
