% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{calc.cluster.averages}
\alias{calc.cluster.averages}
\title{calc.cluster.averages}
\usage{
calc.cluster.averages(
  col_name = "Score.GO.0006096",
  plot.UMAP.too = TRUE,
  return.plot = F,
  obj = combined.obj,
  split_by = GetNamedClusteringRuns()[1],
  scale.zscore = FALSE,
  simplify = T,
  plotit = T,
  histogram = FALSE,
  nbins = 50,
  suffix = NULL,
  stat = c("mean", "median")[2],
  quantile.thr = 0.9,
  absolute.thr = FALSE,
  filter = c(FALSE, "above", "below")[1],
  ylab.text = paste("Cluster", stat, "score"),
  title = paste("Cluster", stat, col_name),
  prefix.cl.names = FALSE,
  report = TRUE,
  subtitle = NULL,
  width = 8,
  height = 6,
  ...,
  xlb = if (absolute.thr) paste("Threshold at", absolute.thr) else
    paste("Black lines: ", kppd(Stringendo::percentage_formatter(c(1 - quantile.thr,
    quantile.thr))), "quantiles |", "Cl. >",
    Stringendo::percentage_formatter(quantile.thr), "are highlighted. |", split_by),
  fname = ppp(col_name, split_by, "cluster.average.barplot.pdf", ...)
)
}
\arguments{
\item{col_name}{PARAM_DESCRIPTION, Default: 'Score.GO.0006096'}

\item{plot.UMAP.too}{PARAM_DESCRIPTION, Default: TRUE}

\item{return.plot}{PARAM_DESCRIPTION, Default: F}

\item{obj}{Seurat object, Default: combined.obj}

\item{split_by}{PARAM_DESCRIPTION, Default: GetNamedClusteringRuns()\link{1}}

\item{scale.zscore}{PARAM_DESCRIPTION, Default: FALSE}

\item{simplify}{PARAM_DESCRIPTION, Default: T}

\item{plotit}{Plot results (& show it), Default: T}

\item{histogram}{PARAM_DESCRIPTION, Default: FALSE}

\item{nbins}{PARAM_DESCRIPTION, Default: 50}

\item{suffix}{A suffix added to the filename, Default: NULL}

\item{stat}{PARAM_DESCRIPTION, Default: c("mean", "median")\link{2}}

\item{quantile.thr}{PARAM_DESCRIPTION, Default: 0.9}

\item{absolute.thr}{PARAM_DESCRIPTION, Default: FALSE}

\item{filter}{PARAM_DESCRIPTION, Default: c(FALSE, "above", "below")\link{1}}

\item{ylab.text}{PARAM_DESCRIPTION, Default: paste("Cluster", stat, "score")}

\item{title}{Title of the plot, Default: paste("Cluster", stat, col_name)}

\item{report}{Summary sentence, Default: F}

\item{subtitle}{PARAM_DESCRIPTION, Default: NULL}

\item{width}{PARAM_DESCRIPTION, Default: 8}

\item{height}{PARAM_DESCRIPTION, Default: 6}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}

\item{xlb}{PARAM_DESCRIPTION, Default: if (absolute.thr) paste("Threshold at", absolute.thr) else paste("Black lines: ",
kppd(Stringendo::percentage_formatter(c(1 - quantile.thr, quantile.thr))),
"quantiles |", "Cl. >", Stringendo::percentage_formatter(quantile.thr),
"are highlighted. |", split_by)}

\item{fname}{File name, Default: ppp(col_name, split_by, "cluster.average.barplot.pdf", ...)}
}
\description{
Calculate the average of a metadata column (numeric) per cluster.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
