% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Calc.Cor.Seurat}
\alias{Calc.Cor.Seurat}
\title{Calc.Cor.Seurat}
\usage{
Calc.Cor.Seurat(
  assay.use = "RNA",
  slot.use = "data",
  quantileX = 0.95,
  max.cells = 40000,
  seed = p$seed,
  digits = 2,
  obj = combined.obj
)
}
\arguments{
\item{assay.use}{The assay to use from the Seurat object. Default: 'RNA'}

\item{slot.use}{The slot to use from the assay in the Seurat object. Default: 'data'}

\item{quantileX}{The quantile level for the calculation. Default: \code{0.95}}

\item{max.cells}{Maximum number of cells to be used in the calculation. Default: \code{40000}}

\item{seed}{The random seed used for the calculation. Default: \code{p$seed}}

\item{digits}{The number of decimal places to round the correlation and covariance values. Default: \code{2}}

\item{obj}{The Seurat object to perform calculations on. Default: \code{combined.obj}}
}
\description{
Calculate gene correlation on a Seurat object.
}
\examples{
\dontrun{
if (interactive()) {
  combined.obj <- calc.q99.Expression.and.set.all.genes(combined.obj, quantileX = 0.99, max.cells = 400000, set.all.genes = FALSE)
  combined.obj <- Calc.Cor.Seurat(assay.use = "RNA", slot.use = "data", digits = 2, obj = combined.obj, quantile = 0.99, max.cells = 40000)
}
}
}
