% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{calc.q99.Expression.and.set.all.genes}
\alias{calc.q99.Expression.and.set.all.genes}
\title{calc.q99.Expression.and.set.all.genes}
\usage{
calc.q99.Expression.and.set.all.genes(
  obj = combined.obj,
  quantileX = 0.99,
  max.cells = 1e+05,
  slot = "data",
  assay = c("RNA", "integrated", "SCT")[1],
  set.misc = TRUE,
  assign_to_global_env = TRUE,
  plot = TRUE,
  suffix = substitute(obj),
  show = TRUE,
  obj.version = obj@version,
  ...
)
}
\arguments{
\item{obj}{Seurat object, Default: \code{combined.obj}}

\item{quantileX}{Quantile level, Default: 0.9}

\item{max.cells}{Max number of cells to do the calculation on. Downsample if excdeeded. Default: 1e+05}

\item{slot}{slot in the Seurat object. Default: 'data'}

\item{assay}{RNA or integrated assay, Default: \code{c("RNA", "integrated")[1]}}

\item{set.misc}{Create the "all.genes" variable in @misc? Default: \code{TRUE}.}

\item{assign_to_global_env}{Create the "all.genes" variable in the global env?, Default: \code{TRUE}.}

\item{plot}{Plot the expression distribution? Default: \code{TRUE}.}

\item{suffix}{Suffix appended to the object name used e.g. in messages. Default: \code{substitute(obj)}.}

\item{show}{Show the distribution plot? Default: \code{TRUE}.}

\item{obj.version}{Manuallyoverride the Version of the Seurat object. Useful when you used the
problematic \code{SeuratObject::UpdateSeuratObject()}.Default: no override \code{obj@version}.}

\item{...}{Additional arguments, not used.}
}
\description{
Calculate the gene expression of the e.g.: 99th quantile (expression in the top 1\% cells).
}
\examples{
\dontrun{
if (interactive()) {
  combined.obj <- calc.q99.Expression.and.set.all.genes(
    obj = combined.obj, quantileX = 0.9,
    max.cells = 25000
  )
  head(sort(as.numeric.wNames(obj@misc$expr.q90), decreasing = TRUE))
  combined.obj <- calc.q99.Expression.and.set.all.genes(
    obj = combined.obj, quantileX = 0.95,
    max.cells = 25000, set.all.genes = FALSE
  )
}
}
}
\seealso{
\code{sparseMatrixStats::rowQuantiles()}
}
