% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{calc.q99.Expression.and.set.all.genes}
\alias{calc.q99.Expression.and.set.all.genes}
\title{calc.q99.Expression.and.set.all.genes}
\usage{
calc.q99.Expression.and.set.all.genes(
  obj = combined.obj,
  quantileX = 0.99,
  max.cells = 1e+05,
  slot = "data",
  assay = c("RNA", "integrated")[1],
  set.all.genes = TRUE,
  set.misc = TRUE,
  assign_to_global_env = TRUE,
  show = TRUE
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{quantileX}{Quantile level, Default: 0.9}

\item{max.cells}{Max number of cells to do the calculation on. Downsample if excdeeded. Default: 1e+05}

\item{slot}{slot in the Seurat object. Default: 'data'}

\item{assay}{RNA or integrated assay, Default: c("RNA", "integrated")\link{1}}

\item{set.misc}{Create the "all.genes" variable in @misc? Default: TRUE}

\item{assign_to_global_env}{Create the "all.genes" variable in the global env?, Default: TRUE}

\item{show}{Show plot? Default: TRUE}
}
\description{
Calculate the gene expression of the e.g.: 90th quantile (expression in the top 10\% cells). #
}
\examples{
\dontrun{
if(interactive()){
 combined.obj <- calc.q99.Expression.and.set.all.genes(obj = combined.obj, quantileX = 0.9,
 max.cells =  25000)
 head(sort(as.numeric.wNames(obj@misc$expr.q90), decreasing = T))
 combined.obj <- calc.q99.Expression.and.set.all.genes(obj = combined.obj, quantileX = 0.95,
 max.cells =  25000, set.all.genes = FALSE)
 }
}
}
\seealso{
\code{\link[sparseMatrixStats]{character(0)}}
}
