% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{calculate.observable.multiplet.rate.10X.LT}
\alias{calculate.observable.multiplet.rate.10X.LT}
\title{calculate.observable.multiplet.rate.10X.LT}
\usage{
calculate.observable.multiplet.rate.10X.LT(
  fractions = c(`176` = 0.07, h9 = 0.93),
  multiplet.rate = 0.12
)
}
\arguments{
\item{fractions}{Genotype fractions observed or loaded on the machine.}

\item{multiplet.rate}{Expected multiplet rate for your number of recovered cells.}
}
\description{
calculate.observable.multiplet.rate.10X.LT
}
