% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{CalculateFractionInTrome}
\alias{CalculateFractionInTrome}
\title{CalculateFractionInTrome}
\usage{
CalculateFractionInTrome(
  genesCalc.Cor.Seuratet = c("MALAT1"),
  obj = combined.obj,
  dataslot = c("counts", "data")[2]
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{dataslot}{PARAM_DESCRIPTION, Default: c("counts", "data")\link{2}}

\item{geneset}{PARAM_DESCRIPTION, Default: c("MALAT1")}
}
\description{
Calculate the fraction of a set of genes within the full Transcriptome of each cell. #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
