% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slingshot.Utils.R
\name{cell_pal}
\alias{cell_pal}
\title{cell_pal}
\usage{
cell_pal(cell_vars, pal_fun, ...)
}
\arguments{
\item{cell_vars}{PARAM_DESCRIPTION}

\item{pal_fun}{PARAM_DESCRIPTION}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Generate palette
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
