% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{check.genes}
\alias{check.genes}
\title{check.genes}
\usage{
check.genes(
  list.of.genes = ClassicMarkers,
  makeuppercase = FALSE,
  verbose = TRUE,
  HGNC.lookup = FALSE,
  obj = combined.obj,
  assay.slot = c("RNA", "integrated")[1],
  dataslot = c("counts", "data")[2]
)
}
\arguments{
\item{list.of.genes}{A vector of gene names to check for existence in the Seurat object or HGNC. Default: ClassicMarkers}

\item{makeuppercase}{Logical, if TRUE, transforms all gene names to uppercase. Default: FALSE}

\item{verbose}{Logical, if TRUE, prints out information about missing genes. Default: TRUE}

\item{HGNC.lookup}{Logical, if TRUE, looks up missing genes in HGNC database. Default: FALSE}

\item{obj}{Seurat object that the function uses to check the gene names. Default: combined.obj}

\item{assay.slot}{The assay slot of the Seurat object to consider. Default: 'RNA'}

\item{dataslot}{The data slot of the Seurat object to consider. Default: 'data'}
}
\description{
Check if a gene name exists in a Seurat object, or in HGNC?
}
\examples{
\dontrun{
if (interactive()) {
  check.genes("top2a", makeuppercase = TRUE)
  check.genes("VGLUT2", verbose = F, HGNC.lookup = T)
}
}
}
