% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{check.genes}
\alias{check.genes}
\title{check.genes}
\usage{
check.genes(
  list.of.genes = ClassicMarkers,
  makeuppercase = FALSE,
  verbose = TRUE,
  HGNC.lookup = FALSE,
  obj = combined.obj,
  assay.slot = c("RNA", "integrated")[1],
  dataslot = c("counts", "data")[2]
)
}
\arguments{
\item{list.of.genes}{PARAM_DESCRIPTION, Default: ClassicMarkers}

\item{makeuppercase}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{HGNC.lookup}{PARAM_DESCRIPTION, Default: FALSE}

\item{obj}{Seurat object, Default: combined.obj}

\item{assay.slot}{PARAM_DESCRIPTION, Default: c("RNA", "integrated")\link{1}}

\item{dataslot}{PARAM_DESCRIPTION, Default: c("counts", "data")\link{2}}
}
\description{
check.genes
}
\examples{
\dontrun{
if(interactive()){
 check.genes("top2a", makeuppercase = TRUE); check.genes("VGLUT2", verbose = F, HGNC.lookup = T)
 }
}
}
