% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{clip10Xcellname}
\alias{clip10Xcellname}
\title{Clip Suffixes from 10X Cell Names}
\usage{
clip10Xcellname(cellnames)
}
\arguments{
\item{cellnames}{A vector of cell names with potential suffixes.}
}
\value{
A vector of cell names with suffixes removed.
}
\description{
Removes suffixes from cell names that are added by 10X technology and Seurat during data processing.
}
\examples{
cellnames <- c("cell1_1", "cell2_2")
clip10Xcellname(cellnames)
}
