% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{clUMAP}
\alias{clUMAP}
\title{clUMAP - Quick Visualization of Clustering Results with UMAP and automatically save the plot}
\usage{
clUMAP(
  ident = NULL,
  obj = combined.obj,
  title = ident,
  sub = NULL,
  prefix = NULL,
  suffix = make.names(sub),
  caption = .parseBasicObjStats(obj, simple = TRUE),
  reduction = "umap",
  splitby = NULL,
  label.cex = 7,
  h = 7,
  w = NULL,
  nr.cols = NULL,
  plotname = Stringendo::ppp(toupper(reduction), ident),
  cols = NULL,
  palette = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[4],
  max.cols.for.std.palette = 7,
  highlight.clusters = NULL,
  cells.highlight = NULL,
  cols.highlight = "red",
  sizes.highlight = 1,
  label = TRUE,
  repel = TRUE,
  legend = !label,
  legend.pos = NULL,
  MaxCategThrHP = 200,
  axes = NULL,
  aspect.ratio = c(FALSE, 0.6)[2],
  save.plot = getOption("su.save.plot", TRUE),
  PNG = TRUE,
  check_for_2D = TRUE,
  ...
)
}
\arguments{
\item{ident}{Cluster identity for visualization; Default: 'integrated_snn_res.0.5'.}

\item{obj}{Seurat object containing single-cell data; Default: \code{combined.obj}.}

\item{title}{Main title of the plot; Default: \code{ident}.}

\item{sub}{Subtitle of the plot; optional; Default: NULL.}

\item{prefix}{Prefix for saved filename; optional; Default: NULL.}

\item{suffix}{Suffix for saved filename; defaults to plot subtitle; Default: NULL.}

\item{caption}{Plot caption; optional; Default: dynamically generated from \code{obj}.}

\item{reduction}{Dimension reduction method ('umap', 'tsne', 'pca'); Default: 'umap'.}

\item{splitby}{Metadata column to split cells by; optional; Default: NULL.}

\item{label.cex}{Size of cluster labels; Default: 7.}

\item{h}{Height of plot in inches; Default: 7.}

\item{w}{Width of plot in inches; optional; Default: NULL.}

\item{nr.cols}{Number of columns for facet wrap if \code{splitby} is not NULL; Default: NULL.}

\item{plotname}{Custom plot name for saving; Default: dynamically generated from \code{reduction} and \code{ident}.}

\item{cols}{Custom color vector for clusters; optional; Default: NULL.}

\item{palette}{Color palette for generating cluster colors; Default: 'glasbey'.}

\item{max.cols.for.std.palette}{Maximum number of categories before switching from the standard
\code{palette} to a large, auto-generated color palette; Default: 7.}

\item{highlight.clusters}{Specific clusters to be highlighted; optional; Default: NULL.}

\item{cells.highlight}{Specific cells to be highlighted; optional; Default: NULL.}

\item{cols.highlight}{Color for highlighted cells; Default: 'red'.}

\item{sizes.highlight}{Size of highlighted cells; Default: 1.}

\item{label}{Show cluster labels; Default: \code{TRUE}.}

\item{repel}{Repel labels to avoid overlap; Default: \code{TRUE}.}

\item{legend}{Show legend; Default: opposite of \code{label}.}

\item{legend.pos}{Position of legend; Default: 'NULL'.}

\item{MaxCategThrHP}{Maximum number of categories before simplification; Default: 200.}

\item{axes}{Show axes; Default: \code{FALSE}.}

\item{aspect.ratio}{Fixed aspect ratio for the plot; Default: \code{TRUE}.}

\item{save.plot}{Save plot to file; Default: \code{TRUE}.}

\item{PNG}{Save as PNG (TRUE) or PDF (FALSE); Default: \code{TRUE}.}

\item{check_for_2D}{Ensure UMAP is 2D; Default: \code{TRUE}.}

\item{...}{Additional parameters for \code{DimPlot}.}
}
\description{
Generates a UMAP visualization based on clustering results from a Seurat object,
and automatically saves it. Offers options for custom titles, subtitles, saving, and more. Assumes
default options for custom titles, subtitles, saving, and more.
}
\examples{
\dontrun{
clUMAP(ident = "integrated_snn_res.0.5", obj = yourSeuratObj)
clUMAP(ident = "integrated_snn_res.0.5", obj = yourSeuratObj, cols = RColorBrewer::brewer.pal(8, "Dark2"))
}

}
