% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{clUMAP}
\alias{clUMAP}
\title{clUMAP}
\usage{
clUMAP(
  ident = "integrated_snn_res.0.5",
  obj = combined.obj,
  reduction = "umap",
  splitby = NULL,
  title = ident,
  sub = NULL,
  prefix = NULL,
  suffix = make.names(sub),
  label.cex = 7,
  h = 7,
  w = NULL,
  nr.cols = NULL,
  plotname = ppp(toupper(reduction), ident),
  cols = NULL,
  palette = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[3],
  highlight.clusters = NULL,
  cells.highlight = NULL,
  label = TRUE,
  repel = TRUE,
  legend = !label,
  MaxCategThrHP = 200,
  axes = FALSE,
  aspect.ratio = c(FALSE, 0.6)[2],
  save.plot = MarkdownHelpers::TRUE.unless("b.save.wplots", v = F),
  PNG = TRUE,
  check_for_2D = TRUE,
  caption = .parseKeyParams(obj),
  ...
)
}
\arguments{
\item{ident}{Identity to be used for clustering, Default: 'integrated_snn_res.0.5'.}

\item{obj}{Seurat object containing single-cell RNA seq data, Default: combined.obj.}

\item{reduction}{Dimension reduction technique to be used. Choose from 'umap', 'tsne', or 'pca'. Default: 'umap'.}

\item{splitby}{Column in the metadata to split the cells by, Default: NULL.}

\item{title}{Title of the plot, Default: ident.}

\item{sub}{Subtitle of the plot, Default: NULL.}

\item{prefix}{A prefix added before the filename, Default: NULL.}

\item{suffix}{A suffix added to the end of the filename, Default: sub.}

\item{label.cex}{Scaling factor for label sizes, Default: 7.}

\item{h}{Height of the plot in inches, Default: 7.}

\item{w}{Width of the plot in inches, Default: NULL.}

\item{nr.cols}{Number of columns to combine multiple feature plots, ignored if split.by is not NULL, Default: NULL.}

\item{plotname}{Title of the plot, Default: ppp(toupper(reduction), ident).}

\item{cols}{Colors to be used for the plot, Default: NULL.}

\item{palette}{Color palette to be used, Default: 'glasbey'.}

\item{highlight.clusters}{Specific clusters to be highlighted, Default: NULL.}

\item{cells.highlight}{Specific cells to be highlighted, Default: NULL.}

\item{label}{If TRUE, clusters are labeled, Default: TRUE.}

\item{repel}{If TRUE, labels are repelled to avoid overlap, Default: TRUE.}

\item{legend}{If TRUE, a legend is added to the plot, Default: !label.}

\item{MaxCategThrHP}{Maximum threshold for the number of categories, Default: 200.}

\item{axes}{If TRUE, the axes are shown on the plot. Default: FALSE.}

\item{aspect.ratio}{Ratio of height to width. If TRUE, the ratio is fixed at 0.6. Default: TRUE.}

\item{save.plot}{If TRUE, the plot is saved into a file, Default: TRUE.}

\item{PNG}{If TRUE, the file is saved as a .png, Default: TRUE.}

\item{check_for_2D}{If TRUE, checks if UMAP is 2 dimensional. Default: TRUE.}

\item{caption}{Add caption to the ggplot object (e.g. a description in bottom right).}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
The quickest way to draw a clustering result UMAP.
}
\examples{
\dontrun{
if (interactive()) {
  clUMAP("cl.names.KnownMarkers.0.5")
  clUMAP("cl.names.KnownMarkers.0.5", cols = NULL)
}
}
}
