% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{clUMAP}
\alias{clUMAP}
\title{clUMAP}
\usage{
clUMAP(
  ident = "integrated_snn_res.0.5",
  obj = combined.obj,
  reduction = "umap",
  splitby = NULL,
  title = ident,
  sub = NULL,
  prefix = NULL,
  suffix = make.names(sub),
  label.cex = 7,
  h = 7,
  w = NULL,
  nr.cols = NULL,
  plotname = ppp(toupper(reduction), ident),
  cols = NULL,
  palette = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[3],
  highlight.clusters = NULL,
  cells.highlight = NULL,
  label = T,
  repel = T,
  legend = !label,
  MaxCategThrHP = 200,
  axes = T,
  aspect.ratio = c(FALSE, 0.6)[2],
  save.plot = MarkdownHelpers::TRUE.unless("b.save.wplots"),
  PNG = T,
  ...
)
}
\arguments{
\item{ident}{identity used, Default: 'integrated_snn_res.0.5'}

\item{obj}{Seurat object, Default: combined.obj}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'umap'}

\item{splitby}{PARAM_DESCRIPTION, Default: NULL}

\item{title}{Title of the plot, Default: ident}

\item{sub}{Subtitle of the plot, Default: NULL}

\item{prefix}{A prefix added before the filename, Default: NULL}

\item{suffix}{A suffix added to the filename, Default: sub}

\item{label.cex}{PARAM_DESCRIPTION, Default: 7}

\item{h}{height of the plot, Default: 7}

\item{w}{width of the plot, Default: NULL}

\item{nr.cols}{Number of columns to combine multiple feature plots to, ignored if split.by is not NULL, Default: NULL}

\item{plotname}{Title of the plot, Default: ppp(toupper(reduction), ident)}

\item{cols}{Colors used, Default: getDiscretePalette(ident.used = ident, show.colors = F)}

\item{palette}{Color palette.}

\item{highlight.clusters}{PARAM_DESCRIPTION, Default: NULL}

\item{cells.highlight}{PARAM_DESCRIPTION, Default: NULL}

\item{label}{PARAM_DESCRIPTION, Default: T}

\item{repel}{PARAM_DESCRIPTION, Default: T}

\item{legend}{PARAM_DESCRIPTION, Default: !label}

\item{MaxCategThrHP}{PARAM_DESCRIPTION, Default: 200}

\item{axes}{Show axes? Default: T}

\item{aspect.ratio}{Fix the aspect ratio? Default: F}

\item{save.plot}{Save the plot into a file?, Default: T}

\item{PNG}{Save file as .png?, Default: T}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
The quickest way to draw a clustering result  UMAP #
}
\examples{
\dontrun{
if(interactive()){
 clUMAP('cl.names.KnownMarkers.0.5' ); clUMAP('cl.names.KnownMarkers.0.5', cols = NULL)
 }
}
}
