% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{clUMAP}
\alias{clUMAP}
\title{clUMAP}
\usage{
clUMAP(
  ident = "integrated_snn_res.0.5",
  obj = combined.obj,
  reduction = "umap",
  splitby = NULL,
  title = ident,
  sub = NULL,
  suffix = sub,
  label.cex = 7,
  h = 7,
  w = NULL,
  nr.cols = NULL,
  plotname = ppp(toupper(reduction), ident),
  cols = getDiscretePalette(ident.used = ident, show.colors = F),
  highlight.clusters = NULL,
  cells.highlight = NULL,
  label = T,
  repel = T,
  legend = !label,
  MaxCategThrHP = 200,
  save.plot = T,
  PNG = T,
  save.object = F,
  ...
)
}
\arguments{
\item{ident}{identity used, Default: 'integrated_snn_res.0.5'}

\item{obj}{Seurat object, Default: combined.obj}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'umap'}

\item{splitby}{PARAM_DESCRIPTION, Default: NULL}

\item{title}{Title of the plot, Default: ident}

\item{sub}{PARAM_DESCRIPTION, Default: NULL}

\item{suffix}{A suffix added to the filename, Default: sub}

\item{label.cex}{PARAM_DESCRIPTION, Default: 7}

\item{h}{height of the plot, Default: 7}

\item{w}{width of the plot, Default: NULL}

\item{nr.cols}{PARAM_DESCRIPTION, Default: NULL}

\item{plotname}{Title of the plot, Default: ppp(toupper(reduction), ident)}

\item{cols}{PARAM_DESCRIPTION, Default: getDiscretePalette(ident.used = ident, show.colors = F)}

\item{highlight.clusters}{PARAM_DESCRIPTION, Default: NULL}

\item{cells.highlight}{PARAM_DESCRIPTION, Default: NULL}

\item{label}{PARAM_DESCRIPTION, Default: T}

\item{repel}{PARAM_DESCRIPTION, Default: T}

\item{legend}{PARAM_DESCRIPTION, Default: !label}

\item{MaxCategThrHP}{PARAM_DESCRIPTION, Default: 200}

\item{save.plot}{PARAM_DESCRIPTION, Default: T}

\item{PNG}{PARAM_DESCRIPTION, Default: T}

\item{save.object}{PARAM_DESCRIPTION, Default: F}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
The quickest way to draw a clustering result  UMAP #
}
\examples{
\dontrun{
if(interactive()){
 clUMAP('cl.names.KnownMarkers.0.5' ); clUMAP('cl.names.KnownMarkers.0.5', cols = NULL)
 }
}
}
