% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.utils.less.used.R
\name{Convert10Xfolders.old}
\alias{Convert10Xfolders.old}
\title{Convert10Xfolders.old}
\usage{
Convert10Xfolders.old(
  InputDir,
  folderPattern = c("filtered", "SoupX_decont")[1],
  min.cells = 10,
  min.features = 200,
  updateHGNC = T,
  ShowStats = T
)
}
\arguments{
\item{InputDir}{A character string specifying the input directory.}

\item{folderPattern}{A character vector specifying the pattern of folder names to be searched. Default is 'filtered'.}

\item{min.cells}{An integer value specifying the minimum number of cells. Default is 10.}

\item{min.features}{An integer value specifying the minimum number of features. Default is 200.}

\item{updateHGNC}{A logical value indicating whether to update the HGNC. Default is TRUE.}

\item{ShowStats}{A logical value indicating whether to show statistics. Default is TRUE.}
}
\description{
This function takes a parent directory with a number of subfolders, each containing the standard output of 10X Cell Ranger. It (1) loads the filtered data matrices, (2) converts them to Seurat objects, and (3) saves them as .RDS files.
}
\examples{
\dontrun{
if (interactive()) {
  Convert10Xfolders.old(InputDir = InputDir)
}
}
}
