% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.utils.less.used.R
\name{Convert10Xfolders.old}
\alias{Convert10Xfolders.old}
\title{Convert10Xfolders.old}
\usage{
Convert10Xfolders.old(
  InputDir,
  folderPattern = c("filtered", "SoupX_decont")[1],
  min.cells = 10,
  min.features = 200,
  updateHGNC = TRUE,
  ShowStats = TRUE
)
}
\arguments{
\item{InputDir}{A character string specifying the input directory.}

\item{folderPattern}{A character vector specifying the pattern of folder names to be searched. Default: 'filtered'.}

\item{min.cells}{An integer value specifying the minimum number of cells. Default: 10.}

\item{min.features}{An integer value specifying the minimum number of features. Default: 200.}

\item{updateHGNC}{A logical value indicating whether to update the HGNC. Default: \code{TRUE}.}

\item{ShowStats}{A logical value indicating whether to show statistics. Default: \code{TRUE}.}
}
\description{
This function takes a parent directory with a number of subfolders, each containing the standard output of 10X Cell Ranger. It (1) loads the filtered data matrices, (2) converts them to Seurat objects, and (3) saves them as .RDS files.
}
\examples{
\dontrun{
if (interactive()) {
  Convert10Xfolders.old(InputDir = InputDir)
}
}
}
