% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Convert10Xfolders}
\alias{Convert10Xfolders}
\title{Convert10Xfolders}
\usage{
Convert10Xfolders(
  InputDir,
  regex = F,
  folderPattern = c("filtered_feature", "SoupX_decont")[1],
  depth = 4,
  min.cells = 5,
  min.features = 200,
  updateHGNC = T,
  ShowStats = T,
  writeCBCtable = TRUE,
  sample.barcoding = F,
  ...
)
}
\arguments{
\item{InputDir}{Input directory}

\item{regex}{PARAM_DESCRIPTION, Default: F}

\item{folderPattern}{PARAM_DESCRIPTION, Default: c("filtered_feature", "SoupX_decont")\link{1}}

\item{depth}{Depth of scan (How many levels below InputDir). Def 2}

\item{min.cells}{PARAM_DESCRIPTION, Default: 5}

\item{min.features}{PARAM_DESCRIPTION, Default: 200}

\item{updateHGNC}{PARAM_DESCRIPTION, Default: T}

\item{ShowStats}{PARAM_DESCRIPTION, Default: T}

\item{writeCBCtable}{write out a list of cell barcodes (CBC) as tsv, Default: T}

\item{sample.barcoding}{Cell Ranger run with sample barcoding. The folder structure is different.}
}
\description{
Take a parent directory with a number of subfolders, each containing the standard output of 10X Cell Ranger. (1.) It loads the filtered data matrices; (2.) converts them to Seurat objects, and (3.) saves them as *.RDS files. #
}
\examples{
\dontrun{
if(interactive()){
 Convert10Xfolders(InputDir)
 }
}
}
