% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Convert10Xfolders}
\alias{Convert10Xfolders}
\title{Convert10Xfolders}
\usage{
Convert10Xfolders(
  InputDir,
  regex = FALSE,
  folderPattern = c("filtered_feature", "raw_feature", "SoupX_decont")[1],
  suffix = strsplit(folderPattern, "_")[[1]][1],
  depth = 4,
  min.cells = 5,
  min.features = 200,
  updateHGNC = TRUE,
  ShowStats = TRUE,
  writeCBCtable = TRUE,
  nthreads = .getNrCores(),
  preset = "high",
  ext = "qs",
  sort_alphanumeric = TRUE,
  save_empty_droplets = TRUE,
  ...
)
}
\arguments{
\item{InputDir}{A character string specifying the input directory.}

\item{regex}{A logical value. If TRUE, the folderPattern is treated as a regular expression. Default: \code{FALSE}.}

\item{folderPattern}{A character vector specifying the pattern of folder names to be searched. Default: 'filtered_feature'.}

\item{suffix}{A character string specifying the suffix of the files saved.}

\item{depth}{An integer value specifying the depth of scan (i.e., how many levels below the InputDir). Default: 2.}

\item{min.cells}{An integer value specifying the minimum number of cells. Default: 5.}

\item{min.features}{An integer value specifying the minimum number of features. Default: 200.}

\item{updateHGNC}{A logical value indicating whether to update the HGNC. Default: \code{TRUE}.}

\item{ShowStats}{A logical value indicating whether to show statistics. Default: \code{TRUE}.}

\item{writeCBCtable}{A logical value indicating whether to write out a list of cell barcodes (CBC) as a tsv file. Default: \code{TRUE}.}

\item{sort_alphanumeric}{sort files alphanumeric? Default: \code{TRUE}.}

\item{save_empty_droplets}{save empty droplets? Default: \code{TRUE}.}

\item{save}{Save .qs object? Default: \code{TRUE}.}
}
\description{
This function takes a parent directory with a number of subfolders, each
containing the standard output of 10X Cell Ranger. It (1) loads the (filtered) data matrices,
(2) converts them to Seurat objects, and (3) saves them as .qs files
}
\examples{
\dontrun{
if (interactive()) Convert10Xfolders(InputDir)
}
}
