% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Convert10Xfolders}
\alias{Convert10Xfolders}
\title{Convert10Xfolders}
\usage{
Convert10Xfolders(
  InputDir,
  regex = FALSE,
  folderPattern = c("filtered_feature", "SoupX_decont")[1],
  depth = 4,
  min.cells = 5,
  min.features = 200,
  updateHGNC = TRUE,
  ShowStats = TRUE,
  writeCBCtable = TRUE,
  sample.barcoding = FALSE,
  nthreads = .getNrCores(),
  preset = "high",
  ext = "qs",
  sort_alphanumeric = TRUE,
  ...
)
}
\arguments{
\item{InputDir}{A character string specifying the input directory.}

\item{regex}{A logical value. If TRUE, the folderPattern is treated as a regular expression. Default is FALSE.}

\item{folderPattern}{A character vector specifying the pattern of folder names to be searched. Default is 'filtered_feature'.}

\item{depth}{An integer value specifying the depth of scan (i.e., how many levels below the InputDir). Default is 2.}

\item{min.cells}{An integer value specifying the minimum number of cells. Default is 5.}

\item{min.features}{An integer value specifying the minimum number of features. Default is 200.}

\item{updateHGNC}{A logical value indicating whether to update the HGNC. Default is TRUE.}

\item{ShowStats}{A logical value indicating whether to show statistics. Default is TRUE.}

\item{writeCBCtable}{A logical value indicating whether to write out a list of cell barcodes (CBC) as a tsv file. Default is TRUE.}

\item{sample.barcoding}{A logical value indicating whether Cell Ranger was run with sample barcoding. Default is FALSE.}

\item{sort_alphanumeric}{sort files alphanumeric? Default: TRUE.}
}
\description{
This function takes a parent directory with a number of subfolders, each
containing the standard output of 10X Cell Ranger. It (1) loads the filtered data matrices,
(2) converts them to Seurat objects, and (3) saves them as .RDS files.
}
\examples{
\dontrun{
if (interactive()) Convert10Xfolders(InputDir)
}
}
