% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Create.MiscSlot}
\alias{Create.MiscSlot}
\title{Create.MiscSlot}
\usage{
Create.MiscSlot(obj, NewSlotName = "UVI.tables")
}
\arguments{
\item{obj}{Seurat object}

\item{NewSlotName}{Name of the new element inside obj@misc.}
}
\description{
Create.MiscSlot
}
