% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{DiscretePaletteSafe}
\alias{DiscretePaletteSafe}
\title{Safely generate a Discrete color palette.}
\usage{
DiscretePaletteSafe(
  n,
  palette.used = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[2],
  show.colors = FALSE,
  seed = 1989
)
}
\arguments{
\item{n}{Number of colors to generate.}

\item{palette.used}{Palette name to use for generating colors. Options include "alphabet",
"alphabet2", "glasbey", "polychrome", "stepped". Default: "alphabet2".}

\item{show.colors}{If TRUE, displays the generated color palette. Default: FALSE.}

\item{seed}{Seed value for reproducibility, especially when random color generation is involved.
Default: 1989.}
}
\value{
Character vector of HEX color values.
}
\description{
Generates a discrete color palette, ensuring no NA values are included, suitable
for visualizations where a specific number of distinct, reproducible colors is needed.
}
\examples{
\dontrun{
colors <- DiscretePaletteSafe(n = 10)
print(colors)
}

}
