% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{.metaColnames}
\alias{.metaColnames}
\title{Extract meta.data Column Names Matching a Pattern}
\usage{
.metaColnames(obj = combined.obj, pattern, perl = T, ...)
}
\arguments{
\item{obj}{A dataframe from which to extract column names.}

\item{pattern}{A regular expression pattern to match column names against.}
}
\value{
A character vector of column names matching the pattern.
}
\description{
Extract meta.data Column Names Matching a Pattern
}
\examples{
# Assuming 'df' is a dataframe with column names "azi.one", "azi.two", "other"
extract_matching_columns(df, "^azi\\\\.")
}
