% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{.parseBasicObjStats}
\alias{.parseBasicObjStats}
\title{Parse basic obj stats}
\usage{
.parseBasicObjStats(obj, sep = " ", simple = FALSE)
}
\arguments{
\item{obj}{An object to extract information from.}
}
\value{
A character string summarizing the key parameters.
}
\description{
Parse cell and feature count to a string.
}
