% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{.parseKeyParams}
\alias{.parseKeyParams}
\title{Parse key parameters from an object and format as a string}
\usage{
.parseKeyParams(
  obj,
  regressionVariables = NULL,
  nrVarFeatures = NULL,
  return.as.name = FALSE,
  assay = Seurat::DefaultAssay(obj),
  suffix = NULL,
  v = T
)
}
\arguments{
\item{obj}{An object to extract information from.}

\item{regressionVariables}{A list or vector containing variables for regression. Default: NULL.
If NULL, the function will attempt to extract the variables from the \code{object@commands$ScaleData}.}

\item{nrVarFeatures}{You can provide this number manually. Default: NULL.}

\item{return.as.name}{If TRUE, returns the name of the object. Default: FALSE.}

\item{assay}{The assay to extract scaled features from. Default: "RNA".}

\item{suffix}{A suffix string to add.}
}
\value{
A character string summarizing the key parameters.
}
\description{
This function extracts the number of scaled features, the number of principal components,
and formats additional information including regression variables.
}
