% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Downsample.Seurat.Objects.PC}
\alias{Downsample.Seurat.Objects.PC}
\title{Downsample.Seurat.Objects.PC}
\usage{
Downsample.Seurat.Objects.PC(ls.obj = ls.Seurat, fraction = 0.1)
}
\arguments{
\item{ls.obj}{List of Seurat objects, Default: ls.Seurat}

\item{NrCells}{Number of cells to downsample to. Default: p$dSample.Organoids}
}
\description{
Downsample a list of Seurat objects, by fraction
}
\examples{
\dontrun{
if(interactive()){
 Downsample.Seurat.Objects.PC()
 }
}
}
