% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Downsample.Seurat.Objects}
\alias{Downsample.Seurat.Objects}
\title{Downsample.Seurat.Objects}
\usage{
Downsample.Seurat.Objects(
  ls.obj = ls.Seurat,
  NrCells = p$dSample.Organoids,
  save_object = FALSE
)
}
\arguments{
\item{ls.obj}{List of Seurat objects. Default: ls.Seurat}

\item{NrCells}{Number of cells to downsample to.}

\item{save_object}{save object by isaveRDS, otherwise return it. Default: FALSE}
}
\description{
Downsample a list of Seurat objects
}
\examples{
\dontrun{
if (interactive()) {
  Downsample.Seurat.Objects(NrCells = 2000)
  Downsample.Seurat.Objects(NrCells = 200)
}
}
}
