% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{downsampleListSeuObjsNCells}
\alias{downsampleListSeuObjsNCells}
\title{Downsample a List of Seurat Objects to a Specific Number of Cells}
\usage{
downsampleListSeuObjsNCells(
  ls.obj = ls.Seurat,
  NrCells = p$dSample.Organoids,
  save_object = FALSE
)
}
\arguments{
\item{ls.obj}{List of Seurat objects to be downsampled; Default: \code{ls.Seurat}.}

\item{NrCells}{Target number of cells to downsample each Seurat object to.}

\item{save_object}{Logical indicating whether to save the downsampled Seurat objects using \code{isaveRDS}
or to return them; Default: FALSE.}
}
\description{
Downsampling each Seurat object in a list to a specified number of cells. This function is
particularly useful for creating smaller, more manageable subsets of large single-cell datasets for
preliminary analyses or testing.
}
\examples{
\dontrun{
  if (interactive()) {
    downsampledSeuratList <- downsampleListSeuObjsNCells(ls.obj =
                                      list(yourSeuratObj1, yourSeuratObj2), NrCells = 2000)
    downsampledSeuratList <- downsampleListSeuObjsNCells(NrCells = 200)
  }
}

}
