% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{downsampleListSeuObjsPercent}
\alias{downsampleListSeuObjsPercent}
\title{downsampleListSeuObjsPercent}
\usage{
downsampleListSeuObjsPercent(
  ls.obj = ls.Seurat,
  fraction = 0.1,
  save_object = FALSE
)
}
\arguments{
\item{ls.obj}{List of Seurat objects, Default: ls.Seurat}

\item{save_object}{save object by isaveRDS, otherwise return it. Default: FALSE}

\item{NrCells}{Number of cells to downsample to. Default: p$dSample.Organoids}
}
\description{
downsample a list of Seurat objects, by fraction
}
\examples{
\dontrun{
if (interactive()) {
  downsampleListSeuObjsPercent()
}
}
}
