% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{downsampleSeuObj.and.Save}
\alias{downsampleSeuObj.and.Save}
\title{downsampleSeuObj.and.Save}
\usage{
downsampleSeuObj.and.Save(
  obj = ORC,
  fraction = 0.25,
  seed = 1989,
  dir = OutDir,
  min.features = p$min.features,
  suffix = fraction,
  nthreads = .getNrCores()
)
}
\arguments{
\item{obj}{Seurat object, Default: ORC}

\item{fraction}{Fractional size to downsample to. Default: 0.25}

\item{seed}{random seed used, Default: 1989}

\item{dir}{Directory to save to. Default: OutDir}

\item{min.features}{Minimum features}

\item{suffix}{A suffix added to the filename, Default: ''}
}
\description{
Subset a Seurat object to a target fraction of cells and save the downsampled
object to the working directory.
}
