% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{downsampleSeuObjByIdentAndMaxcells}
\alias{downsampleSeuObjByIdentAndMaxcells}
\title{Sample Cells From Identifiers in Seurat Object}
\usage{
downsampleSeuObjByIdentAndMaxcells(
  obj,
  ident,
  max.cells,
  verbose = TRUE,
  seed = 1989
)
}
\arguments{
\item{obj}{A Seurat object from which cells are to be sampled.}

\item{ident}{A character vector specifying the identity class from which cells are to be sampled.}

\item{max.cells}{A positive integer indicating the maximum number of cells to sample from each identity class.}

\item{verbose}{Logical indicating if messages about the sampling process should be printed to the console. Defaults to TRUE.}
}
\value{
Returns a Seurat object containing only the sampled cells.
}
\description{
This function samples a specified maximum number of cells from each identity class
in a Seurat object, in the meta.data. It ensures that the sampling does not exceed the total
number of cells available per identity.
}
\details{
This function checks for the presence of the specified identity class within the object's metadata.
If the number of cells within any identity class is less than or equal to the \code{max.cells} parameter,
all cells from that class are retained. Otherwise, a random sample of \code{max.cells} is taken from the class.
The function updates the identity of the cells in the returned Seurat object to reflect the sampled cells.
If \code{verbose} is TRUE, it prints the total number of cells sampled and provides a visual summary of the fraction
of cells retained per identity class.
}
\examples{
# Assuming `seuratObj` is a Seurat object with identities stored in its metadata
sampledSeuratObj <- downsampleSeuObjByIdentAndMaxcells(obj = seuratObj, ident = "cellType", max.cells = 100)

}
