% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{dropLevelsSeurat}
\alias{dropLevelsSeurat}
\title{dropLevelsSeurat}
\usage{
dropLevelsSeurat(obj = combined.obj, verbose = TRUE)
}
\arguments{
\item{obj}{A Seurat object.}

\item{verbose}{Logical. Whether to print a message indicating which levels are being dropped.}
}
\description{
Drop unused levels from factor variables in a Seurat object.
}
