% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{fixZeroIndexing.seurat}
\alias{fixZeroIndexing.seurat}
\title{fixZeroIndexing.seurat}
\usage{
fixZeroIndexing.seurat(ColName.metadata = "res.0.6", obj = org)
}
\arguments{
\item{ColName.metadata}{PARAM_DESCRIPTION, Default: 'res.0.6'}

\item{obj}{Seurat object, Default: org}
}
\description{
Fix zero indexing in seurat clustering, to 1-based indexing. replace zero indexed clusternames.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
