% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{FlipReductionCoordinates}
\alias{FlipReductionCoordinates}
\title{FlipReductionCoordinates}
\usage{
FlipReductionCoordinates(
  obj = combined.obj,
  dim = 2,
  reduction = "umap",
  flip = c("x", "y", "xy", NULL)[1],
  FlipReductionBackupToo = TRUE
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{dim}{Numer of dimensions used, Default: 2}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'umap'}

\item{flip}{PARAM_DESCRIPTION, Default: c("x", "y", "xy", NULL)\link{1}}

\item{FlipReductionBackupToo}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
FlipReductionCoordinates
}
\examples{
\dontrun{
if(interactive()){
 clUMAP(); combined.obj <- FlipReductionCoordinates(combined.obj); clUMAP()
 }
}
}
