% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{FlipReductionCoordinates}
\alias{FlipReductionCoordinates}
\title{FlipReductionCoordinates}
\usage{
FlipReductionCoordinates(
  obj = combined.obj,
  dim = 2,
  reduction = "umap",
  flip = c("x", "y", "xy", NULL)[1],
  FlipReductionBackupToo = TRUE
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{dim}{Numer of dimensions used, Default: 2}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'umap'}

\item{flip}{The axis (axes) to flip around. Default: c("x", "y", "xy", NULL)\link{1}}

\item{FlipReductionBackupToo}{Flip coordinates in backup slot too? Default: TRUE}
}
\description{
Flip reduction coordinates (like UMAP upside down).
}
\examples{
\dontrun{
if (interactive()) {
  clUMAP()
  combined.obj <- FlipReductionCoordinates(combined.obj)
  clUMAP()
}
}
}
