% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{gene.expression.level.plots}
\alias{gene.expression.level.plots}
\title{gene.expression.level.plots}
\usage{
gene.expression.level.plots(
  gene = "TOP2A",
  obj = ls.Seurat[[1]],
  slot = c("counts", "data")[2],
  ...
)
}
\arguments{
\item{gene}{gene of interest, Default: 'TOP2A'}

\item{obj}{Seurat object, Default: ls.Seurat[\link{1}]}

\item{slot}{slot in the Seurat object. Default: c("counts", "data")\link{2}}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Histogram of gene expression levels.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
