% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{gene.name.check}
\alias{gene.name.check}
\title{gene.name.check}
\usage{
gene.name.check(Seu.obj)
}
\arguments{
\item{Seu.obj}{A Seurat object.}
}
\description{
Check gene names in a seurat object, for naming conventions (e.g.: mitochondrial reads have - or .). Use for reading .mtx & writing .rds files. #
}
