% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{get.clustercomposition}
\alias{get.clustercomposition}
\title{get.clustercomposition}
\usage{
get.clustercomposition(
  obj = combined.obj,
  x = "integrated_snn_res.0.3",
  y = "project",
  color = y,
  ...
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{x}{PARAM_DESCRIPTION, Default: 'integrated_snn_res.0.3'}

\item{y}{PARAM_DESCRIPTION, Default: 'project'}

\item{color}{PARAM_DESCRIPTION, Default: y}

\item{...}{Pass any other parameter to the internally called functions (most of them should  work).}
}
\description{
get.clustercomposition
}
\examples{
\dontrun{
if(interactive()){
 get.clustercomposition(); get.clustercomposition(, ylim=c(0,10000))
 }
}
}
