% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{get.clustercomposition}
\alias{get.clustercomposition}
\title{get.clustercomposition}
\usage{
get.clustercomposition(
  obj = combined.obj,
  ident = "integrated_snn_res.0.3",
  splitby = "ShortNames",
  color = y,
  plot = TRUE,
  ScaleTo100pc = TRUE,
  ...
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{color}{Color, Default: y}

\item{plot}{Show plot, Default: TRUE}

\item{ScaleTo100pc}{Scale the Y Axis, Default: TRUE}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}

\item{x}{Bars along the X axis, Default: 'integrated_snn_res.0.3'}

\item{y}{Vertical split of each bar, Default: 'project'}
}
\description{
Get cluster composition: which datasets contribute to each cluster?
}
\examples{
get.clustercomposition()
get.clustercomposition()
}
