% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{getCellIDs.from.meta}
\alias{getCellIDs.from.meta}
\title{getCellIDs.from.meta}
\usage{
getCellIDs.from.meta(
  ColName.meta = "res.0.6",
  values = NA,
  obj = combined.obj,
  inverse = F
)
}
\arguments{
\item{ColName.meta}{PARAM_DESCRIPTION, Default: 'res.0.6'}

\item{values}{PARAM_DESCRIPTION, Default: NA}

\item{obj}{Seurat object, Default: combined.obj}

\item{inverse}{PARAM_DESCRIPTION, Default: F}
}
\description{
Get cellIDs from a metadata column, matching a list of values (using \%in\%). #
}
\examples{
\dontrun{
if(interactive()){
 getCellIDs.from.meta()
 }
}
}
