% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{getCellIDs.from.meta}
\alias{getCellIDs.from.meta}
\title{getCellIDs.from.meta}
\usage{
getCellIDs.from.meta(
  ident = GetClusteringRuns()[1],
  ident_values = NA,
  obj = combined.obj,
  inverse = FALSE
)
}
\arguments{
\item{ident}{A string specifying the name of the metadata column from which to retrieve cell IDs. Default: 'res.0.6'.}

\item{ident_values}{A vector of values to match in the metadata column. Default: \code{NA}.}

\item{obj}{The Seurat object from which to retrieve the cell IDs. Default: combined.obj.}

\item{inverse}{A boolean value indicating whether to inverse the match, i.e., retrieve cell IDs that do not match the provided list of ident_values. Default: \code{FALSE}.}
}
\value{
A vector of cell IDs that match (or don't match, if \code{inverse = TRUE}) the provided list of values.
}
\description{
Retrieves cell IDs from a specified metadata column of a Seurat object, where the cell ID matches a provided list of values. The matching operation uses the \code{\%in\%} operator.
}
\examples{
\dontrun{
if (interactive()) {
  # Example usage:
  getCellIDs.from.meta()
}
}
}
