% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{getClusterColors}
\alias{getClusterColors}
\title{getClusterColors}
\usage{
getClusterColors(
  obj = combined.obj,
  use_new_palettes = TRUE,
  palette = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[3],
  ident = GetClusteringRuns()[1],
  show = T
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{ident}{identity used, Default: GetClusteringRuns()\link{1}}

\item{show}{Show plot of colors? Default: T}
}
\description{
get Seurat's cluster colors.
}
\examples{
\dontrun{
if (interactive()) {
  getClusterColors(obj = combined.obj, ident = GetClusteringRuns()[2])
}
}
}
\seealso{
\code{\link[scales]{hue_pal}}
}
