% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{GetClusteringRuns}
\alias{GetClusteringRuns}
\title{GetClusteringRuns}
\usage{
GetClusteringRuns(
  obj = combined.obj,
  res = FALSE,
  pat = "*snn_res.[0-9].[0-9]+$",
  v = TRUE
)
}
\arguments{
\item{obj}{Seurat object, Default: \code{combined.obj}}

\item{res}{Clustering resoluton to use, Default: \code{FALSE}.}

\item{pat}{Pattern to match, Default: \verb{*snn_res.*[0-9]$}}

\item{v}{verbose, Default: \code{TRUE}.}
}
\value{
Prints and returns the sorted unique cluster names as a character vector.
}
\description{
The \code{GetClusteringRuns} function retrieves metadata column names associated with
clustering runs, based on a pattern to match, \code{"*snn_res.[0-9].[0-9]$"}, by default.
}
\examples{
\dontrun{
if (interactive()) {
  GetClusteringRuns(obj = combined.obj, pat = "*snn_res.*[0-9]$")
}
}
}
