% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{GetClusteringRuns}
\alias{GetClusteringRuns}
\title{GetClusteringRuns}
\usage{
GetClusteringRuns(obj = combined.obj, res = FALSE, pat = "*snn_res.*[0-9]$")
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{res}{Clustering resoluton to use, Default: FALSE}

\item{pat}{Pettern to match, Default: '\emph{snn_res.}\link{0-9}$'}
}
\description{
Get Clustering Runs: metadata column names.
}
\examples{
\dontrun{
if (interactive()) {
  GetClusteringRuns()
}
}
}
