% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{getClusterNames}
\alias{getClusterNames}
\title{Retrieve Cluster Names}
\usage{
getClusterNames(obj = combined.obj, ident = GetClusteringRuns(obj)[2])
}
\arguments{
\item{obj}{A Seurat object. Default: \code{combined.obj}.}

\item{ident}{The identity class from which to retrieve cluster names.
Default uses the second clustering run from \code{GetClusteringRuns(obj)}.}
}
\value{
Prints and returns the sorted unique cluster names as a character vector.
}
\description{
Extracts cluster names based on a specified identity class from a Seurat object.
}
\examples{
\dontrun{
getClusterNames(obj = combined.obj, ident = GetClusteringRuns(obj)[2])
}
}
