% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{getDiscretePalette}
\alias{getDiscretePalette}
\title{getDiscretePalette}
\usage{
getDiscretePalette(
  ident.used = GetClusteringRuns()[1],
  obj = combined.obj,
  palette.used = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[1],
  show.colors = F
)
}
\arguments{
\item{ident.used}{PARAM_DESCRIPTION, Default: GetClusteringRuns()\link{1}}

\item{obj}{Seurat object, Default: combined.obj}

\item{palette.used}{PARAM_DESCRIPTION, Default: c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")\link{1}}

\item{show.colors}{PARAM_DESCRIPTION, Default: F}
}
\description{
Generate a Discrete color Palette.
}
\examples{
\dontrun{
if(interactive()){
 getDiscretePalette()
 }
}
}
