% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{getDiscretePaletteObj}
\alias{getDiscretePaletteObj}
\title{Generate a Discrete Color Palette for Seurat Clusters}
\usage{
getDiscretePaletteObj(
  ident.used,
  obj,
  palette.used = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[2],
  show.colors = FALSE,
  seed = 1989
)
}
\arguments{
\item{ident.used}{Identity column in the Seurat object to base the color palette on.}

\item{obj}{Seurat object containing clustering information.}

\item{palette.used}{The palette name to use for color generation. Options include "alphabet",
"alphabet2", "glasbey", "polychrome", and "stepped". Default: "alphabet2".}

\item{show.colors}{If TRUE, displays the generated colors. Default: FALSE.}

\item{seed}{Seed for random color generation, ensuring reproducibility. Default: 1989.}
}
\value{
A character vector of color values corresponding to the number of clusters.
}
\description{
Generates a discrete color palette for visualizing clusters in a Seurat object,
using a specified identity column to determine the number of unique clusters.
}
\examples{
\dontrun{
if (interactive()) {
  ident.used <- "resolution_1"
  obj <- YourSeuratObject
  colors <- getDiscretePaletteObj(ident.used = ident.used, obj = obj)
  print(colors)
}
}

}
