% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{getMedianMetric}
\alias{getMedianMetric}
\title{getMedianMetric}
\usage{
getMedianMetric(
  ls.obj = ls.Seurat,
  n.datasets = length(ls.Seurat),
  mColname = "percent.mito"
)
}
\arguments{
\item{ls.obj}{List of Seurat objects, Default: ls.Seurat}

\item{n.datasets}{lenght of list (n objects), Default: length(ls.Seurat)}

\item{mColname}{Metadata column name to calculate on. Default: 'percent.mito'}
}
\description{
Get the median values of different columns in meta.data, can iterate over a list of Seurat objects.
}
\examples{
\dontrun{
if(interactive()){
 ls.Seurat <- getMedianMetric(ls.obj = ls.Seurat, n.datasets = length(ls.Seurat),
 mColname = "percent.mito")
 }
}
}
