% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{getMetadataColumn}
\alias{getMetadataColumn}
\title{getMetadataColumn}
\usage{
getMetadataColumn(
  col = "batch",
  obj = combined.obj,
  as_numeric = FALSE,
  v = TRUE
)
}
\arguments{
\item{col}{A string specifying the name of the metadata column to be retrieved. Default: 'batch'.}

\item{obj}{A Seurat object from which the metadata column will be retrieved. Default: combined.obj.}

\item{as_numeric}{A logical flag indicating whether the returned values should be converted to numeric format. Default: \code{FALSE}.}

\item{v}{verbose. Default: \code{TRUE}.}
}
\value{
A named vector containing the values from the specified metadata column. If 'as_numeric' is TRUE, the values are converted to numeric format.
}
\description{
Retrieves a specified metadata column from a Seurat object and returns it as a named vector.
}
\examples{
\dontrun{
if (interactive()) {
  # Example usage:
  batch_metadata <- getMetadataColumn(col = "batch", obj = combined.obj, as_numeric = TRUE)
}
}
}
