% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{getMetadataColumn}
\alias{getMetadataColumn}
\title{getMetadataColumn <- mmeta}
\usage{
getMetadataColumn(
  ColName.metadata = "batch",
  obj = combined.obj,
  as_numeric = F
)
}
\arguments{
\item{ColName.metadata}{PARAM_DESCRIPTION, Default: 'batch'}

\item{obj}{Seurat object, Default: combined.obj}

\item{as_numeric}{PARAM_DESCRIPTION, Default: F}
}
\description{
Get a metadata column from a Seurat object as a named vector. get Cells from metadata.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
