% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{GetMostVarGenes}
\alias{GetMostVarGenes}
\title{GetMostVarGenes}
\usage{
GetMostVarGenes(obj = org, nGenes = p$nVarGenes)
}
\arguments{
\item{obj}{Seurat object, Default: org}

\item{nGenes}{PARAM_DESCRIPTION, Default: p$nVarGenes}
}
\description{
Get the most variable rGenes #
}
