% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{GetNamedClusteringRuns}
\alias{GetNamedClusteringRuns}
\title{GetNamedClusteringRuns}
\usage{
GetNamedClusteringRuns(
  obj = combined.obj,
  res = list(FALSE, 0.5)[[1]],
  topgene = FALSE,
  pat = c("^cl.names.top.gene.+[0-9]\\\\.[0-9]", "Name|name")[2],
  find.alternatives = TRUE,
  v = TRUE
)
}
\arguments{
\item{obj}{Seurat object, Default: \code{combined.obj}}

\item{res}{Clustering resoluton to use, Default: c(FALSE, 0.5)\link{1}}

\item{topgene}{Match clustering named after top expressed gene (see vertesy/Seurat.pipeline/~Diff gene expr.), Default: \code{FALSE}.}

\item{pat}{Pattern to match, Default: '^cl.names.Known.*\link{0,1}\.\link{0-9}$'}

\item{find.alternatives}{If TRUE, tries to find alternative clustering runs with
the same resolution, Default: \code{TRUE}.}

\item{v}{Verbose output, Default: \code{TRUE}.}
}
\description{
The \code{GetNamedClusteringRuns} function retrieves metadata column names associated with
non-numeric ("named") clustering runs, based on a pattern to match, \code{"Name|name"}, by default.
}
\examples{
\dontrun{
if (interactive()) {
  GetNamedClusteringRuns()
}
}
}
