% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{GetOrderedClusteringRuns}
\alias{GetOrderedClusteringRuns}
\title{GetOrderedClusteringRuns}
\usage{
GetOrderedClusteringRuns(
  obj = combined.obj,
  res = F,
  pat = "*snn_res.*[0,1]\\\\.[0-9]\\\\.ordered$"
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{res}{Clustering resoluton to use, Default: F}

\item{pat}{Pettern to match, Default: '\emph{snn_res.}\link{0,1}\.\link{0-9}\.ordered$'}
}
\description{
Get Clustering Runs: metadata column names #
}
\examples{
\dontrun{
if (interactive()) {
  GetOrderedClusteringRuns()
  GetOrderedClusteringRuns(res = 0.5)
}
}
}
