% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{GetTopMarkersDF}
\alias{GetTopMarkersDF}
\title{Get Top Differential Expression Genes Data Frame}
\usage{
GetTopMarkersDF(
  dfDE = df.markers,
  n = p$n.markers,
  order.by = c("avg_log2FC", "p_val_adj")[1],
  exclude = c("^A[CFLP][0-9]{6}", "^Z[0-9]{5}", "^LINC0[0-9]{4}", "^C[1-9]+orf[1-9]+",
    "[-|\\\\.]AS[1-9]*$", "[-|\\\\.]DT[1-9]*$", "^MIR[1-9]", "^SNHG[1-9]")
)
}
\arguments{
\item{dfDE}{Data frame containing the results of differential gene expression analysis
(e.g., output from \code{FindAllMarkers()}); Default: \code{df.markers}.}

\item{n}{Number of top markers to retrieve per cluster; Default: \code{p$n.markers}.}

\item{order.by}{Priority column for sorting markers before selection, such as \code{"avg_log2FC"};
Default: \code{"avg_log2FC"}.}

\item{exclude}{Vector of regex patterns to exclude genes from the top markers list;
Default: \code{c("^AL*|^AC*|^LINC*")}.}
}
\description{
Retrieves a data frame of the top N differentially expressed genes from
differential gene expression analysis results, offering an option to exclude certain genes
based on patterns.
}
\examples{
\dontrun{
if (interactive()) {
  topMarkersDF <- GetTopMarkersDF(dfDE = df.markers, n = 3)
}
}

}
\seealso{
\code{\link[Seurat]{FindAllMarkers}}, \code{\link[dplyr]{arrange}},
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}
}
