% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg_color_hue}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}
}
\description{
reproduce the ggplot2 default color palette #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
