% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg_color_hue}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{The number of colors to generate.}
}
\value{
A vector of colors emulating the default color palette of ggplot2.
}
\description{
Emulates the default color palette of ggplot2. Source:  https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
\examples{
\dontrun{
gg_color_hue(5)
}
}
