% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{gg_plot}
\alias{gg_plot}
\title{gg_plot}
\usage{
gg_plot(
  sds,
  col = NULL,
  title = NULL,
  lineSize = 1,
  reduction = "UMAP",
  titleFsize = 20,
  line.colors = gray.colors(n = length(sds@curves), start = 0, end = 0.6, alpha = 1),
  ...
)
}
\arguments{
\item{sds}{PARAM_DESCRIPTION}

\item{col}{PARAM_DESCRIPTION, Default: NULL}

\item{title}{Title of the plot, Default: NULL}

\item{lineSize}{PARAM_DESCRIPTION, Default: 1}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'UMAP'}

\item{titleFsize}{PARAM_DESCRIPTION, Default: 20}

\item{line.colors}{PARAM_DESCRIPTION, Default: gray.colors(n = length(sds@curves), start = 0, end = 0.6, alpha = 1)}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Adjusted gg_plot for slingshot
}
\examples{
\dontrun{
if(interactive()){
 plotFittedGenePseudotime(data = sds, gene ="SST", expr = EXPR, loessCI=T, col = colz, pch = 20,
 panel_first = grid(NULL) )
 }
}
}
