% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.less.used.R
\name{ggplotColours}
\alias{ggplotColours}
\title{ggplotColours}
\usage{
ggplotColours(n = 6, h = c(0, 360) + 15)
}
\arguments{
\item{n}{PARAM_DESCRIPTION, Default: 6}

\item{h}{height of the plot, Default: c(0, 360) + 15}
}
\description{
Generate ggplot colours for slingshot
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
