% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{HGNC.EnforceUnique}
\alias{HGNC.EnforceUnique}
\title{HGNC.EnforceUnique}
\usage{
HGNC.EnforceUnique(updatedSymbols)
}
\arguments{
\item{updatedSymbols}{Gene symbols, it is the output of HGNChelper's checkGeneSymbols)_.}
}
\description{
Enforce Unique names after HGNC symbol update. While "make.unique" is not the ideal
solution, because it generates mismatches, in my integration example it does reduce the
mismatching genes from ~800 to 4
}
\examples{
\dontrun{
if (interactive()) {
  x <- HGNC.EnforceUnique(updatedSymbols = SymUpd)
}
}
}
