% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{IntersectGeneLsWithObject}
\alias{IntersectGeneLsWithObject}
\title{Intersect Genes with Seurat Object}
\usage{
IntersectGeneLsWithObject(
  genes,
  obj = combined.obj,
  n_genes_shown = 10,
  strict = TRUE
)
}
\arguments{
\item{genes}{A vector of gene names to be intersected with the Seurat object.}

\item{obj}{A Seurat object containing gene expression data.}

\item{n_genes_shown}{Number of missing genes to be printed. Default: 10.}

\item{strict}{All genes to be present in the Seurat object?  Default: TRUE.}
}
\value{
A vector of gene names that are found both in the input 'genes' vector and the
Seurat object.
}
\description{
Intersects a set of gene names with those found in a Seurat object.
}
