% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{IntersectWithExpressed}
\alias{IntersectWithExpressed}
\title{IntersectWithExpressed}
\usage{
IntersectWithExpressed(genes, obj = combined.obj, genes.shown = 10)
}
\arguments{
\item{genes}{A vector of gene names.}

\item{obj}{Seurat object}

\item{genes.shown}{Number of genes printed (head).}
}
\description{
IntersectWithExpressed
}
