% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{isave.image}
\alias{isave.image}
\title{isave.image}
\usage{
isave.image(
  ...,
  path_rdata = paste0("~/Dropbox/Abel.IMBA/AnalysisD/_Rdata.files/", basename(OutDir)),
  showMemObject = T,
  options = c("--force", NULL)[1]
)
}
\arguments{
\item{...}{Additional parameters passed to the idate() function in the creation of the file name.}

\item{path_rdata}{A string that specifies the path for storing the image of the workspace. Default is "~/Dropbox/Abel.IMBA/AnalysisD/_Rdata.files/" appended with the basename of OutDir.}

\item{showMemObject}{A boolean flag, if TRUE the function will print out the memory size of the largest objects in the workspace. Default is TRUE.}

\item{options}{A string for gzip options. Default is "--force".}
}
\description{
Save an image of the current workspace using a faster and efficient compression method that runs in the background.
}
\examples{
\dontrun{
if (interactive()) {
  isave.image(my.R.image)
}
}
}
