% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{isave.RDS}
\alias{isave.RDS}
\title{isave.RDS}
\usage{
isave.RDS(
  obj,
  prefix = NULL,
  suffix = NULL,
  inOutDir = F,
  project = getProject(),
  alternative_path_rdata = paste0("~/Dropbox (VBC)/Abel.IMBA/AnalysisD/_RDS.files/",
    basename(OutDir)),
  homepath = "/Users/abel.vertesy/",
  showMemObject = T,
  saveParams = T
)
}
\arguments{
\item{obj}{Seurat object}

\item{prefix}{PARAM_DESCRIPTION, Default: NULL}

\item{suffix}{A suffix added to the filename, Default: NULL}

\item{inOutDir}{PARAM_DESCRIPTION, Default: F}

\item{project}{project code appended to the saved file name. Default: try(basename(rstudioapi::getActiveProject()), silent=T) using  getProject().}

\item{alternative_path_rdata}{PARAM_DESCRIPTION, Default: paste0("~/Dropbox/Abel.IMBA/AnalysisD/_RDS.files/", basename(OutDir))}

\item{homepath}{homepath to replace '~', Default: paste0("~/Dropbox/Abel.IMBA/AnalysisD/_RDS.files/", basename(OutDir))}

\item{showMemObject}{PARAM_DESCRIPTION, Default: T}

\item{saveParams}{PARAM_DESCRIPTION, Default: T}
}
\description{
Save and RDS object.
}
\examples{
\dontrun{
if(interactive()){
 isave.RDS(my.R.object)
 }
}
}
