% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{jJaccardIndexBinary}
\alias{jJaccardIndexBinary}
\title{jJaccardIndexBinary}
\usage{
jJaccardIndexBinary(x, y)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}
}
\description{
Calculate Jaccard Index. Modified from: https://www.displayr.com/how-to-calculate-jaccard-coefficients-in-displayr-using-r/ #
}
\examples{
\dontrun{
if(interactive()){
 JaccardSimilarity <- jJaccardIndexBinary(  x=sample(x = 0:1, size = 100, replace = T), y=sample(x = 0:1, size = 100, replace = T))
 }
}
}
