% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{jJaccardIndexVec}
\alias{jJaccardIndexVec}
\title{jJaccardIndexVec}
\usage{
jJaccardIndexVec(A = 1:3, B = 2:4)
}
\arguments{
\item{A}{Set A, Default: 1:3}

\item{B}{Set B, Default: 2:4}
}
\description{
Calculate jaccard similarity for 2 vectors. Helper to jPairwiseJaccardIndexList.
}
