% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{jPairwiseJaccardIndexList}
\alias{jPairwiseJaccardIndexList}
\title{jPairwiseJaccardIndexList}
\usage{
jPairwiseJaccardIndexList(lsG = ls_genes)
}
\arguments{
\item{lsG}{PARAM_DESCRIPTION, Default: ls_genes}
}
\description{
Create a pairwise jaccard similarity matrix across all combinations of columns in binary.presence.matrix. Modified from: https://www.displayr.com/how-to-calculate-jaccard-coefficients-in-displayr-using-r/ #
}
\examples{
\dontrun{
if(interactive()){
 jPairwiseJaccardIndexList(lsG = ls_genes)
 }
}
}
