% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{load10Xv3}
\alias{load10Xv3}
\title{load10Xv3}
\usage{
load10Xv3(
  dataDir,
  cellIDs = NULL,
  channelName = NULL,
  readArgs = list(),
  includeFeatures = c("Gene Expression"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dataDir}{A character string specifying the directory containing the 10X output folders.}

\item{cellIDs}{A vector specifying the cell IDs. Default: NULL.}

\item{channelName}{A character string specifying the channel name. Default: NULL.}

\item{readArgs}{A list of arguments to pass to the internal Read10X function. Default: list().}

\item{includeFeatures}{A character vector specifying which features to include. Default: c("Gene Expression").}

\item{verbose}{A logical flag indicating whether to print status messages. Default: TRUE.}

\item{...}{Additional arguments to pass to the internally called functions.}
}
\value{
An object of class "SoupChannel" representing the loaded 10X data.
}
\description{
Load 10X output folders.
}
\examples{
\dontrun{
if (interactive()) {
  channel <- load10Xv3(dataDir = dataDir)
}
}
}
\seealso{
\code{\link[SoupX]{SoupChannel}}
}
