% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.utils.less.used.R
\name{load10Xv3}
\alias{load10Xv3}
\title{Load 10X Genomics Version 3 Data}
\usage{
load10Xv3(
  dataDir,
  cellIDs = NULL,
  channelName = NULL,
  readArgs = list(),
  includeFeatures = c("Gene Expression"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dataDir}{A string specifying the directory that contains the 10X Genomics output folders.
This directory should include subdirectories for raw and filtered data, typically named starting with
\code{raw_} and \code{filt_}, respectively.}

\item{cellIDs}{An optional vector of cell IDs to include in the loaded data. Default is \code{NULL},
indicating that all available cells will be included. This is useful for subsetting the data based
on specific cell IDs.}

\item{channelName}{An optional string specifying the channel name for the data being loaded.
This can be used to label the data according to the experimental condition or sample name. Default is \code{NULL}.}

\item{readArgs}{A list of additional arguments to pass to the internal \code{Read10X} function used for
loading the data. Default is an empty list.}

\item{includeFeatures}{A character vector specifying which features to include in the loaded data.
Common values include "Gene Expression", "Antibody Capture", and "CRISPR Guide Capture".
Default is \code{c("Gene Expression")}.}

\item{verbose}{A logical flag indicating whether to print progress messages and status updates as the
data is loaded. Default is \code{TRUE}.}

\item{...}{Additional arguments passed to other internally called functions, if applicable.}
}
\value{
An object of class \code{SoupChannel}, representing the loaded 10X data. This object includes
raw counts, filtered counts, and optionally, additional metadata and dimensionality reduction coordinates
(e.g., t-SNE).
}
\description{
Loads 10X Genomics data from a specified directory containing output folders for raw and filtered data.
This function is designed to handle data from 10X Genomics Chromium Single Cell technologies (version 3).
}
\details{
This function provides a comprehensive approach to loading and organizing 10X Genomics data
for further analysis. It accommodates the data structure commonly found in 10X Genomics version 3 outputs
and allows for the inclusion of various types of molecular data as well as optional cell and channel
specifications.
}
\examples{
\dontrun{
if (interactive()) {
  # Assuming `dataDir` is the path to your 10X data directory
  channel <- load10Xv3(dataDir = "path/to/10X/data")
  # Now `channel` contains the loaded 10X data as a `SoupChannel` object
}
}

}
\seealso{
\code{\link[SoupX]{SoupChannel}} for the structure and utilities of the \code{SoupChannel} class.
}
