% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{LoadAllSeurats}
\alias{LoadAllSeurats}
\title{LoadAllSeurats}
\usage{
LoadAllSeurats(
  InputDir,
  file.pattern = "^filtered.+Rds$",
  string.remove1 = c(F, "filtered_feature_bc_matrix.", "raw_feature_bc_matrix.")[2],
  string.remove2 = c(F, ".min.cells.10.min.features.200.Rds")[2]
)
}
\arguments{
\item{InputDir}{A character string specifying the input directory.}

\item{file.pattern}{A character string specifying the pattern of file names to be searched. Default is '^filtered.+Rds$'.}

\item{string.remove1}{A character string or FALSE. If a string is provided, it is removed from file names. Default is "filtered_feature_bc_matrix.".}

\item{string.remove2}{A character string or FALSE. If a string is provided, it is removed from file names. Default is ".min.cells.10.min.features.200.Rds".}
}
\description{
This function loads all Seurat objects found in a directory. It also works with symbolic links (but not with aliases).
}
\examples{
\dontrun{
if (interactive()) {
  ls.Seurat <- LoadAllSeurats(InputDir = InputDir)
}
}
}
