% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{LoadAllSeurats}
\alias{LoadAllSeurats}
\title{LoadAllSeurats}
\usage{
LoadAllSeurats(
  InputDir,
  file.pattern = "^filtered.+Rds$",
  string.remove1 = list(FALSE, "filtered_feature_bc_matrix.",
    "raw_feature_bc_matrix.")[[2]],
  string.replace1 = "",
  string.remove2 = list(FALSE, ".min.cells.10.min.features.200.Rds")[[2]],
  sort_alphanumeric = TRUE
)
}
\arguments{
\item{InputDir}{A character string specifying the input directory.}

\item{file.pattern}{A character string specifying the pattern of file names to be searched.
Default: '^filtered.+Rds$'.}

\item{string.remove1}{A character string or FALSE. If a string is provided, it is removed from
file names. Default: "filtered_feature_bc_matrix.".}

\item{string.replace1}{A character string of the new text instead of "string.remove1".}

\item{string.remove2}{A character string or FALSE. If a string is provided, it is removed from
file names. Default: ".min.cells.10.min.features.200.Rds".}

\item{sort_alphanumeric}{sort files alphanumeric? Default: \code{TRUE}.}
}
\description{
This function loads all Seurat objects found in a directory. It also works with
symbolic links (but not with aliases).
}
\examples{
\dontrun{
if (interactive()) {
  ls.Seurat <- LoadAllSeurats(InputDir = InputDir)
}
}
}
