% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{m3.export.umap.2.Seurat}
\alias{m3.export.umap.2.Seurat}
\title{m3.export.umap.2.Seurat}
\usage{
m3.export.umap.2.Seurat(
  mobj = cds_from_seurat,
  sobj = combined.obj,
  def.dim = 2
)
}
\arguments{
\item{mobj}{PARAM_DESCRIPTION, Default: cds_from_seurat}

\item{sobj}{PARAM_DESCRIPTION, Default: combined.obj}

\item{def.dim}{PARAM_DESCRIPTION, Default: 2}
}
\description{
m3.export.umap.2.Seurat
}
\examples{
\dontrun{
if(interactive()){
 combined.obj.bac <- combined.obj; # combined.obj <- combined.obj.bac
 combined.obj <- m3.export.umap.2.Seurat(mobj = cds_from_seurat, sobj = combined.obj); qUMAP("DDIT4")
 }
}
}
