% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{make10Xcellname}
\alias{make10Xcellname}
\title{Add Suffix to Cell Names (e.g. lane suffix: _1)}
\usage{
make10Xcellname(cellnames, suffix = "_1")
}
\arguments{
\item{cellnames}{A vector of cell names without numeric suffixes.}

\item{suffix}{The suffix to add to each cell name. Default: '_1'.}
}
\value{
A vector of cell names with the specified suffix appended.
}
\description{
Appends a specified suffix to cell names to mimic lane suffixes used in 10X datasets.
}
\examples{
cellnames <- c("cell1", "cell2")
make10Xcellname(cellnames)
}
