% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{matchBestIdentity}
\alias{matchBestIdentity}
\title{Match and Translate Best Identity}
\usage{
matchBestIdentity(
  obj,
  ident_to_rename,
  reference_ident = GetOrderedClusteringRuns(obj)[1],
  prefix = reference_ident,
  suffix = gsub(prefix, "", x = reference_ident),
  new_ident_name = kpp(prefix, ident_to_rename, "match.to", suffix),
  plot_suffix = prefix,
  barplot_match = TRUE,
  barplot_fractions = TRUE,
  rnd_colors = TRUE,
  w = 12,
  h = 9,
  ...
)
}
\arguments{
\item{obj}{The object to update. This object must have a \code{meta.data} attribute which is a data frame
containing columns named as \code{ident_to_rename} and \code{reference_ident}.}

\item{ident_to_rename}{A string. The name of the column in \code{obj@meta.data} that is used as the source of identities.
There is no default value for this parameter.}

\item{reference_ident}{A string. The name of the column in \code{obj@meta.data} that is used as the target of identities.
There is no default value for this parameter.}

\item{prefix}{A string to add to the new identity column name. Default is prefix = Reference.}

\item{suffix}{...}

\item{new_ident_name}{A string. The name for the newly created identity column in \code{obj@meta.data}.
Default is a concatenation: kpp(prefix, ident_to_rename, "match.to", reference_ident) .}

\item{plot_suffix}{A string. The suffix to add to the final UMAP.}

\item{barplot_match}{Draw a barplot to show the \% match between the original and new identities.
Default: TRUE}

\item{barplot_fractions}{Draw a barplot to show the \% of each category in the original and new
identities, using \code{scBarplot.CellFractions()}. Default: TRUE}

\item{rnd_colors}{Use randomly shuffled colors for the barplot of fractions? Default: \code{TRUE}.}

\item{w}{Width for the saved image. Default: 9}

\item{h}{Height for the saved image. Default: 12}

\item{...}{Additional parameters to be passed to \code{.replace_by_most_frequent_categories} function.}
}
\value{
An updated version of \code{obj} with an additional column in \code{obj@meta.data} named as \code{new_ident_name}
representing the new identity. The function also generates a UMAP plot based on this new identity.
}
\description{
This function matches the best identity from \code{ident_to_rename} to \code{reference_ident} in an object,
in other words, it replaces original categories with the most frequent ones from the reference,
hence helps to filter out less important categories.
}
\examples{
\dontrun{
updated_obj <- matchBestIdentity(my_obj, "origin_identity", "target_identity")
}
}
\seealso{
\code{clUMAP()}, \code{Stringendo::kpp()}, \code{.replace_by_most_frequent_categories()}
}
