% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{matchBestIdentity}
\alias{matchBestIdentity}
\title{Match and Translate Best Identity}
\usage{
matchBestIdentity(
  obj,
  ident_to_rename,
  reference_ident = GetOrderedClusteringRuns(obj)[1],
  to_suffix = "matched",
  new_ident_name = kpp(ident_to_rename, "best.match", to_suffix),
  plot_suffix = "R44",
  ...
)
}
\arguments{
\item{obj}{The object to update. This object must have a \code{meta.data} attribute which is a data frame
containing columns named as \code{ident_to_rename} and \code{reference_ident}.}

\item{ident_to_rename}{A string. The name of the column in \code{obj@meta.data} that is used as the source of identities.
There is no default value for this parameter.}

\item{reference_ident}{A string. The name of the column in \code{obj@meta.data} that is used as the target of identities.
There is no default value for this parameter.}

\item{to_suffix}{A string. The suffix to add to the new identity name. Default is the output of the \code{FixPlotName}
function applied to the \code{ident_to_rename} string, with all alphabetical and underscore characters removed.}

\item{new_ident_name}{A string. The name for the newly created identity column in \code{obj@meta.data}.
Default is a concatenation of \code{ident_to_rename}, "best.match", and \code{to_suffix} using \code{kpp} function.}

\item{plot_suffix}{A string. The suffix to add to the final UMAP.}

\item{...}{Additional parameters to be passed to \code{.replace_by_most_frequent_categories} function.}
}
\value{
An updated version of \code{obj} with an additional column in \code{obj@meta.data} named as \code{new_ident_name}
representing the new identity. The function also generates a UMAP plot based on this new identity.
}
\description{
This function matches the best identity from \code{ident_to_rename} to \code{reference_ident} in an object,
in other words, it replaces original categories with the most frequent ones from the reference,
hence helps to filter out less important categories.
}
\examples{
\dontrun{
updated_obj <- matchBestIdentity(my_obj, "origin_identity", "target_identity")
}
}
\seealso{
\code{\link[clUMAP]{clUMAP}}, \code{\link[kpp]{kpp}}, \code{\link[FixPlotName]{FixPlotName}},
\code{\link[.replace_by_most_frequent_categories]{.replace_by_most_frequent_categories}}
}
