% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MULTI-seq.Utils.R
\name{mSeq.map.all96.BCs}
\alias{mSeq.map.all96.BCs}
\title{mSeq.map.all96.BCs}
\usage{
mSeq.map.all96.BCs(
  readTable = readTable,
  CellIDs = CellIDs,
 
    path2allBCs = "~/Google_Drive/Science/IMBA/MULTI.seq/from.US/All.MULTI-seq_barcodes.Mar2019.tsv"
)
}
\arguments{
\item{readTable}{PARAM_DESCRIPTION, Default: readTable}

\item{CellIDs}{PARAM_DESCRIPTION, Default: CellIDs}

\item{path2allBCs}{PARAM_DESCRIPTION, Default: '~/Google_Drive/Science/IMBA/MULTI.seq/from.US/All.MULTI-seq_barcodes.Mar2019.tsv'}
}
\description{
mSeq.map.all96.BCs
}
\examples{
\dontrun{
if(interactive()){
 bar.table <-mSeq.map.all96.BCs(readTable = readTable, CellIDs = CellIDs)
 }
}
}
