% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{multiFeatureHeatmap.A4}
\alias{multiFeatureHeatmap.A4}
\title{multiFeatureHeatmap.A4}
\usage{
multiFeatureHeatmap.A4(
  obj = combined.obj,
  list.of.genes,
  gene.per.page = 5,
  group.cells.by = "batch",
  plot.reduction = "umap",
  cex = iround(3/gene.per.page),
  sep_scale = F,
  gene.min.exp = "q5",
  gene.max.exp = "q95",
  jpeg.res = 225,
  jpeg.q = 90,
  ...
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{list.of.genes}{A list of genes to plot. No default.}

\item{gene.per.page}{Number of genes to plot per page. Default: 5}

\item{group.cells.by}{Cell grouping variable for the heatmap. Default: 'batch'}

\item{plot.reduction}{Dimension reduction technique to use for plots. Default: 'umap'}

\item{cex}{Point size in the plot. Default: iround(3/gene.per.page)}

\item{sep_scale}{Logical, whether to scale the features separately. Default: FALSE}

\item{gene.min.exp}{Minimum gene expression level for plotting. Default: 'q5'}

\item{gene.max.exp}{Maximum gene expression level for plotting. Default: 'q95'}

\item{jpeg.res}{Resolution of the jpeg output. Default: 225}

\item{jpeg.q}{Quality of the jpeg output. Default: 90}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Save multiple FeatureHeatmaps from a list of genes on A4 jpeg.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tictoc]{tic}}
}
