% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{multiFeaturePlot.A4}
\alias{multiFeaturePlot.A4}
\title{multiFeaturePlot.A4}
\usage{
multiFeaturePlot.A4(
  list.of.genes,
  obj = combined.obj,
  foldername = substitute(list.of.genes),
  plot.reduction = "umap",
  intersectionAssay = c("RNA", "integrated")[1],
  layout = c("tall", "wide", FALSE)[2],
  colors = c("grey", "red"),
  nr.Col = 2,
  nr.Row = 4,
  cex = round(0.1/(nr.Col * nr.Row), digits = 2),
  gene.min.exp = "q01",
  gene.max.exp = "q99",
  subdir = T,
  prefix = NULL,
  suffix = NULL,
  background_col = "white",
  aspect.ratio = c(FALSE, 0.6)[2],
  saveGeneList = FALSE,
  w = wA4,
  h = hA4,
  scaling = 1,
  format = c("jpg", "pdf", "png")[1],
  ...
)
}
\arguments{
\item{list.of.genes}{PARAM_DESCRIPTION}

\item{obj}{Seurat object, Default: combined.obj}

\item{foldername}{PARAM_DESCRIPTION, Default: substitute(list.of.genes)}

\item{plot.reduction}{PARAM_DESCRIPTION, Default: 'umap'}

\item{intersectionAssay}{PARAM_DESCRIPTION, Default: c("RNA", "integrated")\link{1}}

\item{layout}{PARAM_DESCRIPTION, Default: c("tall", "wide", FALSE)\link{2}}

\item{colors}{PARAM_DESCRIPTION, Default: c("grey", "red")}

\item{nr.Col}{PARAM_DESCRIPTION, Default: 2}

\item{nr.Row}{PARAM_DESCRIPTION, Default: 4}

\item{cex}{Point size, Default: round(0.1/(nr.Col * nr.Row), digits = 2)}

\item{gene.min.exp}{PARAM_DESCRIPTION, Default: 'q01'}

\item{gene.max.exp}{PARAM_DESCRIPTION, Default: 'q99'}

\item{subdir}{PARAM_DESCRIPTION, Default: T}

\item{prefix}{PARAM_DESCRIPTION, Default: NULL}

\item{suffix}{A suffix added to the filename, Default: NULL}

\item{background_col}{background color def: White}

\item{aspect.ratio}{Fix the aspect ratio? Default: Yes, at 0.6}

\item{saveGeneList}{PARAM_DESCRIPTION, Default: FALSE}

\item{w}{width of the plot, Default: wA4}

\item{h}{height of the plot, Default: hA4}

\item{scaling}{PARAM_DESCRIPTION, Default: 1}

\item{format}{PARAM_DESCRIPTION, Default: c("jpg", "pdf", "png")\link{1}}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Save multiple FeaturePlots, as jpeg, on A4 for each gene, which are stored as a list of gene names. #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tictoc]{tic}}
\code{\link[cowplot]{plot_grid}}
}
