% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{PctCellsAboveX}
\alias{PctCellsAboveX}
\title{Percentage of Cells Above Threshold}
\usage{
PctCellsAboveX(
  obj = combined.obj,
  feature = "TOP2A",
  ident = GetNamedClusteringRuns(obj = obj, v = FALSE)[1],
  threshold = 2,
  suffix = ppp(substitute(obj), ncol(obj), "thr", threshold),
  box = FALSE,
  ident.box = NULL,
  subset_ident = NULL,
  subset_values = NULL,
  omit.na = TRUE,
  assay = "RNA",
  plot = TRUE,
  caption = NULL,
  ylab = "\% cells above threshold",
  ...
)
}
\arguments{
\item{obj}{A Seurat object. Default: combined.obj.}

\item{feature}{The feature to evaluate.}

\item{ident}{The identity class to split the data by. Default: GetNamedClusteringRuns(obj)\link{1}.}

\item{threshold}{The threshold value to evaluate the feature against. Default: 2.}

\item{box}{Logical value indicating whether to plot the boxplot. Default: \code{TRUE}.}

\item{subset_ident}{The identity class to subset the data by. Default: NULL.}

\item{subset_values}{The values of the identity class to keep in the subset. Default: NULL.}

\item{omit.na}{Logical value indicating whether to omit NA values. Default: \code{TRUE}.}

\item{assay}{The assay to use for feature extraction. Default: 'RNA'.}

\item{plot}{Logical value indicating whether to plot the results. Default: \code{TRUE}.}

\item{ylab}{The label for the y-axis of the plot. Default: "\% cells above threshold".}

\item{...}{Additional parameters to pass to the plotting function.}

\item{box.ident}{The identity class to split the data by for individual dots in the boxplot.
Ident will be used for the boxes displayed (matching the barplot).  Default: NULL.}
}
\value{
A named vector with the percentage of cells above the threshold for each identity class.
}
\description{
This function calculates the percentage of cells above a specified threshold for a given
feature in a Seurat object. It can subset the data based on a specified identity and values.
}
\examples{
\dontrun{
PctCellsAboveX(obj = seurat_object, feature = "GeneA", ident = "CellType", threshold = 1.5)
}
}
