% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{PctCellsExpressingGenes}
\alias{PctCellsExpressingGenes}
\title{PctCellsExpressingGenes}
\usage{
PctCellsExpressingGenes(
  genes,
  obj,
  assay = "RNA",
  min.expr = 1,
  ident = NULL,
  max.idents = 100
)
}
\arguments{
\item{genes}{A character vector specifying the genes of interest. Must be a non-empty character vector.}

\item{obj}{A Seurat object containing single-cell data.}

\item{assay}{The assay to use for expression data. Default: "RNA".}

\item{min.expr}{The minimum expression level to consider a gene as "expressed". Default: 1.}

\item{ident}{A categorical variable from the metadata of the Seurat object. If NULL, returns overall
proportions. Default: NULL.}

\item{max.idents}{Maximum number of unique values allowed in the \code{ident} variable. Default: 100.}
}
\value{
A named vector if \code{ident} is NULL, containing the proportion of cells co-expressing all genes
(AND), the proportion expressing any gene (OR), and the proportion expressing each gene individually.
If \code{ident} is provided, returns a matrix with rows representing categories and columns representing
expression proportions.
}
\description{
Calculates the proportion of cells expressing one or more specified genes using a Seurat
object as input.
}
\examples{
\dontrun{
# Load the Seurat object (example)
library(Seurat)
`combined.obj <- readRDS("path/to/your/seurat_object.rds")`

# Define genes of interest
genes <- c("TOP2A", "MAP2")
# Call the function
PctCellsExpressingGenes(genes = genes, obj = combined.obj)
# Call the function with ident
PctCellsExpressingGenes(genes = genes, obj = combined.obj, ident = "cluster")
}

}
