% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot.clust.size.distr}
\alias{plot.clust.size.distr}
\title{plot.clust.size.distr}
\usage{
\method{plot}{clust.size.distr}(
  obj = combined.obj,
  ident = GetClusteringRuns()[2],
  plot = T,
  thr.hist = 30,
  ...
)
}
\arguments{
\item{obj}{The Seurat object to be used for the plot. Default: combined.obj.}

\item{ident}{The identity or clustering to be used for the plot. Default: The second result from GetClusteringRuns().}

\item{plot}{A logical value indicating whether to plot the data. If FALSE, a vector of the cluster size distribution will be returned. Default: TRUE.}

\item{thr.hist}{A threshold for the number of clusters above which a histogram will be plotted instead of a bar plot. Default: 30.}

\item{...}{Additional arguments to be passed to the internally called plotting functions.}
}
\description{
Creates a bar plot or histogram of the cluster size distribution from a given Seurat object.
}
\examples{
\dontrun{
if (interactive()) {
  plot.clust.size.distr()
}
}
}
