% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot.Gene.Cor.Heatmap}
\alias{plot.Gene.Cor.Heatmap}
\title{plot.Gene.Cor.Heatmap}
\usage{
\method{plot}{Gene.Cor.Heatmap}(
  genes = WU.2017.139.IEGsf,
  assay.use = "RNA",
  slot.use = c("data", "scale.data", "data.imputed")[1],
  quantileX = 0.95,
  min.g.cor = 0.3,
  calc.COR = FALSE,
  cutRows = NULL,
  cutCols = cutRows,
  obj = combined.obj,
  ...
)
}
\arguments{
\item{genes}{Genes of iinterest, Default: WU.2017.139.IEGsf}

\item{assay.use}{PARAM_DESCRIPTION, Default: 'RNA'}

\item{slot.use}{PARAM_DESCRIPTION, Default: c("data", "scale.data", "data.imputed")\link{1}}

\item{quantileX}{Quantile level, Default: 0.95}

\item{min.g.cor}{PARAM_DESCRIPTION, Default: 0.3}

\item{calc.COR}{PARAM_DESCRIPTION, Default: FALSE}

\item{cutRows}{PARAM_DESCRIPTION, Default: NULL}

\item{cutCols}{PARAM_DESCRIPTION, Default: cutRows}

\item{obj}{Seurat object, Default: combined.obj}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot a gene correlation heatmap.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
