% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot.GeneExpHist}
\alias{plot.GeneExpHist}
\title{plot.GeneExpHist}
\usage{
\method{plot}{GeneExpHist}(
  obj = cobj.H9.L92,
  genes = c("MALAT1", "MT-CO1", "MT-CO2", "MT-CYB", "TMSB4X", "KAZN"),
  assay = "RNA",
  slot_ = "data",
  thr_expr = 10,
  suffix = NULL,
  xlab = p0("log10(Summed UMI count @", slot_, ")"),
  return_cells_passing = TRUE,
  quantile_thr = 0.95,
  return_quantile,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
plot.GeneExpHist
}
