% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot.Metadata.Cor.Heatmap}
\alias{plot.Metadata.Cor.Heatmap}
\title{plot.Metadata.Cor.Heatmap}
\usage{
\method{plot}{Metadata.Cor.Heatmap}(
  columns = c("nCount_RNA", "nFeature_RNA", "percent.mito", "percent.ribo"),
  cormethod = c("pearson", "spearman")[1],
  main = paste("Metadata", cormethod, "correlations"),
  obj = combined.obj,
  w = 10,
  h = w,
  ...
)
}
\arguments{
\item{columns}{A vector of column names for which to calculate correlations. Default: c("nCount_RNA", "nFeature_RNA", "percent.mito", "percent.ribo").}

\item{cormethod}{The method to calculate correlations, either "pearson" or "spearman". Default: "pearson".}

\item{main}{The main title for the plot. Default: "Metadata correlations" followed by the correlation method.}

\item{obj}{The main Seurat object used for calculations. Default: combined.obj.}

\item{w}{The width of the plot. Default: 10.}

\item{h}{The height of the plot. Default: width of the plot (w).}

\item{...}{Additional parameters passed to the internally called functions.}
}
\description{
Plots a heatmap of metadata correlation values.
}
\seealso{
\code{\link[ggcorrplot]{ggcorrplot}}
}
