% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot.UMAP.tSNE.sidebyside}
\alias{plot.UMAP.tSNE.sidebyside}
\title{plot.UMAP.tSNE.sidebyside}
\usage{
\method{plot}{UMAP.tSNE.sidebyside}(
  obj = combined.obj,
  grouping = "res.0.6",
  no_legend = F,
  do_return = TRUE,
  do_label = T,
  label_size = 10,
  vector_friendly = TRUE,
  cells_use = NULL,
  no_axes = T,
  pt_size = 0.5,
  name.suffix = NULL,
  width = hA4,
  heigth = 1.75 * wA4,
  filetype = "pdf",
  ...
)
}
\arguments{
\item{obj}{Seurat object. Default: combined.obj}

\item{grouping}{Variable to group cells by. Default: 'res.0.6'}

\item{no_legend}{Logical, whether to display legend. Default: FALSE}

\item{do_return}{Logical, whether to return plot object. Default: TRUE}

\item{do_label}{Logical, whether to display labels. Default: TRUE}

\item{label_size}{Size of labels. Default: 10}

\item{vector_friendly}{Logical, whether to optimize for vector outputs. Default: TRUE}

\item{cells_use}{A vector of cell names to use for the plot. Default: NULL}

\item{no_axes}{Logical, whether to hide axes. Default: TRUE}

\item{pt_size}{Size of points. Default: 0.5}

\item{name.suffix}{Suffix to append to the plot's name. Default: NULL}

\item{width}{Width of the plot. Default: hA4}

\item{heigth}{Height of the plot. Default: 1.75 * wA4}

\item{filetype}{Filetype to save plot as. Default: 'pdf'}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot a UMAP and tSNE side by side.
}
\seealso{
\code{\link[cowplot]{save_plot}}
}
