% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Plot3D.ListOfCategories}
\alias{Plot3D.ListOfCategories}
\title{Plot3D.ListOfCategories}
\usage{
Plot3D.ListOfCategories(
  obj = combined.obj,
  annotate.by = "integrated_snn_res.0.7",
  cex = 1.25,
  default.assay = c("integrated", "RNA")[2],
  ListOfCategories = c("v.project", "experiment", "Phase", "integrated_snn_res.0.7"),
  SubFolderName = ppp("plot3D", substitute(ListOfCategories))
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{annotate.by}{PARAM_DESCRIPTION, Default: 'integrated_snn_res.0.7'}

\item{cex}{Point size, Default: 1.25}

\item{default.assay}{PARAM_DESCRIPTION, Default: c("integrated", "RNA")\link{2}}

\item{ListOfCategories}{PARAM_DESCRIPTION, Default: c("v.project", "experiment", "Phase", "integrated_snn_res.0.7")}

\item{SubFolderName}{PARAM_DESCRIPTION, Default: ppp("plot3D", substitute(ListOfCategories))}
}
\description{
Plot and save list of 3D UMAP ot tSNE plots using plotly. #
}
\examples{
\dontrun{
if(interactive()){
 categ3Dplots <- c("v.project","experiment", "Phase", "integrated_snn_res.0.7", "Area", "Individual", "Type")
 Plot3D.ListOfCategories(obj = combined.obj, ListOfCategories = categ3Dplots)
 }
}
}
