% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{Plot3D.ListOfGenes}
\alias{Plot3D.ListOfGenes}
\title{Plot3D.ListOfGenes}
\usage{
Plot3D.ListOfGenes(
  obj = combined.obj,
  annotate.by = "integrated_snn_res.0.7",
  opacity = 0.5,
  cex = 1.25,
  default.assay = c("integrated", "RNA")[2],
  ListOfGenes = c("BCL11B", "FEZF2", "EOMES", "DLX6-AS1", "HOPX", "DDIT4"),
  SubFolderName = ppp("plot3D", substitute(ListOfGenes))
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{annotate.by}{PARAM_DESCRIPTION, Default: 'integrated_snn_res.0.7'}

\item{opacity}{PARAM_DESCRIPTION, Default: 0.5}

\item{cex}{Point size, Default: 1.25}

\item{default.assay}{PARAM_DESCRIPTION, Default: c("integrated", "RNA")\link{2}}

\item{ListOfGenes}{PARAM_DESCRIPTION, Default: c("BCL11B", "FEZF2", "EOMES", "DLX6-AS1", "HOPX", "DDIT4")}

\item{SubFolderName}{PARAM_DESCRIPTION, Default: ppp("plot3D", substitute(ListOfGenes))}
}
\description{
Plot and save list of 3D UMAP ot tSNE plots using plotly. #
}
\examples{
\dontrun{
if(interactive()){
 CellTypeMarkers <- c(  "PGK1", "CTIP2" = "BCL11B" , "FEZF2", "EOMES", "DLX6-AS1", "HOPX", "DDIT4","TOP2A", "PTGDS", "EDNRB", "EGFR", "SCGN", "NR2F2", "EMX2", "GAD2", "DLX2", "SATB2")
 Plot3D.ListOfGenes(obj = combined.obj, ListOfGenes = CellTypeMarkers)
 }
}
}
