% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{Plot3D.ListOfGenes}
\alias{Plot3D.ListOfGenes}
\title{Plot3D.ListOfGenes}
\usage{
Plot3D.ListOfGenes(
  obj = combined.obj,
  annotate.by = "integrated_snn_res.0.7",
  opacity = 0.5,
  cex = 1.25,
  default.assay = c("integrated", "RNA")[2],
  ListOfGenes = c("BCL11B", "FEZF2", "EOMES", "DLX6-AS1", "HOPX", "DDIT4"),
  SubFolderName = ppp("plot3D", substitute_deparse(ListOfGenes))
)
}
\arguments{
\item{obj}{Seurat object to be used for the plot. Default: combined.obj}

\item{annotate.by}{Variable to annotate the clusters by. Default: 'integrated_snn_res.0.7'}

\item{opacity}{Opacity for the plot points. Default: 0.5}

\item{cex}{Point size for the plot. Default: 1.25}

\item{default.assay}{Default assay to be used from the Seurat object. Default: second entry from c("integrated", "RNA")}

\item{ListOfGenes}{List of genes to be plotted. Default: c("BCL11B", "FEZF2", "EOMES", "DLX6-AS1", "HOPX", "DDIT4")}

\item{SubFolderName}{Name of the subfolder where the plots will be saved. Default: a subfolder named 'plot3D' concatenated with the list of genes.}
}
\description{
Plot and save list of 3D UMAP or tSNE plots using plotly.
}
\examples{
\dontrun{
if (interactive()) {
  CellTypeMarkers <- c("PGK1", "CTIP2" = "BCL11B", "FEZF2", "EOMES", "DLX6-AS1", "HOPX", "DDIT4", "TOP2A", "PTGDS", "EDNRB", "EGFR", "SCGN", "NR2F2", "EMX2", "GAD2", "DLX2", "SATB2")
  Plot3D.ListOfGenes(obj = combined.obj, ListOfGenes = CellTypeMarkers)
}
}
}
