% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot3D.umap.gene}
\alias{plot3D.umap.gene}
\title{plot3D.umap.gene}
\usage{
plot3D.umap.gene(
  gene = "TOP2A",
  obj = combined.obj,
  quantileCutoff = 0.99,
  def.assay = c("integrated", "RNA")[2],
  suffix = NULL,
  AutoAnnotBy = GetNamedClusteringRuns(obj)[1],
  alpha = 0.5,
  dotsize = 1.25
)
}
\arguments{
\item{gene}{gene of interest, Default: 'TOP2A'}

\item{obj}{Seurat object, Default: combined.obj}

\item{quantileCutoff}{PARAM_DESCRIPTION, Default: 0.99}

\item{def.assay}{PARAM_DESCRIPTION, Default: c("integrated", "RNA")\link{2}}

\item{suffix}{A suffix added to the filename, Default: NULL}

\item{AutoAnnotBy}{PARAM_DESCRIPTION, Default: GetNamedClusteringRuns(obj)\link{1}}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.5}

\item{dotsize}{PARAM_DESCRIPTION, Default: 1.25}
}
\description{
Plot a 3D umap with gene expression. Uses plotly. Based on github.com/Dragonmasterx87. #
}
\examples{
\dontrun{
if(interactive()){
 plot3D.umap.gene(obj = combined.obj, gene = "DDIT4", quantileCutoff = .95)
 plot3D.umap.gene(obj = combined.obj, gene = "percent.mito", quantileCutoff = .95) # for continous meta variables
 plot3D.umap.gene(obj = combined.obj, gene = "nFeature_RNA", quantileCutoff = .95) # for continous meta variables
 }
}
}
