% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plot3D.umap}
\alias{plot3D.umap}
\title{plot3D.umap}
\usage{
plot3D.umap(
  category = "v.project",
  obj = combined.obj,
  suffix = NULL,
  AutoAnnotBy = GetNamedClusteringRuns(obj)[1],
  dotsize = 1.25
)
}
\arguments{
\item{category}{PARAM_DESCRIPTION, Default: 'v.project'}

\item{obj}{Seurat object, Default: combined.obj}

\item{suffix}{A suffix added to the filename, Default: NULL}

\item{AutoAnnotBy}{PARAM_DESCRIPTION, Default: GetNamedClusteringRuns(obj)\link{1}}

\item{dotsize}{PARAM_DESCRIPTION, Default: 1.25}
}
\description{
Plot a 3D umap based on one of the metadata columns. Uses plotly. Based on github.com/Dragonmasterx87. #
}
\examples{
\dontrun{
if(interactive()){
 plot3D.umap(combined.obj, category = "Phase")
 }
}
}
