% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{plot3D.umap}
\alias{plot3D.umap}
\title{plot3D.umap}
\usage{
plot3D.umap(
  category = GetNamedClusteringRuns(obj)[1],
  annotate.by = GetNamedClusteringRuns(obj)[1],
  obj = combined.obj,
  suffix = NULL,
  dotsize = 1.25,
  col.names = c("umap_1", "umap_2", "umap_3"),
  ...
)
}
\arguments{
\item{category}{The metadata column based on which the 3D UMAP will be plotted.
Default: First returned result from GetNamedClusteringRuns(obj) function.}

\item{annotate.by}{The cluster or grouping to be used for automatic annotation.
Default: First returned result from GetNamedClusteringRuns(obj) function.}

\item{obj}{The Seurat object for which the 3D umap plot will be generated. Default: combined.obj}

\item{suffix}{A suffix added to the filename. Default: NULL}

\item{dotsize}{The size of the dots in the plot. Default: 1.25}

\item{...}{Pass any other parameter to the internally called \code{plotly::plot_ly}.}
}
\description{
Plot a 3D umap based on one of the metadata columns. Uses plotly. Based on github.com/Dragonmasterx87.
}
\examples{
\dontrun{
if (interactive()) {
  plot3D.umap(combined.obj, category = "Phase")
}
}
}
