% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{PlotFilters}
\alias{PlotFilters}
\title{Plot filtering thresholds and distributions}
\usage{
PlotFilters(
  ls.obj,
  par.ls = p,
  parentdir = OutDirOrig,
  suffices = names(ls.obj),
  filetype = ".jpg",
  below.mito = par.ls$thr.lp.mito,
  above.mito = par.ls$thr.hp.mito,
  below.ribo = par.ls$thr.lp.ribo,
  above.ribo = par.ls$thr.hp.ribo,
  below.nFeature_RNA = if ("quantile.thr.lp.nFeature_RNA" \%in\% names(par.ls))
    par.ls$quantile.thr.lp.nFeature_RNA else par.ls$thr.lp.nFeature_RNA,
  above.nFeature_RNA = par.ls$thr.hp.nFeature_RNA,
  subdir = FixPlotName("Filtering.plots", "mito", par.ls$thr.hp.mito, par.ls$thr.lp.mito,
    "ribo", par.ls$thr.hp.ribo, par.ls$thr.lp.ribo, "nFeature", below.nFeature_RNA,
    above.nFeature_RNA),
  transparency = 0.25,
  cex = 0.75,
  theme.used = theme_bw(base_size = 18)
)
}
\arguments{
\item{ls.obj}{A list of Seurat objects to be analyzed. Default: \code{ls.Seurat}.}

\item{par.ls}{A list of parameters used to source default thresholds (e.g. \code{thr.lp.mito},
\code{thr.hp.mito}, \code{thr.lp.ribo}, \code{thr.hp.ribo}, \code{thr.lp.nFeature_RNA}, \code{thr.hp.nFeature_RNA}). Default: \code{p}.}

\item{parentdir}{The parent directory where the plots will be stored. Default: \code{OutDirOrig}.}

\item{suffices}{Suffixes used in the output plot file names. Default: Names of the Seurat objects in \code{ls.obj}.}

\item{filetype}{The file type of the output plot images. Default: \code{'.png'}.}

\item{below.mito}{Lower bound of mitochondrial content. Default: \code{p$thr.lp.mito}.}

\item{above.mito}{Upper bound of mitochondrial content. Default: \code{p$thr.hp.mito}.}

\item{below.ribo}{Lower bound of ribosomal content. Default: \code{p$thr.lp.ribo}.}

\item{above.ribo}{Upper bound of ribosomal content. Default: \code{p$thr.hp.ribo}.}

\item{below.nFeature_RNA}{Lower bound of RNA features. Default: \code{p$thr.lp.nFeature_RNA}.}

\item{above.nFeature_RNA}{Upper bound of RNA features. Default: \code{p$thr.hp.nFeature_RNA}.}

\item{subdir}{Subdirectory within \code{parentdir} where plots will be stored. Default: generated using a call to \code{kpp()}.}

\item{transparency}{Point transparency on scatter plots. Default: 0.25.}

\item{cex}{Size of points on scatter plots. Default: 0.75.}

\item{theme.used}{A \code{ggplot2} theme for all plots. Default: \code{theme_bw(base_size = 18)}.}
}
\description{
This function plots the filtering thresholds and distributions for Seurat objects,
using four panels to highlight the relationship between gene- and UMI-counts, and the
ribosomal- and mitochondrial-content.  !! Default arguments assume that \code{p} is a list of
parameters, present in the global environment, with elements \code{thr.lp.mito}, \code{thr.hp.mito},
\code{thr.lp.ribo}, \code{thr.hp.ribo}, \code{thr.lp.nFeature_RNA}, and \code{thr.hp.nFeature_RNA}.
}
\examples{
\dontrun{
if (interactive()) {
  # !! Default arguments assume that `p` is a list of parameters, present in the global
  # environment, with elements `thr.lp.mito`, `thr.hp.mito`, `thr.lp.ribo`, `thr.hp.ribo`,
  # `thr.lp.nFeature_RNA`, and `thr.hp.nFeature_RNA`.
  PlotFilters(ls.Seurat)
}
}

}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{labs}},
\code{\link[ggplot2]{geom_point}}
}
