% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{PlotFilters}
\alias{PlotFilters}
\title{PlotFilters}
\usage{
PlotFilters(
  ls.obj = ls.Seurat,
  parentdir = OutDirOrig,
  suffices = names(ls.obj),
  filetype = ".png",
  below.mito = p$thr.lp.mito,
  above.mito = p$thr.hp.mito,
  below.ribo = p$thr.lp.ribo,
  above.ribo = p$thr.hp.ribo,
  below.nFeature_RNA = if ("quantile.thr.lp.nFeature_RNA" \%in\% names(p))
    p$quantile.thr.lp.nFeature_RNA else p$thr.lp.nFeature_RNA,
  above.nFeature_RNA = p$thr.hp.nFeature_RNA,
  subdir = kpp("Filtering.plots", "mito", p$thr.hp.mito, p$thr.lp.mito, "ribo",
    p$thr.hp.ribo, p$thr.lp.ribo, "nFeature", p$thr.hp.nFeature_RNA, below.nFeature_RNA,
    "/"),
  transparency = 0.25,
  cex = 0.75,
  theme.used = theme_bw(base_size = 18),
  LabelDistFromTop = 200
)
}
\arguments{
\item{ls.obj}{List of Seurat objects, Default: ls.Seurat}

\item{parentdir}{PARAM_DESCRIPTION, Default: OutDirOrig}

\item{suffices}{PARAM_DESCRIPTION, Default: names(ls.obj)}

\item{filetype}{PARAM_DESCRIPTION, Default: '.png'}

\item{below.mito}{PARAM_DESCRIPTION, Default: p$thr.lp.mito}

\item{above.mito}{PARAM_DESCRIPTION, Default: p$thr.hp.mito}

\item{below.ribo}{PARAM_DESCRIPTION, Default: p$thr.lp.ribo}

\item{above.ribo}{PARAM_DESCRIPTION, Default: p$thr.hp.ribo}

\item{below.nFeature_RNA}{PARAM_DESCRIPTION, Default: p$thr.lp.nFeature_RNA}

\item{above.nFeature_RNA}{PARAM_DESCRIPTION, Default: p$thr.hp.nFeature_RNA}

\item{subdir}{PARAM_DESCRIPTION, Default: kpp("Filtering.plots", "mito", p$thr.hp.mito, p$thr.lp.mito,
"ribo", p$thr.hp.ribo, p$thr.lp.ribo, "nFeature", p$thr.hp.nFeature_RNA,
p$thr.lp.nFeature_RNA, "/")}

\item{transparency}{PARAM_DESCRIPTION, Default: 0.25}

\item{cex}{Point size, Default: 0.75}

\item{theme.used}{PARAM_DESCRIPTION, Default: theme_bw(base_size = 18)}

\item{LabelDistFromTop}{PARAM_DESCRIPTION, Default: 200}
}
\description{
Plot filtering threshold and distributions, using four panels to highlight the relation between Gene- and UMI-count, ribosomal- and mitochondrial-content. #
}
\examples{
\dontrun{
if(interactive()){
 PlotFilters(ls.Seurat)
 }
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{labs}}, \code{\link[ggplot2]{geom_point}}
}
