% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{PlotFilters}
\alias{PlotFilters}
\title{PlotFilters}
\usage{
PlotFilters(
  ls.obj = ls.Seurat,
  parentdir = OutDirOrig,
  suffices = names(ls.obj),
  filetype = ".png",
  below.mito = p$thr.lp.mito,
  above.mito = p$thr.hp.mito,
  below.ribo = p$thr.lp.ribo,
  above.ribo = p$thr.hp.ribo,
  below.nFeature_RNA = if ("quantile.thr.lp.nFeature_RNA" \%in\% names(p))
    p$quantile.thr.lp.nFeature_RNA else p$thr.lp.nFeature_RNA,
  above.nFeature_RNA = p$thr.hp.nFeature_RNA,
  subdir = kpp("Filtering.plots", "mito", p$thr.hp.mito, p$thr.lp.mito, "ribo",
    p$thr.hp.ribo, p$thr.lp.ribo, "nFeature", p$thr.hp.nFeature_RNA, below.nFeature_RNA,
    "/"),
  transparency = 0.25,
  cex = 0.75,
  theme.used = theme_bw(base_size = 18),
  LabelDistFromTop = 200
)
}
\arguments{
\item{ls.obj}{A list of Seurat objects to be analyzed. Default is ls.Seurat.}

\item{parentdir}{A string representing the parent directory where the plots will be stored. Default is OutDirOrig.}

\item{suffices}{A vector of strings that will be used as suffixes in the output plot file names. Default is the names of the Seurat objects in ls.obj.}

\item{filetype}{A string indicating the file type of the output plot images. Default is '.png'.}

\item{below.mito}{Numeric threshold for the lower bound of mitochondrial content. Default is p$thr.lp.mito.}

\item{above.mito}{Numeric threshold for the upper bound of mitochondrial content. Default is p$thr.hp.mito.}

\item{below.ribo}{Numeric threshold for the lower bound of ribosomal content. Default is p$thr.lp.ribo.}

\item{above.ribo}{Numeric threshold for the upper bound of ribosomal content. Default is p$thr.hp.ribo.}

\item{below.nFeature_RNA}{Numeric threshold for the lower bound of RNA features. Default is p$thr.lp.nFeature_RNA.}

\item{above.nFeature_RNA}{Numeric threshold for the upper bound of RNA features. Default is p$thr.hp.nFeature_RNA.}

\item{subdir}{A string specifying the subdirectory within the parent directory where the plots will be stored. Default is generated using a call to kpp().}

\item{transparency}{Numeric value controlling the transparency of points on the scatter plots. Default is 0.25.}

\item{cex}{Numeric value controlling the size of points on the scatter plots. Default is 0.75.}

\item{theme.used}{A ggplot2 theme to be used for all plots. Default is theme_bw(base_size = 18).}

\item{LabelDistFromTop}{Numeric value specifying the distance from the top of the plot for the label placement. Default is 200.}
}
\description{
Plot filtering threshold and distributions, using four panels to highlight the relation between Gene- and UMI-count, ribosomal- and mitochondrial-content.
}
\examples{
\dontrun{
if (interactive()) {
  PlotFilters(ls.Seurat)
}
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{labs}}, \code{\link[ggplot2]{geom_point}}
}
