% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plotTheSoup}
\alias{plotTheSoup}
\title{plotTheSoup}
\usage{
plotTheSoup(
  CellRanger_outs_Dir = "~/Data/114593/114593",
  SeqRun = str_extract(CellRanger_outs_Dir, "[[:alnum:]_]+(?=/outs/)"),
  ls.Alpha = 1
)
}
\arguments{
\item{SeqRun}{PARAM_DESCRIPTION, Default: gsub("\emph{(\link{0-9}+).}", "\\1", x = basename(CellRangerOutputDir))}

\item{CellRangerOutputDir}{PARAM_DESCRIPTION, Default: '~/Data/114593/114593'}
}
\description{
Plot stats about the ambient RNA content in a 10X experiment.
}
\seealso{
\code{\link[Matrix]{colSums}}
\code{\link[tibble]{rownames}}
\code{\link[ggrepel]{geom_label_repel}}
}
