% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{plotTheSoup}
\alias{plotTheSoup}
\title{plotTheSoup}
\usage{
plotTheSoup(
  CellRangerOutputDir = "~/Data/114593/114593",
  SeqRun = gsub("*([0-9]+).*", "\\\\1", x = basename(CellRangerOutputDir))
)
}
\arguments{
\item{CellRangerOutputDir}{PARAM_DESCRIPTION, Default: '~/Data/114593/114593'}

\item{SeqRun}{PARAM_DESCRIPTION, Default: gsub("\emph{(\link{0-9}+).}", "\\1", x = basename(CellRangerOutputDir))}
}
\description{
plotTheSoup
}
\examples{
\dontrun{
if(interactive()){
 plotTheSoup(CellRangerOutputDir = "~/Data/114593/114593" , SeqRun = gsub('*([0-9]+).*','\\\\1', x = basename(CellRangerOutputDir)))
 }
}
}
\seealso{
\code{\link[Matrix]{colSums}}
\code{\link[tibble]{rownames}}
\code{\link[ggrepel]{geom_label_repel}}
}
