% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{PlotTopGenes}
\alias{PlotTopGenes}
\title{PlotTopGenes}
\usage{
PlotTopGenes(obj = combined.obj, n = 32)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{n}{PARAM_DESCRIPTION, Default: 32}
}
\description{
Plot the highest expressed genes on umaps, in a subfolder. Requires calling calc.q90.Expression.and.set.all.genes before. #
}
\examples{
\dontrun{
if(interactive()){
 PlotTopGenes()
 }
}
}
