% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{PlotTopGenesPerCluster}
\alias{PlotTopGenesPerCluster}
\title{PlotTopGenesPerCluster}
\usage{
PlotTopGenesPerCluster(
  obj = combined.obj,
  cl_res = res,
  nrGenes = p$n.markers,
  order.by = c("combined.score", "avg_log2FC", "p_val_adj")[1],
  df_markers = obj@misc$df.markers[[paste0("res.", cl_res)]]
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{cl_res}{PARAM_DESCRIPTION, Default: res}

\item{nrGenes}{PARAM_DESCRIPTION, Default: p$n.markers}

\item{order.by}{Sort output tibble by which column, Default: c("combined.score", "avg_log2FC", "p_val_adj")\link{1}}

\item{df_markers}{PARAM_DESCRIPTION, Default: combined.obj@misc$df.markers[\link{paste0("res.", cl_res)}]}
}
\description{
Plot the top N diff. exp. genes in each cluster
}
\examples{
\dontrun{
if(interactive()){
 PlotTopGenesPerCluster(obj = combined.obj, cl_res = 0.5, nrGenes = p$'n.markers')
 }
}
}
