% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{PlotUpdateStats}
\alias{PlotUpdateStats}
\title{PlotUpdateStats}
\usage{
PlotUpdateStats(
  mat = UpdateStatMat,
  column.names = c("Updated (\%)", "Updated (Nr.)")
)
}
\arguments{
\item{mat}{PARAM_DESCRIPTION, Default: UpdateStatMat}

\item{column.names}{PARAM_DESCRIPTION, Default: c("Updated (\%)", "Updated (Nr.)")}
}
\description{
Scatter plot of update stats. #
}
\examples{
\dontrun{
if(interactive()){
 PlotUpdateStats(mat = result.of.GetUpdateStats)
 }
}
}
