% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{points_on_curve.principal_curve}
\alias{points_on_curve.principal_curve}
\title{points_on_curve.principal_curve}
\usage{
\method{points_on_curve}{principal_curve}(curve, lambda, ...)
}
\arguments{
\item{curve}{PARAM_DESCRIPTION}

\item{lambda}{PARAM_DESCRIPTION}

\item{...}{Pass any other parameter to the internally called functions (most of them should  work).}
}
\description{
points_on_curve.principal_curve
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
