% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slingshot.Utils.R
\name{points_on_curve}
\alias{points_on_curve}
\title{points_on_curve}
\usage{
points_on_curve(curve, lambda, ...)
}
\arguments{
\item{curve}{The curve on which to visualize points. Default: PARAM_DEFAULT_VALUE.}

\item{lambda}{The parametric location of points along the curve. Default: PARAM_DEFAULT_VALUE.}
}
\value{
A function call to points_on_curve based on the class of the curve.
}
\description{
Helper function to visualize points on a curve in Slingshot.
}
\details{
This function uses method dispatch based on the class of the 'curve' argument to visualize points on the curve.
}
