% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slingshot.Utils.R
\name{points_on_curve.SlingshotDataSet}
\alias{points_on_curve.SlingshotDataSet}
\title{points_on_curve.SlingshotDataSet}
\usage{
\method{points_on_curve}{SlingshotDataSet}(curve, lambda, ...)
}
\arguments{
\item{curve}{PARAM_DESCRIPTION}

\item{lambda}{PARAM_DESCRIPTION}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Helper to visualize points_on_curve in slingshot.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
