% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{PrctCellExpringGene}
\alias{PrctCellExpringGene}
\title{PrctCellExpringGene}
\usage{
PrctCellExpringGene(genes, group.by = "all", obj = combined.obj)
}
\arguments{
\item{genes}{Genes of iinterest}

\item{group.by}{PARAM_DESCRIPTION, Default: 'all'}

\item{obj}{Seurat object, Default: combined.obj}
}
\description{
From Github/Ryan-Zhu https://github.com/satijalab/seurat/issues/371 #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
