% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{PrctCellExpringGene}
\alias{PrctCellExpringGene}
\title{Proportion of Cells Expressing Given Genes}
\source{
Adapted from Ryan-Zhu on GitHub.
}
\usage{
PrctCellExpringGene(genes, group.by = "all", obj = combined.obj)
}
\arguments{
\item{genes}{Character vector of gene names of interest.}

\item{group.by}{Optional grouping variable for analysis (e.g., cell type). Default: 'all'.}

\item{obj}{Seurat object to analyze. Default: \code{combined.obj}.}
}
\value{
Data frame with genes and their cell expression proportion, optionally grouped.
}
\description{
Calculates the proportion of cells expressing one or more specified genes.
}
\examples{
\dontrun{
PrctCellExpringGene(genes = c("LTB", "GNLY"), obj = combined.obj)
}

}
