% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{prefix_cells_seurat}
\alias{prefix_cells_seurat}
\title{prefix_cells_seurat}
\usage{
prefix_cells_seurat(ls_obj, obj_IDs)
}
\arguments{
\item{ls_obj}{List. A list of Seurat S4 objects}

\item{obj_IDs}{Character vector. A vector of sample IDs corresponding to each Seurat S4 object}
}
\description{
This function adds prefixes from 'obj_IDs' to cell names in Seurat S4 objects from 'ls_obj'
}
\examples{
# ls_obj <- list(seurat_obj1, seurat_obj2)
# obj_IDs <- c("sample1", "sample2")
# ls_obj_prefixed <- prefix_cells_seurat(ls_obj = ls_obj, obj_IDs = obj_IDs)
}
