% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{processSeuratObject}
\alias{processSeuratObject}
\title{Process Seurat Objects in Parallel}
\usage{
processSeuratObject(obj, param.list = p)
}
\arguments{
\item{obj}{A Seurat object to be processed.}

\item{param.list}{A list of parameters used in the processing steps.}
}
\value{
A Seurat object after applying scaling, PCA, UMAP, neighbor finding, and clustering.
}
\description{
Applies a series of Seurat processing steps to each Seurat object in a list.
The operations include scaling data, running PCA, UMAP, finding neighbors, and finding clusters.
This is done in parallel using multiple cores.
}
\examples{
# Assuming ls.Seurat is a list of Seurat objects and params is a list of parameters
# results <- mclapply(ls.Seurat, processSeuratObject, params, mc.cores = 4)
}
