% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{processSeuratObject}
\alias{processSeuratObject}
\title{Process Seurat Objects in Parallel}
\usage{
processSeuratObject(
  obj,
  param.list = p,
  update_gene_symbols = FALSE,
  add.meta.fractions = FALSE,
  precompute = TRUE,
  compute = TRUE,
  save = TRUE,
  plot = TRUE,
  nfeatures = param.list$n.var.genes,
  variables.2.regress = param.list$variables.2.regress.combined,
  harmony.covariates = variables.2.regress,
  n.PC = param.list$n.PC,
  resolutions = param.list$snn_res,
  reduction_input = "pca",
  WorkingDir = getwd(),
  harmony.seurat.implementation = FALSE,
  ...
)
}
\arguments{
\item{obj}{A Seurat object to be processed.}

\item{param.list}{A list of parameters used in the processing steps.}

\item{update_gene_symbols}{A boolean indicating whether to update gene symbols from HGNC. Default: \code{FALSE}.}

\item{add.meta.fractions}{A boolean indicating whether to add metadata for fractions of cells in each cluster. Default: \code{FALSE}.}

\item{precompute}{A boolean indicating whether to compute steps: \code{FindVariableFeatures()},
\code{calc.q99.Expression.and.set.all.genes()},  \code{ScaleData()} and \code{RunPCA()} Default: \code{TRUE}.}

\item{compute}{A boolean indicating whether to compute the steps: \code{IntegrateLayers() / RunHarmony()},
\code{RunUMAP()}, \code{FindNeighbors()}, and \code{FindClusters()}. Default: \code{TRUE}.}

\item{save}{A boolean indicating whether to save the results. Default: \code{TRUE}.}

\item{plot}{A boolean indicating whether to plot the results. Default: \code{TRUE}.}

\item{nfeatures}{The number of variable genes to use. Default: 2000.}

\item{variables.2.regress}{A list of variables to regress out. Default: NULL.}

\item{harmony.covariates}{A list of covariates to use for Harmony. Default: variables.2.regress.}

\item{n.PC}{The number of principal components to use. Default: 30.}

\item{resolutions}{A list of resolutions to use for clustering. Default: c(0.1, 0.2, 0.3, 0.4, 0.5).}

\item{reduction_input}{The reduction method to use as input for clustering & UMAP. Default: "pca".}

\item{WorkingDir}{The working directory to save the results. Default: getwd().}

\item{harmony.seurat.implementation}{A boolean indicating whether to use the Seurat implementation
of Harmony. Default: \code{FALSE}.}

\item{...}{Additional parameters to be passed to \code{ScaleData()}.}
}
\value{
A Seurat object after applying scaling, PCA, UMAP, neighbor finding, and clustering.
}
\description{
Applies a series of Seurat processing steps to each Seurat object in a list.
The operations include scaling data, running PCA, UMAP, finding neighbors, and finding clusters.
This is done in parallel using multiple cores.
}
\examples{
# Assuming ls.Seurat is a list of Seurat objects and params is a list of parameters
# results <- mclapply(ls.Seurat, processSeuratObject, params, mc.cores = 4)

}
