% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{qClusteringUMAPS}
\alias{qClusteringUMAPS}
\title{qClusteringUMAPS}
\usage{
qClusteringUMAPS(
  obj = combined.obj,
  dims = na.omit.strip(GetClusteringRuns(obj)[1:4]),
  prefix = "Clustering.UMAP.Res",
  suffix = "",
  title = sppu(prefix, as.numeric(str_extract(dims, "\\\\d+\\\\.\\\\d+$")), suffix),
  nrow = 2,
  ncol = 2,
  ...
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{dims}{Any numeric metadata columns}
}
\description{
Quickly plot 1-4 clustering resolutions on an a4 page
}
\examples{
qQC.plots.BrainOrg.RV()
}
