% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qFeatureScatter}
\alias{qFeatureScatter}
\title{qFeatureScatter}
\usage{
qFeatureScatter(
  feature1 = "M.RabV.N2c",
  feature2 = "P.RabV.N2c",
  obj = combined.obj,
  ext = "png",
  plot = TRUE,
  ...
)
}
\arguments{
\item{feature1}{PARAM_DESCRIPTION, Default: 'M.RabV.N2c'}

\item{feature2}{PARAM_DESCRIPTION, Default: 'P.RabV.N2c'}

\item{obj}{Seurat object, Default: combined.obj}

\item{ext}{PARAM_DESCRIPTION, Default: 'png'}

\item{plot}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{Pass any other parameter to the internally called functions (most of them should  work).}
}
\description{
qFeatureScatter
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
