% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{qFeatureScatter}
\alias{qFeatureScatter}
\title{Scatter Plot of Two Features in Seurat Object}
\usage{
qFeatureScatter(
  feature1 = "TOP2A",
  feature2 = "ID2",
  obj = combined.obj,
  plot = TRUE,
  logX = FALSE,
  logY = FALSE,
  ...
)
}
\arguments{
\item{feature1}{The first feature for the scatter plot's x-axis.}

\item{feature2}{The second feature for the scatter plot's y-axis.}

\item{obj}{Seurat object containing the data for features.}

\item{plot}{Flag to display the plot within the R session.}

\item{logX}{Apply logarithmic transformation to x-axis values.}

\item{logY}{Apply logarithmic transformation to y-axis values.}

\item{...}{Additional parameters passed to Seurat's \code{FeatureScatter}.}

\item{ext}{File extension for saving the plot, if enabled.}
}
\value{
A \code{ggplot} object of the feature scatter plot if \code{plot} is TRUE.
}
\description{
Generates a scatter plot comparing two features (genes or metrics) from a Seurat
object and optionally saves it. The function wraps around Seurat's \code{FeatureScatter} for
enhanced usability, including optional logarithmic transformations and saving capabilities.
}
\examples{
\dontrun{
# Generate and display a scatter plot for features TOP2A and ID2
qFeatureScatter(feature1 = "TOP2A", feature2 = "ID2", obj = yourSeuratObject)
}

}
\seealso{
\code{\link[Seurat]{FeatureScatter}}, \code{\link[ggplot2]{ggplot}}
}
