% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qFeatureScatter}
\alias{qFeatureScatter}
\title{qFeatureScatter}
\usage{
qFeatureScatter(
  feature1 = "TOP2A",
  feature2 = "ID2",
  obj = combined.obj,
  ext = "png",
  plot = TRUE,
  ...
)
}
\arguments{
\item{feature1}{PARAM_DESCRIPTION, Default: 'TOP2A'}

\item{feature2}{PARAM_DESCRIPTION, Default: 'ID2'}

\item{obj}{Seurat object, Default: combined.obj}

\item{ext}{PARAM_DESCRIPTION, Default: 'png'}

\item{plot}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Quickly plot and save a FeatureScatter plot.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
