% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qMarkerCheck.BrainOrg}
\alias{qMarkerCheck.BrainOrg}
\title{qMarkerCheck.BrainOrg}
\usage{
qMarkerCheck.BrainOrg(
  obj = combined.obj,
  custom.genes = F,
  suffix = "",
  raster = FALSE,
  raster.dpi = c(512, 256)/2
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{custom.genes}{Use custom gene set? Default: F}

\item{suffix}{Folder name suffix, Default: ""}
}
\description{
Quickly plot key markers in brain organoids
}
\examples{
\dontrun{
if(interactive()){
 qMarkerCheck.BrainOrg(combined.obj)
 }
}
}
