% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qMarkerCheck.BrainOrg}
\alias{qMarkerCheck.BrainOrg}
\title{qMarkerCheck.BrainOrg}
\usage{
qMarkerCheck.BrainOrg(obj = combined.obj, custom.genes = F)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{custom.genes}{PARAM_DESCRIPTION, Default: F}
}
\description{
Quickly plot key markers in brain organoids
}
\examples{
\dontrun{
if(interactive()){
 qMarkerCheck.BrainOrg(combined.obj)
 }
}
}
