% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{qMarkerCheck.BrainOrg}
\alias{qMarkerCheck.BrainOrg}
\title{qMarkerCheck.BrainOrg}
\usage{
qMarkerCheck.BrainOrg(obj = combined.obj, custom.genes = FALSE, suffix = "")
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{custom.genes}{Use custom gene set? Default: FALSE}

\item{suffix}{Folder name suffix, Default: ""}
}
\description{
Quickly plot key markers in brain organoids
}
\examples{
\dontrun{
if (interactive()) {
  qMarkerCheck.BrainOrg(combined.obj)
}
}
}
