% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{qQC.plots.BrainOrg}
\alias{qQC.plots.BrainOrg}
\title{qQC.plots.BrainOrg}
\usage{
qQC.plots.BrainOrg(
  obj = combined.obj,
  QC.Features = c("nFeature_RNA", "percent.ribo", "percent.mito", "nuclear.fraction"),
  prefix = "QC.markers.4.UMAP",
  suffix = "",
  title = sppu(prefix, QC.Features, suffix),
  nrow = 2,
  ncol = 2,
  ...
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{QC.Features}{Any numeric metadata columns}
}
\description{
Quickly plot key QC markers in brain organoids
}
\examples{
qQC.plots.BrainOrg()
}
