% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qQC.plots.BrainOrg}
\alias{qQC.plots.BrainOrg}
\title{qQC.plots.BrainOrg}
\usage{
qQC.plots.BrainOrg(
  obj = combined.obj,
  title = "Top 4 QC markers on UMAP",
  nrow = 2,
  ncol = 2,
  QC.Features = c("nFeature_RNA", "percent.ribo", "percent.mito", "log10.HGA_Markers")
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{QC.Features}{Any numeric metadata columns}
}
\description{
Quickly plot key QC markers in brain organoids
}
\examples{
qQC.plots.BrainOrg.RV()
}
