% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qqSaveGridA4}
\alias{qqSaveGridA4}
\title{qqSaveGridA4}
\usage{
qqSaveGridA4(
  plotlist = pl,
  plots = 1:2,
  NrPlots = length(plots),
  height = hA4,
  width = wA4,
  fname = "Fractions.Organoid-to-organoid variation.png",
  ...
)
}
\arguments{
\item{plotlist}{PARAM_DESCRIPTION, Default: pl}

\item{plots}{PARAM_DESCRIPTION, Default: 1:2}

\item{NrPlots}{PARAM_DESCRIPTION, Default: length(plots)}

\item{height}{PARAM_DESCRIPTION, Default: hA4}

\item{width}{PARAM_DESCRIPTION, Default: wA4}

\item{fname}{File name, Default: 'Fractions.Organoid-to-organoid variation.png'}
}
\description{
Save 2 or 4 ggplot objects using plot_grid() on an A4 page #
}
\examples{
\dontrun{
if(interactive()){
 qqSaveGridA4(plotlist= pl, plots = 1:2, fname = "Fractions.per.Cl.png"); qqSaveGridA4(plotlist= pl, plots = 1:4, fname = "Fractions.per.Cl.4.png")
 }
}
}
