% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qsave.image}
\alias{qsave.image}
\title{Save workspace - qsave.image}
\usage{
qsave.image(..., showMemObject = TRUE, options = c("--force", NULL)[1])
}
\arguments{
\item{...}{Pass any other parameter to the internally called functions (most of them should work).}

\item{options}{Options passed on to gzip, via CLI. Default: c("--force", NULL)\link{1}}
}
\description{
Faster saving of workspace, and compression outside R, when it can run in the background. Seemingly quite CPU hungry and not very efficient compression. #
}
\seealso{
\code{\link[Stringendo]{kollapse}}, \code{\link[function]{iprint}}
}
