% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qSeuViolin}
\alias{qSeuViolin}
\title{qSeuViolin}
\usage{
qSeuViolin(
  object = ls.Seurat[[1]],
  suffix = GEX_library,
  features = "nFeature_RNA",
  split.by = "orig.ident",
  logY = TRUE
)
}
\arguments{
\item{logY}{}
}
\description{
qSeuViolin
}
