% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qSeuViolin}
\alias{qSeuViolin}
\title{qSeuViolin}
\usage{
qSeuViolin(
  object = ls.Seurat[[1]],
  suffix = GEX_library,
  features = "nFeature_RNA",
  split.by = "orig.ident",
  logY = TRUE
)
}
\arguments{
\item{object}{A Seurat object.}

\item{suffix}{A string to append to the title of the plot.}

\item{features}{The name of the feature to plot.}

\item{split.by}{The grouping variable to split the plot by.}

\item{logY}{Whether to plot the y-axis on a log scale.}
}
\value{
A ggplot object.
}
\description{
This function creates a violin plot of a single feature in a Seurat object, split by a grouping variable.
}
