% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{qSeuViolin}
\alias{qSeuViolin}
\title{qSeuViolin}
\usage{
qSeuViolin(
  object = ls.Seurat[[1]],
  features = "nFeature_RNA",
  idents = GetNamedClusteringRuns(object)[1],
  split.by = NULL,
  replace.na = FALSE,
  suffix = NULL,
  suffix.2.title = F,
  logY = TRUE,
  hline = FALSE,
  caption = FALSE,
  show_plot = T,
  w = 9,
  h = 5,
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{features}{The name of the feature to plot.}

\item{idents}{...}

\item{split.by}{The grouping variable to split the plot by.}

\item{suffix}{A string to append to the title of the plot.}

\item{suffix.2.title}{...}

\item{logY}{Whether to plot the y-axis on a log scale.}

\item{hline}{hline}

\item{caption}{caption}

\item{show_plot}{...}
}
\value{
A ggplot object.
}
\description{
This function creates a violin plot of a single feature in a Seurat object,
split by a grouping variable.
}
