% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{qSeuViolin}
\alias{qSeuViolin}
\title{Create a Violin Plot for a Seurat Object Feature and save the file.}
\usage{
qSeuViolin(
  object = ls.Seurat[[1]],
  features = "nFeature_RNA",
  idents = GetNamedClusteringRuns(object)[1],
  split.by = NULL,
  replace.na = FALSE,
  suffix = NULL,
  suffix.2.title = FALSE,
  logY = TRUE,
  hline = FALSE,
  caption = FALSE,
  show_plot = TRUE,
  w = 9,
  h = 5,
  ...
)
}
\arguments{
\item{object}{A Seurat object to be plotted.}

\item{features}{A character string specifying the name of the feature to plot.}

\item{idents}{A character vector specifying the identities to be used in the plot.}

\item{split.by}{A character string specifying the grouping variable for splitting the plot.}

\item{replace.na}{A logical indicating whether NA values should be replaced.}

\item{suffix}{An optional string to append to the title of the plot.}

\item{suffix.2.title}{A logical indicating whether to append the suffix to the plot title.}

\item{logY}{A logical indicating whether to use a logarithmic scale for the y-axis.}

\item{hline}{A numeric or logical value; if numeric, the value where a horizontal line should be drawn.}

\item{caption}{A character string or logical for the plot caption. If FALSE, no caption is displayed.}

\item{show_plot}{A logical indicating whether to display the plot.}

\item{w}{Width of the plot.}

\item{h}{Height of the plot.}

\item{...}{Additional arguments passed to \code{VlnPlot}.}
}
\value{
A ggplot object representing the violin plot.
}
\description{
Generates a violin plot for a specified feature in a Seurat object,
allowing for the data to be split by a specified grouping variable.
The function supports customization options such as logarithmic scaling, custom titles, and more.
}
\examples{
# Assuming `seurat_obj` is a valid Seurat object
qSeuViolin(object = seurat_obj, features = "nFeature_RNA")

}
