% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qUMAP}
\alias{qUMAP}
\title{qUMAP}
\usage{
qUMAP(
  feature = "TOP2A",
  obj = combined.obj,
  title = feature,
  sub = NULL,
  reduction = "umap",
  splitby = NULL,
  prefix = NULL,
  suffix = make.names(sub),
  save.plot = MarkdownHelpers::TRUE.unless("b.save.wplots"),
  PNG = T,
  h = 7,
  w = NULL,
  nr.cols = NULL,
  assay = c("RNA", "integrated")[1],
  axes = FALSE,
  aspect.ratio = c(FALSE, 0.6)[1],
  HGNC.lookup = TRUE,
  make.uppercase = FALSE,
  check_for_2D = TRUE,
  qlow = "q10",
  qhigh = "q90",
  caption = FALSE,
  raster = MarkdownHelpers::FALSE.unless("b.raster"),
  ...
)
}
\arguments{
\item{feature}{Feature to be visualized on the UMAP, Default: 'TOP2A'.}

\item{obj}{Seurat object containing single-cell RNA seq data, Default: combined.obj.}

\item{title}{Title of the plot, Default: feature.}

\item{sub}{Subtitle of the plot, Default: NULL.}

\item{reduction}{Dimension reduction technique to be used. Choose from 'umap', 'tsne', or 'pca'. Default: 'umap'.}

\item{splitby}{Column in the metadata to split the cells by, Default: NULL.}

\item{prefix}{A prefix added before the filename, Default: NULL.}

\item{suffix}{A suffix added to the end of the filename, Default: subtitle.}

\item{save.plot}{If TRUE, the plot is saved into a file, Default: T.}

\item{PNG}{If TRUE, the file is saved as a .png, Default: T.}

\item{h}{Height of the plot in inches, Default: 7.}

\item{w}{Width of the plot in inches, Default: NULL.}

\item{nr.cols}{Number of columns to combine multiple feature plots, ignored if split.by is not NULL, Default: NULL.}

\item{assay}{Which assay to use, 'RNA' or 'integrated', Default: 'RNA'.}

\item{axes}{If TRUE, the axes are shown on the plot. Default: FALSE.}

\item{aspect.ratio}{Ratio of height to width. If TRUE, the ratio is fixed at 0.6. Default: FALSE.}

\item{HGNC.lookup}{If TRUE, the HGNC gene symbol lookup is performed. Default: TRUE.}

\item{make.uppercase}{If TRUE, feature names are converted to uppercase. Default: FALSE.}

\item{check_for_2D}{If TRUE, checks if UMAP is 2 dimensional. Default: TRUE.}

\item{qlow}{Lower quantile for the color scale, Default: 'q10'.}

\item{qhigh}{Upper quantile for the color scale, Default: 'q90'.}

\item{caption}{Add caption to the ggplot object (e.g. a description in bottom right).}

\item{raster}{Use raster image for single cells. Necessary for large objects. Default: Only true when 'b.raster'  global variable is defined as TRUE. Code: MarkdownHelpers::FALSE.unless('b.raster')}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
The quickest way to draw a gene expression UMAP.
}
\examples{
\dontrun{
if(interactive()){
 qUMAP('nFeature_RNA'); qUMAP('TOP2A')
 }
}
}
