% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{qUMAP}
\alias{qUMAP}
\title{qUMAP}
\usage{
qUMAP(
  feature = "TOP2A",
  obj = combined.obj,
  title = feature,
  sub = NULL,
  reduction = "umap",
  splitby = NULL,
  prefix = NULL,
  suffix = make.names(sub),
  save.plot = TRUE.unless(b.save.wplots),
  PNG = T,
  h = 7,
  w = NULL,
  nr.cols = NULL,
  assay = c("RNA", "integrated")[1],
  axes = T,
  HGNC.lookup = TRUE,
  make.uppercase = FALSE,
  qlow = "q10",
  qhigh = "q90",
  ...
)
}
\arguments{
\item{feature}{PARAM_DESCRIPTION, Default: 'TOP2A'}

\item{obj}{Seurat object, Default: combined.obj}

\item{title}{Title of the plot, Default: feature}

\item{sub}{Subtitle of the plot, Default: NULL}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'umap'}

\item{splitby}{PARAM_DESCRIPTION, Default: NULL}

\item{prefix}{A prefix added before the filename, Default: NULL}

\item{suffix}{A suffix added to the end of the filename, Default: subtitle}

\item{save.plot}{Save the plot into a file?, Default: T}

\item{PNG}{Save file as .png?, Default: T}

\item{h}{height of the plot, Default: 7}

\item{w}{width of the plot, Default: NULL}

\item{nr.cols}{Number of columns to combine multiple feature plots to, ignored if split.by is not NULL, Default: NULL}

\item{assay}{RNA or integrated assay, Default: c("RNA", "integrated")\link{1}}

\item{axes}{Show axes?}

\item{HGNC.lookup}{PARAM_DESCRIPTION, Default: TRUE}

\item{make.uppercase}{PARAM_DESCRIPTION, Default: TRUE}

\item{qlow}{PARAM_DESCRIPTION, Default: 'q10'}

\item{qhigh}{PARAM_DESCRIPTION, Default: 'q90'}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
The quickest way to draw a gene expression UMAP #
}
\examples{
\dontrun{
if(interactive()){
 qUMAP('nFeature_RNA'); qUMAP('TOP2A')
 }
}
}
