% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{read10x}
\alias{read10x}
\title{read10x}
\usage{
read10x(dir)
}
\arguments{
\item{dir}{A character string specifying the directory where the gzipped files are located.}
}
\value{
A Seurat object constructed from the 10X dataset.
}
\description{
This function reads a 10X dataset from gzipped matrix.mtx, features.tsv and barcodes.tsv files.
}
\examples{
\dontrun{
if(interactive()){
 seuratObject <- read10x(dir = dir)
 }
}
}
\seealso{
\code{\link[tictoc]{tic}}
\code{\link[R.utils]{compressFile}}
\code{\link[Seurat]{Read10X}}
}
