% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{recall.genes.ls}
\alias{recall.genes.ls}
\title{recall.genes.l}
\usage{
recall.genes.ls(obj = combined.obj)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}
}
\description{
genes.ls #
}
\examples{
\dontrun{
if(interactive()){
 recall.genes.ls(); genes.ls
 }
}
}
