% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{recall.meta.tags.n.datasets}
\alias{recall.meta.tags.n.datasets}
\title{recall.meta.tags.n.datasets}
\usage{
recall.meta.tags.n.datasets(obj = combined.obj)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}
}
\description{
recall.meta.tags.n.datasets
}
\examples{
\dontrun{
if(interactive()){
 recall.n.datasets(); n.datasets
 }
}
}
