% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R, R/Seurat.Utils.Visualization.R
\name{RecallReduction}
\alias{RecallReduction}
\title{RecallReduction}
\usage{
RecallReduction(obj = combined.obj, dim = 2, reduction = "umap")

RecallReduction(obj = combined.obj, dim = 2, reduction = "umap")
}
\arguments{
\item{obj}{Seurat object from which the backup will be restored; Default: \code{combined.obj}.}

\item{dim}{Number of dimensions of the reduction data to restore; Default: 2.}

\item{reduction}{Type of dimensionality reduction to be restored ('umap', 'tsne', 'pca'); Default: 'umap'.}
}
\description{
Set active UMAP to \code{obj@reductions$umap} from \code{obj@misc$reductions.backup}.

Restores dimensionality reduction data (e.g., UMAP, tSNE, PCA) from a backup
stored within \code{obj@misc$reductions.backup} to the active \code{obj@reductions} slot.
}
\examples{
\dontrun{
if (interactive()) {
  combined.obj <- RecallReduction(obj = combined.obj, dim = 2, reduction = "umap")
  qUMAP()
  combined.obj <- RecallReduction(obj = combined.obj, dim = 3, reduction = "umap")
  qUMAP()
}
}
\dontrun{
if (interactive()) {
  combined.obj <- RecallReduction(obj = combined.obj, dim = 2, reduction = "umap")
  qUMAP()
  combined.obj <- RecallReduction(obj = combined.obj, dim = 3, reduction = "umap")
  qUMAP()
}
}

}
