% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{RelabelSmallCategories}
\alias{RelabelSmallCategories}
\title{Relabel Small Categories / Clusters}
\usage{
RelabelSmallCategories(
  obj,
  col_in,
  backup_col_name = ppp(col_in, "orig"),
  min_count = 100,
  small_label = "Other",
  v = TRUE
)
}
\arguments{
\item{obj}{Seurat object. The Seurat object containing the metadata.}

\item{col_in}{Character string. Name of the metadata column to process.}

\item{backup_col_name}{Character string. Name of the new metadata column where to backup the original values.
Default: \code{ppp(col_in, "orig")}.}

\item{min_count}{Numeric. Minimum number of cells required for a category to be retained.
Categories with counts less than this number will be relabeled. Default: 100}

\item{small_label}{Character string. Label to assign to small categories. Default: "Other".}

\item{v}{Logical. If \code{TRUE}, prints verbose output. Default: \code{TRUE}.}
}
\value{
Seurat object. The modified Seurat object with the new metadata column added.
}
\description{
Relabels small categories in a specified metadata column of a Seurat object. Categories with
cell counts less than a minimum count are relabeled to a specified label. The function adds
a new metadata column with the updated labels.
}
\examples{
# Assuming 'seurat_obj' is your Seurat object
seurat_obj <- RelabelSmallCategories(
  obj = seurat_obj,
  col_in = "cell_type",
  min_count = 50,
  small_label = "MinorType",
  v = TRUE
)

}
