% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{remove.cells.by.UMAP}
\alias{remove.cells.by.UMAP}
\title{remove.cells.by.UMAP}
\usage{
remove.cells.by.UMAP(
  reduction = "umap",
  umap_dim = 1,
  obj = combined.obj,
  cutoff = 0,
  cut_below = TRUE,
  ...
)
}
\arguments{
\item{reduction}{Dim. reduction: umap or pca or tsne}

\item{umap_dim}{Which axis to remove by?}

\item{obj}{Seurat object, Default: combined.obj}

\item{cutoff}{Horizontal or vertical cutoff value}

\item{cut_below}{Shall we remove below or above the line}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
remove.cells.by.UMAP
}
