% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{remove.residual.small.clusters}
\alias{remove.residual.small.clusters}
\title{remove.residual.small.clusters}
\usage{
remove.residual.small.clusters(
  identitites = GetOrderedClusteringRuns(),
  obj = combined.obj,
  max.cells = round((ncol(obj))/2000)
)
}
\arguments{
\item{identitites}{Identities to scan for residual clusters}

\item{obj}{Seurat object, Default: combined.obj}

\item{max.cells}{Max number of cells in cluster to be removed. Default: 0.5\% of the dataset}
}
\description{
E.g.: after subsetting often some residual cells remain in clusters originally denfined in the full dataset.
}
