% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{RemoveGenesSeurat}
\alias{RemoveGenesSeurat}
\title{RemoveGenesSeurat}
\usage{
RemoveGenesSeurat(obj = ls.Seurat[[i]], symbols2remove = c("TOP2A"))
}
\arguments{
\item{obj}{Seurat object, Default: ls.Seurat[\link{i}]}

\item{symbols2remove}{PARAM_DESCRIPTION, Default: c("TOP2A")}
}
\description{
Replace gene names in different slots of a Seurat object. Run this before integration. Run this before integration. It only changes metadata; obj@assays$RNA@counts, @data and @scale.data. #
}
\examples{
\dontrun{
if(interactive()){
 RemoveGenesSeurat(obj = SeuratObj, symbols2remove = "TOP2A")
 }
}
}
