% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{RenameClustering}
\alias{RenameClustering}
\title{RenameClustering}
\usage{
RenameClustering(
  namedVector = ManualNames,
  orig.ident = "RNA_snn_res.0.3",
  new.ident = ppp(orig.ident, "ManualNames"),
  obj = combined.obj
)
}
\arguments{
\item{namedVector}{named vector, where values = new, names(vec) = old}

\item{orig.ident}{meta.data colname original}

\item{new.ident}{meta.data colname new}

\item{obj}{Seurat object}
}
\description{
RenameClustering
}
