% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{RenameClustering}
\alias{RenameClustering}
\title{RenameClustering}
\usage{
RenameClustering(
  namedVector = ManualNames,
  orig.ident = "RNA_snn_res.0.3",
  suffix.new.ident = "ManualNames",
  new.ident = ppp(orig.ident, suffix.new.ident),
  obj = combined.obj,
  suffix = substitute(obj),
  ...
)
}
\arguments{
\item{namedVector}{named vector, where values = new, names(vec) = old}

\item{orig.ident}{meta.data colname original}

\item{suffix.new.ident}{How to name (suffix) the new identity. Default: "ManualNames"}

\item{new.ident}{meta.data colname new}

\item{obj}{Seurat object}
}
\description{
RenameClustering
}
