% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{RenameGenesSeurat}
\alias{RenameGenesSeurat}
\title{Rename Gene Symbols in a Seurat Object}
\usage{
RenameGenesSeurat(
  obj = ls.Seurat[[i]],
  newnames = HGNC.updated[[i]]$Suggested.Symbol,
  assay = "RNA",
  slots = c("data", "counts", "meta.features")
)
}
\arguments{
\item{obj}{A Seurat object containing the assay and slots to be updated; Default: \code{ls.Seurat[[i]]}
(replace \code{i} with the appropriate index).}

\item{newnames}{A character vector containing the new gene names intended to replace the
existing ones; Default: \code{HGNC.updated[[i]]$Suggested.Symbol}. Ensure this matches the order
and length of the genes in the specified assay.}

\item{assay}{The name of the assay within the Seurat object where gene names will be updated;
Default: \code{"RNA"}. This function assumes simple objects containing only an RNA assay.}

\item{slots}{A character vector specifying which slots within the assay to update. Possible
values include \code{"data"}, \code{"counts"}, and \code{"meta.features"}; other layers can be specified if present.}
}
\description{
This function replaces gene names across various slots within a specified assay
of a Seurat object. It is designed to be run prior to any data integration or downstream analysis
processes. The function targets the \verb{@counts}, \verb{@data}, and \verb{@meta.features} slots within
the specified assay, ensuring consistency in gene nomenclature across the object.
}
\details{
It is crucial to run this function before any data integration or further analysis
to ensure gene symbol consistency. The function does not support complex objects with multiple
assays where dependencies between assays might lead to inconsistencies. Use with caution and
verify the results.
}
\note{
This function modifies the Seurat object in place, changing gene symbols directly within
the specified slots. Be sure to have a backup of your Seurat object if needed before applying
this function.
}
\examples{
\dontrun{
if (interactive()) {
  # Assuming `SeuratObj` is your Seurat object
  # and `HGNC.updated.genes` contains the updated gene symbols
  SeuratObj <- RenameGenesSeurat(
    obj = SeuratObj,
    newnames = HGNC.updated.genes$Suggested.Symbol
  )
  # `SeuratObj` now has updated gene symbols in the specified assay and slots
}
}

}
