% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{RenameGenesSeurat}
\alias{RenameGenesSeurat}
\title{RenameGenesSeurat}
\usage{
RenameGenesSeurat(
  obj = ls.Seurat[[i]],
  newnames = HGNC.updated[[i]]$Suggested.Symbol,
  assay = "RNA"
)
}
\arguments{
\item{obj}{Seurat object, Default: ls.Seurat[\link{i}]}

\item{newnames}{A vector of new gene names. Default: HGNC.updated[\link{i}]$Suggested.Symbol}

\item{assay}{Which Seurat assay to replace. Default: RNA. Disclaimer: Intended use on simple objects that ONLY contain an RNA object. I highly advise against selectively replacing name in other assays that may have slots that cannot be updated by this function.}
}
\description{
Replace gene names in different slots of a Seurat object. Run this before integration. Run this before integration. It only changes obj@assays$RNA@counts, @data and @scale.data. #
}
\examples{
\dontrun{
if(interactive()){
 RenameGenesSeurat(obj = SeuratObj, newnames = HGNC.updated.genes$Suggested.Symbol)
 }
}
}
