% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{RunDGEA}
\alias{RunDGEA}
\title{Run Differential Gene Expression Analysis (DGEA)}
\usage{
RunDGEA(
  obj = obj.RG,
  param.list = p,
  reorder.clusters = TRUE,
  reorder.dimension = 1,
  ordering = "ordered",
  res.analyzed.DE = c(0.1),
  add.combined.score = TRUE,
  save.obj = TRUE,
  directory = OutDirOrig,
  calculte.DGEA = TRUE,
  plot.DGEA = TRUE,
  Cluster.Labels.Automatic = TRUE,
  subdirectory = "DGEA"
)
}
\arguments{
\item{obj}{Seurat object, assumed to be pre-configured with necessary data. Default: obj.RG.}

\item{param.list}{List of parameters for DE analysis. Default: p.}

\item{reorder.clusters}{Logical indicating whether to reorder clusters based on dimension.
Default: TRUE.}

\item{reorder.dimension}{Integer specifying the dimension for reordering (1 for x, -1 for y).
Default: 1.}

\item{ordering}{Character string specifying the cluster annotation method; can be "ordered"
or "simple". Default: "ordered".}

\item{res.analyzed.DE}{Vector of numeric values specifying the resolutions to analyze.
Default: c(0.1).}

\item{add.combined.score}{Logical indicating whether to add a combined score to the markers.
Default: TRUE.}

\item{save.obj}{Logical indicating whether to save the modified Seurat object.
Default: TRUE.}

\item{directory}{Character string specifying the base directory for saving results.
Default: OutDirOrig.}

\item{calculte.DGEA}{Logical determining if the DE analysis should be calculated.
Default: TRUE.}

\item{plot.DGEA}{Logical determining if results should be plotted.
Default: TRUE.}

\item{Cluster.Labels.Automatic}{Logical indicating automatic labeling of clusters.
Default: TRUE.}

\item{subdirectory}{Character string specifying the subdirectory for saving outputs within
the base directory. Default: "DGEA".}
}
\value{
Modified Seurat object and markers list.
}
\description{
Runs a differential gene expression analysis based on specified parameters, reorders clusters
if needed, and optionally saves results. Supports output and plotting configurations.
}
\examples{
RunDGEA(obj = mySeuratObject, param.list = myListParams, directory = "Results/MyAnalysis")

}
