% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{runDGEA}
\alias{runDGEA}
\title{Run Differential Gene Expression Analysis (DGEA)}
\usage{
runDGEA(
  obj,
  param.list = p,
  ident = NULL,
  res.analyzed.DE = if (is.null(ident)) c(0.1) else ident,
  reorder.clusters = if (is.null(ident)) TRUE else FALSE,
  reorder.dimension = 1,
  directory,
  dir_suffix,
  subdirectory = ppp("DGEA_res", idate()),
  add.combined.score = TRUE,
  save.obj = TRUE,
  calculate.DGEA = TRUE,
  plot.DGEA = TRUE,
  umap_caption = "",
  plot.av.enrichment.hist = TRUE,
  plot.log.top.gene.stats = TRUE,
  auto.cluster.naming = TRUE,
  clean.misc.slot = TRUE,
  clean.meta.data = TRUE,
  n.cores = 1,
  presto = TRUE,
  WorkingDir = getwd()
)
}
\arguments{
\item{obj}{Seurat object, assumed to be pre-configured with necessary data.}

\item{param.list}{List of parameters for DE analysis. Default: p.}

\item{ident}{Use this to specify a non-standard cluster identity, such as named clusters.
\code{runDGEA} will use this ident explicitly for the DE analysis. Default: NULL.}

\item{res.analyzed.DE}{Vector of numeric values specifying the resolutions to analyze.
Default: c(0.1).}

\item{reorder.clusters}{Logical indicating whether to reorder clusters based on dimension.
Default: \code{TRUE}.}

\item{reorder.dimension}{Integer specifying the dimension for reordering (1 for x, -1 for y).
Default: 1.}

\item{directory}{Character string specifying the base directory for saving results.
Default: OutDir}

\item{dir_suffix}{Character string specifying the suffix for the subdirecotry directory.}

\item{subdirectory}{Character string specifying the subdirectory for saving outputs within
the base directory. Default: "DGEA + date".}

\item{add.combined.score}{Logical indicating whether to add a combined score to the markers.
Default: \code{TRUE}.}

\item{save.obj}{Logical indicating whether to save the modified Seurat object.
Default: \code{TRUE}.}

\item{calculate.DGEA}{Logical determining if the DE analysis should be calculated.
Default: \code{TRUE}.}

\item{plot.DGEA}{Logical determining if results should be plotted.
Default: \code{TRUE}.}

\item{umap_caption}{Character string specifying the caption for the UMAP plot. Default: "".}

\item{plot.av.enrichment.hist}{Logical indicating whether to plot the average enrichment histogram.
Default: \code{TRUE}.}

\item{plot.log.top.gene.stats}{Logical indicating whether to plot the log top gene statistics.}

\item{auto.cluster.naming}{Logical indicating automatic labeling of clusters.
Default: \code{TRUE}.}

\item{clean.misc.slot}{Logical indicating whether to clean the misc slots of previous
clustering results. Default: \code{TRUE}.}

\item{clean.meta.data}{Logical indicating whether to clean the metadata slots of
previous clustering results. Default: \code{TRUE}.}

\item{n.cores}{Integer specifying the number of cores to use for parallel processing (multisession).
Default: 1.}

\item{presto}{Logical indicating whether to use presto for DE analysis. Default: \code{TRUE}.}

\item{WorkingDir}{Character string specifying the working directory. Default: getwd().}
}
\value{
Modified Seurat object and markers list.
}
\description{
Runs a differential gene expression analysis based on specified parameters,
reorders clusters if needed, and optionally saves results. Supports output and plotting configurations.
}
\examples{
runDGEA(obj = mySeuratObject, param.list = myListParams, directory = "Results/MyAnalysis")

}
