% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{runDGEA}
\alias{runDGEA}
\title{Run Differential Gene Expression Analysis (DGEA)}
\usage{
runDGEA(
  obj = obj.RG,
  param.list = p,
  reorder.clusters = TRUE,
  clean.misc.slot = TRUE,
  clean.meta.data = TRUE,
  reorder.dimension = 1,
  ordering = "ordered",
  res.analyzed.DE = c(0.1),
  add.combined.score = TRUE,
  save.obj = TRUE,
  directory = OutDir,
  subdirectory = "DGEA_res",
  calculate.DGEA = TRUE,
  plot.DGEA = TRUE,
  auto.cluster.naming = TRUE,
  plot.av.enrichment.hist = TRUE,
  plot.log.top.gene.stats = TRUE,
  ...
)
}
\arguments{
\item{obj}{Seurat object, assumed to be pre-configured with necessary data. Default: obj.RG.}

\item{param.list}{List of parameters for DE analysis. Default: p.}

\item{reorder.clusters}{Logical indicating whether to reorder clusters based on dimension.
Default: TRUE.}

\item{clean.misc.slot}{Logical indicating whether to clean the misc slots of previous
clustering results. Default: TRUE.}

\item{clean.meta.data}{Logical indicating whether to clean the metadata slots of
previous clustering results. Default: TRUE.}

\item{reorder.dimension}{Integer specifying the dimension for reordering (1 for x, -1 for y).
Default: 1.}

\item{ordering}{Character string specifying the cluster annotation method; can be "ordered"
or "simple". Default: "ordered".}

\item{res.analyzed.DE}{Vector of numeric values specifying the resolutions to analyze.
Default: c(0.1).}

\item{add.combined.score}{Logical indicating whether to add a combined score to the markers.
Default: TRUE.}

\item{save.obj}{Logical indicating whether to save the modified Seurat object.
Default: TRUE.}

\item{directory}{Character string specifying the base directory for saving results.
Default: OutDirOrig.}

\item{subdirectory}{Character string specifying the subdirectory for saving outputs within
the base directory. Default: "DGEA".}

\item{calculate.DGEA}{Logical determining if the DE analysis should be calculated.
Default: TRUE.}

\item{plot.DGEA}{Logical determining if results should be plotted.
Default: TRUE.}

\item{auto.cluster.naming}{Logical indicating automatic labeling of clusters.
Default: TRUE.}

\item{plot.av.enrichment.hist}{Logical indicating whether to plot the average enrichment histogram.
Default: TRUE.}

\item{plot.log.top.gene.stats}{Logical indicating whether to plot the log top gene statistics.}
}
\value{
Modified Seurat object and markers list.
}
\description{
Runs a differential gene expression analysis based on specified parameters,
reorders clusters if needed, and optionally saves results. Supports output and plotting configurations.
}
\examples{
runDGEA(obj = mySeuratObject, param.list = myListParams, directory = "Results/MyAnalysis")

}
