% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{sampleNpc}
\alias{sampleNpc}
\title{sampleNpc}
\usage{
sampleNpc(metaDF = MetaData[which(Pass), ], pc = 0.1)
}
\arguments{
\item{metaDF}{A dataframe representing a subset of the metadata of a Seurat object. Default: Subset of 'MetaData' for which 'Pass' is TRUE.}

\item{pc}{The percentage of the dataframe to sample, expressed as a decimal. Default: 0.1.}
}
\value{
A vector of sampled cell IDs.
}
\description{
This function samples a specified percentage of a dataframe (specifically a subset of the metadata of a Seurat object) and returns the corresponding cell IDs.
}
\examples{
\dontrun{
if(interactive()){
 # Example usage:
 # Suppose 'MetaData' is a dataframe and 'Pass' is a boolean vector with the same length.
 # The following example will sample 10\% of the rows of 'MetaData' where 'Pass' is TRUE.
 sampleNpc(metaDF = MetaData[which(Pass),], pc = 0.1)
 }
}
}
