% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{save.parameters}
\alias{save.parameters}
\title{save.parameters}
\usage{
save.parameters(obj = combined.obj, params = p, overwrite = TRUE)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{params}{List of parameters, Default: p}
}
\description{
Save parameters to obj@misc$p
}
\examples{
\dontrun{
if (interactive()) {
  save.parameters(obj = combined.obj, params = p)
}
}
}
