% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{save.parameters}
\alias{save.parameters}
\title{save.parameters}
\usage{
save.parameters(obj = combined.obj, params = p)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{params}{PARAM_DESCRIPTION, Default: p}
}
\description{
save.parameters
}
\examples{
\dontrun{
if(interactive()){
 save.parameters(obj = combined.obj, params = p);
 }
}
}
