% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{save2umaps.A4}
\alias{save2umaps.A4}
\title{save2umaps.A4}
\usage{
save2umaps.A4(
  plot_list,
  pname = F,
  suffix = NULL,
  scale = 1,
  nrow = 2,
  ncol = 1,
  h = hA4 * scale,
  w = wA4 * scale,
  ...
)
}
\arguments{
\item{plot_list}{PARAM_DESCRIPTION}

\item{pname}{PARAM_DESCRIPTION, Default: F}

\item{suffix}{A suffix added to the filename, Default: NULL}

\item{scale}{PARAM_DESCRIPTION, Default: 1}

\item{nrow}{PARAM_DESCRIPTION, Default: 2}

\item{ncol}{PARAM_DESCRIPTION, Default: 1}

\item{h}{height of the plot, Default: hA4 * scale}

\item{w}{width of the plot, Default: wA4 * scale}

\item{...}{Pass any other parameter to the internally called functions (most of them should  work).}
}
\description{
save2umaps.A4
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
