% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{SavePlotlyAsHtml}
\alias{SavePlotlyAsHtml}
\title{SavePlotlyAsHtml}
\usage{
SavePlotlyAsHtml(plotly_obj, category. = category, suffix. = NULL)
}
\arguments{
\item{plotly_obj}{PARAM_DESCRIPTION}

\item{category.}{PARAM_DESCRIPTION, Default: category}

\item{suffix.}{PARAM_DESCRIPTION, Default: NULL}
}
\description{
Save Plotly 3D scatterplot as an html file. #
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[htmlwidgets]{saveWidget}}
}
