% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{SavePlotlyAsHtml}
\alias{SavePlotlyAsHtml}
\title{SavePlotlyAsHtml}
\usage{
SavePlotlyAsHtml(plotly_obj, category. = category, suffix. = NULL)
}
\arguments{
\item{plotly_obj}{The Plotly object to save.}

\item{category}{The category of the plot.}

\item{suffix}{A suffix to add to the filename.}

\item{OutputDir}{The output directory.}
}
\description{
Save a Plotly 3D scatterplot as an HTML file.
}
\examples{
\dontrun{
plt <- plotly::plot_ly("some stuff")
SavePlotlyAsHtml(plt, category. = "label.categ", suffix. = "test")
}

}
\seealso{
\code{\link[htmlwidgets]{saveWidget}}
}
