% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{scBarplot.CellFractions}
\alias{scBarplot.CellFractions}
\title{scBarplot.CellFractions}
\usage{
scBarplot.CellFractions(
  obj = combined.obj,
  group.by = "integrated_snn_res.0.5.ordered",
  fill.by = "age",
  downsample = T,
  plotname = paste(tools::toTitleCase(fill.by), "proportions"),
  hlines = c(0.25, 0.5, 0.75),
  return_table = F,
  seedNr = 1989
)
}
\arguments{
\item{obj}{The Seurat object to use for the analysis. Default: combined.obj.}

\item{group.by}{A string specifying the variable to group by. Default: 'integrated_snn_res.0.5.ordered'.}

\item{fill.by}{A string specifying the variable to fill by. Default: 'age'.}

\item{downsample}{Logical flag indicating whether to downsample the data. If TRUE, the number of cells in each group is equal to the number of cells in the smallest group. Default: TRUE.}

\item{plotname}{A string specifying the title of the plot. Default: paste(tools::toTitleCase(fill.by), "proportions").}

\item{hlines}{A numeric vector specifying the y-intercepts for horizontal lines on the plot. Default: c(0.25, 0.5, 0.75).}

\item{return_table}{Logical flag indicating whether to return a contingency table instead of a bar plot. Default: FALSE.}

\item{seedNr}{Seed for random number generation to ensure reproducibility. Default: 1989.}
}
\description{
Generates a bar plot of cell fractions per cluster.
}
\examples{
\dontrun{
if(interactive()){
 scBarplot.CellFractions(obj = combined.obj, group.by = "integrated_snn_res.0.1", fill.by = "Phase", downsample = T)
 scBarplot.CellFractions(obj = combined.obj, group.by = "integrated_snn_res.0.1", fill.by = "Phase", downsample = F)
 }
}
}
\seealso{
\code{\link[tools]{toTitleCase}}
}
