% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scBarplot.CellsPerCluster}
\alias{scBarplot.CellsPerCluster}
\title{scBarplot.CellsPerCluster}
\usage{
scBarplot.CellsPerCluster(
  ident = GetOrderedClusteringRuns(obj = obj)[1],
  obj = combined.obj,
  sort = FALSE,
  label = list(TRUE, "percent")[[1]],
  suffix = if (label == "percent") "percent" else NULL,
  palette = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[3],
  return_table = FALSE,
  ylab_adj = 1.1,
  min.cells = round(ncol(obj)/500),
  ...
)
}
\arguments{
\item{ident}{identity used, Default: 'cl.names.KnownMarkers.0.5'}

\item{obj}{Seurat object, Default: combined.obj}

\item{sort}{Sort by cluster size? Default: FALSE}

\item{label}{True: displays cell count, but you can provide anything in a vector.}

\item{suffix}{File name suffix}

\item{palette}{Color palette. Default: glasbey.}

\item{return_table}{Should it return the plotting data instead of the plot?}

\item{min.cells}{Threshold for too small categories.}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Barplot the Fraction of cells per cluster. (dupl?)
}
\examples{
\dontrun{
if (interactive()) {
  scBarplot.CellsPerCluster()
  scBarplot.CellsPerCluster(sort = TRUE)
}
}
}
