% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scBarplot.CellsPerCluster}
\alias{scBarplot.CellsPerCluster}
\title{Barplot of Fraction of Cells per Cluster}
\usage{
scBarplot.CellsPerCluster(
  obj = combined.obj,
  ident = GetOrderedClusteringRuns(obj = obj)[1],
  sort = FALSE,
  title = "Cells per Identity Group",
  sub = ident,
  label = list(TRUE, "percent")[[1]],
  suffix = if (label == "percent") "percent" else NULL,
  palette = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[3],
  return_table = FALSE,
  ylab_adj = 1.1,
  min.cells = round(ncol(obj)/100),
  ...
)
}
\arguments{
\item{obj}{Seurat object for analysis. Default: \code{combined.obj}.}

\item{ident}{Cluster identity. Used to specify which clustering results to visualize.
Default: First entry from ordered clustering runs.}

\item{sort}{If TRUE, sorts clusters by size. Default: FALSE.}

\item{title}{Title for the plot. Default: "Cells per Identity Group".}

\item{sub}{Subtitle for the plot. Default: "identity".}

\item{label}{If TRUE, shows cell count or percentage based on the label vector. Default: TRUE.}

\item{suffix}{Optional suffix for file naming. Used in conjunction with \code{kpp}.}

\item{palette}{Color palette for the barplot. Default: 'glasbey'.}

\item{return_table}{If TRUE, returns the data used for plotting instead of the plot itself. Default: FALSE.}

\item{ylab_adj}{Adjustment factor for y-axis label positioning. Default: 1.1.}

\item{min.cells}{Minimum cell count threshold for categories. Adjusted by data size.}

\item{...}{Additional parameters for internal function calls.}
}
\description{
Visualizes the fraction of cells within each cluster through a barplot.
}
\examples{
\dontrun{
scBarplot.CellsPerCluster()
scBarplot.CellsPerCluster(sort = TRUE)
}
}
