% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{scBarplot.CellsPerCluster}
\alias{scBarplot.CellsPerCluster}
\title{scBarplot.CellsPerCluster}
\usage{
scBarplot.CellsPerCluster(
  ident = GetOrderedClusteringRuns()[1],
  sort = F,
  label = T,
  palette = c("alphabet", "alphabet2", "glasbey", "polychrome", "stepped")[3],
  obj = combined.obj,
  return_table = F,
  ...
)
}
\arguments{
\item{ident}{identity used, Default: 'cl.names.KnownMarkers.0.5'}

\item{sort}{PARAM_DESCRIPTION, Default: F}

\item{obj}{Seurat object, Default: combined.obj}
}
\description{
Barplot the Fraction of cells per cluster. (dupl?)
}
\examples{
\dontrun{
if(interactive()){
 scBarplot.CellsPerCluster(); scBarplot.CellsPerCluster(sort = T)
 }
}
}
