% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scBarplot.CellsPerObject}
\alias{scBarplot.CellsPerObject}
\title{Barplot of Cells Per Seurat Object}
\usage{
scBarplot.CellsPerObject(
  ls.obj = ls.Seurat,
  plotname = "Nr.Cells.After.Filtering",
  xlab.angle = 45,
  names = FALSE,
  ...
)
}
\arguments{
\item{ls.obj}{List of Seurat objects to analyze. Default: \code{ls.Seurat}.}

\item{plotname}{Title for the plot. Default: 'Nr.Cells.After.Filtering'.}

\item{xlab.angle}{Angle for x-axis labels, enhancing readability. Default: 45.}

\item{names}{Optionally provide custom names for x-axis labels. If FALSE, uses object names
from \code{ls.obj}. Default: \code{FALSE}.}

\item{...}{Extra parameters passed to \code{qbarplot}.}
}
\description{
Visualizes the number of cells in each Seurat object within a list, showing the
distribution of cell counts across different datasets or experimental conditions.
}
\examples{
\dontrun{
ls.obj <- list(obj, obj)
scBarplot.CellsPerObject(ls.obj)
}

}
