% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{scBarplot.CellsPerObject}
\alias{scBarplot.CellsPerObject}
\title{scBarplot.CellsPerObject}
\usage{
scBarplot.CellsPerObject(
  ls.Seu = ls.Seurat,
  plotname = "Nr.Cells.After.Filtering",
  xlab.angle = 45,
  names = FALSE,
  ...
)
}
\arguments{
\item{ls.Seu}{A list of Seurat objects. Default: ls.Seurat.}

\item{plotname}{A string specifying the title of the plot. Default: 'Nr.Cells.After.Filtering'.}

\item{xlab.angle}{The angle at which the x-axis labels should be displayed. Default: 45.}

\item{names}{A logical value indicating whether to use the provided names as labels on the x-axis. If FALSE, the names of the Seurat objects will be used. Default: FALSE.}

\item{...}{Additional arguments to be passed to the qbarplot function.}
}
\description{
Creates a bar plot for the number of cells per object from a list of Seurat objects.
}
