% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scBarplot.FractionAboveThr}
\alias{scBarplot.FractionAboveThr}
\title{Barplot the Fraction of Cells Above Threshold per Cluster}
\usage{
scBarplot.FractionAboveThr(
  thrX = 0.2,
  value.col = "percent.ribo",
  id.col = getClusterNames()[1],
  obj = combined.obj,
  return.df = FALSE,
  subtitle = id.col,
  label = NULL,
  suffix = NULL,
  above = TRUE,
  ylim = c(0, 100),
  ...
)
}
\arguments{
\item{thrX}{Threshold for calculating the fraction of cells. Default: 0.3.}

\item{value.col}{Column in metadata with values to assess against \code{thrX}. Default: 'percent.ribo'.}

\item{id.col}{Cluster identity column in metadata. Default: 'cl.names.top.gene.res.0.3'.}

\item{obj}{Seurat object with single-cell data. Default: \code{combined.obj}.}

\item{return.df}{Whether to return the underlying data frame instead of the plot. Default: FALSE.}

\item{subtitle}{Optional subtitle for the plot.}

\item{label}{Whether to add labels to the bar plot. Default: NULL.}

\item{suffix}{Suffix for the output file name.}

\item{above}{Whether to calculate the fraction of cells above or below the threshold. Default: TRUE.}

\item{...}{Additional parameters for plotting functions.}
}
\description{
Generates a bar plot depicting the percentage of cells within each cluster that
exceed a specified threshold, based on a selected metadata column.
}
\examples{
\dontrun{
scBarplot.FractionAboveThr(id.col = "cl.names.top.gene.res.0.3", value.col = "percent.ribo", thrX = 0.2)
}

}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{group_by}}
}
