% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scBarplot.FractionAboveThr}
\alias{scBarplot.FractionAboveThr}
\title{scBarplot.FractionAboveThr}
\usage{
scBarplot.FractionAboveThr(
  thrX = 0.3,
  value.col = "percent.ribo",
  id.col = "cl.names.top.gene.res.0.3",
  obj = combined.obj,
  return.df = FALSE,
  label = FALSE,
  suffix = NULL,
  subtitle = id.col,
  ...
)
}
\arguments{
\item{thrX}{The threshold value to determine the fraction of cells. Default: 0.3}

\item{value.col}{Column name from metadata which holds the values for calculating the fraction of cells. Default: 'percent.ribo'}

\item{id.col}{Column name from metadata to be used for identifying clusters. Default: 'cl.names.top.gene.res.0.3'}

\item{obj}{A Seurat object holding single cell data. Default: combined.obj}

\item{return.df}{A logical indicating if the function should return the data frame used to create the plot. Default: FALSE}

\item{label}{A logical indicating if labels should be added to the bar plot. Default: FALSE}

\item{suffix}{An optional suffix for the filename.}

\item{subtitle}{Subtitle}

\item{...}{Additional parameters to pass to internally called functions.}
}
\description{
Create a bar plot showing the fraction of cells, within each cluster, that exceed a certain threshold based on a metadata column.
}
\examples{
\dontrun{
if (interactive()) {
  scBarplot.FractionAboveThr(id.col = "cl.names.top.gene.res.0.3", value.col = "percent.ribo", thrX = 0)
}
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{se-deprecated}}
}
