% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{scBarplot.FractionBelowThr}
\alias{scBarplot.FractionBelowThr}
\title{scBarplot.FractionBelowThr}
\usage{
scBarplot.FractionBelowThr(
  thrX = 0.01,
  value.col = "percent.ribo",
  id.col = "cl.names.top.gene.res.0.3",
  obj = combined.obj,
  return.df = F
)
}
\arguments{
\item{thrX}{The threshold value to determine the fraction of cells. Default: 0.01}

\item{value.col}{Column name from metadata which holds the values for calculating the fraction of cells. Default: 'percent.ribo'}

\item{id.col}{Column name from metadata to be used for identifying clusters. Default: 'cl.names.top.gene.res.0.3'}

\item{obj}{A Seurat object holding single cell data. Default: combined.obj}

\item{return.df}{A logical indicating if the function should return the data frame used to create the plot. Default: FALSE}
}
\description{
Create a bar plot showing the fraction of cells, within each cluster, that are below a certain threshold based on a metadata column.
}
\examples{
\dontrun{
if (interactive()) {
  scBarplot.FractionBelowThr(id.col = "cl.names.top.gene.res.0.3", value.col = "percent.ribo", thrX = 0.01, return.df = TRUE)
}
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{se-deprecated}}
}
