% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scBarplot.FractionBelowThr}
\alias{scBarplot.FractionBelowThr}
\title{Fraction of Cells Below Threshold per Cluster}
\usage{
scBarplot.FractionBelowThr(
  thrX = 0.2,
  value.col = "percent.ribo",
  id.col = "cl.names.top.gene.res.0.3",
  obj = combined.obj,
  return.df = FALSE,
  subtitle = id.col,
  suffix = NULL,
  ...
)
}
\arguments{
\item{thrX}{Threshold value for assessing cell counts. Default: 0.01.}

\item{value.col}{Metadata column with values for threshold comparison. Default: 'percent.ribo'.}

\item{id.col}{Cluster identifier in metadata. Default: 'cl.names.top.gene.res.0.3'.}

\item{obj}{Seurat object with cell data. Default: \code{combined.obj}.}

\item{return.df}{If TRUE, returns the data frame instead of the plot. Default: FALSE.}
}
\description{
Generates a bar plot to visualize the percentage of cells within each cluster that
fall below a specified threshold, according to a metadata column value.
Inherits all parameters from \code{scBarplot.FractionAboveThr} with the exception that \code{above} is set to FALSE.
}
\examples{
\dontrun{
scBarplot.FractionBelowThr(id.col = "cl.names.top.gene.res.0.3", value.col = "percent.ribo", thrX = 0.01)
}

}
\seealso{
\code{scBarplot.FractionAboveThr}

\code{\link[dplyr]{select}}, \code{\link[dplyr]{group_by}}
}
