% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scBarplotEnrichr}
\alias{scBarplotEnrichr}
\title{Barplot GO Enrichment Results by enrichplot}
\usage{
scBarplotEnrichr(
  df.enrichment,
  showCategory = 20,
  label_format = 30,
  tag = "...",
  universe = df.enrichment@universe,
  title = paste("GO Enriched Terms", tag),
  subtitle = Stringendo::kppws("Input: ", Stringendo::substitute_deparse(df.enrichment)),
  caption = paste0("Input genes: ", length(df.enrichment@gene), " | Enriched terms: ",
    nrow(df.enrichment), " | Shown: ", min(showCategory, nrow(df.enrichment)),
    " | background genes: ", length(universe)),
  save = TRUE,
  also.pdf = FALSE,
  save.obj = FALSE,
  w = 10,
  h = 10,
  ...
)
}
\arguments{
\item{df.enrichment}{Data frame. Enrichment results from GO analysis. Default: NULL.}

\item{showCategory}{Integer. Number of categories (GO-terms as bars) to show in the plot. Default: 20.}

\item{label_format}{Integer. Wrap length (in characters) for category labels. Default: 30.}

\item{tag}{Character. Tag to be added to the title of the plot. Default: "in ...".}

\item{universe}{Character. Background gene list (universe). Default: \code{df.enrichment@universe}.}

\item{title}{Character. Title of the plot. Default: "GO Enrichment Analysis" followed by \code{tag}.}

\item{subtitle}{Character. Subtitle of the plot. Default: NULL.}

\item{caption}{Character. Caption of the plot. Default: constructed from input parameters.}

\item{save}{Logical. Whether to save the plot to a file. Default: \code{TRUE}.}

\item{also.pdf}{Save plot in both png and pdf formats.}

\item{save.obj}{Logical. Whether to save the ggplot object. Default: FALSE.}

\item{w}{Width of the plot canvas, calculated as the width of an A4 page times \code{scale}; Default: \code{11.69 * scale}.}

\item{h}{Height of the plot canvas, calculated as the height of an A4 page times \code{scale}; Default: \code{8.27 * scale}.}

\item{...}{Additional arguments passed to \code{enrichplot::barplot.enrichResult}.}
}
\value{
None. The function prints the plot and optionally saves it.
}
\description{
This function creates a bar plot of GO enrichment analysis results using the
\code{enrichplot::barplot.enrichResult} function. It also allows saving the plot to a file.
}
\examples{
\dontrun{
df.enrichment <- data.frame() # Example enrichment results data frame
plotGOEnrichment(df.enrichment)
}
}
