% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scBarplotEnrichr}
\alias{scBarplotEnrichr}
\title{Barplot GO Enrichment Results by enrichplot}
\usage{
scBarplotEnrichr(
  df.enrichment,
  tag = "...",
  universe = NULL,
  title = paste(tag, "GO Enrichment Analysis"),
  subtitle = kppws("Input: ", substitute(df.enrichment)),
  caption = paste0("genes: ", nrow(df.enrichment), " | background: ", length(universe)),
  save = TRUE,
  ...
)
}
\arguments{
\item{df.enrichment}{Data frame. Enrichment results from GO analysis. Default: NULL.}

\item{tag}{Character. Tag to be added to the title of the plot. Default: "in ...".}

\item{universe}{Character. Background gene list (universe). Default: NULL.}

\item{title}{Character. Title of the plot. Default: "GO Enrichment Analysis" followed by \code{tag}.}

\item{subtitle}{Character. Subtitle of the plot. Default: NULL.}

\item{caption}{Character. Caption of the plot. Default: constructed from input parameters.}

\item{save}{Logical. Whether to save the plot to a file. Default: TRUE.}

\item{...}{Additional arguments passed to \code{enrichplot::barplot.enrichResult}.}
}
\value{
None. The function prints the plot and optionally saves it.
}
\description{
This function creates a bar plot of GO enrichment analysis results using the
\code{enrichplot::barplot.enrichResult} function. It also allows saving the plot to a file.
}
\examples{
\dontrun{
df.enrichment <- data.frame() # Example enrichment results data frame
plotGOEnrichment(df.enrichment)
}
}
