% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scCalcPCAVarExplained}
\alias{scCalcPCAVarExplained}
\title{PCA percent of variation associated with each PC}
\usage{
scCalcPCAVarExplained(obj = combined.obj)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}
}
\description{
Determine percent of variation associated with each PC. For normal prcomp objects, see: PCA.percent.var.explained().
}
