% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{scEnhancedVolcano}
\alias{scEnhancedVolcano}
\title{scEnhancedVolcano}
\usage{
scEnhancedVolcano(
  toptable = df.markers.X,
  lab = rownames(toptable),
  suffix = "",
  title = kppd("DGEA", suffix),
  subtitle = paste("min FC:", iround(2^abs(min(toptable$avg_log2FC)))),
  x = "avg_log2FC",
  y = "p_val_adj",
  selectLab = lab[1:3],
  h = 8,
  w = h
)
}
\description{
scEnhancedVolcano
}
