% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{scEnhancedVolcano}
\alias{scEnhancedVolcano}
\title{scEnhancedVolcano}
\usage{
scEnhancedVolcano(
  toptable,
  lab = rownames(toptable),
  min.p = 1e-50,
  suffix = "",
  title = paste("DGEA"),
  subtitle = paste("min FC:", iround(2^abs(min(toptable$avg_log2FC)))),
  caption = paste("min p cutoff:", min.p),
  x = "avg_log2FC",
  y = "p_val_adj",
  selectLab = trail(lab, 10),
  pCutoffCol = "p_val_adj",
  pCutoff = 0.001,
  h = 8,
  w = h,
  ...
)
}
\arguments{
\item{toptable}{A data frame with the results of differential gene expression analysis.}

\item{lab}{A vector of gene symbols to label on the plot.}

\item{suffix}{A string to append to the title of the plot.}

\item{title}{The title of the plot.}

\item{subtitle}{The subtitle of the plot.}

\item{x}{The x-axis, which is typically the average log2 fold change.}

\item{y}{The y-axis, which is typically the adjusted p-value.}

\item{selectLab}{A vector of gene symbols to select for labeling.}

\item{h}{The height of the plot.}

\item{w}{The width of the plot.}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\value{
A ggplot object.
}
\description{
This function creates an enhanced volcano plot.
}
