% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{SelectHighlyExpressedGenesq99}
\alias{SelectHighlyExpressedGenesq99}
\title{Intersect Genes with the List of Noticeably Expressed Genes}
\usage{
SelectHighlyExpressedGenesq99(genes, obj = combined.obj, above = 0, sort = F)
}
\arguments{
\item{genes}{A vector of gene names to be intersected with the Seurat object.}

\item{obj}{A Seurat object containing gene expression data. Default: \code{combined.obj}.}

\item{above}{The expression level threshold above which genes are considered noticeably
expressed. Default: 0.}

\item{sort}{A logical flag indicating whether to sort the filtered genes by their expression
levels in decreasing order. Default: FALSE.}
}
\value{
A vector of gene names that are found both in the input 'genes' vector and the Seurat
object, and have expression levels above the specified 'above' threshold. If \code{sort} is TRUE,
these genes are returned in decreasing order of their expression levels.
}
\description{
Intersects a vector of gene names with a Seurat object to find genes that are both
in the input list and have expression levels in the top quantiles as defined by the object's
q99 expression data. It aims to filter genes based on their expression levels being above a
specified threshold. Additionally, it offers an option to sort the genes by their expression
levels in decreasing order.
}
\examples{
# Assuming `genes` is a vector of gene names and `
}
