% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{set.all.genes}
\alias{set.all.genes}
\title{set.all.genes}
\usage{
set.all.genes(obj = combined.obj)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}
}
\description{
It is just a reminder to use calc.q99.Expression.and.set.all.genes to create the all.genes variable
}
\examples{
\dontrun{
if (interactive()) {
  set.all.genes()
  all.genes
}
}
}
