% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R, R/Seurat.Utils.Visualization.R
\name{SetupReductionsNtoKdimensions}
\alias{SetupReductionsNtoKdimensions}
\title{SetupReductionsNtoKdimensions}
\usage{
SetupReductionsNtoKdimensions(
  obj = combined.obj,
  nPCs = p$n.PC,
  dimensions = 3:2,
  reduction = "umap",
  ...
)

SetupReductionsNtoKdimensions(
  obj = combined.obj,
  nPCs = p$n.PC,
  dimensions = 3:2,
  reduction = "umap",
  ...
)
}
\arguments{
\item{obj}{Seurat object to process; Default: \code{combined.obj}.}

\item{nPCs}{Number of principal components to consider in the reduction; Default: \code{p$n.PC}.}

\item{dimensions}{Numeric vector specifying target dimensions for the reductions; Default: \code{3:2}.}

\item{reduction}{Type of dimensionality reduction to apply ('umap', 'tsne', or 'pca'); Default: 'umap'.}

\item{...}{Additional parameters to pass to the dimensionality reduction functions.}
}
\value{
The input Seurat object with computed dimensionality reductions and backups of these reductions.

Modified Seurat object with added dimensionality reduction data and backups.
}
\description{
Function to calculate N-to-K dimensional umaps (default = 2:3); and back them up to
slots \code{obj@misc$reductions.backup} from @reductions$umap

Executes specified dimensionality reduction (UMAP, tSNE, or PCA) over a range of dimensions
and backs up the results within a Seurat object. This function allows for exploration of data structure
at varying levels of granularity and ensures that reduction results are preserved for future reference.
}
\examples{
\dontrun{
if (interactive()) {
  combined.obj <- SetupReductionsNtoKdimensions(obj = combined.obj, nPCs = 10, dimensions = 2:3, reduction = "umap")
}
}
\dontrun{
  if (interactive()) {
    combined.obj <- SetupReductionsNtoKdimensions(obj = combined.obj, nPCs = 10,
    dimensions = 2:3, reduction = "umap")
  }
}

}
