% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{SetupReductionsNtoKdimensions}
\alias{SetupReductionsNtoKdimensions}
\title{SetupReductionsNtoKdimensions}
\usage{
SetupReductionsNtoKdimensions(
  obj = combined.obj,
  nPCs = p$n.PC,
  dimensions = 3:2,
  reduction = "umap",
  ...
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{nPCs}{PARAM_DESCRIPTION, Default: p$n.PC}

\item{dimensions}{PARAM_DESCRIPTION, Default: 3:2}

\item{reduction}{UMAP, tSNE, or PCA (Dim. reduction to use), Default: 'umap'}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Calculate N-to-K dimensional umaps (default = 2:3); and back them up UMAP to \code{obj@misc$reductions.backup} from @reductions$umap #
}
\examples{
\dontrun{
if(interactive()){
 combined.obj <- SetupReductionsNtoKdimensions(obj = combined.obj, nPCs = p$'n.PC', dimensions = 2:3, reduction="umap"); qUMAP()
 }
}
}
