% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{SetupReductionsNtoKdimensions}
\alias{SetupReductionsNtoKdimensions}
\title{SetupReductionsNtoKdimensions}
\usage{
SetupReductionsNtoKdimensions(
  obj = combined.obj,
  nPCs = p$n.PC,
  dimensions = 3:2,
  reduction = "umap",
  ...
)
}
\arguments{
\item{obj}{A Seurat object. Default: combined.obj}

\item{nPCs}{A numeric value representing the number of principal components to use. Default: p$n.PC}

\item{dimensions}{A numeric vector specifying the dimensions to use for the dimensionality reductions. Default: 3:2}

\item{reduction}{A character string specifying the type of dimensionality reduction to perform. Can be "umap", "tsne", or "pca". Default: 'umap'}
}
\value{
The input Seurat object with computed dimensionality reductions and backups of these reductions.
}
\description{
Function to compute dimensionality reductions for a given Seurat object and backup the computed reductions.
}
\examples{
\dontrun{
if (interactive()) {
  combined.obj <- SetupReductionsNtoKdimensions(obj = combined.obj, nPCs = 10, dimensions = 2:3, reduction = "umap")
}
}
}
