% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{seu.add.meta.from.vector}
\alias{seu.add.meta.from.vector}
\title{seu.add.meta.from.vector}
\usage{
seu.add.meta.from.vector(
  obj = combined.obj,
  metaD.colname = metaD.colname.labeled,
  Label.per.cell = Cl.Label.per.cell
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{metaD.colname}{PARAM_DESCRIPTION, Default: metaD.colname.labeled}

\item{Label.per.cell}{PARAM_DESCRIPTION, Default: Cl.Label.per.cell}
}
\description{
Add a new metadata column to a Seurat  object #
}
\examples{
\dontrun{
if(interactive()){
 combined.obj <- add.Cl.Label.2.Metadata(obj = combined.obj, metaD.colname = metaD.colname.labeled, Label.per.cell=Cl.Label.per.cell ) # formerly add.Cl.Label.2.Metadata
 }
}
}
