% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{seu.Make.Cl.Label.per.cell}
\alias{seu.Make.Cl.Label.per.cell}
\title{seu.Make.Cl.Label.per.cell}
\usage{
seu.Make.Cl.Label.per.cell(TopGenes, clID.per.cell)
}
\arguments{
\item{TopGenes}{A named vector, where values are gene names and names are cluster IDs.}

\item{clID.per.cell}{A vector of cell-IDs used to create the output vector.}
}
\description{
Take a named vector (of e.g. values ="gene names", names = clusterID), and a vector of cell-IDs and make a vector of "GeneName.ClusterID".
}
\examples{
\dontrun{
if (interactive()) {
  seu.Make.Cl.Label.per.cell(TopGenes = TopGenes.Classic, clID.per.cell = getMetadataColumn(ColName.metadata = metaD.CL.colname))
}
}
}
