% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{seu.map.and.add.new.ident.to.meta}
\alias{seu.map.and.add.new.ident.to.meta}
\title{seu.map.and.add.new.ident.to.meta}
\usage{
seu.map.and.add.new.ident.to.meta(
  obj = combined.obj,
  ident.table = clusterIDs.GO.process,
  metaD.colname = substitute(ident.table)
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{ident.table}{PARAM_DESCRIPTION, Default: clusterIDs.GO.process}

\item{metaD.colname}{PARAM_DESCRIPTION, Default: substitute(ident.table)}
}
\description{
seu.map.and.add.new.ident.to.meta
}
\examples{
\dontrun{
if(interactive()){
 combined.obj <- seu.map.and.add.new.ident.to.meta(obj = combined.obj, ident.table = clusterIDs.GO.process)
 }
}
}
