% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{seu.plot.PC.var.explained}
\alias{seu.plot.PC.var.explained}
\title{seu.plot.PC.var.explained}
\usage{
seu.plot.PC.var.explained(obj = combined.obj, use.MarkdownReports = F)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{use.MarkdownReports}{Use MarkdownReports for plotting, Default: F}
}
\description{
Plot the percent of variation associated with each PC.
}
