% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{seu.RemoveMetadata}
\alias{seu.RemoveMetadata}
\title{seu.RemoveMetadata}
\usage{
seu.RemoveMetadata(
  obj = combined.obj,
  cols_remove = grepv(colnames(obj@meta.data), pattern = "^integr|^cl.names", perl = T)
)
}
\arguments{
\item{obj}{A Seurat object from which metadata columns will be removed. Default: combined.obj}

\item{cols_remove}{A character vector specifying metadata column names to remove. By default, it will remove all columns that do not start with "integr" or "cl.names".}
}
\value{
A Seurat object with specified metadata columns removed.
}
\description{
Remove specified metadata columns from a Seurat object.
}
\examples{
\dontrun{
combined.obj <- seu.RemoveMetadata(obj = combined.obj, cols_remove = c("column1", "column2"))
}
}
