% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{seu.RemoveMetadata}
\alias{seu.RemoveMetadata}
\title{seu.RemoveMetadata}
\usage{
seu.RemoveMetadata(
  obj = combined.obj,
  cols_remove = grepv(colnames(obj@meta.data), pattern = "^integr|^cl.names", perl = T)
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{cols_remove}{columns to remove}
}
\description{
seu.RemoveMetadata
}
