% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{SeuratColorVector}
\alias{SeuratColorVector}
\title{SeuratColorVector}
\usage{
SeuratColorVector(
  ident = NULL,
  obj = combined.obj,
  plot.colors = F,
  simple = F
)
}
\arguments{
\item{ident}{identity used, Default: NULL}

\item{obj}{Seurat object, Default: combined.obj}

\item{plot.colors}{PARAM_DESCRIPTION, Default: F}

\item{simple}{Return simply the unique colors, in order? Default: F}
}
\description{
Recall a Seurat color vector.
}
\examples{
\dontrun{
if(interactive()){
 SeuratColorVector(); SeuratColorVector(ident = GetNamedClusteringRuns()[2], plot.colors = T)
 }
}
}
\seealso{
\code{\link[scales]{hue_pal}}
}
