% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{showToolsSlots}
\alias{showToolsSlots}
\title{Display Slots in the @tools of an Seurat Object}
\usage{
showToolsSlots(obj, max.level = 1, subslot = NULL, ...)
}
\arguments{
\item{obj}{An object whose \verb{@tools} slot needs to be examined.}
}
\description{
\code{showToolsSlots} prints the names of slots in the \verb{@tools} of a given object.
It specifically targets list elements, skipping over data frames and other non-list objects.
}
\details{
The function iterates over the slots in the \verb{@tools} of \code{obj}. If a slot is a list
(and not a data frame), it prints the names of elements within this list. If the slot
is not a list or is a data frame, it skips printing the names. The function currently
does not use the \code{indent} parameter but it could be incorporated in future enhancements
to control the formatting of the output.
}
\examples{
showToolsSlots(obj)

}
