% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{SmallestNonAboveX}
\alias{SmallestNonAboveX}
\title{SmallestNonAboveX}
\usage{
SmallestNonAboveX(vec, X = 0)
}
\arguments{
\item{vec}{Numeric input vector}

\item{X}{Threshold, Default: 0}
}
\description{
replace small values with the next smallest value found, which is >X. #
}
\examples{
\dontrun{
if(interactive()){
 SmallestNonZero(vec = df.markers$"p_val")
 }
}
}
