% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{SNP.demux.fix.GT.table}
\alias{SNP.demux.fix.GT.table}
\title{SNP.demux.fix.GT.table}
\usage{
SNP.demux.fix.GT.table(
  GT.table,
  obj,
  suffix,
  col1 = "assignment.named",
  col1.new.name = "Genotype",
  col2 = "status",
  col2.new.name = "Singlet.status",
  cellname_prefix = FALSE,
  return_tbl_for_cells_found_in_object = TRUE,
  min_cells_overlap = 10
)
}
\arguments{
\item{GT.table}{A 2-column dataframe representing the Genotype assignment table with singlet status. This is a mandatory input.}

\item{obj}{A Seurat object to which the corrections will be applied. This is a mandatory input.}

\item{suffix}{A character string that will be added to the 'Singlet.status' column and used to overwrite the 'Genotype' column. This is a mandatory input.}

\item{col1}{The column to use for the 'Genotype' information in the input table. Defaults to 'assignment.named'.}

\item{col1.new.name}{The new name to use for the 'Genotype' column. Defaults to 'Genotype'.}

\item{col2}{The column to use for the 'Singlet.status' information in the input table. Defaults to 'status'.}

\item{col2.new.name}{The new name to use for the 'Singlet.status' column. Defaults to 'Singlet.status'.}

\item{cellname_prefix}{If TRUE, the function assumes that Seurat has added a prefix to cell names to distinguish 10X lanes and adjusts accordingly. Defaults to FALSE.}

\item{return_tbl_for_cells_found_in_object}{If TRUE, the function returns the subset of the GT table corresponding to cells found in both the GT table and the Seurat object. Defaults to TRUE.}

\item{min_cells_overlap}{min_cells_overlap}
}
\value{
The cleaned and standardized GT table, with specific transformations applied as defined by the input parameters.
}
\description{
This function cleans and standardizes a Genotype assignment table obtained from the SoupOrCell tool.
}
\examples{
# Assuming 'GT.table' is your Genotype assignment table, 'obj' is your Seurat object and 'suffix' is your chosen suffix
# GT.table.cleaned <- SNP.demux.fix.GT.table(GT.table = GT.table, obj = obj, suffix = suffix)
}
