% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{sparse.cor}
\alias{sparse.cor}
\title{sparse.cor}
\usage{
sparse.cor(smat)
}
\arguments{
\item{smat}{A sparse matrix.}
}
\value{
A list with two elements:
\itemize{
\item \code{cov}: The covariance matrix.
\item \code{cor}: The correlation matrix.
}
}
\description{
Calculate a sparse correlation matrix.
}
