% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{StoreAllMarkers}
\alias{StoreAllMarkers}
\title{StoreAllMarkers}
\usage{
StoreAllMarkers(
  obj = combined.obj,
  df_markers = df.markers,
  res = 0.5,
  digit = c(0, 3)[2]
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{df_markers}{Data frame, result of DGEA analysis (FindAllMarkers), Default: df.markers}

\item{res}{Clustering resoluton to use, Default: 0.5}

\item{digit}{Number of digits to keep, Default: c(0, 3)\link{2}}
}
\description{
Save the output table of \code{FindAllMarkers()} (df_markers) under \verb{@misc$df.markers$res...}. By default, it rounds up insignificant digits up to 3. #
}
\examples{
\dontrun{
if(interactive()){
 combined.obj <- StoreAllMarkers(df_markers = df.markers, res = 0.5)
 }
}
}
