% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{subsetMonocleObject}
\alias{subsetMonocleObject}
\title{subsetMonocleObject}
\usage{
subsetMonocleObject(
  obj = cds_from_seurat,
  fraction_ = 0.1,
  nCells = F,
  seed_ = 1989
)
}
\arguments{
\item{obj}{Seurat object, Default: cds_from_seurat}

\item{fraction_}{PARAM_DESCRIPTION, Default: 0.1}

\item{nCells}{PARAM_DESCRIPTION, Default: F}

\item{seed_}{PARAM_DESCRIPTION, Default: 1989}
}
\description{
Subset a compressed Seurat Obj and save it in wd. #
}
\examples{
\dontrun{
if(interactive()){
 cds.10pc <- subsetMonocleObject(cds_from_seurat, fraction_ = .1);  cds.25pc <- subsetMonocleObject(cds_from_seurat, fraction_ = .25)
 }
}
}
