% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{subsetSeuObj.and.Save}
\alias{subsetSeuObj.and.Save}
\title{subsetSeuObj.and.Save}
\usage{
subsetSeuObj.and.Save(
  obj = ORC,
  fraction = 0.25,
  seed = 1989,
  dir = OutDir,
  min.features = p$min.features,
  suffix = ""
)
}
\arguments{
\item{obj}{Seurat object, Default: ORC}

\item{fraction}{PARAM_DESCRIPTION, Default: 0.25}

\item{seed}{random seed used, Default: 1989}

\item{dir}{PARAM_DESCRIPTION, Default: OutDir}

\item{min.features}{Minimum features}

\item{suffix}{A suffix added to the filename, Default: ''}
}
\description{
Subset a compressed Seurat Obj and save it in wd. #
}
