% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{subsetSeuObj}
\alias{subsetSeuObj}
\title{subsetSeuObj}
\usage{
subsetSeuObj(obj = ls.Seurat[[i]], fraction_ = 0.25, nCells = F, seed_ = 1989)
}
\arguments{
\item{obj}{A Seurat object to subset. Default: the i-th element of the list 'ls.Seurat'.}

\item{fraction_}{The fraction of the object's data to keep. Default: 0.25.}

\item{nCells}{If set to a number greater than 1, indicates the absolute number of cells to keep. If FALSE, the function uses 'fraction_' to determine the number of cells. Default: FALSE.}

\item{seed_}{A seed for random number generation to ensure reproducible results. Default: 1989.}
}
\description{
Subset a compressed Seurat object and save it in the working directory.
}
