% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{suPlotVariableFeatures}
\alias{suPlotVariableFeatures}
\title{suPlotVariableFeatures for Single Seurat Object}
\usage{
suPlotVariableFeatures(
  obj = combined.obj,
  NrVarGenes = 15,
  repel = TRUE,
  plotWidth = 7,
  plotHeight = 5,
  save = TRUE,
  suffix = kpp("nVF", .getNrScaledFeatures(obj)),
  ...
)
}
\arguments{
\item{obj}{A single Seurat object.}

\item{NrVarGenes}{A vector containing the top 20 variable genes for the Seurat object.}

\item{repel}{A logical value indicating whether to repel the labels to avoid overlap. Default: TRUE.}

\item{plotWidth}{Numeric value specifying the width of the plot when saved. Default: 7.}

\item{plotHeight}{Numeric value specifying the height of the plot when saved. Default: 5.}

\item{sampleName}{A string specifying the sample name, used to generate the filename for saving
the plot.}

\item{ppp}{A function for constructing the path and filename for saving the plot. It takes three
arguments: a prefix for the filename, a sample name, and the file extension ('pdf').}
}
\description{
Generates a Variable Feature Plot for a specified Seurat object, labels points with
the top 20 variable genes, and saves the plot to a PDF file.
}
\examples{
\dontrun{
suPlotVariableFeatures(combined.obj)
}
}
