% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Metadata.R
\name{transferLabelsSeurat}
\alias{transferLabelsSeurat}
\title{Transfer labels from a reference Seurat object to a query Seurat object}
\usage{
transferLabelsSeurat(
  query_obj,
  reference_obj,
  reference_path = NULL,
  reference_ident,
  anchors = NULL,
  new_ident = gsub(pattern = "ordered", replacement = "transferred", x = reference_ident),
  predictions_col = "predicted.id",
  save_anchors = TRUE,
  reference_suffix = "reference",
  new_ident_suffix = "R44",
  plot_reference = TRUE,
  ...
)
}
\arguments{
\item{query_obj}{A Seurat object for which the labels are to be transferred.}

\item{reference_obj}{Alternative to \code{reference_path}. If provided, the path is not used to load the reference data.}

\item{reference_path}{A character string indicating the file path to the reference Seurat object. The path must exist.}

\item{new_ident}{A character string specifying the name of the new identity class to be created in the query Seurat object. Default is obtained by replacing 'ordered' with 'transferred' in named_ident.}

\item{predictions_col}{A character string specifying the column in the metadata of the transferred Seurat object containing the transferred labels. Default is 'predicted.id'.}

\item{save_anchors}{save anchors as RDS file.}

\item{new_ident_suffix}{A string to added to the UMAP with the new identity.}

\item{...}{Additional arguments passed to the Seurat.utils::clUMAP function.}

\item{named_ident}{A character string specifying the name of the identity class to be used from the reference Seurat object. Default is 'RNA_snn_res.0.3.ordered.ManualNames'.}

\item{suffix}{A character string to be used as a suffix in the visualization. Default is 'NEW'.}
}
\value{
The modified query Seurat object with the transferred labels as a new identity class.
}
\description{
Function to transfer labels from a reference Seurat object to a query Seurat object
using anchoring and transfer data methods from the Seurat package. It then visualizes the
reference and the combined objects using Uniform Manifold Approximation and Projection (UMAP).
}
\examples{
# combined.objX <- transferLabelsSeurat(named_ident = 'RNA_snn_res.0.3.ordered.ManualNames',
#                                     reference_obj = reference_obj,
#                                     query_obj = combined.obj)

}
