% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{umapHiLightSel}
\alias{umapHiLightSel}
\title{umapHiLightSel}
\usage{
umapHiLightSel(
  obj = combined.obj,
  COI = c("0", "2", "4", "5", "11"),
  res.cl = "integrated_snn_res.0.3"
)
}
\arguments{
\item{obj}{A Seurat object. Default: combined.obj.}

\item{COI}{A vector of cluster IDs to highlight in the UMAP plot. Default: c("0", "2", "4", "5",  "11").}

\item{res.cl}{Name of the column in the Seurat object metadata that contains the cluster IDs. Default: 'integrated_snn_res.0.3'.}
}
\value{
This function does not return a value. It saves a UMAP plot to the current working directory.
}
\description{
Generates a UMAP plot from a Seurat object with a subset of cells highlighted.
}
\examples{
\dontrun{
if(interactive()){
  umapHiLightSel(obj = seuratObject, COI = c("0", "1"), res.cl = "resolution_0.8")
}
}
}
\seealso{
\code{\link[Seurat]{DimPlot}}
}
