% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{umapHiLightSel}
\alias{umapHiLightSel}
\title{umapHiLightSel}
\usage{
umapHiLightSel(
  obj = combined.obj,
  COI = c("0", "2", "4", "5", "11"),
  res.cl = "integrated_snn_res.0.3"
)
}
\arguments{
\item{obj}{Seurat object, Default: combined.obj}

\item{COI}{PARAM_DESCRIPTION, Default: c("0", "2", "4", "5", "11")}

\item{res.cl}{PARAM_DESCRIPTION, Default: 'integrated_snn_res.0.3'}
}
\description{
Highlight a set of cells based on clusterIDs provided. #
}
