% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{UpdateGenesSeurat}
\alias{UpdateGenesSeurat}
\title{UpdateGenesSeurat}
\usage{
UpdateGenesSeurat(
  obj = ls.Seurat[[i]],
  species_ = "human",
  EnforceUnique = T,
  ShowStats = F
)
}
\arguments{
\item{obj}{Seurat object, Default: ls.Seurat[\link{i}]}

\item{species_}{PARAM_DESCRIPTION, Default: 'human'}

\item{EnforceUnique}{PARAM_DESCRIPTION, Default: T}

\item{ShowStats}{PARAM_DESCRIPTION, Default: F}
}
\description{
Update genes symbols that are stored in a Seurat object. It returns a data frame. The last column are the updated gene names. #
}
\examples{
\dontrun{
if(interactive()){
 UpdateGenesSeurat()
 }
}
}
\seealso{
\code{\link[HGNChelper]{checkGeneSymbols}}
}
