% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{UpdateGenesSeurat}
\alias{UpdateGenesSeurat}
\title{UpdateGenesSeurat}
\usage{
UpdateGenesSeurat(
  obj = ls.Seurat[[i]],
  species_ = "human",
  EnforceUnique = TRUE,
  ShowStats = FALSE
)
}
\arguments{
\item{obj}{Seurat object to update gene symbols in. Default: ls.Seurat[\link{i}]}

\item{species_}{Species to which the gene symbols correspond, used to check gene symbols. Default: 'human'}

\item{EnforceUnique}{Logical, if TRUE, enforces unique gene symbols. Default: TRUE}

\item{ShowStats}{Logical, if TRUE, displays statistics of gene symbols update. Default: FALSE}
}
\description{
Update genes symbols that are stored in a Seurat object. It returns a data frame. The last column are the updated gene names.
}
\seealso{
\code{\link[HGNChelper]{checkGeneSymbols}}
}
