% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{whitelist.subset.ls.Seurat}
\alias{whitelist.subset.ls.Seurat}
\title{whitelist.subset.ls.Seurat}
\usage{
whitelist.subset.ls.Seurat(
  ls.obj = ls.Seurat,
  metadir = p$cellWhiteList,
  whitelist.file = "NonStressedCellIDs.2020.10.21_18h.tsv"
)
}
\arguments{
\item{ls.obj}{List of Seurat objects, Default: ls.Seurat}

\item{metadir}{PARAM_DESCRIPTION, Default: p$cellWhiteList}

\item{whitelist.file}{PARAM_DESCRIPTION, Default: 'NonStressedCellIDs.2020.10.21_18h.tsv'}
}
\description{
Subset cells in a (list of) Seurat objects, based on an externally provided list of cell IDs.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
