% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{whitelist.subset.ls.Seurat}
\alias{whitelist.subset.ls.Seurat}
\title{whitelist.subset.ls.Seurat}
\usage{
whitelist.subset.ls.Seurat(
  ls.obj = ls.Seurat,
  metadir = p$cellWhiteList,
  whitelist.file = "NonStressedCellIDs.2020.10.21_18h.tsv"
)
}
\arguments{
\item{ls.obj}{A list of Seurat objects. Default: ls.Seurat.}

\item{metadir}{Directory for the metadata. Default: p$cellWhiteList.}

\item{whitelist.file}{Filename of the whitelist containing cell IDs. Default: "NonStressedCellIDs.2020.10.21_18h.tsv".}
}
\value{
A list of Seurat objects containing only the cells specified in the whitelist.
}
\description{
Subsets cells in a list of Seurat objects based on an externally provided list of cell IDs.
}
\details{
The function first validates the presence of all identities from the metadata in the
Seurat objects. If all identities are present, the function subsets each Seurat object based on
the whitelist of cell IDs.
}
\examples{
\dontrun{
if (interactive()) {
  ls.Seurat.subset <- whitelist.subset.ls.Seurat(
    ls.obj = ls.Seurat, metadir = p$"cellWhiteList",
    whitelist.file = "NonStressedCellIDs.2020.10.21_18h.tsv"
  )
}
}
}
\seealso{
\code{\link[Seurat]{subset}}
}
