% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.Visualization.R
\name{ww.calc_helper}
\alias{ww.calc_helper}
\title{Helper to calculate Cell Expression Proportion for Gene}
\source{
Adapted from Ryan-Zhu on GitHub.
}
\usage{
ww.calc_helper(obj, genes)
}
\arguments{
\item{obj}{Seurat object with cell data.}

\item{genes}{Single gene name as a character string.}
}
\value{
Proportion of cells expressing the gene. Returns \code{NA} if the gene is not found.
}
\description{
Computes the proportion of cells expressing a specific gene within a Seurat object.
}
\examples{
\dontrun{
ww.calc_helper(obj = seurat_object, genes = "Gene1")
}

}
