% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{ww.check.quantile.cutoff.and.clip.outliers}
\alias{ww.check.quantile.cutoff.and.clip.outliers}
\title{ww.check.quantile.cutoff.and.clip.outliers}
\usage{
ww.check.quantile.cutoff.and.clip.outliers(
  expr.vec = plotting.data[, gene],
  quantileCutoffX = quantileCutoff,
  min.cells.expressing = 10
)
}
\arguments{
\item{expr.vec}{PARAM_DESCRIPTION, Default: plotting.data\link{, gene}}

\item{quantileCutoffX}{PARAM_DESCRIPTION, Default: quantileCutoff}

\item{min.cells.expressing}{PARAM_DESCRIPTION, Default: 10}
}
\description{
Helper function.
}
