% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{ww.check.quantile.cutoff.and.clip.outliers}
\alias{ww.check.quantile.cutoff.and.clip.outliers}
\title{ww.check.quantile.cutoff.and.clip.outliers}
\usage{
ww.check.quantile.cutoff.and.clip.outliers(
  expr.vec = plotting.data[, gene],
  quantileCutoffX = quantileCutoff,
  min.cells.expressing = 10
)
}
\arguments{
\item{expr.vec}{A numeric vector representing gene expression data. Default: plotting.data\link{, gene}}

\item{quantileCutoffX}{A numeric value representing the quantile at which to clip outliers. Default: quantileCutoff}

\item{min.cells.expressing}{A numeric value representing the minimum number of cells expressing a gene that should remain after clipping outliers. Default: 10}
}
\value{
The input expression vector with outliers clipped.
}
\description{
Function to check a specified quantile cutoff and clip outliers from a given expression vector.
}
\examples{
\dontrun{
ww.check.quantile.cutoff.and.clip.outliers(expr.vec = expr.data, quantileCutoffX = 0.99, min.cells.expressing = 10)
}
}
