% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{ww.get.1st.Seur.element}
\alias{ww.get.1st.Seur.element}
\title{Get the First Seurat Object from a List of Seurat Objects}
\usage{
ww.get.1st.Seur.element(obj)
}
\arguments{
\item{obj}{A Seurat object, a list of Seurat objects, or any other list.}
}
\value{
The first Seurat object from the list or the Seurat object itself.
If the input is not a Seurat object or a list containing at least one Seurat
object, the function will throw an error.
}
\description{
Return the first Seurat object from a list, or the object itself
if a single Seurat object is supplied.
}
