% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{xread}
\alias{xread}
\title{Read an R Object Using 'qs' Package for Fast Decompression}
\usage{
xread(
  file,
  nthreads = 4,
  loadParamsAndAllGenes = TRUE,
  overwriteParams = FALSE,
  overwriteAllGenes = FALSE,
  ...
)
}
\arguments{
\item{file}{A character string specifying the path to the file where the R object is saved.}

\item{nthreads}{The number of threads to use when reading the object, defaults to 4.}

\item{...}{Further arguments passed on to the 'qs::qread' function.}
}
\value{
The R object that was saved in the specified file.
}
\description{
This function reads an R object from a file saved in a format specific to the 'qs' package,
which is designed for quick and efficient compression and decompression of R objects.
It also times the read operation, providing feedback on the duration of the operation.
}
\note{
The function uses the 'qs' package for fast and efficient deserialization of objects
and includes a timing feature from the 'tictoc' package.
}
\seealso{
\code{\link[qs]{qread}} for the underlying read function used.
}
