% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.Utils.R
\name{xsave}
\alias{xsave}
\title{Save an R Object Using 'qs' Package for Fast Compressed Saving}
\usage{
xsave(
  obj,
  prefix = NULL,
  suffix = NULL,
  nthreads = 12,
  preset = "high",
  project = getProject(),
  out_dir = if (exists("OutDir")) OutDir else getwd(),
  background_job = FALSE,
  showMemObject = TRUE,
  saveParams = TRUE
)
}
\arguments{
\item{obj}{The R object to be saved.}

\item{prefix}{Optional; a prefix to add to the filename.}

\item{suffix}{Optional; a suffix to add to the filename.}

\item{nthreads}{Number of threads to use when saving, defaults to 12.}

\item{preset}{Compression preset, defaults to 'high'.}

\item{project}{The project name to be included in the filename, defaults to the result of \code{getProject()}.}

\item{out_dir}{Output Directory}

\item{background_job}{Logical; if TRUE save runs as "background job"}

\item{showMemObject}{Logical; if TRUE, displays the memory size of the largest objects.}

\item{saveParams}{Logical; if TRUE and if the object is a Seurat object, additional parameters are saved within it.}
}
\value{
Invisible; The function is called for its side effects (saving a file) and does not return anything.
}
\description{
This function saves an R object to a file in a quick and efficient format using the 'qs' package.
It constructs the file name based on various inputs and stores additional metadata if the object is a Seurat object.
The saving path can be adjusted by the presence of 'OutDir' in the global environment or defaults to the working directory.
}
\note{
The function uses the 'qs' package for quick and efficient serialization of objects and includes a timing feature from the 'tictoc' package.
}
\seealso{
\code{\link[qs]{qsave}} for the underlying save function used.
}
